Created favourite settings functionality #727
