+++ CookieMonster.js
@@ -1,2 +1,2 @@
 /*! For license information please see CookieMonster.js.LICENSE.txt */
-(()=>{var e={877:function(e){!function(t,o){"use strict";"object"!=typeof e.exports?o(t):e.exports=t.document?o(t):function(e){if(!e.document)throw new Error("jscolor needs a window with document");return o(e)}}("undefined"!=typeof window?window:this,(function(e){"use strict";var t,o,n,i,a=((i={initialized:!1,instances:[],readyQueue:[],register:function(){void 0!==e&&e.document&&e.document.addEventListener("DOMContentLoaded",i.pub.init,!1)},installBySelector:function(t,o){if(!(o=o?i.node(o):e.document))throw new Error("Missing root node");for(var n=o.querySelectorAll(t),a=new RegExp("(^|\\s)("+i.pub.lookupClass+")(\\s*(\\{[^}]*\\})|\\s|$)","i"),r=0;r<n.length;r+=1){var l,s;if(!(n[r].jscolor&&n[r].jscolor instanceof i.pub)&&(void 0===n[r].type||"color"!=n[r].type.toLowerCase()||!i.isColorAttrSupported)&&(null!==(l=i.getDataAttr(n[r],"jscolor"))||n[r].className&&(s=n[r].className.match(a)))){var c=n[r],d="";null!==l?d=l:s&&(console.warn('Installation using class name is DEPRECATED. Use data-jscolor="" attribute instead.'+i.docsRef),s[4]&&(d=s[4]));var p=null;if(d.trim())try{p=i.parseOptionsStr(d)}catch(e){console.warn(e+"\n"+d)}try{new i.pub(c,p)}catch(e){console.warn(e)}}}},parseOptionsStr:function(e){var t=null;try{t=JSON.parse(e)}catch(o){if(!i.pub.looseJSON)throw new Error("Could not parse jscolor options as JSON: "+o);try{t=new Function("var opts = ("+e+'); return typeof opts === "object" ? opts : {};')()}catch(e){throw new Error("Could not evaluate jscolor options: "+e)}}return t},getInstances:function(){for(var e=[],t=0;t<i.instances.length;t+=1)i.instances[t]&&i.instances[t].targetElement&&e.push(i.instances[t]);return e},createEl:function(t){var o=e.document.createElement(t);return i.setData(o,"gui",!0),o},node:function(t){if(!t)return null;if("string"==typeof t){var o=t,n=null;try{n=e.document.querySelector(o)}catch(e){return console.warn(e),null}return n||console.warn("No element matches the selector: %s",o),n}return i.isNode(t)?t:(console.warn("Invalid node of type %s: %s",typeof t,t),null)},isNode:function(e){return"object"==typeof Node?e instanceof Node:e&&"object"==typeof e&&"number"==typeof e.nodeType&&"string"==typeof e.nodeName},nodeName:function(e){return!(!e||!e.nodeName)&&e.nodeName.toLowerCase()},removeChildren:function(e){for(;e.firstChild;)e.removeChild(e.firstChild)},isTextInput:function(e){return e&&"input"===i.nodeName(e)&&"text"===e.type.toLowerCase()},isButton:function(e){if(!e)return!1;var t=i.nodeName(e);return"button"===t||"input"===t&&["button","submit","reset"].indexOf(e.type.toLowerCase())>-1},isButtonEmpty:function(e){switch(i.nodeName(e)){case"input":return!e.value||""===e.value.trim();case"button":return""===e.textContent.trim()}return null},isPassiveEventSupported:function(){var t=!1;try{var o=Object.defineProperty({},"passive",{get:function(){t=!0}});e.addEventListener("testPassive",null,o),e.removeEventListener("testPassive",null,o)}catch(e){}return t}(),isColorAttrSupported:(n=e.document.createElement("input"),!(!n.setAttribute||(n.setAttribute("type","color"),"color"!=n.type.toLowerCase()))),dataProp:"_data_jscolor",setData:function(){var e=arguments[0];if(3===arguments.length){var t=e.hasOwnProperty(i.dataProp)?e[i.dataProp]:e[i.dataProp]={},o=arguments[2];return t[a=arguments[1]]=o,!0}if(2===arguments.length&&"object"==typeof arguments[1]){t=e.hasOwnProperty(i.dataProp)?e[i.dataProp]:e[i.dataProp]={};var n=arguments[1];for(var a in n)n.hasOwnProperty(a)&&(t[a]=n[a]);return!0}throw new Error("Invalid arguments")},removeData:function(){var e=arguments[0];if(!e.hasOwnProperty(i.dataProp))return!0;for(var t=1;t<arguments.length;t+=1){var o=arguments[t];delete e[i.dataProp][o]}return!0},getData:function(e,t,o){if(!e.hasOwnProperty(i.dataProp)){if(void 0===o)return;e[i.dataProp]={}}var n=e[i.dataProp];return n.hasOwnProperty(t)||void 0===o||(n[t]=o),n[t]},getDataAttr:function(e,t){var o="data-"+t;return e.getAttribute(o)},setDataAttr:function(e,t,o){var n="data-"+t;e.setAttribute(n,o)},_attachedGroupEvents:{},attachGroupEvent:function(e,t,o,n){i._attachedGroupEvents.hasOwnProperty(e)||(i._attachedGroupEvents[e]=[]),i._attachedGroupEvents[e].push([t,o,n]),t.addEventListener(o,n,!1)},detachGroupEvents:function(e){if(i._attachedGroupEvents.hasOwnProperty(e)){for(var t=0;t<i._attachedGroupEvents[e].length;t+=1){var o=i._attachedGroupEvents[e][t];o[0].removeEventListener(o[1],o[2],!1)}delete i._attachedGroupEvents[e]}},preventDefault:function(e){e.preventDefault&&e.preventDefault(),e.returnValue=!1},captureTarget:function(e){e.setCapture&&(i._capturedTarget=e,i._capturedTarget.setCapture())},releaseTarget:function(){i._capturedTarget&&(i._capturedTarget.releaseCapture(),i._capturedTarget=null)},triggerEvent:function(t,o,n,a){if(t){var r=null;return"function"==typeof Event?r=new Event(o,{bubbles:n,cancelable:a}):(r=e.document.createEvent("Event")).initEvent(o,n,a),!!r&&(i.setData(r,"internal",!0),t.dispatchEvent(r),!0)}},triggerInputEvent:function(e,t,o,n){e&&i.isTextInput(e)&&i.triggerEvent(e,t,o,n)},eventKey:function(e){var t={9:"Tab",13:"Enter",27:"Escape"};return"string"==typeof e.code?e.code:void 0!==e.keyCode&&t.hasOwnProperty(e.keyCode)?t[e.keyCode]:null},strList:function(e){return e?e.replace(/^\s+|\s+$/g,"").split(/\s+/):[]},hasClass:function(e,t){return!!t&&(void 0!==e.classList?e.classList.contains(t):-1!=(" "+e.className.replace(/\s+/g," ")+" ").indexOf(" "+t+" "))},addClass:function(e,t){var o=i.strList(t);if(void 0===e.classList)for(n=0;n<o.length;n+=1)i.hasClass(e,o[n])||(e.className+=(e.className?" ":"")+o[n]);else for(var n=0;n<o.length;n+=1)e.classList.add(o[n])},removeClass:function(e,t){var o=i.strList(t);if(void 0===e.classList)for(a=0;a<o.length;a+=1){var n=new RegExp("^\\s*"+o[a]+"\\s*|\\s*"+o[a]+"\\s*$|\\s+"+o[a]+"(\\s+)","g");e.className=e.className.replace(n,"$1")}else for(var a=0;a<o.length;a+=1)e.classList.remove(o[a])},getCompStyle:function(t){return(e.getComputedStyle?e.getComputedStyle(t):t.currentStyle)||{}},setStyle:function(e,t,o,n){var a=o?"important":"",r=null;for(var l in t)if(t.hasOwnProperty(l)){var s=null;null===t[l]?(r||(r=i.getData(e,"origStyle")),r&&r.hasOwnProperty(l)&&(s=r[l])):(n&&(r||(r=i.getData(e,"origStyle",{})),r.hasOwnProperty(l)||(r[l]=e.style[l])),s=t[l]),null!==s&&e.style.setProperty(l,s,a)}},hexColor:function(e,t,o){return"#"+(("0"+Math.round(e).toString(16)).substr(-2)+("0"+Math.round(t).toString(16)).substr(-2)+("0"+Math.round(o).toString(16)).substr(-2)).toUpperCase()},hexaColor:function(e,t,o,n){return"#"+(("0"+Math.round(e).toString(16)).substr(-2)+("0"+Math.round(t).toString(16)).substr(-2)+("0"+Math.round(o).toString(16)).substr(-2)+("0"+Math.round(255*n).toString(16)).substr(-2)).toUpperCase()},rgbColor:function(e,t,o){return"rgb("+Math.round(e)+","+Math.round(t)+","+Math.round(o)+")"},rgbaColor:function(e,t,o,n){return"rgba("+Math.round(e)+","+Math.round(t)+","+Math.round(o)+","+Math.round(100*(null==n?1:n))/100+")"},linearGradient:(o=function(){for(var t="linear-gradient",o=["","-webkit-","-moz-","-o-","-ms-"],n=e.document.createElement("div"),i=0;i<o.length;i+=1){var a=o[i]+t,r=a+"(to right, rgba(0,0,0,0), rgba(0,0,0,0))";if(n.style.background=r,n.style.background)return a}return t}(),function(){return o+"("+Array.prototype.join.call(arguments,", ")+")"}),setBorderRadius:function(e,t){i.setStyle(e,{"border-radius":t||"0"})},setBoxShadow:function(e,t){i.setStyle(e,{"box-shadow":t||"none"})},getElementPos:function(e,t){var o=0,n=0,a=e.getBoundingClientRect();if(o=a.left,n=a.top,!t){var r=i.getViewPos();o+=r[0],n+=r[1]}return[o,n]},getElementSize:function(e){return[e.offsetWidth,e.offsetHeight]},getAbsPointerPos:function(e){var t=0,o=0;return void 0!==e.changedTouches&&e.changedTouches.length?(t=e.changedTouches[0].clientX,o=e.changedTouches[0].clientY):"number"==typeof e.clientX&&(t=e.clientX,o=e.clientY),{x:t,y:o}},getRelPointerPos:function(e){var t=(e.target||e.srcElement).getBoundingClientRect(),o=0,n=0;return void 0!==e.changedTouches&&e.changedTouches.length?(o=e.changedTouches[0].clientX,n=e.changedTouches[0].clientY):"number"==typeof e.clientX&&(o=e.clientX,n=e.clientY),{x:o-t.left,y:n-t.top}},getViewPos:function(){var t=e.document.documentElement;return[(e.pageXOffset||t.scrollLeft)-(t.clientLeft||0),(e.pageYOffset||t.scrollTop)-(t.clientTop||0)]},getViewSize:function(){var t=e.document.documentElement;return[e.innerWidth||t.clientWidth,e.innerHeight||t.clientHeight]},RGB_HSV:function(e,t,o){e/=255,t/=255,o/=255;var n=Math.min(Math.min(e,t),o),i=Math.max(Math.max(e,t),o),a=i-n;if(0===a)return[null,0,100*i];var r=e===n?3+(o-t)/a:t===n?5+(e-o)/a:1+(t-e)/a;return[60*(6===r?0:r),a/i*100,100*i]},HSV_RGB:function(e,t,o){var n=o/100*255;if(null===e)return[n,n,n];e/=60,t/=100;var i=Math.floor(e),a=n*(1-t),r=n*(1-t*(i%2?e-i:1-(e-i)));switch(i){case 6:case 0:return[n,r,a];case 1:return[r,n,a];case 2:return[a,n,r];case 3:return[a,r,n];case 4:return[r,a,n];case 5:return[n,a,r]}},parseColorString:function(e){var t,o={rgba:null,format:null};if(t=e.match(/^\W*([0-9A-F]{3,8})\W*$/i)){if(8===t[1].length)o.format="hexa",o.rgba=[parseInt(t[1].substr(0,2),16),parseInt(t[1].substr(2,2),16),parseInt(t[1].substr(4,2),16),parseInt(t[1].substr(6,2),16)/255];else if(6===t[1].length)o.format="hex",o.rgba=[parseInt(t[1].substr(0,2),16),parseInt(t[1].substr(2,2),16),parseInt(t[1].substr(4,2),16),null];else{if(3!==t[1].length)return!1;o.format="hex",o.rgba=[parseInt(t[1].charAt(0)+t[1].charAt(0),16),parseInt(t[1].charAt(1)+t[1].charAt(1),16),parseInt(t[1].charAt(2)+t[1].charAt(2),16),null]}return o}if(t=e.match(/^\W*rgba?\(([^)]*)\)\W*$/i)){var n,i,a,r,l=t[1].split(","),s=/^\s*(\d+|\d*\.\d+|\d+\.\d*)\s*$/;if(l.length>=3&&(n=l[0].match(s))&&(i=l[1].match(s))&&(a=l[2].match(s)))return o.format="rgb",o.rgba=[parseFloat(n[1])||0,parseFloat(i[1])||0,parseFloat(a[1])||0,null],l.length>=4&&(r=l[3].match(s))&&(o.format="rgba",o.rgba[3]=parseFloat(r[1])||0),o}return!1},parsePaletteValue:function(e){var t=[];"string"==typeof e?e.replace(/#[0-9A-F]{3}([0-9A-F]{3})?|rgba?\(([^)]*)\)/gi,(function(e){t.push(e)})):Array.isArray(e)&&(t=e);for(var o=[],n=0;n<t.length;n++){var a=i.parseColorString(t[n]);a&&o.push(a)}return o},containsTranparentColor:function(e){for(var t=0;t<e.length;t++){var o=e[t].rgba[3];if(null!==o&&o<1)return!0}return!1},isAlphaFormat:function(e){switch(e.toLowerCase()){case"hexa":case"rgba":return!0}return!1},scaleCanvasForHighDPR:function(t){var o=e.devicePixelRatio||1;t.width*=o,t.height*=o,t.getContext("2d").scale(o,o)},genColorPreviewCanvas:function(e,t,o,n){var a=Math.round(i.pub.previewSeparator.length),r=i.pub.chessboardSize,l=i.pub.chessboardColor1,s=i.pub.chessboardColor2,c=o||2*r,d=2*r,p=i.createEl("canvas"),u=p.getContext("2d");p.width=c,p.height=d,n&&i.scaleCanvasForHighDPR(p),u.fillStyle=l,u.fillRect(0,0,c,d),u.fillStyle=s;for(var m=0;m<c;m+=2*r)u.fillRect(m,0,r,r),u.fillRect(m+r,r,r,r);e&&(u.fillStyle=e,u.fillRect(0,0,c,d));var h=null;switch(t){case"left":h=0,u.clearRect(0,0,a/2,d);break;case"right":h=c-a,u.clearRect(c-a/2,0,a/2,d)}if(null!==h){u.lineWidth=1;for(var g=0;g<i.pub.previewSeparator.length;g+=1)u.beginPath(),u.strokeStyle=i.pub.previewSeparator[g],u.moveTo(.5+h+g,0),u.lineTo(.5+h+g,d),u.stroke()}return{canvas:p,width:c,height:d}},genColorPreviewGradient:function(e,t,o){var n;return n=t&&o?["to "+{left:"right",right:"left"}[t],e+" 0%",e+" "+o+"px","rgba(0,0,0,0) "+(o+1)+"px","rgba(0,0,0,0) 100%"]:["to right",e+" 0%",e+" 100%"],i.linearGradient.apply(this,n)},redrawPosition:function(){if(i.picker&&i.picker.owner){var e,t,o=i.picker.owner;o.fixed?(e=i.getElementPos(o.targetElement,!0),t=[0,0]):(e=i.getElementPos(o.targetElement),t=i.getViewPos());var n,a,r,l=i.getElementSize(o.targetElement),s=i.getViewSize(),c=i.getPickerDims(o),d=[c.borderW,c.borderH];switch(o.position.toLowerCase()){case"left":n=1,a=0,r=-1;break;case"right":n=1,a=0,r=1;break;case"top":n=0,a=1,r=-1;break;default:n=0,a=1,r=1}var p=(l[a]+d[a])/2;if(o.smartPosition)u=[-t[n]+e[n]+d[n]>s[n]&&-t[n]+e[n]+l[n]/2>s[n]/2&&e[n]+l[n]-d[n]>=0?e[n]+l[n]-d[n]:e[n],-t[a]+e[a]+l[a]+d[a]-p+p*r>s[a]?-t[a]+e[a]+l[a]/2>s[a]/2&&e[a]+l[a]-p-p*r>=0?e[a]+l[a]-p-p*r:e[a]+l[a]-p+p*r:e[a]+l[a]-p+p*r>=0?e[a]+l[a]-p+p*r:e[a]+l[a]-p-p*r];else var u=[e[n],e[a]+l[a]-p+p*r];var m=u[n],h=u[a],g=o.fixed?"fixed":"absolute",f=(u[0]+d[0]>e[0]||u[0]<e[0]+l[0])&&u[1]+d[1]<e[1]+l[1];i._drawPosition(o,m,h,g,f)}},_drawPosition:function(e,t,o,n,a){var r=a?0:e.shadowBlur;i.picker.wrap.style.position=n,i.picker.wrap.style.left=t+"px",i.picker.wrap.style.top=o+"px",i.setBoxShadow(i.picker.boxS,e.shadow?new i.BoxShadow(0,r,e.shadowBlur,0,e.shadowColor):null)},getPickerDims:function(e){var t=2*e.controlBorderWidth+e.width,o=2*e.controlBorderWidth+e.height,n=2*e.controlBorderWidth+2*i.getControlPadding(e)+e.sliderSize;i.getSliderChannel(e)&&(t+=n),e.hasAlphaChannel()&&(t+=n);var a=i.getPaletteDims(e,t);a.height&&(o+=a.height+e.padding),e.closeButton&&(o+=2*e.controlBorderWidth+e.padding+e.buttonHeight);var r=t+2*e.padding,l=o+2*e.padding;return{contentW:t,contentH:o,paddedW:r,paddedH:l,borderW:r+2*e.borderWidth,borderH:l+2*e.borderWidth,palette:a}},getPaletteDims:function(e,t){var o=0,n=0,i=0,a=0,r=0,l=e._palette?e._palette.length:0;return l&&(n=(o=e.paletteCols)>0?Math.ceil(l/o):0,i=Math.max(1,Math.floor((t-(o-1)*e.paletteSpacing)/o)),a=e.paletteHeight?Math.min(e.paletteHeight,i):i),n&&(r=n*a+(n-1)*e.paletteSpacing),{cols:o,rows:n,cellW:i,cellH:a,width:t,height:r}},getControlPadding:function(e){return Math.max(e.padding/2,2*e.pointerBorderWidth+e.pointerThickness-e.controlBorderWidth)},getPadYChannel:function(e){switch(e.mode.charAt(1).toLowerCase()){case"v":return"v"}return"s"},getSliderChannel:function(e){if(e.mode.length>2)switch(e.mode.charAt(2).toLowerCase()){case"s":return"s";case"v":return"v"}return null},triggerCallback:function(e,t){if(e[t]){var o=null;if("string"==typeof e[t])try{o=new Function(e[t])}catch(e){console.error(e)}else o=e[t];o&&o.call(e)}},triggerGlobal:function(e){for(var t=i.getInstances(),o=0;o<t.length;o+=1)t[o].trigger(e)},_pointerMoveEvent:{mouse:"mousemove",touch:"touchmove"},_pointerEndEvent:{mouse:"mouseup",touch:"touchend"},_pointerOrigin:null,_capturedTarget:null,onDocumentKeyUp:function(e){-1!==["Tab","Escape"].indexOf(i.eventKey(e))&&i.picker&&i.picker.owner&&i.picker.owner.tryHide()},onWindowResize:function(e){i.redrawPosition()},onWindowScroll:function(e){i.redrawPosition()},onParentScroll:function(e){i.picker&&i.picker.owner&&i.picker.owner.tryHide()},onDocumentMouseDown:function(e){var t=e.target||e.srcElement;t.jscolor&&t.jscolor instanceof i.pub?t.jscolor.showOnClick&&!t.disabled&&t.jscolor.show():i.getData(t,"gui")?i.getData(t,"control")&&i.onControlPointerStart(e,t,i.getData(t,"control"),"mouse"):i.picker&&i.picker.owner&&i.picker.owner.tryHide()},onPickerTouchStart:function(e){var t=e.target||e.srcElement;i.getData(t,"control")&&i.onControlPointerStart(e,t,i.getData(t,"control"),"touch")},onControlPointerStart:function(t,o,n,a){var r=i.getData(o,"instance");i.preventDefault(t),i.captureTarget(o);var l=function(e,r){i.attachGroupEvent("drag",e,i._pointerMoveEvent[a],i.onDocumentPointerMove(t,o,n,a,r)),i.attachGroupEvent("drag",e,i._pointerEndEvent[a],i.onDocumentPointerEnd(t,o,n,a))};if(l(e.document,[0,0]),e.parent&&e.frameElement){var s=e.frameElement.getBoundingClientRect(),c=[-s.left,-s.top];l(e.parent.window.document,c)}var d=i.getAbsPointerPos(t),p=i.getRelPointerPos(t);switch(i._pointerOrigin={x:d.x-p.x,y:d.y-p.y},n){case"pad":"v"===i.getSliderChannel(r)&&0===r.channels.v&&r.fromHSVA(null,null,100,null),i.setPad(r,t,0,0);break;case"sld":i.setSld(r,t,0);break;case"asld":i.setASld(r,t,0)}r.trigger("input")},onDocumentPointerMove:function(e,t,o,n,a){return function(e){var n=i.getData(t,"instance");switch(o){case"pad":i.setPad(n,e,a[0],a[1]);break;case"sld":i.setSld(n,e,a[1]);break;case"asld":i.setASld(n,e,a[1])}n.trigger("input")}},onDocumentPointerEnd:function(e,t,o,n){return function(e){var o=i.getData(t,"instance");i.detachGroupEvents("drag"),i.releaseTarget(),o.trigger("input"),o.trigger("change")}},onPaletteSampleClick:function(e){var t=e.currentTarget,o=i.getData(t,"instance"),n=i.getData(t,"color");"any"===o.format.toLowerCase()&&(o._setFormat(n.format),i.isAlphaFormat(o.getFormat())||(n.rgba[3]=1)),null===n.rgba[3]&&(!0===o.paletteSetsAlpha||"auto"===o.paletteSetsAlpha&&o._paletteHasTransparency)&&(n.rgba[3]=1),o.fromRGBA.apply(o,n.rgba),o.trigger("input"),o.trigger("change"),o.hideOnPaletteClick&&o.hide()},setPad:function(e,t,o,n){var a=i.getAbsPointerPos(t),r=o+a.x-i._pointerOrigin.x-e.padding-e.controlBorderWidth,l=n+a.y-i._pointerOrigin.y-e.padding-e.controlBorderWidth,s=r*(360/(e.width-1)),c=100-l*(100/(e.height-1));switch(i.getPadYChannel(e)){case"s":e.fromHSVA(s,c,null,null);break;case"v":e.fromHSVA(s,null,c,null)}},setSld:function(e,t,o){var n=100-(o+i.getAbsPointerPos(t).y-i._pointerOrigin.y-e.padding-e.controlBorderWidth)*(100/(e.height-1));switch(i.getSliderChannel(e)){case"s":e.fromHSVA(null,n,null,null);break;case"v":e.fromHSVA(null,null,n,null)}},setASld:function(e,t,o){var n=1-(o+i.getAbsPointerPos(t).y-i._pointerOrigin.y-e.padding-e.controlBorderWidth)*(1/(e.height-1));if(n<1){var a=e.getFormat();"any"!==e.format.toLowerCase()||i.isAlphaFormat(a)||e._setFormat("hex"===a?"hexa":"rgba")}e.fromHSVA(null,null,null,n)},createPadCanvas:function(){var e={elm:null,draw:null},t=i.createEl("canvas"),o=t.getContext("2d");return e.elm=t,e.draw=function(e,n,i){t.width=e,t.height=n,o.clearRect(0,0,t.width,t.height);var a=o.createLinearGradient(0,0,t.width,0);a.addColorStop(0,"#F00"),a.addColorStop(1/6,"#FF0"),a.addColorStop(2/6,"#0F0"),a.addColorStop(.5,"#0FF"),a.addColorStop(4/6,"#00F"),a.addColorStop(5/6,"#F0F"),a.addColorStop(1,"#F00"),o.fillStyle=a,o.fillRect(0,0,t.width,t.height);var r=o.createLinearGradient(0,0,0,t.height);switch(i.toLowerCase()){case"s":r.addColorStop(0,"rgba(255,255,255,0)"),r.addColorStop(1,"rgba(255,255,255,1)");break;case"v":r.addColorStop(0,"rgba(0,0,0,0)"),r.addColorStop(1,"rgba(0,0,0,1)")}o.fillStyle=r,o.fillRect(0,0,t.width,t.height)},e},createSliderGradient:function(){var e={elm:null,draw:null},t=i.createEl("canvas"),o=t.getContext("2d");return e.elm=t,e.draw=function(e,n,i,a){t.width=e,t.height=n,o.clearRect(0,0,t.width,t.height);var r=o.createLinearGradient(0,0,0,t.height);r.addColorStop(0,i),r.addColorStop(1,a),o.fillStyle=r,o.fillRect(0,0,t.width,t.height)},e},createASliderGradient:function(){var e={elm:null,draw:null},t=i.createEl("canvas"),o=t.getContext("2d");return e.elm=t,e.draw=function(e,n,a){t.width=e,t.height=n,o.clearRect(0,0,t.width,t.height);var r=t.width/2,l=i.pub.chessboardColor1,s=i.pub.chessboardColor2;if(o.fillStyle=l,o.fillRect(0,0,t.width,t.height),r>0)for(var c=0;c<t.height;c+=2*r)o.fillStyle=s,o.fillRect(0,c,r,r),o.fillRect(r,c+r,r,r);var d=o.createLinearGradient(0,0,0,t.height);d.addColorStop(0,a),d.addColorStop(1,"rgba(0,0,0,0)"),o.fillStyle=d,o.fillRect(0,0,t.width,t.height)},e},BoxShadow:(t=function(e,t,o,n,i,a){this.hShadow=e,this.vShadow=t,this.blur=o,this.spread=n,this.color=i,this.inset=!!a},t.prototype.toString=function(){var e=[Math.round(this.hShadow)+"px",Math.round(this.vShadow)+"px",Math.round(this.blur)+"px",Math.round(this.spread)+"px",this.color];return this.inset&&e.push("inset"),e.join(" ")},t),flags:{leaveValue:1,leaveAlpha:2,leavePreview:4},enumOpts:{format:["auto","any","hex","hexa","rgb","rgba"],previewPosition:["left","right"],mode:["hsv","hvs","hs","hv"],position:["left","right","top","bottom"],alphaChannel:["auto",!0,!1],paletteSetsAlpha:["auto",!0,!1]},deprecatedOpts:{styleElement:"previewElement",onFineChange:"onInput",overwriteImportant:"forceStyle",closable:"closeButton",insetWidth:"controlBorderWidth",insetColor:"controlBorderColor",refine:null},docsRef:" See https://jscolor.com/docs/",pub:function(t,o){var n=this;function a(e,t){if("string"!=typeof e)throw new Error("Invalid value for option name: "+e);if(i.enumOpts.hasOwnProperty(e)&&("string"==typeof t&&(t=t.toLowerCase()),-1===i.enumOpts[e].indexOf(t)))throw new Error("Option '"+e+"' has invalid value: "+t);if(i.deprecatedOpts.hasOwnProperty(e)){var o=e,a=i.deprecatedOpts[e];if(!a)throw new Error("Option '"+e+"' is DEPRECATED");console.warn("Option '%s' is DEPRECATED, using '%s' instead."+i.docsRef,o,a),e=a}var r="set__"+e;if("function"==typeof n[r])return n[r](t),!0;if(e in n)return n[e]=t,!0;throw new Error("Unrecognized configuration option: "+e)}function r(e){if("string"!=typeof e)throw new Error("Invalid value for option name: "+e);if(i.deprecatedOpts.hasOwnProperty(e)){var t=e,o=i.deprecatedOpts[e];if(!o)throw new Error("Option '"+e+"' is DEPRECATED");console.warn("Option '%s' is DEPRECATED, using '%s' instead."+i.docsRef,t,o),e=o}var a="get__"+e;if("function"==typeof n[a])return n[a](value);if(e in n)return n[e];throw new Error("Unrecognized configuration option: "+e)}function l(){n._processParentElementsInDOM(),i.picker||(i.picker={owner:null,wrap:i.createEl("div"),box:i.createEl("div"),boxS:i.createEl("div"),boxB:i.createEl("div"),pad:i.createEl("div"),padB:i.createEl("div"),padM:i.createEl("div"),padCanvas:i.createPadCanvas(),cross:i.createEl("div"),crossBY:i.createEl("div"),crossBX:i.createEl("div"),crossLY:i.createEl("div"),crossLX:i.createEl("div"),sld:i.createEl("div"),sldB:i.createEl("div"),sldM:i.createEl("div"),sldGrad:i.createSliderGradient(),sldPtrS:i.createEl("div"),sldPtrIB:i.createEl("div"),sldPtrMB:i.createEl("div"),sldPtrOB:i.createEl("div"),asld:i.createEl("div"),asldB:i.createEl("div"),asldM:i.createEl("div"),asldGrad:i.createASliderGradient(),asldPtrS:i.createEl("div"),asldPtrIB:i.createEl("div"),asldPtrMB:i.createEl("div"),asldPtrOB:i.createEl("div"),pal:i.createEl("div"),btn:i.createEl("div"),btnT:i.createEl("span")},i.picker.pad.appendChild(i.picker.padCanvas.elm),i.picker.padB.appendChild(i.picker.pad),i.picker.cross.appendChild(i.picker.crossBY),i.picker.cross.appendChild(i.picker.crossBX),i.picker.cross.appendChild(i.picker.crossLY),i.picker.cross.appendChild(i.picker.crossLX),i.picker.padB.appendChild(i.picker.cross),i.picker.box.appendChild(i.picker.padB),i.picker.box.appendChild(i.picker.padM),i.picker.sld.appendChild(i.picker.sldGrad.elm),i.picker.sldB.appendChild(i.picker.sld),i.picker.sldB.appendChild(i.picker.sldPtrOB),i.picker.sldPtrOB.appendChild(i.picker.sldPtrMB),i.picker.sldPtrMB.appendChild(i.picker.sldPtrIB),i.picker.sldPtrIB.appendChild(i.picker.sldPtrS),i.picker.box.appendChild(i.picker.sldB),i.picker.box.appendChild(i.picker.sldM),i.picker.asld.appendChild(i.picker.asldGrad.elm),i.picker.asldB.appendChild(i.picker.asld),i.picker.asldB.appendChild(i.picker.asldPtrOB),i.picker.asldPtrOB.appendChild(i.picker.asldPtrMB),i.picker.asldPtrMB.appendChild(i.picker.asldPtrIB),i.picker.asldPtrIB.appendChild(i.picker.asldPtrS),i.picker.box.appendChild(i.picker.asldB),i.picker.box.appendChild(i.picker.asldM),i.picker.box.appendChild(i.picker.pal),i.picker.btn.appendChild(i.picker.btnT),i.picker.box.appendChild(i.picker.btn),i.picker.boxB.appendChild(i.picker.box),i.picker.wrap.appendChild(i.picker.boxS),i.picker.wrap.appendChild(i.picker.boxB),i.picker.wrap.addEventListener("touchstart",i.onPickerTouchStart,!!i.isPassiveEventSupported&&{passive:!1}));var t,o,a=i.picker,r=!!i.getSliderChannel(n),l=n.hasAlphaChannel(),p=i.getPickerDims(n),u=2*n.pointerBorderWidth+n.pointerThickness+2*n.crossSize,m=i.getControlPadding(n),h=Math.min(n.borderRadius,Math.round(n.padding*Math.PI));a.wrap.className="jscolor-picker-wrap",a.wrap.style.clear="both",a.wrap.style.width=p.borderW+"px",a.wrap.style.height=p.borderH+"px",a.wrap.style.zIndex=n.zIndex,a.box.className="jscolor-picker",a.box.style.width=p.paddedW+"px",a.box.style.height=p.paddedH+"px",a.box.style.position="relative",a.boxS.className="jscolor-picker-shadow",a.boxS.style.position="absolute",a.boxS.style.left="0",a.boxS.style.top="0",a.boxS.style.width="100%",a.boxS.style.height="100%",i.setBorderRadius(a.boxS,h+"px"),a.boxB.className="jscolor-picker-border",a.boxB.style.position="relative",a.boxB.style.border=n.borderWidth+"px solid",a.boxB.style.borderColor=n.borderColor,a.boxB.style.background=n.backgroundColor,i.setBorderRadius(a.boxB,h+"px"),a.padM.style.background="rgba(255,0,0,.2)",a.sldM.style.background="rgba(0,255,0,.2)",a.asldM.style.background="rgba(0,0,255,.2)",a.padM.style.opacity=a.sldM.style.opacity=a.asldM.style.opacity="0",a.pad.style.position="relative",a.pad.style.width=n.width+"px",a.pad.style.height=n.height+"px",a.padCanvas.draw(n.width,n.height,i.getPadYChannel(n)),a.padB.style.position="absolute",a.padB.style.left=n.padding+"px",a.padB.style.top=n.padding+"px",a.padB.style.border=n.controlBorderWidth+"px solid",a.padB.style.borderColor=n.controlBorderColor,a.padM.style.position="absolute",a.padM.style.left="0px",a.padM.style.top="0px",a.padM.style.width=n.padding+2*n.controlBorderWidth+n.width+m+"px",a.padM.style.height=2*n.controlBorderWidth+2*n.padding+n.height+"px",a.padM.style.cursor="crosshair",i.setData(a.padM,{instance:n,control:"pad"}),a.cross.style.position="absolute",a.cross.style.left=a.cross.style.top="0",a.cross.style.width=a.cross.style.height=u+"px",a.crossBY.style.position=a.crossBX.style.position="absolute",a.crossBY.style.background=a.crossBX.style.background=n.pointerBorderColor,a.crossBY.style.width=a.crossBX.style.height=2*n.pointerBorderWidth+n.pointerThickness+"px",a.crossBY.style.height=a.crossBX.style.width=u+"px",a.crossBY.style.left=a.crossBX.style.top=Math.floor(u/2)-Math.floor(n.pointerThickness/2)-n.pointerBorderWidth+"px",a.crossBY.style.top=a.crossBX.style.left="0",a.crossLY.style.position=a.crossLX.style.position="absolute",a.crossLY.style.background=a.crossLX.style.background=n.pointerColor,a.crossLY.style.height=a.crossLX.style.width=u-2*n.pointerBorderWidth+"px",a.crossLY.style.width=a.crossLX.style.height=n.pointerThickness+"px",a.crossLY.style.left=a.crossLX.style.top=Math.floor(u/2)-Math.floor(n.pointerThickness/2)+"px",a.crossLY.style.top=a.crossLX.style.left=n.pointerBorderWidth+"px",a.sld.style.overflow="hidden",a.sld.style.width=n.sliderSize+"px",a.sld.style.height=n.height+"px",a.sldGrad.draw(n.sliderSize,n.height,"#000","#000"),a.sldB.style.display=r?"block":"none",a.sldB.style.position="absolute",a.sldB.style.left=n.padding+n.width+2*n.controlBorderWidth+2*m+"px",a.sldB.style.top=n.padding+"px",a.sldB.style.border=n.controlBorderWidth+"px solid",a.sldB.style.borderColor=n.controlBorderColor,a.sldM.style.display=r?"block":"none",a.sldM.style.position="absolute",a.sldM.style.left=n.padding+n.width+2*n.controlBorderWidth+m+"px",a.sldM.style.top="0px",a.sldM.style.width=n.sliderSize+2*m+2*n.controlBorderWidth+(l?0:Math.max(0,n.padding-m))+"px",a.sldM.style.height=2*n.controlBorderWidth+2*n.padding+n.height+"px",a.sldM.style.cursor="default",i.setData(a.sldM,{instance:n,control:"sld"}),a.sldPtrIB.style.border=a.sldPtrOB.style.border=n.pointerBorderWidth+"px solid "+n.pointerBorderColor,a.sldPtrOB.style.position="absolute",a.sldPtrOB.style.left=-(2*n.pointerBorderWidth+n.pointerThickness)+"px",a.sldPtrOB.style.top="0",a.sldPtrMB.style.border=n.pointerThickness+"px solid "+n.pointerColor,a.sldPtrS.style.width=n.sliderSize+"px",a.sldPtrS.style.height=i.pub.sliderInnerSpace+"px",a.asld.style.overflow="hidden",a.asld.style.width=n.sliderSize+"px",a.asld.style.height=n.height+"px",a.asldGrad.draw(n.sliderSize,n.height,"#000"),a.asldB.style.display=l?"block":"none",a.asldB.style.position="absolute",a.asldB.style.left=n.padding+n.width+2*n.controlBorderWidth+m+(r?n.sliderSize+3*m+2*n.controlBorderWidth:0)+"px",a.asldB.style.top=n.padding+"px",a.asldB.style.border=n.controlBorderWidth+"px solid",a.asldB.style.borderColor=n.controlBorderColor,a.asldM.style.display=l?"block":"none",a.asldM.style.position="absolute",a.asldM.style.left=n.padding+n.width+2*n.controlBorderWidth+m+(r?n.sliderSize+2*m+2*n.controlBorderWidth:0)+"px",a.asldM.style.top="0px",a.asldM.style.width=n.sliderSize+2*m+2*n.controlBorderWidth+Math.max(0,n.padding-m)+"px",a.asldM.style.height=2*n.controlBorderWidth+2*n.padding+n.height+"px",a.asldM.style.cursor="default",i.setData(a.asldM,{instance:n,control:"asld"}),a.asldPtrIB.style.border=a.asldPtrOB.style.border=n.pointerBorderWidth+"px solid "+n.pointerBorderColor,a.asldPtrOB.style.position="absolute",a.asldPtrOB.style.left=-(2*n.pointerBorderWidth+n.pointerThickness)+"px",a.asldPtrOB.style.top="0",a.asldPtrMB.style.border=n.pointerThickness+"px solid "+n.pointerColor,a.asldPtrS.style.width=n.sliderSize+"px",a.asldPtrS.style.height=i.pub.sliderInnerSpace+"px",a.pal.className="jscolor-palette",a.pal.style.display=p.palette.rows?"block":"none",a.pal.style.position="absolute",a.pal.style.left=n.padding+"px",a.pal.style.top=2*n.controlBorderWidth+2*n.padding+n.height+"px",a.pal.innerHTML="";for(var g=i.genColorPreviewCanvas("rgba(0,0,0,0)"),f=0,C=0;C<p.palette.rows;C++)for(var y=0;y<p.palette.cols&&f<n._palette.length;y++,f++){var b=n._palette[f],k=i.rgbaColor.apply(null,b.rgba),G=i.createEl("div");G.style.width=p.palette.cellW-2*n.controlBorderWidth+"px",G.style.height=p.palette.cellH-2*n.controlBorderWidth+"px",G.style.backgroundColor=k;var w=i.createEl("div");w.className="jscolor-palette-sample",w.style.display="block",w.style.position="absolute",w.style.left=(p.palette.cols<=1?0:Math.round(y*((p.contentW-p.palette.cellW)/(p.palette.cols-1))*10)/10)+"px",w.style.top=C*(p.palette.cellH+n.paletteSpacing)+"px",w.style.border=n.controlBorderWidth+"px solid",w.style.borderColor=n.controlBorderColor,w.style.cursor="pointer",null!==b.rgba[3]&&b.rgba[3]<1&&(w.style.backgroundImage="url('"+g.canvas.toDataURL()+"')",w.style.backgroundRepeat="repeat",w.style.backgroundPosition="center center"),i.setData(w,{instance:n,control:"palette-sample",color:b}),w.addEventListener("click",i.onPaletteSampleClick,!1),w.appendChild(G),a.pal.appendChild(w)}a.btn.className="jscolor-btn-close",a.btn.style.display=n.closeButton?"block":"none",a.btn.style.position="absolute",a.btn.style.left=n.padding+"px",a.btn.style.bottom=n.padding+"px",a.btn.style.padding="0 15px",a.btn.style.maxWidth=p.contentW-2*n.controlBorderWidth-30+"px",a.btn.style.overflow="hidden",a.btn.style.height=n.buttonHeight+"px",a.btn.style.whiteSpace="nowrap",a.btn.style.border=n.controlBorderWidth+"px solid",o=(t=n.controlBorderColor.split(/\s+/)).length<2?t[0]:t[1]+" "+t[0]+" "+t[0]+" "+t[1],a.btn.style.borderColor=o,a.btn.style.color=n.buttonColor,a.btn.style.font="12px sans-serif",a.btn.style.textAlign="center",a.btn.style.cursor="pointer",a.btn.onmousedown=function(){n.hide()},a.btnT.style.lineHeight=n.buttonHeight+"px",a.btnT.innerHTML="",a.btnT.appendChild(e.document.createTextNode(n.closeText)),s(),c(),d(),i.picker.owner&&i.picker.owner!==n&&i.removeClass(i.picker.owner.targetElement,i.pub.activeClassName),i.picker.owner=n,n.container===e.document.body?i.redrawPosition():i._drawPosition(n,0,0,"relative",!1),a.wrap.parentNode!==n.container&&n.container.appendChild(a.wrap),i.addClass(n.targetElement,i.pub.activeClassName)}function s(){var e=i.getPadYChannel(n),t=Math.round(n.channels.h/360*(n.width-1)),o=Math.round((1-n.channels[e]/100)*(n.height-1)),a=2*n.pointerBorderWidth+n.pointerThickness+2*n.crossSize,r=-Math.floor(a/2);switch(i.picker.cross.style.left=t+r+"px",i.picker.cross.style.top=o+r+"px",i.getSliderChannel(n)){case"s":var l=i.HSV_RGB(n.channels.h,100,n.channels.v),s=i.HSV_RGB(n.channels.h,0,n.channels.v),c="rgb("+Math.round(l[0])+","+Math.round(l[1])+","+Math.round(l[2])+")",d="rgb("+Math.round(s[0])+","+Math.round(s[1])+","+Math.round(s[2])+")";i.picker.sldGrad.draw(n.sliderSize,n.height,c,d);break;case"v":var p=i.HSV_RGB(n.channels.h,n.channels.s,100);c="rgb("+Math.round(p[0])+","+Math.round(p[1])+","+Math.round(p[2])+")",d="#000",i.picker.sldGrad.draw(n.sliderSize,n.height,c,d)}i.picker.asldGrad.draw(n.sliderSize,n.height,n.toHEXString())}function c(){var e=i.getSliderChannel(n);if(e){var t=Math.round((1-n.channels[e]/100)*(n.height-1));i.picker.sldPtrOB.style.top=t-(2*n.pointerBorderWidth+n.pointerThickness)-Math.floor(i.pub.sliderInnerSpace/2)+"px"}i.picker.asldGrad.draw(n.sliderSize,n.height,n.toHEXString())}function d(){var e=Math.round((1-n.channels.a)*(n.height-1));i.picker.asldPtrOB.style.top=e-(2*n.pointerBorderWidth+n.pointerThickness)-Math.floor(i.pub.sliderInnerSpace/2)+"px"}function p(){return i.picker&&i.picker.owner===n}if(o||(o={}),this.channels={r:255,g:255,b:255,h:0,s:0,v:100,a:1},this.format="auto",this.value=void 0,this.alpha=void 0,this.onChange=void 0,this.onInput=void 0,this.valueElement=void 0,this.alphaElement=void 0,this.previewElement=void 0,this.previewPosition="left",this.previewSize=32,this.previewPadding=8,this.required=!0,this.hash=!0,this.uppercase=!0,this.forceStyle=!0,this.width=181,this.height=101,this.mode="HSV",this.alphaChannel="auto",this.position="bottom",this.smartPosition=!0,this.showOnClick=!0,this.hideOnLeave=!0,this.palette=[],this.paletteCols=10,this.paletteSetsAlpha="auto",this.paletteHeight=16,this.paletteSpacing=4,this.hideOnPaletteClick=!1,this.sliderSize=16,this.crossSize=8,this.closeButton=!1,this.closeText="Close",this.buttonColor="rgba(0,0,0,1)",this.buttonHeight=18,this.padding=12,this.backgroundColor="rgba(255,255,255,1)",this.borderWidth=1,this.borderColor="rgba(187,187,187,1)",this.borderRadius=8,this.controlBorderWidth=1,this.controlBorderColor="rgba(187,187,187,1)",this.shadow=!0,this.shadowBlur=15,this.shadowColor="rgba(0,0,0,0.2)",this.pointerColor="rgba(76,76,76,1)",this.pointerBorderWidth=1,this.pointerBorderColor="rgba(255,255,255,1)",this.pointerThickness=2,this.zIndex=5e3,this.container=void 0,this.minS=0,this.maxS=100,this.minV=0,this.maxV=100,this.minA=0,this.maxA=1,this.option=function(){if(!arguments.length)throw new Error("No option specified");if(1===arguments.length&&"string"==typeof arguments[0]){try{return r(arguments[0])}catch(e){console.warn(e)}return!1}if(arguments.length>=2&&"string"==typeof arguments[0]){try{if(!a(arguments[0],arguments[1]))return!1}catch(e){return console.warn(e),!1}return this.redraw(),this.exposeColor(),!0}if(1===arguments.length&&"object"==typeof arguments[0]){var e=arguments[0],t=!0;for(var o in e)if(e.hasOwnProperty(o))try{a(o,e[o])||(t=!1)}catch(e){console.warn(e),t=!1}return this.redraw(),this.exposeColor(),t}throw new Error("Invalid arguments")},this.channel=function(e,t){if("string"!=typeof e)throw new Error("Invalid value for channel name: "+e);if(void 0===t)return this.channels.hasOwnProperty(e.toLowerCase())?this.channels[e.toLowerCase()]:(console.warn("Getting unknown channel: "+e),!1);var o=!1;switch(e.toLowerCase()){case"r":o=this.fromRGBA(t,null,null,null);break;case"g":o=this.fromRGBA(null,t,null,null);break;case"b":o=this.fromRGBA(null,null,t,null);break;case"h":o=this.fromHSVA(t,null,null,null);break;case"s":o=this.fromHSVA(null,t,null,null);break;case"v":o=this.fromHSVA(null,null,t,null);break;case"a":o=this.fromHSVA(null,null,null,t);break;default:return console.warn("Setting unknown channel: "+e),!1}return!!o&&(this.redraw(),!0)},this.trigger=function(e){for(var t=i.strList(e),o=0;o<t.length;o+=1){var n=t[o].toLowerCase(),a=null;switch(n){case"input":a="onInput";break;case"change":a="onChange"}a&&i.triggerCallback(this,a),i.triggerInputEvent(this.valueElement,n,!0,!0)}},this.fromHSVA=function(e,t,o,n,a){if(void 0===e&&(e=null),void 0===t&&(t=null),void 0===o&&(o=null),void 0===n&&(n=null),null!==e){if(isNaN(e))return!1;this.channels.h=Math.max(0,Math.min(360,e))}if(null!==t){if(isNaN(t))return!1;this.channels.s=Math.max(0,Math.min(100,this.maxS,t),this.minS)}if(null!==o){if(isNaN(o))return!1;this.channels.v=Math.max(0,Math.min(100,this.maxV,o),this.minV)}if(null!==n){if(isNaN(n))return!1;this.channels.a=this.hasAlphaChannel()?Math.max(0,Math.min(1,this.maxA,n),this.minA):1}var r=i.HSV_RGB(this.channels.h,this.channels.s,this.channels.v);return this.channels.r=r[0],this.channels.g=r[1],this.channels.b=r[2],this.exposeColor(a),!0},this.fromRGBA=function(e,t,o,n,a){if(void 0===e&&(e=null),void 0===t&&(t=null),void 0===o&&(o=null),void 0===n&&(n=null),null!==e){if(isNaN(e))return!1;e=Math.max(0,Math.min(255,e))}if(null!==t){if(isNaN(t))return!1;t=Math.max(0,Math.min(255,t))}if(null!==o){if(isNaN(o))return!1;o=Math.max(0,Math.min(255,o))}if(null!==n){if(isNaN(n))return!1;this.channels.a=this.hasAlphaChannel()?Math.max(0,Math.min(1,this.maxA,n),this.minA):1}var r=i.RGB_HSV(null===e?this.channels.r:e,null===t?this.channels.g:t,null===o?this.channels.b:o);null!==r[0]&&(this.channels.h=Math.max(0,Math.min(360,r[0]))),0!==r[2]&&(this.channels.s=Math.max(0,this.minS,Math.min(100,this.maxS,r[1]))),this.channels.v=Math.max(0,this.minV,Math.min(100,this.maxV,r[2]));var l=i.HSV_RGB(this.channels.h,this.channels.s,this.channels.v);return this.channels.r=l[0],this.channels.g=l[1],this.channels.b=l[2],this.exposeColor(a),!0},this.fromHSV=function(e,t,o,n){return console.warn("fromHSV() method is DEPRECATED. Using fromHSVA() instead."+i.docsRef),this.fromHSVA(e,t,o,null,n)},this.fromRGB=function(e,t,o,n){return console.warn("fromRGB() method is DEPRECATED. Using fromRGBA() instead."+i.docsRef),this.fromRGBA(e,t,o,null,n)},this.fromString=function(e,t){if(!this.required&&""===e.trim())return this.setPreviewElementBg(null),this.setValueElementValue(""),!0;var o=i.parseColorString(e);return!!o&&("any"===this.format.toLowerCase()&&(this._setFormat(o.format),i.isAlphaFormat(this.getFormat())||(o.rgba[3]=1)),this.fromRGBA(o.rgba[0],o.rgba[1],o.rgba[2],o.rgba[3],t),!0)},this.toString=function(e){switch(void 0===e&&(e=this.getFormat()),e.toLowerCase()){case"hex":return this.toHEXString();case"hexa":return this.toHEXAString();case"rgb":return this.toRGBString();case"rgba":return this.toRGBAString()}return!1},this.toHEXString=function(){return i.hexColor(this.channels.r,this.channels.g,this.channels.b)},this.toHEXAString=function(){return i.hexaColor(this.channels.r,this.channels.g,this.channels.b,this.channels.a)},this.toRGBString=function(){return i.rgbColor(this.channels.r,this.channels.g,this.channels.b)},this.toRGBAString=function(){return i.rgbaColor(this.channels.r,this.channels.g,this.channels.b,this.channels.a)},this.toGrayscale=function(){return.213*this.channels.r+.715*this.channels.g+.072*this.channels.b},this.toCanvas=function(){return i.genColorPreviewCanvas(this.toRGBAString()).canvas},this.toDataURL=function(){return this.toCanvas().toDataURL()},this.toBackground=function(){return i.pub.background(this.toRGBAString())},this.isLight=function(){return this.toGrayscale()>127.5},this.hide=function(){p()&&(i.removeClass(n.targetElement,i.pub.activeClassName),i.picker.wrap.parentNode.removeChild(i.picker.wrap),delete i.picker.owner)},this.show=function(){l()},this.redraw=function(){p()&&l()},this.getFormat=function(){return this._currentFormat},this._setFormat=function(e){this._currentFormat=e.toLowerCase()},this.hasAlphaChannel=function(){return"auto"===this.alphaChannel?"any"===this.format.toLowerCase()||i.isAlphaFormat(this.getFormat())||void 0!==this.alpha||void 0!==this.alphaElement:this.alphaChannel},this.processValueInput=function(e){this.fromString(e)||this.exposeColor()},this.processAlphaInput=function(e){this.fromHSVA(null,null,null,parseFloat(e))||this.exposeColor()},this.exposeColor=function(e){var t=this.toString(),o=this.getFormat();if(i.setDataAttr(this.targetElement,"current-color",t),e&i.flags.leaveValue||!this.valueElement||("hex"!==o&&"hexa"!==o||(this.uppercase||(t=t.toLowerCase()),this.hash||(t=t.replace(/^#/,""))),this.setValueElementValue(t)),!(e&i.flags.leaveAlpha)&&this.alphaElement){var n=Math.round(100*this.channels.a)/100;this.setAlphaElementValue(n)}e&i.flags.leavePreview||!this.previewElement||((i.isTextInput(this.previewElement)||i.isButton(this.previewElement)&&!i.isButtonEmpty(this.previewElement))&&this.previewPosition,this.setPreviewElementBg(this.toRGBAString())),p()&&(s(),c(),d())},this.setPreviewElementBg=function(e){if(this.previewElement){var t=null,o=null;(i.isTextInput(this.previewElement)||i.isButton(this.previewElement)&&!i.isButtonEmpty(this.previewElement))&&(t=this.previewPosition,o=this.previewSize);var n=[];if(e){n.push({image:i.genColorPreviewGradient(e,t,o?o-i.pub.previewSeparator.length:null),position:"left top",size:"auto",repeat:t?"repeat-y":"repeat",origin:"padding-box"});var a=i.genColorPreviewCanvas("rgba(0,0,0,0)",t?{left:"right",right:"left"}[t]:null,o,!0);n.push({image:"url('"+a.canvas.toDataURL()+"')",position:(t||"left")+" top",size:a.width+"px "+a.height+"px",repeat:t?"repeat-y":"repeat",origin:"padding-box"})}else n.push({image:"none",position:"left top",size:"auto",repeat:"no-repeat",origin:"padding-box"});for(var r={image:[],position:[],size:[],repeat:[],origin:[]},l=0;l<n.length;l+=1)r.image.push(n[l].image),r.position.push(n[l].position),r.size.push(n[l].size),r.repeat.push(n[l].repeat),r.origin.push(n[l].origin);var s={"background-image":r.image.join(", "),"background-position":r.position.join(", "),"background-size":r.size.join(", "),"background-repeat":r.repeat.join(", "),"background-origin":r.origin.join(", ")};i.setStyle(this.previewElement,s,this.forceStyle);var c={left:null,right:null};t&&(c[t]=this.previewSize+this.previewPadding+"px"),s={"padding-left":c.left,"padding-right":c.right},i.setStyle(this.previewElement,s,this.forceStyle,!0)}},this.setValueElementValue=function(e){this.valueElement&&("input"===i.nodeName(this.valueElement)?this.valueElement.value=e:this.valueElement.innerHTML=e)},this.setAlphaElementValue=function(e){this.alphaElement&&("input"===i.nodeName(this.alphaElement)?this.alphaElement.value=e:this.alphaElement.innerHTML=e)},this._processParentElementsInDOM=function(){if(!this._parentElementsProcessed){this._parentElementsProcessed=!0;var e=this.targetElement;do{var t=i.getCompStyle(e);t.position&&"fixed"===t.position.toLowerCase()&&(this.fixed=!0),e!==this.targetElement&&(i.getData(e,"hasScrollListener")||(e.addEventListener("scroll",i.onParentScroll,!1),i.setData(e,"hasScrollListener",!0)))}while((e=e.parentNode)&&"body"!==i.nodeName(e))}},this.tryHide=function(){this.hideOnLeave&&this.hide()},this.set__palette=function(e){this.palette=e,this._palette=i.parsePaletteValue(e),this._paletteHasTransparency=i.containsTranparentColor(this._palette)},i.pub.options)for(var u in i.pub.options)if(i.pub.options.hasOwnProperty(u))try{a(u,i.pub.options[u])}catch(e){console.warn(e)}var m=[];o.preset&&("string"==typeof o.preset?m=o.preset.split(/\s+/):Array.isArray(o.preset)?m=o.preset.slice():console.warn("Unrecognized preset value")),-1===m.indexOf("default")&&m.push("default");for(var h=m.length-1;h>=0;h-=1){var g=m[h];if(g)if(i.pub.presets.hasOwnProperty(g)){for(var u in i.pub.presets[g])if(i.pub.presets[g].hasOwnProperty(u))try{a(u,i.pub.presets[g][u])}catch(e){console.warn(e)}}else console.warn("Unknown preset: %s",g)}var f=["preset"];for(var u in o)if(o.hasOwnProperty(u)&&-1===f.indexOf(u))try{a(u,o[u])}catch(e){console.warn(e)}if(void 0===this.container?this.container=e.document.body:this.container=i.node(this.container),!this.container)throw new Error("Cannot instantiate color picker without a container element");if(this.targetElement=i.node(t),!this.targetElement){if("string"==typeof t&&/^[a-zA-Z][\w:.-]*$/.test(t))throw new Error("If '"+t+"' is supposed to be an ID, please use '#"+t+"' or any valid CSS selector.");throw new Error("Cannot instantiate color picker without a target element")}if(this.targetElement.jscolor&&this.targetElement.jscolor instanceof i.pub)throw new Error("Color picker already installed on this element");if(this.targetElement.jscolor=this,i.addClass(this.targetElement,i.pub.className),i.instances.push(this),i.isButton(this.targetElement)&&("button"!==this.targetElement.type.toLowerCase()&&(this.targetElement.type="button"),i.isButtonEmpty(this.targetElement))){i.removeChildren(this.targetElement),this.targetElement.appendChild(e.document.createTextNode(" "));var C=i.getCompStyle(this.targetElement);(parseFloat(C["min-width"])||0)<this.previewSize&&i.setStyle(this.targetElement,{"min-width":this.previewSize+"px"},this.forceStyle)}if(void 0===this.valueElement?i.isTextInput(this.targetElement)&&(this.valueElement=this.targetElement):null===this.valueElement||(this.valueElement=i.node(this.valueElement)),this.alphaElement&&(this.alphaElement=i.node(this.alphaElement)),void 0===this.previewElement?this.previewElement=this.targetElement:null===this.previewElement||(this.previewElement=i.node(this.previewElement)),this.valueElement&&i.isTextInput(this.valueElement)){var y={onInput:this.valueElement.oninput};this.valueElement.oninput=null,this.valueElement.addEventListener("keydown",(function(e){"Enter"===i.eventKey(e)&&(n.valueElement&&n.processValueInput(n.valueElement.value),n.tryHide())}),!1),this.valueElement.addEventListener("change",(function(e){if(!i.getData(e,"internal")){var t=n.valueElement.value;n.processValueInput(n.valueElement.value),i.triggerCallback(n,"onChange"),n.valueElement.value!==t&&i.triggerInputEvent(n.valueElement,"change",!0,!0)}}),!1),this.valueElement.addEventListener("input",(function(e){i.getData(e,"internal")||(n.valueElement&&n.fromString(n.valueElement.value,i.flags.leaveValue),i.triggerCallback(n,"onInput"))}),!1),y.onInput&&this.valueElement.addEventListener("input",y.onInput,!1),this.valueElement.setAttribute("autocomplete","off"),this.valueElement.setAttribute("autocorrect","off"),this.valueElement.setAttribute("autocapitalize","off"),this.valueElement.setAttribute("spellcheck",!1)}this.alphaElement&&i.isTextInput(this.alphaElement)&&(this.alphaElement.addEventListener("keydown",(function(e){"Enter"===i.eventKey(e)&&(n.alphaElement&&n.processAlphaInput(n.alphaElement.value),n.tryHide())}),!1),this.alphaElement.addEventListener("change",(function(e){if(!i.getData(e,"internal")){var t=n.alphaElement.value;n.processAlphaInput(n.alphaElement.value),i.triggerCallback(n,"onChange"),i.triggerInputEvent(n.valueElement,"change",!0,!0),n.alphaElement.value!==t&&i.triggerInputEvent(n.alphaElement,"change",!0,!0)}}),!1),this.alphaElement.addEventListener("input",(function(e){i.getData(e,"internal")||(n.alphaElement&&n.fromHSVA(null,null,null,parseFloat(n.alphaElement.value),i.flags.leaveAlpha),i.triggerCallback(n,"onInput"),i.triggerInputEvent(n.valueElement,"input",!0,!0))}),!1),this.alphaElement.setAttribute("autocomplete","off"),this.alphaElement.setAttribute("autocorrect","off"),this.alphaElement.setAttribute("autocapitalize","off"),this.alphaElement.setAttribute("spellcheck",!1));var b="FFFFFF";void 0!==this.value?b=this.value:this.valueElement&&void 0!==this.valueElement.value&&(b=this.valueElement.value);var k=void 0;if(void 0!==this.alpha?k=""+this.alpha:this.alphaElement&&void 0!==this.alphaElement.value&&(k=this.alphaElement.value),this._currentFormat=null,["auto","any"].indexOf(this.format.toLowerCase())>-1){var G=i.parseColorString(b);this._currentFormat=G?G.format:"hex"}else this._currentFormat=this.format.toLowerCase();this.processValueInput(b),void 0!==k&&this.processAlphaInput(k)}}).pub.className="jscolor",i.pub.activeClassName="jscolor-active",i.pub.looseJSON=!0,i.pub.presets={},i.pub.presets.default={},i.pub.presets.light={backgroundColor:"rgba(255,255,255,1)",controlBorderColor:"rgba(187,187,187,1)",buttonColor:"rgba(0,0,0,1)"},i.pub.presets.dark={backgroundColor:"rgba(51,51,51,1)",controlBorderColor:"rgba(153,153,153,1)",buttonColor:"rgba(240,240,240,1)"},i.pub.presets.small={width:101,height:101,padding:10,sliderSize:14,paletteCols:8},i.pub.presets.medium={width:181,height:101,padding:12,sliderSize:16,paletteCols:10},i.pub.presets.large={width:271,height:151,padding:12,sliderSize:24,paletteCols:15},i.pub.presets.thin={borderWidth:1,controlBorderWidth:1,pointerBorderWidth:1},i.pub.presets.thick={borderWidth:2,controlBorderWidth:2,pointerBorderWidth:2},i.pub.sliderInnerSpace=3,i.pub.chessboardSize=8,i.pub.chessboardColor1="#666666",i.pub.chessboardColor2="#999999",i.pub.previewSeparator=["rgba(255,255,255,.65)","rgba(128,128,128,.65)"],i.pub.init=function(){if(!i.initialized)for(e.document.addEventListener("mousedown",i.onDocumentMouseDown,!1),e.document.addEventListener("keyup",i.onDocumentKeyUp,!1),e.addEventListener("resize",i.onWindowResize,!1),e.addEventListener("scroll",i.onWindowScroll,!1),i.pub.install(),i.initialized=!0;i.readyQueue.length;)i.readyQueue.shift()()},i.pub.install=function(e){var t=!0;try{i.installBySelector("[data-jscolor]",e)}catch(e){t=!1,console.warn(e)}if(i.pub.lookupClass)try{i.installBySelector("input."+i.pub.lookupClass+", button."+i.pub.lookupClass,e)}catch(e){}return t},i.pub.ready=function(e){return"function"!=typeof e?(console.warn("Passed value is not a function"),!1):(i.initialized?e():i.readyQueue.push(e),!0)},i.pub.trigger=function(e){var t=function(){i.triggerGlobal(e)};i.initialized?t():i.pub.ready(t)},i.pub.hide=function(){i.picker&&i.picker.owner&&i.picker.owner.hide()},i.pub.chessboard=function(e){return e||(e="rgba(0,0,0,0)"),i.genColorPreviewCanvas(e).canvas.toDataURL()},i.pub.background=function(e){var t=[];t.push(i.genColorPreviewGradient(e));var o=i.genColorPreviewCanvas();return t.push(["url('"+o.canvas.toDataURL()+"')","left top","repeat"].join(" ")),t.join(", ")},i.pub.options={},i.pub.lookupClass="jscolor",i.pub.installByClassName=function(){return console.error('jscolor.installByClassName() is DEPRECATED. Use data-jscolor="" attribute instead of a class name.'+i.docsRef),!1},i.register(),i.pub);return void 0===e.jscolor&&(e.jscolor=e.JSColor=a),a}))}},t={};function o(n){var i=t[n];if(void 0!==i)return i.exports;var a=t[n]={exports:{}};return e[n].call(a.exports,a,a.exports,o),a.exports}o.n=e=>{var t=e&&e.__esModule?()=>e.default:()=>e;return o.d(t,{a:t}),t},o.d=(e,t)=>{for(var n in t)o.o(t,n)&&!o.o(e,n)&&Object.defineProperty(e,n,{enumerable:!0,get:t[n]})},o.o=(e,t)=>Object.prototype.hasOwnProperty.call(e,t),(()=>{"use strict";let e={};function t(){e.WrinklerButtons&&Game.elderWrath?(l("PopAllNormalWrinklerButton").style.display="",l("PopFattestWrinklerButton").style.display=""):(l("PopAllNormalWrinklerButton").style.display="none",l("PopFattestWrinklerButton").style.display="none")}const n=["","","M","G","T","P","E","Z","Y"],i=["","","M","B","Tr","Quadr","Quint","Sext","Sept","Oct","Non","Dec","Undec","Duodec","Tredec","Quattuordec","Quindec","Sexdec","Septendec","Octodec","Novemdec","Vigint","Unvigint","Duovigint","Trevigint","Quattuorvigint"],a=["","K","M","B","T","Qa","Qi","Sx","Sp","Oc","No","De","UDe","DDe","TDe","QaDe","QiDe","SxDe","SpDe","ODe","NDe","Vi","UVi","DVi","TVi","QaVi","QiVi","SxVi","SpVi","OVi","NVi","Tr","UTr","DTr","TTr","QaTr","QiTr","SxTr","SpTr","OTr","NTr","Qaa","UQa","DQa","TQa","QaQa","QiQa","SxQa","SpQa","OQa","NQa","Qia","UQi","DQi","TQi","QaQi","QiQi","SxQi","SpQi","OQi","NQi","Sxa","USx","DSx","TSx","QaSx","QiSx","SxSx","SpSx","OSx","NSx","Spa","USp","DSp","TSp","QaSp","QiSp","SxSp","SpSp","OSp","NSp","Oco","UOc","DOc","TOc","QaOc","QiOc","SxOc","SpOc","OOc","NOc","Noa","UNo","DNo","TNo","QaNo","QiNo","SxNo","SpNo","ONo","NNo","Ct","UCt"];let r,s,c,d,p,u,m,h,g,f,C,y,b,k,G,w,v,x=[],T=[],M=[],S={};const P="CMText",B="CMBack",E="Blue",N="Green",O="Yellow",F="Orange",W="Red",L="Purple",A="Gray",j="Pink",D="Brown",H=[A,E,N,O,F,W,L,j,D],U=[["GoldCookTooltipPlaceholder","Calculated with Golden Switch off","200px"],["GoldCookDragonsFortuneTooltipPlaceholder","Calculated with Golden Switch off and at least one golden cookie on-screen","240px"],["PrestMaxTooltipPlaceholder","The MAX prestige is calculated with the cookies gained from popping all wrinklers with Skruuia god in Diamond slot, selling all stock market goods, selling all buildings with Earth Shatterer and Reality Bending auras, and buying Chocolate egg","320px"],["NextPrestTooltipPlaceholder","Calculated with cookies gained from wrinklers and Chocolate egg","200px"],["HeavenChipMaxTooltipPlaceholder","The MAX heavenly chips is calculated with the cookies gained from popping all wrinklers with Skruuia god in Diamond slot, selling all stock market goods, selling all buildings with Earth Shatterer and Reality Bending auras, and buying Chocolate egg","330px"],["ResetTooltipPlaceholder","The bonus income you would get from new prestige levels unlocked at 100% of its potential and from ascension achievements if you have the same buildings/upgrades after reset","370px"],["ChoEggTooltipPlaceholder","The amount of cookies you would get from popping all wrinklers with Skruuia god in Diamond slot, selling all stock market goods, selling all buildings with Earth Shatterer and Reality Bending auras, and then buying Chocolate egg","300px"],["ChainNextLevelPlaceholder","Cheated cookies might break this formula","250px"]],R={};let $,I,z=0,V=-1,_=[],X=1,Q=1,q=1;const Y=[10,15,30,60,300,600,900,1800],K=[1,5,10,15,30],J={Frenzy:O,"Dragon Harvest":D,"Elder frenzy":N,Clot:W,"Click frenzy":E,Dragonflight:j};let Z,ee,te,oe,ne,ie,ae,re={},le="";function se(t,o,r){const l=e.ScaleDecimals+1;if(t===1/0)return"Infinity";if(void 0===t)return"0";if(0===e.Scale)return S.Beautify(t,o);if(Number.isFinite(t)){if(t<0)return`-${se(Math.abs(t))}`;let s="";if(0===t)return t.toString();if(t>.001&&t<e.ScaleCutoff)return s=e.ScaleSeparator?t.toLocaleString("nl"):t.toLocaleString("en"),s;if(4===e.Scale&&!r||4===r)s=t.toExponential(l).toString().replace("e","E");else{const o=t.toExponential().toString(),c=Math.floor(o.slice(o.indexOf("e")+1)/3);s=(t/Number("1e"+3*c)).toFixed(l),1===e.Scale&&!r||1===r?t>=.01&&t<Number("1e"+3*n.length)?s+=` ${n[c]}`:s=se(t,0,4):2===e.Scale&&!r||2===r?t>=.01&&t<Number("1e"+3*i.length)?s+=` ${i[c]}`:s=se(t,0,4):3===e.Scale&&!r||3===r?t>=.01&&t<Number("1e"+3*a.length)?s+=` ${a[c]}`:s=se(t,0,4):(5===e.Scale&&!r||5===r)&&(s+="E"+3*c)}return""===s&&(console.log(`Could not beautify number with Cookie Monster Beautify: ${t}`),s=S.Beautify(t,o)),e.ScaleSeparator&&(s=s.replace(".",",")),s}return console.log(`Could not beautify number with Cookie Monster Beautify: ${t}`),S.Beautify(t,o)}function ce(t,o){let n=t;if(t===1/0)return t;if(t<0)return"Negative time period";n=Math.ceil(t);const i=Math.floor(n/31557600),a=Math.floor(n%31557600/86400),r=Math.floor(n%86400/3600),l=Math.floor(n%3600/60),s=Math.floor(n%60);let c="";if(e.TimeFormat){if(n>315576e4)return"XX:XX:XX:XX:XX";c+=(i<10?"0":"")+i+":",c+=(a<10?"0":"")+a+":",c+=(r<10?"0":"")+r+":",c+=(l<10?"0":"")+l+":",c+=(s<10?"0":"")+s}else{if(n>7776e5)return o?"Over 9000 days!":">9000d";c+=i>0?i+(o?1===i?" year":" years":"y")+", ":"",c+=a>0?a+(o?1===a?" day":" days":"d")+", ":"",(c.length>0||r>0)&&(c+=r+(o?1===r?" hour":" hours":"h")+", "),(c.length>0||l>0)&&(c+=l+(o?1===l?" minute":" minutes":"m")+", "),c+=s+(o?1===s?" second":" seconds":"s")}return c}function de(t){let o,n;return t<=0?(n=e.TimeFormat?"00:00:00:00:00":"Done!",o=N):(n=ce(t),o=t>300?W:t>60?F:O),{text:n,color:o}}let pe,ue,me,he,ge,fe,Ce,ye,be,ke,Ge,we,ve,xe,Te,Me,Se,Pe,Be,Ee,Ne,Oe,Fe,We,Le,Ae,je,De,He,Ue,Re,$e,Ie,ze,Ve,_e=0,Xe=0,Qe=0,qe=0,Ye=0,Ke=0,Je=0,Ze=0,et=0,tt=0,ot=0,nt=0,it=0,at=0,rt=0,lt=1,st=1,ct=1,dt=0,pt=0,ut=0,mt=0,ht={},gt={},ft={},Ct=0,yt=0,bt=[0,null],kt=0,Gt=0,wt={},vt={},xt={},Tt={},Mt={},St=0,Pt=0,Bt=[],Et=0,Nt=0,Ot=[],Ft=0,Wt=0,Lt=[],At=0,jt=0,Dt=[],Ht=0,Ut=0,Rt={},$t={},It=0,zt=0,Vt=[],_t={},Xt=0,Qt={0:[0,0,0],1:[0,0,0],2:[0,0,0],3:[0,0,0],4:[0,0,0],5:[0,0,0],6:[0,0,0],7:[0,0,0],8:[0,0,0],9:[0,0,0],10:[0,0,0]},qt=[],Yt=[],Kt=[];function Jt(e){let t=0;return Game.dragonAuras[je].name!==e&&Game.dragonAuras[De].name!==e||(t=1),"Reality Bending"!==Game.dragonAuras[je].name&&"Reality Bending"!==Game.dragonAuras[De].name||(t+=.1),t}function Zt(){let e=.25;return e*=1+Jt("Earth Shatterer"),e}function eo(e,t){return void 0===Ie[e]?void 0===t?1:t:Ie[e]}function to(e){const t=Yt[e];return(1!==Game.ascensionMode||"prestige"!==t.pool&&"fortune"!==t.tier)&&t?t.bought:0}function oo(e){if(Game.hasGod){void 0===qt.Temple.minigame&&(qt.Temple.minigame=Game.Objects.Temple.minigame);const t=qt.Temple.minigame.gods[e];if(He===t.id)return 1;if(Ue===t.id)return 2;if(Re===t.id)return 3}return!1}function no(e,t){let o=t;if(to("Season savings")&&(o*=.99),to("Santa's dominion")&&(o*=.99),to("Faberge egg")&&(o*=.99),to("Divine discount")&&(o*=.99),to("Fortune #100")&&(o*=.99),o*=1-.02*Jt("Fierce Hoarder"),Game.hasBuff("Everything must go")&&(o*=.95),Game.hasBuff("Crafty pixies")&&(o*=.98),Game.hasBuff("Nasty goblins")&&(o*=1.02),e.fortune&&to(e.fortune.name)&&(o*=.93),o*=eo("buildingCost"),qt.Temple.minigameLoaded){const e=oo("creation");1===e?o*=.93:2===e?o*=.95:3===e&&(o*=.98)}return o}function io(e,t,o,n,i,a){const r=void 0===a?0:a;let l=i,s=o,c=0;-1===i&&(l=s),i||(l=Game.buyBulk);for(let o=0;o<l;o++){let o=t*Game.priceIncrease**Math.max(0,s-n);o=r?Game.modifyBuildingPrice(e,o):no(e,o),o=Math.ceil(o);const i=r?e.getSellMultiplier():Zt();o=Math.floor(o*i),s>0&&(c+=o,s-=1)}return c}function ao(){let t=Game.buyBulk;if(1===Game.buyMode?Q=t:t=Q,1===t?t=ht:10===t?t=gt:100===t&&(t=ft),1===Game.buyMode?1===e.BuildColour?Object.keys(t).forEach((o=>{l(`productPrice${Game.Objects[o].id}`).style.color=e[`Colour${t[o].color}`]})):Object.keys(Game.Objects).forEach((e=>{l(`productPrice${Game.Objects[e].id}`).style.removeProperty("color")})):-1===Game.buyMode&&Object.keys(ht).forEach((e=>{const t=Game.Objects[e];l(`productPrice${t.id}`).style.color="",l(`productPrice${t.id}`).innerHTML=se(io(t,t.basePrice,t.amount,t.free,Game.buyBulk,1))})),1===Game.buyMode&&e.SortBuildings){let o;1===e.SortBuildings?(o=Object.keys(ht).map((e=>{const t=ht[e];return t.name=e,t.id=Game.Objects[e].id,t})),o.sort((function(e,t){return H.indexOf(e.color)>H.indexOf(t.color)?1:H.indexOf(e.color)<H.indexOf(t.color)||e.pp<t.pp?-1:0}))):2===e.SortBuildings&&(o=Object.keys(t).map((e=>{const o=t[e];return o.name=e,o.id=Game.Objects[e].id,o})),o.sort((function(e,t){return H.indexOf(e.color)>H.indexOf(t.color)?1:H.indexOf(e.color)<H.indexOf(t.color)||e.pp<t.pp?-1:0})));for(let e=0;e<o.length;e++)Game.Objects[o[e].name].l.style.gridRow=`${e+2}/${e+2}`}else{const e=Object.keys(ht).map((e=>{const t=ht[e];return t.name=e,t.id=Game.Objects[e].id,t}));e.sort(((e,t)=>e.id-t.id));for(let t=0;t<e.length;t++)Game.Objects[e[t].name].l.style.gridRow=`${t+2}/${t+2}`}}function ro(){if(e.UpBarColour>0){let e=0,t=0,o=0,n=0,i=0,a=0,r=0;Object.keys(Game.UpgradesInStore).forEach((s=>{const c=Game.UpgradesInStore[s];let d=!1;for(let e=0;e<l(`upgrade${s}`).childNodes.length;e+=1)if(-1!==l(`upgrade${s}`).childNodes[e].className.indexOf(B)){l(`upgrade${s}`).childNodes[e].className=B+wt[c.name].color,d=!0;break}if(!d){const e=document.createElement("div");e.style.width="10px",e.style.height="10px",e.className=B+wt[c.name].color,l(`upgrade${s}`).appendChild(e)}wt[c.name].color===E?e+=1:wt[c.name].color===N?t+=1:wt[c.name].color===O?o+=1:wt[c.name].color===F?n+=1:wt[c.name].color===W?i+=1:wt[c.name].color===L?a+=1:wt[c.name].color===A&&(r+=1)})),l("CMUpgradeBarBlue").textContent=e,l("CMUpgradeBarGreen").textContent=t,l("CMUpgradeBarYellow").textContent=o,l("CMUpgradeBarOrange").textContent=n,l("CMUpgradeBarRed").textContent=i,l("CMUpgradeBarPurple").textContent=a,l("CMUpgradeBarGray").textContent=r}const t=[];for(let e=0;e<Game.UpgradesInStore.length;e+=1){const o={};o.name=Game.UpgradesInStore[e].name,o.price=Game.UpgradesInStore[e].basePrice,o.pp=wt[o.name].pp,t.push(o)}e.SortUpgrades?t.sort((function(e,t){return H.indexOf(e.color)>H.indexOf(t.color)?1:H.indexOf(e.color)<H.indexOf(t.color)||e.pp<t.pp?-1:0})):t.sort(((e,t)=>e.price-t.price));const o=function(e,t){return e.findIndex((e=>e.name===t.name))};for(let e=0;e<Game.UpgradesInStore.length;e+=1)l(`upgrade${e}`).style.order=o(t,Game.UpgradesInStore[e])+1}function lo(){return e.CPSMode?Me:0===e.CalcWrink?Game.cookiesPs*(1-Game.cpsSucked):1===e.CalcWrink?Game.cookiesPs*(kt+(1-.05*Gt)):null!==bt[1]&&2===e.CalcWrink&&1===Game.wrinklers[bt[1]].type?Game.cookiesPs*(3*kt/Gt+(1-.05*Gt)):Game.cookiesPs*(kt/Gt+(1-.05*Gt))}function so(){return 1===e.CalcWrink?Ct:2===e.CalcWrink?bt[0]:0}function co(e,t,o){const n=document.createElement("div");n.id=e,n.style.height="12px",n.style.margin="0px 10px",n.style.position="relative";const i=document.createElement("div");i.style.width="100%",i.style.height="10px",i.style.margin="auto",i.style.position="absolute",i.style.left="0px",i.style.top="0px",i.style.right="0px",i.style.bottom="0px";const a=document.createElement("span");a.style.display="inline-block",a.style.textAlign="right",a.style.fontSize="10px",a.style.width="108px",a.style.marginRight="5px",a.style.verticalAlign="text-top",a.textContent=t,i.appendChild(a);for(let e=0;e<o.length;e++){const t=document.createElement("span");t.id=o[e].id,t.style.display="inline-block",t.style.height="10px",t.style.verticalAlign="text-top",t.style.textAlign="center",o.length-1===e&&(t.style.borderTopRightRadius="10px",t.style.borderBottomRightRadius="10px"),void 0!==o[e].color&&(t.className=B+o[e].color),i.appendChild(t)}const r=document.createElement("span");return r.id=`${e}Time`,r.style.marginLeft="5px",r.style.verticalAlign="text-top",i.appendChild(r),n.appendChild(i),n}function po(e){if(null!==l("CMBotBar")){const t=l("CMBotBar").firstChild.firstChild.childNodes[0],o=l("CMBotBar").firstChild.firstChild.childNodes[1],n=l("CMBotBar").firstChild.firstChild.childNodes[2],i=l("CMBotBar").firstChild.firstChild.childNodes[3],a=e,r=t.appendChild(document.createElement("td"));r.appendChild(document.createTextNode(`${-1!==a.indexOf(" ")?a.substring(0,a.indexOf(" ")):a} (`)),r.appendChild(document.createElement("span")).className="CMTextBlue",r.appendChild(document.createTextNode(")")),t.lastChild.style.paddingLeft="8px",o.appendChild(document.createElement("td")),o.lastChild.style.paddingLeft="8px",n.appendChild(document.createElement("td")),n.lastChild.style.paddingLeft="8px",i.appendChild(document.createElement("td")),i.lastChild.style.paddingLeft="2px"}}function uo(){if(1===e.BotBar&&ht&&1===Game.buyMode){let t=0;Object.keys(ht).forEach((o=>{let n,i=Game.buyBulk;1===Game.buyMode?X=i:i=X,1===i&&(i=ht),10===i&&(i=gt),100===i&&(i=ft),t+=1,l("CMBotBar").firstChild.firstChild.childNodes[0].childNodes[t].childNodes[1].textContent=Game.Objects[o].amount,l("CMBotBar").firstChild.firstChild.childNodes[1].childNodes[t].textContent=se(i[o].bonus,2),l("CMBotBar").firstChild.firstChild.childNodes[2].childNodes[t].className=P+i[o].color,n=e.PPDisplayTime?ce(Math.round(i[o].pp)):se(Math.round(i[o].pp),2),l("CMBotBar").firstChild.firstChild.childNodes[2].childNodes[t].textContent=n;const a=de((Game.Objects[o].bulkPrice-(Game.cookies+so()))/lo());l("CMBotBar").firstChild.firstChild.childNodes[3].childNodes[t].className=P+a.color,"Done!"===a.text&&Game.cookies<Game.Objects[o].bulkPrice?l("CMBotBar").firstChild.firstChild.childNodes[3].childNodes[t].textContent=`${a.text} (with Wrink)`:l("CMBotBar").firstChild.firstChild.childNodes[3].childNodes[t].textContent=a.text}))}}function mo(){Game.Background.canvas.width=Game.Background.canvas.parentNode.offsetWidth,Game.Background.canvas.height=Game.Background.canvas.parentNode.offsetHeight,Game.LeftBackground.canvas.width=Game.LeftBackground.canvas.parentNode.offsetWidth,Game.LeftBackground.canvas.height=Game.LeftBackground.canvas.parentNode.offsetHeight}function ho(){1===e.BotBar&&1===e.TimerBar&&1===e.TimerBarPos?(l("CMBotBar").style.bottom=l("CMTimerBar").style.height,l("game").style.bottom=`${Number(l("CMTimerBar").style.height.replace("px",""))+70}px`):1===e.BotBar?(l("CMBotBar").style.bottom="0px",l("game").style.bottom="70px"):1===e.TimerBar&&1===e.TimerBarPos?l("game").style.bottom=l("CMTimerBar").style.height:l("game").style.bottom="0px",1===e.TimerBar&&0===e.TimerBarPos?l("sectionLeft").style.top=l("CMTimerBar").style.height:l("sectionLeft").style.top="",mo()}function go(){1===e.TimerBar?l("CMTimerBar").style.display="":l("CMTimerBar").style.display="none",ho()}function fo(){_e=Game.dragonAura,Xe=Game.dragonAura2}function Co(e){const t=Game.Achievements[e],o={};return o.name=t.name,o}function yo(e){let t=1;if(Object.keys(e.tieredUpgrades).forEach((o=>{!Game.Tiers[e.tieredUpgrades[o].tier].special&&to(e.tieredUpgrades[o].name)&&(t*=2)})),Object.keys(e.synergies).forEach((o=>{if(to(e.synergies[o].name)){const n=e.synergies[o];n.buildingTie1.name===e.name?t*=1+.05*n.buildingTie2.amount:n.buildingTie2.name===e.name&&(t*=1+.001*n.buildingTie1.amount)}})),e.fortune&&to(e.fortune.name)&&(t*=1.07),e.grandma&&to(e.grandma.name)&&(t*=1+.01*qt.Grandma.amount*(1/(e.id-1))),"object"==typeof e.tieredUpgrades.misfortune&&1===e.vanilla&&to(e.tieredUpgrades.misfortune.name))switch(Game.elderWrath){default:t*=1;break;case 1:t*=1.02;break;case 2:t*=1.04;break;case 3:t*=1.06}return t}function bo(e){const t=Game.Objects[e],o={};return"Cursor"===t.name?o.cps=function(e){let t=0;to("Thousand fingers")&&(t+=.1),to("Million fingers")&&(t*=5),to("Billion fingers")&&(t*=10),to("Trillion fingers")&&(t*=20),to("Quadrillion fingers")&&(t*=20),to("Quintillion fingers")&&(t*=20),to("Sextillion fingers")&&(t*=20),to("Septillion fingers")&&(t*=20),to("Octillion fingers")&&(t*=20),to("Nonillion fingers")&&(t*=20);let o=1,n=0;return Object.keys(qt).forEach((e=>{"Cursor"!==qt[e].name&&(n+=qt[e].amount)})),t*=n,o*=yo(e),o*=Game.magicCpS("Cursor"),o*=eo("cursorCps"),Game.ComputeCps(.1,to("Reinforced index finger")+to("Carpal tunnel prevention cream")+to("Ambidextrous"),t)*o}:"Grandma"===t.name?o.cps=function(e){let o=1;if(Object.keys(Game.GrandmaSynergies).forEach((e=>{to(Game.GrandmaSynergies[e])&&(o*=2)})),to("Bingo center/Research facility")&&(o*=4),to("Ritual rolling pins")&&(o*=2),to("Naughty list")&&(o*=2),to("Elderwort biscuits")&&(o*=1.02),o*=eo("grandmaCps"),to("Cat ladies"))for(let e=0;e<Game.UpgradesByPool.kitten.length;e++)to(Game.UpgradesByPool.kitten[e].name)&&(o*=1.29);o*=yo(e);let n=0;to("One mind")&&(n+=.02*qt.Grandma.amount),to("Communal brainsweep")&&(n+=.02*qt.Grandma.amount),to("Elder Pact")&&(n+=.05*qt.Portal.amount);let i=0;return Object.keys(qt).forEach((e=>{"Grandma"!==qt[e].name&&(i+=qt[e].amount)})),o*=1+.01*Jt("Elder Battalion")*i,o*=Game.magicCpS(t.name),(t.baseCps+n)*o}:o.cps=function(e){let t=1;return t*=yo(e),t*=Game.magicCpS(e.name),e.baseCPS*t},o.baseCps=t.baseCps,o.name=t.name,o.tieredUpgrades=t.tieredUpgrades,o.synergies=t.synergies,o.fortune=t.fortune,o.grandma=t.grandma,o.baseCPS=t.baseCps,o.id=t.id,o.vanilla=t.vanilla,o}function ko(e){const t=Game.Upgrades[e],o={};return o.power=t.power,"function"==typeof o.power&&("Sugar crystal cookies"===t.name?o.power=function(){let e=5;return Object.keys(qt).forEach((t=>{qt[t].level>=10&&(e+=1)})),e}:o.power=function(){let e=2;if(to("Starlove")&&(e=3),Game.hasGod){const t=oo("seasons");1===t?e*=1.3:2===t?e*=1.2:3===t&&(e*=1.1)}return e}),o.pool=t.pool,o.name=t.name,o}function Go(){Oe=Game.UpgradesOwned,Fe=Game.pledges,We=Game.AchievementsOwned,Le=Game.heavenlyPower,Ae=Game.prestige,Object.keys(Game.Objects).forEach((e=>{const t=Game.Objects[e];let o=qt[e];void 0===o&&(qt[e]=bo(e),o=qt[e],po(e)),o.amount=t.amount,o.level=t.level,o.totalCookies=t.totalCookies,o.basePrice=t.basePrice,o.free=t.free,t.minigameLoaded&&("Temple"===t.name&&(He=t.minigame.slot[0],Ue=t.minigame.slot[1],Re=t.minigame.slot[2]),o.minigameLoaded=t.minigameLoaded,o.minigame=t.minigame),qt[e]=o})),Object.keys(Game.Upgrades).forEach((e=>{const t=Game.Upgrades[e];let o=Yt[e];void 0===o&&(Yt[e]=ko(e),o=Yt[e]),o.bought=t.bought,Yt[e]=o})),Object.keys(Game.Achievements).forEach((e=>{const t=Game.Achievements[e];let o=Kt[e];void 0===o&&(Kt[e]=Co(e),o=Kt[e]),o.won=t.won,Kt[e]=o})),fo(),je=_e,De=Xe}function wo(e){const t=document.createElement("div");return t.style.fontWeight="bold",t.id=`${e}Title`,t.className="CMTextBlue",t.textContent=e,t}function vo(e){e.appendChild(wo("Bonus Income"));const t=document.createElement("div");t.style.marginBottom="4px",t.style.color="white",t.id="CMTooltipIncome",e.appendChild(t),e.appendChild(wo("Bonus Cookies per Click")),e.lastChild.style.display="none";const o=document.createElement("div");o.style.marginBottom="4px",o.style.color="white",o.style.display="none",o.id="CMTooltipCookiePerClick",e.appendChild(o),e.appendChild(wo("Payback Period"));const n=document.createElement("div");n.style.marginBottom="4px",n.id="CMTooltipPP",e.appendChild(n),e.appendChild(wo("Time Left"));const i=document.createElement("div");if(i.id="CMTooltipTime",e.appendChild(i),"b"===te){e.appendChild(wo("Production left till next achievement")),e.lastChild.id="CMTooltipProductionHeader";const t=document.createElement("div");t.id="CMTooltipProduction",e.appendChild(t)}}function xo(e,t,o){let n=e,i=0;for(;n<o;)n+=.002*Math.max(.002,(n/Math.max(t,100))**.5),i+=1;return i/Game.fps}function To(){let e=1;return Object.keys(Game.buffs).forEach((t=>{void 0!==Game.buffs[t].multCpS&&(e*=Game.buffs[t].multCpS)})),e}function Mo(){null!==l("CMDispTooltipWarningParent")&&(0===e.ToolWarnPos?(l("CMDispTooltipWarningParent").style.top="auto",l("CMDispTooltipWarningParent").style.margin="4px -4px",l("CMDispTooltipWarningParent").style.padding="3px 4px"):(l("CMDispTooltipWarningParent").style.right="auto",l("CMDispTooltipWarningParent").style.margin="4px",l("CMDispTooltipWarningParent").style.padding="4px 3px"))}function So(){if(Go(),"none"!==l("tooltipAnchor").style.display&&l("CMTooltipArea")){l("CMTooltipArea").innerHTML="";const t=function(){l("tooltip").firstChild.style.paddingBottom="4px";const e=document.createElement("div");return e.style.border="1px solid",e.style.padding="4px",e.style.margin="0px -4px",e.id="CMTooltipBorder",e.className="CMTextGray",e}();l("CMTooltipArea").appendChild(t),"b"===te?function(){if(1===e.TooltipBuildUpgrade&&1===Game.buyMode){let t;if(vo(l("CMTooltipBorder")),1===Game.buyMode?q=t:t=q,1===Game.buyBulk?t=ht:10===Game.buyBulk?t=gt:100===Game.buyBulk&&(t=ft),Z=Game.Objects[oe].bulkPrice,ee=t[oe].bonus,1===e.TooltipBuildUpgrade&&1===Game.buyMode){l("CMTooltipIncome").textContent=se(ee,2);const o=Math.round(ee/Game.cookiesPs*1e4);Number.isFinite(o)&&0!==o?l("CMTooltipIncome").textContent+=` (${o/100}% of income)`:l("CMTooltipIncome").textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`,l("CMTooltipBorder").className=P+t[oe].color,e.PPDisplayTime?l("CMTooltipPP").textContent=ce(t[oe].pp):l("CMTooltipPP").textContent=se(t[oe].pp,2),l("CMTooltipPP").className=P+t[oe].color;const n=de((Z-(Game.cookies+so()))/lo());l("CMTooltipTime").textContent=n.text,"Done!"===n.text&&Game.cookies<t[oe].price?l("CMTooltipTime").textContent=`${n.text} (with Wrink)`:l("CMTooltipTime").textContent=n.text,l("CMTooltipTime").className=P+n.color}l("CMTooltipProductionHeader").style.display="none",l("CMTooltipTime").style.marginBottom="0px";for(const e of Object.keys(Game.Objects[oe].productionAchievs))if(!Game.HasAchiev(Game.Objects[oe].productionAchievs[e].achiev.name)){const t=Game.Objects[oe].productionAchievs[e];l("CMTooltipTime").style.marginBottom="4px",l("CMTooltipProductionHeader").style.display="",l("CMTooltipProduction").className=`ProdAchievement${oe}`,l("CMTooltipProduction").textContent=se(t.pow-qt[oe].totalCookies,15),l("CMTooltipProduction").style.color="white";break}}else l("CMTooltipArea").style.display="none"}():"u"===te?function(){if(vo(l("CMTooltipBorder")),ee=wt[Game.UpgradesInStore[oe].name].bonus,Z=Game.Upgrades[Game.UpgradesInStore[oe].name].getPrice(),ne=wt[Game.UpgradesInStore[oe].name].bonusMouse,1===e.TooltipBuildUpgrade){l("CMTooltipIncome").textContent=se(ee,2);const t=Math.round(ee/Game.cookiesPs*1e4);"0"!==l("CMTooltipIncome").textContent||"b"!==te&&"u"!==te?(Number.isFinite(t)&&0!==t?l("CMTooltipIncome").textContent+=` (${t/100}% of income)`:l("CMTooltipIncome").textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`,l("CMTooltipBorder").className=P+wt[Game.UpgradesInStore[oe].name].color,ne&&(l("CMTooltipCookiePerClick").textContent=se(ne),l("CMTooltipCookiePerClick").style.display="block",l("CMTooltipCookiePerClick").previousSibling.style.display="block"),0===ee&&ne?(l("CMTooltipPP").textContent=`${se(Z/ne)} Clicks`,l("CMTooltipPP").style.color="white"):(e.PPDisplayTime?l("CMTooltipPP").textContent=ce(wt[Game.UpgradesInStore[oe].name].pp):l("CMTooltipPP").textContent=se(wt[Game.UpgradesInStore[oe].name].pp,2),l("CMTooltipPP").className=P+wt[Game.UpgradesInStore[oe].name].color)):(l("Bonus IncomeTitle").style.display="none",l("CMTooltipIncome").style.display="none",l("Payback PeriodTitle").style.display="none",l("CMTooltipPP").style.display="none");const o=de((Z-(Game.cookies+so()))/lo());if(l("CMTooltipTime").textContent=o.text,"Done!"===o.text&&Game.cookies<Game.UpgradesInStore[oe].getPrice()?l("CMTooltipTime").textContent=`${o.text} (with Wrink)`:l("CMTooltipTime").textContent=o.text,l("CMTooltipTime").className=P+o.color,"Chocolate egg"===Game.UpgradesInStore[oe].name){l("CMTooltipBorder").lastChild.style.marginBottom="4px",l("CMTooltipBorder").appendChild(wo("Cookies to be gained (Currently/Max)"));const e=document.createElement("div");e.style.color="white",e.textContent=`${se(.05*Game.cookies)} / ${se(mt)}`,l("CMTooltipBorder").appendChild(e)}}else l("CMTooltipArea").style.display="none"}():"s"===te?function(){if(1===e.TooltipLump){const e=l("CMTooltipBorder");e.appendChild(wo("Current Sugar Lump"));const o=document.createElement("div");o.id="CMTooltipTime",e.appendChild(o);const n=0===(t=Game.lumpCurrentType)?{text:"Normal",color:A}:1===t?{text:"Bifurcated",color:N}:2===t?{text:"Golden",color:O}:3===t?{text:"Meaty",color:F}:4===t?{text:"Caramelized",color:L}:{text:"Unknown Sugar Lump",color:W};o.textContent=n.text,o.className=P+n.color}else l("CMTooltipArea").style.display="none";var t}():"g"===te?function(){const{minigame:t}=Game.Objects["Wizard tower"],o=t.getSpellCost(t.spellsById[oe]);if(1===e.TooltipGrim&&o<=t.magicM){const e=l("CMTooltipBorder");e.appendChild(wo("Time Left"));const n=document.createElement("div");n.id="CMTooltipTime",e.appendChild(n);const i=de(xo(t.magic,t.magicM,o));if(n.textContent=i.text,n.className=P+i.color,o<=t.magic){e.appendChild(wo("Recover Time"));const n=document.createElement("div");n.id="CMTooltipRecover",e.appendChild(n);const i=de(xo(Math.max(0,t.magic-o),t.magicM,t.magic));n.textContent=i.text,n.className=P+i.color}if("0"===oe){e.appendChild(wo("Cookies to be gained/lost"));const t=document.createElement("div");t.id="x",e.appendChild(t);const o=document.createElement("span");o.style.color="#33FF00",o.textContent=se(Math.min(.15*(Game.cookies+so()),60*rt*30),2),t.appendChild(o);const n=document.createElement("span");n.textContent=" / ",t.appendChild(n);const i=document.createElement("span");i.style.color="red",i.textContent=se(60*rt*15,2),t.appendChild(i)}l("CMTooltipArea").appendChild(e)}else l("CMTooltipArea").style.display="none"}():"p"===te?function(){const{minigame:t}=Game.Objects.Farm;if(e.TooltipPlots&&0!==t.plot[oe[1]][oe[0]][0]){const e=t.plot[oe[1]][oe[0]][1]>t.plantsById[t.plot[oe[1]][oe[0]][0]-1].mature,o=t.plantsById[t.plot[oe[1]][oe[0]][0]-1].name;l("CMTooltipBorder").appendChild(wo("Reward (Current / Maximum)"));const n=document.createElement("div");n.id="CMTooltipPlantReward",l("CMTooltipBorder").appendChild(n),"Bakeberry"===o?l("CMTooltipPlantReward").textContent=`${e?se(Math.min(.03*Game.cookies,60*Game.cookiesPs*30)):"0"} / ${se(60*Game.cookiesPs*30)}`:"Chocoroot"===o||"White chocoroot"===o?l("CMTooltipPlantReward").textContent=`${e?se(Math.min(.03*Game.cookies,60*Game.cookiesPs*3)):"0"} / ${se(60*Game.cookiesPs*3)}`:"Queenbeet"===o?l("CMTooltipPlantReward").textContent=`${e?se(Math.min(.04*Game.cookies,60*Game.cookiesPs*60)):"0"} / ${se(60*Game.cookiesPs*60)}`:"Duketater"===o?l("CMTooltipPlantReward").textContent=`${e?se(Math.min(.08*Game.cookies,60*Game.cookiesPs*120)):"0"} / ${se(60*Game.cookiesPs*120)}`:l("CMTooltipArea").style.display="none"}else l("CMTooltipArea").style.display="none"}():"ha"===te?function(){const{minigame:t}=Game.Objects.Farm;if(e.TooltipLump){l("CMTooltipBorder").appendChild(wo("Cookies gained from harvesting:"));let e=0,o=0;Game.keys[16]&&Game.keys[17]&&(o=1);for(let n=0;n<6;n++)for(let i=0;i<6;i++)if(t.plot[n][i][0]>=1){const a=t.plot[n][i],r=t.plantsById[a[0]-1],l=r.name;let s=!0;o&&r.immortal&&(s=!1),a[1]<r.matureBase&&(s=!1),s&&"Bakeberry"===l?e+=Math.min(.03*Game.cookies,60*Game.cookiesPs*30):s&&"Chocoroot"===l||"White chocoroot"===l?e+=Math.min(.03*Game.cookies,60*Game.cookiesPs*3):s&&"Queenbeet"===l?e+=Math.min(.04*Game.cookies,60*Game.cookiesPs*60):s&&"Duketater"===l&&(e+=Math.min(.08*Game.cookies,60*Game.cookiesPs*120))}l("CMTooltipBorder").appendChild(document.createTextNode(se(e)))}else l("CMTooltipArea").style.display="none"}():"wb"===te?function(){l("tooltip").innerHTML="",l("tooltip").appendChild(wo("Reward:"));const e=document.createElement("div");e.id="CMWrinklerReward","PopAll"===oe?e.textContent=se(Ct):"PopFattest"===oe&&(e.textContent=se(bt[0])),l("tooltip").appendChild(e)}():("pag"===te||"pas"===te&&-1!==oe[1])&&function(){if(1===e.TooltipPantheon){const t=l("CMTooltipBorder");let o;o="pas"===te?oe[1]:oe,t.appendChild(wo("Effect in position 1:"));const n=document.createElement("div");if(n.id="CMPantheonTooltipPosition1",0!==Qt[o][0]){n.textContent=se(Qt[o][0]);const t=Math.round(Qt[o][0]/Game.cookiesPs*1e4);Number.isFinite(t)&&0!==t?n.textContent+=` (${t/100}% of income)`:n.textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`}else n.textContent="No effect to CPS";t.appendChild(n),t.appendChild(wo("Effect in position 2:"));const i=document.createElement("div");if(i.id="CMPantheonTooltipPosition2",0!==Qt[o][1]){i.textContent=se(Qt[o][1]);const t=Math.round(Qt[o][1]/Game.cookiesPs*1e4);Number.isFinite(t)&&0!==t?i.textContent+=` (${t/100}% of income)`:i.textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`}else i.textContent="No effect to CPS";t.appendChild(i),t.appendChild(wo("Effect in position 3:"));const a=document.createElement("div");if(a.id="CMPantheonTooltipPosition2",0!==Qt[o][2]){a.textContent=se(Qt[o][2]);const t=Math.round(Qt[o][2]/Game.cookiesPs*1e4);Number.isFinite(t)&&0!==t?a.textContent+=` (${t/100}% of income)`:a.textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`}else a.textContent="No effect to CPS";t.appendChild(a),l("CMTooltipArea").appendChild(t)}else l("CMTooltipArea").style.display="none"}(),function(){if("b"===te||"u"===te){null===document.getElementById("CMDispTooltipWarningParent")&&(l("tooltipAnchor").appendChild(function(){const t=document.createElement("div");t.style.position="absolute",t.style.display="block",t.style.left="auto",t.style.bottom="auto",t.id="CMDispTooltipWarningParent";const o=function(e,t,o,n,i){const a=document.createElement("div");a.id=e,a.style.display="none",a.style.transition="opacity 0.1s ease-out",a.className="CMBorder"+t,a.style.padding="2px",a.style.background="#000 url(img/darkNoise.png)";const r=document.createElement("div");a.appendChild(r);const l=document.createElement("span");l.className=P+t,l.style.fontWeight="bold",l.textContent=o,r.appendChild(l),r.appendChild(document.createTextNode(n));const s=document.createElement("div");a.appendChild(s);const c=document.createElement("span");return c.id=i,s.appendChild(document.createTextNode("Deficit: ")),s.appendChild(c),a};return t.appendChild(o("CMDispTooltipWarnLucky",W,"Warning: ",'Purchase of this item will put you under the number of Cookies required for "Lucky!"',"CMDispTooltipWarnLuckyText")),t.firstChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnLuckyFrenzy",O,"Warning: ",'Purchase of this item will put you under the number of Cookies required for "Lucky!" (Frenzy)',"CMDispTooltipWarnLuckyFrenzyText")),t.lastChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnConjure",L,"Warning: ",'Purchase of this item will put you under the number of Cookies required for "Conjure Baked Goods"',"CMDispTooltipWarnConjureText")),t.lastChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnConjureFrenzy",L,"Warning: ",'Purchase of this item will put you under the number of Cookies required for "Conjure Baked Goods" (Frenzy)',"CMDispTooltipWarnConjureFrenzyText")),t.lastChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnEdifice",L,"Warning: ",'Purchase of this item will put you under the number of Cookies needed for "Spontaneous Edifice" to possibly give you your most expensive building"',"CMDispTooltipWarnEdificeText")),t.lastChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnUser",W,"Warning: ",`Purchase of this item will put you under the number of Cookies equal to ${e.ToolWarnUser} seconds of CPS`,"CMDispTooltipWarnUserText")),t}()),Mo()),0===e.ToolWarnPos?l("CMDispTooltipWarningParent").style.right="0px":l("CMDispTooltipWarningParent").style.top=`${l("tooltip").offsetHeight}px`,l("CMDispTooltipWarningParent").style.width=l("tooltip").offsetWidth-6+"px";const t=Game.cookies+so()-Z,o=e.ToolWarnBon?ee:0;let n=Ye;if(1===e.ToolWarnBon){let e=ee;e/=To(),n+=60*e*15/.15}if(1===e.ToolWarnLucky&&t<n&&("b"!==te||1===Game.buyMode)?(l("CMDispTooltipWarnLucky").style.display="",l("CMDispTooltipWarnLuckyText").textContent=`${se(n-t)} (${ce((n-t)/(lo()+o))})`):l("CMDispTooltipWarnLucky").style.display="none",1===e.ToolWarnLuckyFrenzy){const e=7*n;t<e&&("b"!==te||1===Game.buyMode)?(l("CMDispTooltipWarnLuckyFrenzy").style.display="",l("CMDispTooltipWarnLuckyFrenzyText").textContent=`${se(e-t)} (${ce((e-t)/(lo()+o))})`):l("CMDispTooltipWarnLuckyFrenzy").style.display="none"}else l("CMDispTooltipWarnLuckyFrenzy").style.display="none";if(1===e.ToolWarnConjure){const e=2*n;t<e&&("b"!==te||1===Game.buyMode)?(l("CMDispTooltipWarnConjure").style.display="",l("CMDispTooltipWarnConjureText").textContent=`${se(e-t)} (${ce((e-t)/(lo()+o))})`):l("CMDispTooltipWarnConjure").style.display="none"}else l("CMDispTooltipWarnConjure").style.display="none";if(1===e.ToolWarnConjureFrenzy){const e=2*n*7;t<e&&("b"!==te||1===Game.buyMode)?(l("CMDispTooltipWarnConjureFrenzy").style.display="",l("CMDispTooltipWarnConjureFrenzyText").textContent=`${se(e-t)} (${ce((e-t)/(lo()+o))})`):l("CMDispTooltipWarnConjureFrenzy").style.display="none"}else l("CMDispTooltipWarnConjureFrenzy").style.display="none";1===e.ToolWarnEdifice&&Game.Objects["Wizard tower"].minigameLoaded&&it&&t<it&&("b"!==te||1===Game.buyMode)?(l("CMDispTooltipWarnEdifice").style.display="",l("CMDispTooltipWarnEdificeText").textContent=`${se(it-t)} (${ce((it-t)/(lo()+o))})`):l("CMDispTooltipWarnEdifice").style.display="none",e.ToolWarnUser>0&&t<e.ToolWarnUser*lo()&&("b"!==te||1===Game.buyMode)?(l("CMDispTooltipWarnUser").style.display="",l("CMDispTooltipWarnUser").children[0].textContent=`Purchase of this item will put you under the number of Cookies equal to ${e.ToolWarnUser} seconds of CPS`,l("CMDispTooltipWarnUserText").textContent=`${se(e.ToolWarnUser*lo()-t)} (${ce((e.ToolWarnUser*lo()-t)/(lo()+o))})`):l("CMDispTooltipWarnUser").style.display="none"}else null!==l("CMDispTooltipWarningParent")&&l("CMDispTooltipWarningParent").remove()}()}else null===l("CMTooltipArea")&&null!==l("CMDispTooltipWarningParent")&&l("CMDispTooltipWarningParent").remove()}function Po(){if(Game.prefs.autosave&&Game.drawT%10==0&&"stats"===Game.onMenu&&e.Stats){const e=document.getElementById("CMStatsAutosaveTimer");e&&(e.innerText=Game.sayTime(60*Game.fps-Game.T%(60*Game.fps),4))}if(ao(),ro(),function(){if(1===e.TimerBar){const t=l("CMTimerBar").offsetWidth-163,o=l("CMTimerBar").offsetWidth-133;let n=0;0!==Game.shimmerTypes.golden.spawned||Game.Has("Golden switch [off]")?l("CMTimerBarGC").style.display="none":(l("CMTimerBarGC").style.display="",l("CMTimerBarGCMinBar").style.width=`${Math.round(Math.max(0,Game.shimmerTypes.golden.minTime-Game.shimmerTypes.golden.time)*t/Game.shimmerTypes.golden.maxTime)}px`,e.TimerBarOverlay>=1?l("CMTimerBarGCMinBar").textContent=Math.ceil((Game.shimmerTypes.golden.minTime-Game.shimmerTypes.golden.time)/Game.fps):l("CMTimerBarGCMinBar").textContent="",Game.shimmerTypes.golden.minTime===Game.shimmerTypes.golden.maxTime?(l("CMTimerBarGCMinBar").style.borderTopRightRadius="10px",l("CMTimerBarGCMinBar").style.borderBottomRightRadius="10px"):(l("CMTimerBarGCMinBar").style.borderTopRightRadius="",l("CMTimerBarGCMinBar").style.borderBottomRightRadius=""),l("CMTimerBarGCBar").style.width=`${Math.round(Math.min(Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.minTime,Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.time)*t/Game.shimmerTypes.golden.maxTime)}px`,e.TimerBarOverlay>=1?l("CMTimerBarGCBar").textContent=Math.ceil(Math.min(Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.minTime,Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.time)/Game.fps):l("CMTimerBarGCBar").textContent="",l("CMTimerBarGCTime").textContent=Math.ceil((Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.time)/Game.fps),n+=1),"christmas"===Game.season&&0===Game.shimmerTypes.reindeer.spawned?(l("CMTimerBarRen").style.display="",l("CMTimerBarRenMinBar").style.width=`${Math.round(Math.max(0,Game.shimmerTypes.reindeer.minTime-Game.shimmerTypes.reindeer.time)*t/Game.shimmerTypes.reindeer.maxTime)}px`,e.TimerBarOverlay>=1?l("CMTimerBarRenMinBar").textContent=Math.ceil((Game.shimmerTypes.reindeer.minTime-Game.shimmerTypes.reindeer.time)/Game.fps):l("CMTimerBarRenMinBar").textContent="",l("CMTimerBarRenBar").style.width=`${Math.round(Math.min(Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.minTime,Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.time)*t/Game.shimmerTypes.reindeer.maxTime)}px`,e.TimerBarOverlay>=1?l("CMTimerBarRenBar").textContent=Math.ceil(Math.min(Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.minTime,Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.time)/Game.fps):l("CMTimerBarRenBar").textContent="",l("CMTimerBarRenTime").textContent=Math.ceil((Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.time)/Game.fps),n+=1):l("CMTimerBarRen").style.display="none";const i={};l("CMTimerBarBuffTimers").innerHTML="",Object.keys(Game.buffs).forEach((t=>{if(Game.buffs[t]){const a=co(Game.buffs[t].name,Game.buffs[t].name,[{id:`${Game.buffs[t].name}Bar`}]);a.style.display="";let r="";r=void 0!==J[Game.buffs[t].name]?J[Game.buffs[t].name]:L,a.lastChild.children[1].className=B+r,a.lastChild.children[1].style.color="black",2===e.TimerBarOverlay?a.lastChild.children[1].textContent=`${Math.round(Game.buffs[t].time/Game.buffs[t].maxTime*100)}%`:a.lastChild.children[1].textContent="",a.lastChild.children[1].style.width=`${Math.round(Game.buffs[t].time*(o-8*Math.ceil(Game.buffs[t].time/Game.fps).toString().length)/Game.buffs[t].maxTime)}px`,a.lastChild.children[2].textContent=Math.ceil(Game.buffs[t].time/Game.fps),n+=1,i[Game.buffs[t].name]=a}})),Object.keys(i).forEach((e=>{l("CMTimerBarBuffTimers").appendChild(i[e])})),0!==n&&(l("CMTimerBar").style.height=12*n+2+"px"),ae!==n&&(ae=n,ho())}}(),uo(),So(),function(){if(1===e.TooltipWrink&&1===z){let e=!1;Object.keys(Game.wrinklers).forEach((t=>{const o=Game.wrinklers[t];if(o.phase>0&&o.selected){if(e=!0,0===_[t]||void 0===_[t]){const e=document.createElement("div"),o=document.createElement("div");o.style.minWidth="120px",o.style.marginBottom="4px";const n=document.createElement("div");n.style.textAlign="center",n.id="CMTooltipWrinkler",o.appendChild(n),e.appendChild(o),Game.tooltip.draw(this,escape(e.innerHTML)),V=t,_[t]=1}}else _[t]=0})),e||Game.tooltip.hide()}}(),function(){if(1===e.TooltipWrink&&null!==l("CMTooltipWrinkler")){let{sucked:e}=Game.wrinklers[V],t=1.1;if(Game.Has("Sacrilegious corruption")&&(t*=1.05),1===Game.wrinklers[V].type&&(t*=3),e*=t,Game.Has("Wrinklerspawn")&&(e*=1.05),qt.Temple.minigameLoaded){const t=Game.hasGod("scorn");1===t?e*=1.15:2===t?e*=1.1:3===t&&(e*=1.05)}l("CMTooltipWrinkler").textContent=se(e)}}(),e.UpStats&&"stats"===Game.onMenu&&(Game.drawT-1)%(5*Game.fps)!=0&&(Game.drawT-1)%Game.fps==0&&Game.UpdateMenu(),t(),e.Scale){let e=l("cookies").innerHTML.replace(/.*(?=<br>)/i,se(Game.cookies));Game.prefs.monospace&&(e=`<span class="monospace">${e}</span>`),l("cookies").innerHTML=e}}class Bo{constructor(e){this.maxLength=e,this.queue=[]}addLatest(e){this.queue.push(e)>this.maxLength&&this.queue.shift()}calcAverage(e){let t=e;t>this.maxLength&&(t=this.maxLength),t>this.queue.length&&(t=this.queue.length);let o=0;for(let e=this.queue.length-1;e>=0&&e>this.queue.length-1-t;e--)o+=this.queue[e];return 0===o?0:o/t}}function Eo(){const t=Math.floor(Date.now()/1e3);if(Game.T/Game.fps%1==0){let o=Game.cookies+Xt;Game.cpsSucked>0&&(o+=Ct),Ht=Math.max(Game.cookiesEarned,o),o*=.05;const n=t-Ce,i=Math.max(0,Game.cookies-ye)/n,a=Math.max(0,Ct-be)/n,r=Math.max(0,bt[0]-ke)/n,l=Math.max(0,o-mt)/n,s=(Game.cookieClicks-Ge)/n;for(let e=0;e<n;e++)pe.addLatest(i),ue.addLatest(a),me.addLatest(r),he.addLatest(l),ge.addLatest(s);Ce=t,ye=Game.cookies,be=Ct,ke=bt[0],mt=o,Ge=Game.cookieClicks;const c=Y[e.AvgCPSHist];we=pe.calcAverage(c),ve=ue.calcAverage(c),xe=me.calcAverage(c),Te=he.calcAverage(c),Me=we,1===e.CalcWrink&&(Me+=ve),2===e.CalcWrink&&(Me+=xe);const d=Game.HasUnlocked("Chocolate egg")&&!Game.Has("Chocolate egg");Ut=d||0===e.CalcWrink?we+ve+(d?Te:0):Me,vt=ge.calcAverage(K[e.AvgClicksHist])}}function No(t,o){let n="";return n=t.pp<=0||t.pp===1/0?A:t.pp<zt?E:t.pp===zt?N:t.pp<Vt[10][0]?O:t.pp<Vt[20][0]?F:t.pp<Vt[30][0]?W:L,0!==Number(e.PPSecondsLowerLimit)&&o/lo()<Number(e.PPSecondsLowerLimit)&&(n=E),e.PPOnlyConsiderBuyable&&o-Game.cookies>0&&(n=W),n}function Oo(t,o){Object.keys(t).forEach((n=>{if(e.PPRigidelMode&&1===o)t[n].color=A;else{t[n].color=No(t[n],Game.Objects[n].getSumPrice(o));for(let o=0;o<e.PPExcludeTop;o++)t[n].pp===Vt[o][0]&&(t[n].color=A)}}))}function Fo(t,o){Object.keys(t).forEach((n=>{const i=Game.Objects[n].getSumPrice(o);Game.cookiesPs?t[n].pp=Math.max(i-(Game.cookies+so()),0)/Game.cookiesPs+i/t[n].bonus:t[n].pp=i/t[n].bonus,e.PPRigidelMode&&1===o||Vt.push([t[n].pp,o])}))}function Wo(){!function(){zt=1/0,Vt=[],void 0===e.PPExcludeTop&&(e.PPExcludeTop=0),Fo(ht,1),Fo(gt,10),Fo(ft,100),Vt.sort(((e,t)=>e[0]-t[0]));let t=0;if(e.PPOnlyConsiderBuyable)for(;Vt[t][1]>Game.cookies&&(t+=1,Vt.length!==t+1););zt=Vt[e.PPExcludeTop][t],Oo(ht,1),Oo(gt,10),Oo(ft,100)}(),Object.keys(wt).forEach((e=>{Game.cookiesPs?wt[e].pp=Math.max(Game.Upgrades[e].getPrice()-(Game.cookies+so()),0)/Game.cookiesPs+Game.Upgrades[e].getPrice()/wt[e].bonus:wt[e].pp=Game.Upgrades[e].getPrice()/wt[e].bonus,Number.isNaN(wt[e].pp)&&(wt[e].pp=1/0),wt[e].color=No(wt[e],Game.Upgrades[e].getPrice())}))}function Lo(e,t,o,n,i){let a=o,r=0;for(let o=0;o<i;o+=1){let o=t*Game.priceIncrease**Math.max(0,a-n);o=Game.modifyBuildingPrice(e,o),o=Math.ceil(o),r+=o,a+=1}return r}function Ao(e){Kt[e]&&0===Kt[e].won&&(Kt[e].won=1,"shadow"!==Game.Achievements[e].pool&&(We+=1))}function jo(){ze=0;let e=1;const t={};Object.keys(Game.Objects).forEach((e=>{if(Game.Objects[e].minigameLoaded&&Game.Objects[e].minigame.effs){const o=Game.Objects[e].minigame.effs;Object.keys(o).forEach((e=>{t[e]?t[e]*=o[e]:t[e]=o[e]}))}})),Ie=t,1!==Game.ascensionMode&&(e+=.01*parseFloat(Ae)*Le*function(){let e=0;if(to("Heavenly chip secret")&&(e+=.05),to("Heavenly cookie stand")&&(e+=.2),to("Heavenly bakery")&&(e+=.25),to("Heavenly confectionery")&&(e+=.25),to("Heavenly key")&&(e+=.25),e*=1+.05*Jt("Dragon God"),to("Lucky digit")&&(e*=1.01),to("Lucky number")&&(e*=1.01),to("Lucky payout")&&(e*=1.01),Game.hasGod){const t=oo("creation");1===t?e*=.7:2===t?e*=.8:3===t&&(e*=.9)}return e}()),e*=eo("cps"),to("Heralds")&&1!==Game.ascensionMode&&(e*=1+.01*Game.heralds),Object.keys(Game.cookieUpgrades).forEach((t=>{const o=Game.cookieUpgrades[t];to(o.name)&&("function"==typeof o.power?e*=1+.01*Yt[o.name].power(Yt[o.name]):e*=1+.01*o.power)})),to("Specialized chocolate chips")&&(e*=1.01),to("Designer cocoa beans")&&(e*=1.02),to("Underworld ovens")&&(e*=1.03),to("Exotic nuts")&&(e*=1.04),to("Arcane sugar")&&(e*=1.05),to("Increased merriness")&&(e*=1.15),to("Improved jolliness")&&(e*=1.15),to("A lump of coal")&&(e*=1.01),to("An itchy sweater")&&(e*=1.01),to("Santa's dominion")&&(e*=1.2),to("Fortune #100")&&(e*=1.01),to("Fortune #101")&&(e*=1.07),to("Dragon scale")&&(e*=1.03);let o=1;if(oo){let t=oo("asceticism");1===t?e*=1.15:2===t?e*=1.1:3===t&&(e*=1.05),t=oo("ages"),1===t?e*=1+.15*Math.sin(Date.now()/1e3/10800*Math.PI*2):2===t?e*=1+.15*Math.sin(Date.now()/1e3/43200*Math.PI*2):3===t&&(e*=1+.15*Math.sin(Date.now()/1e3/86400*Math.PI*2)),t=oo("decadence"),1===t?o*=.93:2===t?o*=.95:3===t&&(o*=.98),t=oo("industry"),1===t?o*=1.1:2===t?o*=1.06:3===t&&(o*=1.03),t=oo("labor"),1===t?o*=.97:2===t?o*=.98:3===t&&(o*=.99)}to("Santa's legacy")&&(e*=1+.03*(Game.santaLevel+1));const n=We/25;let i=1;if(to("Santa's milk and cookies")&&(i*=1.05),i*=1+.05*Jt("Breath of Milk"),oo){const e=oo("mother");1===e?i*=1.1:2===e?i*=1.05:3===e&&(i*=1.03)}i*=eo("milk");let a=1;to("Kitten helpers")&&(a*=1+.1*n*i),to("Kitten workers")&&(a*=1+.125*n*i),to("Kitten engineers")&&(a*=1+.15*n*i),to("Kitten overseers")&&(a*=1+.175*n*i),to("Kitten managers")&&(a*=1+.2*n*i),to("Kitten accountants")&&(a*=1+.2*n*i),to("Kitten specialists")&&(a*=1+.2*n*i),to("Kitten experts")&&(a*=1+.2*n*i),to("Kitten consultants")&&(a*=1+.2*n*i),to("Kitten assistants to the regional manager")&&(a*=1+.175*n*i),to("Kitten marketeers")&&(a*=1+.15*n*i),to("Kitten analysts")&&(a*=1+.125*n*i),to("Kitten executives")&&(a*=1+.115*n*i),to("Kitten angels")&&(a*=1+.1*n*i),to("Fortune #103")&&(a*=1+.05*n*i),Object.keys(qt).forEach((e=>{const t=qt[e];let a=t.cps(t);1!==Game.ascensionMode&&(a*=(1+.01*t.level)*o),"Grandma"===t.name&&to("Milkhelp&reg; lactose intolerance relief tablets")&&(a*=1+.05*n*i),ze+=t.amount*a})),to('"egg"')&&(ze+=9),e*=a;let r=1;if(to("Chicken egg")&&(r*=1.01),to("Duck egg")&&(r*=1.01),to("Turkey egg")&&(r*=1.01),to("Quail egg")&&(r*=1.01),to("Robin egg")&&(r*=1.01),to("Ostrich egg")&&(r*=1.01),to("Cassowary egg")&&(r*=1.01),to("Salmon roe")&&(r*=1.01),to("Frogspawn")&&(r*=1.01),to("Shark egg")&&(r*=1.01),to("Turtle egg")&&(r*=1.01),to("Ant larva")&&(r*=1.01),to("Century egg")){let e=10*Math.floor((Date.now()-Game.startDate)/1e3/10)/60/60/24;e=Math.min(e,100),pt=1+.1*(1-(1-e/100)**3),r*=pt}e*=r,to("Sugar baking")&&(e*=1+.01*Math.min(100,Game.lumps)),e*=1+Jt("Radiant Appetite");const l=ze*e;Object.keys(Game.CpsAchievements).forEach((e=>{l>=Game.CpsAchievements[e].threshold&&Ao(Game.CpsAchievements[e].name)})),Ve=l;const{n:s}=Game.shimmerTypes.golden,c=Jt("Dragon's Fortune");for(let t=0;t<s;t++)e*=1+1.23*c;const d=Game.bakeryName.toLowerCase();if("orteil"===d?e*=.99:"ortiel"===d&&(e*=.98),to("Elder Covenant")&&(e*=.95),to("Golden switch [off]")){let t=1.5;if(to("Residual luck")){const e=Game.goldenCookieUpgrades;Object.keys(e).forEach((o=>{to(e[o])&&(t+=.1)}))}e*=t}if(to("Shimmering veil [off]")){let t=.5;to("Reinforced membrane")&&(t+=.1),e*=1+t}to("Magic shenanigans")&&(e*=1e3),to("Occult obstruction")&&(e*=0),ze=Game.runModHookOnValue("cps",ze),e*=To(),ze*=e}const Do=["Fortune #001","Fortune #002","Fortune #003","Fortune #004","Fortune #005","Fortune #006","Fortune #007","Fortune #008","Fortune #009","Fortune #010","Fortune #011","Fortune #012","Fortune #013","Fortune #014","Fortune #015","Fortune #016","Fortune #017","Fortune #018","Fortune #100","Fortune #101","Fortune #102","Fortune #103","Fortune #104"],Ho=["Skull cookies","Ghost cookies","Bat cookies","Slime cookies","Pumpkin cookies","Eyeball cookies","Spider cookies"],Uo=["Christmas tree biscuits","Snowflake biscuits","Snowman biscuits","Holly biscuits","Candy cane biscuits","Bell biscuits","Present biscuits"],Ro=["Pure heart biscuits","Ardent heart biscuits","Sour heart biscuits","Weeping heart biscuits","Golden heart biscuits","Eternal heart biscuits","Prism heart biscuits"],$o=["Elderwort biscuits","Bakeberry cookies","Duketater cookies","Green yeast digestives","Wheat slims","Fern tea","Ichor syrup"];function Io(e){return Kt[e]?Kt[e].won:0}function zo(){let e=0;Object.keys(Game.GrandmaSynergies).forEach((t=>{to(Game.GrandmaSynergies[t])&&(e+=1)})),!Io("Elder")&&e>=7&&Ao("Elder"),!Io("Veteran")&&e>=14&&Ao("Veteran");let t=0,o=1,n=1,i=1e5;Object.keys(qt).forEach((e=>{t+=qt[e].amount,i=Math.min(qt[e].amount,i),Io("Mathematician")||qt[e].amount<Math.min(128,2**(Game.ObjectsById.length-Game.Objects[e].id-1))&&(o=0),Io("Base 10")||qt[e].amount<10*(Game.ObjectsById.length-Game.Objects[e].id)&&(n=0)})),i>=1&&Ao("One with everything"),1===o&&Ao("Mathematician"),1===n&&Ao("Base 10"),i>=100&&Ao("Centennial"),i>=150&&Ao("Centennial and a half"),i>=200&&Ao("Bicentennial"),i>=250&&Ao("Bicentennial and a half"),i>=300&&Ao("Tricentennial"),i>=350&&Ao("Tricentennial and a half"),i>=400&&Ao("Quadricentennial"),i>=450&&Ao("Quadricentennial and a half"),i>=500&&Ao("Quincentennial"),i>=550&&Ao("Quincentennial and a half"),i>=600&&Ao("Sexcentennial"),t>=100&&Ao("Builder"),t>=500&&Ao("Architect"),t>=1e3&&Ao("Engineer"),t>=2e3&&Ao("Lord of Constructs"),t>=4e3&&Ao("Grand design"),t>=8e3&&Ao("Ecumenopolis"),Oe>=20&&Ao("Enhancer"),Oe>=50&&Ao("Augmenter"),Oe>=100&&Ao("Upgrader"),Oe>=200&&Ao("Lord of Progress"),Oe>=300&&Ao("The full picture"),Oe>=400&&Ao("When there's nothing left to add"),t>=4e3&&Oe>=300&&Ao("Polymath"),t>=8e3&&Oe>=400&&Ao("Renaissance baker"),qt.Cursor.amount+qt.Grandma.amount>=777&&Ao("The elder scrolls");let a=!0;Object.keys(Ho).forEach((e=>{to(Ho[e])||(a=!1)})),a&&Ao("Spooky cookies");let r=!0;if(Object.keys(Uo).forEach((e=>{to(Uo[e])||(r=!1)})),r&&Ao("Let it snow"),to("Fortune cookies")){const e=Game.Tiers.fortune.upgrades;let t=0;Object.keys(e).forEach((o=>{to(e[o].name)&&(t+=1)})),t>=e.length&&Ao("O Fortuna")}}function Vo(e){const t={};return Object.keys(Game.Objects).forEach((o=>{t[o]={},t[o].bonus=function(e,t){Go(),qt[e].amount+=t;const o=qt[e];"Cursor"===e?(o.amount>=1&&Ao("Click"),o.amount>=2&&Ao("Double-click"),o.amount>=50&&Ao("Mouse wheel"),o.amount>=100&&Ao("Of Mice and Men"),o.amount>=200&&Ao("The Digital"),o.amount>=300&&Ao("Extreme polydactyly"),o.amount>=400&&Ao("Dr. T"),o.amount>=500&&Ao("Thumbs, phalanges, metacarpals"),o.amount>=600&&Ao("With her finger and her thumb"),o.amount>=700&&Ao("Gotta hand it to you"),o.amount>=800&&Ao("The devil's workshop")):Object.keys(Game.Objects[o.name].tieredAchievs).forEach((e=>{o.amount>=Game.Tiers[Game.Objects[o.name].tieredAchievs[e].tier].achievUnlock&&Ao(Game.Objects[o.name].tieredAchievs[e].name)}));const n=We;return jo(),zo(),n!==We&&jo(),ze-Game.cookiesPs}(o,e),1!==e&&(Be=1)})),t}function _o(){Object.keys(Game.Objects).forEach((e=>{ht[e].price=Lo(Game.Objects[e],Game.Objects[e].basePrice,Game.Objects[e].amount,Game.Objects[e].free,1),gt[e].price=Lo(Game.Objects[e],Game.Objects[e].basePrice,Game.Objects[e].amount,Game.Objects[e].free,10),ft[e].price=Lo(Game.Objects[e],Game.Objects[e].basePrice,Game.Objects[e].amount,Game.Objects[e].free,100)}))}function Xo(){ht=Vo(1),gt=Vo(10),ft=Vo(100),wt=[],Object.keys(Game.Upgrades).forEach((t=>{const o=function(e){if("toggle"===Game.Upgrades[e].pool||0===Game.Upgrades[e].bought&&Game.Upgrades[e].unlocked&&"prestige"!==Game.Upgrades[e].pool){Go(),"Shimmering veil [on]"===Yt[e].name?Yt["Shimmering veil [off]"].bought=0:"Golden switch [on]"===Yt[e].name?Yt["Golden switch [off]"].bought=0:Yt[e].bought=(Yt[e].bought+1)%2,Yt[e],Game.CountsAsUpgradeOwned(Game.Upgrades[e].pool)&&(Oe+=1),"Elder Pledge"===e?(Fe+=1,Fe>0&&Ao("Elder nap"),Fe>=5&&Ao("Elder slumber")):"Elder Covenant"===e?Ao("Elder calm"):"Prism heart biscuits"===e?Ao("Lovely cookies"):"Heavenly key"===e&&Ao("Wholesome");const t=We;jo(),zo(),t!==We&&jo();const o=function(){let e=0;to("Thousand fingers")&&(e+=.1),to("Million fingers")&&(e*=5),to("Billion fingers")&&(e*=10),to("Trillion fingers")&&(e*=20),to("Quadrillion fingers")&&(e*=20),to("Quintillion fingers")&&(e*=20),to("Sextillion fingers")&&(e*=20),to("Septillion fingers")&&(e*=20),to("Octillion fingers")&&(e*=20),to("Nonillion fingers")&&(e*=20);let t=0;Object.keys(qt).forEach((e=>{t+=qt[e].amount})),t-=qt.Cursor.amount,e*=t,to("Plastic mouse")&&(e+=.01*ze),to("Iron mouse")&&(e+=.01*ze),to("Titanium mouse")&&(e+=.01*ze),to("Adamantium mouse")&&(e+=.01*ze),to("Unobtainium mouse")&&(e+=.01*ze),to("Eludium mouse")&&(e+=.01*ze),to("Wishalloy mouse")&&(e+=.01*ze),to("Fantasteel mouse")&&(e+=.01*ze),to("Nevercrack mouse")&&(e+=.01*ze),to("Armythril mouse")&&(e+=.01*ze),to("Technobsidian mouse")&&(e+=.01*ze),to("Plasmarble mouse")&&(e+=.01*ze),to("Miraculite mouse")&&(e+=.01*ze),to("Fortune #104")&&(e+=.01*ze);let o=1;if(to("Santa's helpers")&&(o*=1.1),to("Cookie egg")&&(o*=1.1),to("Halo gloves")&&(o*=1.1),to("Dragon claw")&&(o*=1.03),to("Aura gloves")&&(o*=1+.05*Math.min(Game.Objects.Cursor.level,to("Luminous gloves")?20:10)),o*=eo("click"),qt.Temple.minigameLoaded&&oo){const e=oo("labor");1===e?o*=1.15:2===e?o*=1.1:3===e&&(o*=1.05)}Object.keys(Game.buffs).forEach((e=>{void 0!==Game.buffs[e].multClick&&(o*=Game.buffs[e].multClick)})),o*=1+.05*Jt("Dragon Cursor");let n=o*Game.ComputeCps(1,to("Reinforced index finger")+to("Carpal tunnel prevention cream")+to("Ambidextrous"),e);return n=Game.runModHookOnValue("cookiesPerClick",n),Game.hasBuff("Cursed finger")&&(n=Game.buffs["Cursed finger"].power),n}()-Game.computedMouseCps;return o?[ze-Game.cookiesPs,o]:[ze-Game.cookiesPs]}return[]}(t);"Elder Pledge"===t?(wt[t]={bonus:Game.cookiesPs-we},1===e.CalcWrink?wt[t].bonus-=ve:2===e.CalcWrink&&(wt[t].bonus-=xe),Number.isFinite(wt[t].bonus)||(wt[t].bonus=0)):(wt[t]={},o[0]&&(wt[t].bonus=o[0]),o[1]&&(wt[t].bonusMouse=o[1]))}))}function Qo(e,t,o){let n=0,i=0,a=0,r=0,l=1+Math.max(0,Math.ceil(Math.log(Game.cookies)/Math.LN10)-10);for(;a<t*o;)i=Math.max(e,Math.min(Math.floor(1/9*10**l*e*o),t*o)),a=Math.max(e,Math.min(Math.floor(1/9*10**(l+1)*e*o),t*o)),r=Math.floor(1/9*10**(l+1)*e*o),n+=i,l+=1;return[n,i,r]}function qo(){let e=60*rt*60*6*lt;const t=To();t>0?e/=t:e=0,Bt=Qo(7,e,st),St=2*Bt[1]/st,Pt=Bt[2]/60/60/6/lt,Ot=Qo(6,e,ct),Et=2*Ot[1]/ct,Nt=Ot[2]/60/60/6/lt,Lt=Qo(7,7*e,st),Ft=2*Lt[1]/st,Wt=Lt[2]/60/60/6/lt,Dt=Qo(6,7*e,ct),At=2*Dt[1]/ct,jt=Dt[2]/60/60/6/lt}function Yo(){const e=Math.floor(Date.now()/1e3);if(Game.T/Game.fps%1==0){const t=Game.HowMuchPrestige(Game.cookiesReset),o=Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned))-Math.floor(t),n=e-Se,i=Math.max(0,o-Pe)/n;for(let e=0;e<n;e++)fe.addLatest(i);Se=e,Pe=o,dt=fe.calcAverage(5)}}function Ko(){xt="",Mt="",Tt="";const e=[];Object.keys(Game.Upgrades).forEach((t=>{e.push(Game.Upgrades[t])})),e.sort((function(e,t){return e.order>t.order?1:e.order<t.order?-1:0})),Object.keys(e).forEach((t=>{const o=e[t];if(0===o.bought){let e="";e+=function(e){let t="crate upgrade missing";"prestige"===e.pool&&(t+=" heavenly");let o=0;Game.prefs.crates||(o=1),o&&(t+=" noFrame");let{icon:n}=e;e.iconFunction&&(n=e.iconFunction());const i=`function() {return Game.crateTooltip(Game.UpgradesById[${e.id}], 'stats');}`;return`<div class="${t}"\n\t${Game.getDynamicTooltip(i,"top",!0)}\n\tstyle = "${n[2]?`background-image: url(${n[2]});`:""}background-position:${48*-n[0]}px ${48*-n[1]}px;">\n\t</div>`}(o),"prestige"===o.pool?Tt+=e:"cookie"===o.pool?Mt+=e:"toggle"!==o.pool&&"unused"!==o.pool&&"debug"!==o.pool&&(xt+=e)}}))}function Jo(){if("christmas"===Game.season){let e=60*Game.cookiesPs;Game.hasBuff("Elder frenzy")&&(e*=.5),Game.hasBuff("Frenzy")&&(e*=.75),ut=Math.max(25,e),Game.Has("Ho ho ho-flavored frosting")&&(ut*=2)}}function Zo(){Ye=900*rt/.15,Ye*=lt;const e=To();e>0?Ye/=e:Ye=0,Ke=st*(.15*Ye)+13,Je=ct*(.15*Ye)+13,Ze=7*Ye,et=st*(.15*Ze)+13,tt=ct*(.15*Ze)+13,ot=2*Ye,nt=.15*ot,it=0;let t=0,o=0;Object.keys(Game.Objects).forEach((e=>{Game.Objects[e].amount>t&&(t=Game.Objects[e].amount),Game.Objects[e].amount>0&&(o+=1)})),Object.keys(Game.Objects).forEach((e=>{(Game.Objects[e].amount<t||1===o)&&Game.Objects[e].amount<400&&2*Game.Objects[e].price>it&&(it=2*Game.Objects[e].price,at=e)}))}function en(){let e=1,t=1,o=1;to("Green yeast digestives")&&(o*=1.01),to("Dragon fang")&&(o*=1.03),e*=1+.1*Game.auraMult("Ancestral Metamorphosis"),e*=Game.eff("goldenCookieGain"),t*=1+.1*Game.auraMult("Unholy Dominion"),t*=Game.eff("wrathCookieGain"),st=o*e,ct=o*t,lt=1,0===Game.shimmerTypes.golden.n&&(lt*=1+1.23*Game.auraMult("Dragon's Fortune"))}function tn(){Ct=0,yt=0,bt=[0,null];for(let e=0;e<Game.wrinklers.length;e++){let{sucked:t}=Game.wrinklers[e],o=1.1;if(Game.Has("Sacrilegious corruption")&&(o*=1.05),1===Game.wrinklers[e].type&&(o*=3),t*=o,Game.Has("Wrinklerspawn")&&(t*=1.05),qt.Temple.minigameLoaded){const e=Game.hasGod("scorn");1===e?t*=1.15:2===e?t*=1.1:3===e&&(t*=1.05)}Ct+=t,0===Game.wrinklers[e].type&&(yt+=t,t>bt[0]&&(bt=[t,e]))}}function on(){fo(),tn(),Zo(),en(),qo(),Ko(),Jo(),pe=new Bo(Y[Y.length-1]),ue=new Bo(Y[Y.length-1]),me=new Bo(Y[Y.length-1]),he=new Bo(Y[Y.length-1]),ge=new Bo(K[K.length-1]),fe=new Bo(5),Yo(),Eo(),Xo(),_o(),Wo()}const nn={CPSMode:1,AvgCPSHist:3,AvgClicksHist:0,CalcWrink:0,Scale:2,ScaleDecimals:2,ScaleSeparator:0,ScaleCutoff:999999,TimeFormat:0,DetailedTime:1,PPDisplayTime:0,BuildColour:1,PPOnlyConsiderBuyable:0,PPExcludeTop:0,PPRigidelMode:0,PPSecondsLowerLimit:0,ColourBlue:"#4bb8f0",ColourGreen:"#00ff00",ColourYellow:"#ffff00",ColourOrange:"#ff7f00",ColourRed:"#ff0000",ColourPurple:"#ff00ff",ColourGray:"#b3b3b3",ColourPink:"#ff1493",ColourBrown:"#8b4513",BotBar:1,TimerBar:1,TimerBarPos:0,TimerBarOverlay:2,UpBarColour:1,UpgradeBarFixedPos:1,SortBuildings:0,SortUpgrades:0,DragonAuraInfo:1,GrimoireBar:1,GCTimer:1,Favicon:1,WrinklerButtons:1,HideSectionsButtons:0,TooltipBuildUpgrade:1,TooltipAmor:0,ToolWarnLucky:1,ToolWarnLuckyFrenzy:1,ToolWarnConjure:1,ToolWarnConjureFrenzy:1,ToolWarnEdifice:1,ToolWarnUser:0,ToolWarnBon:1,ToolWarnPos:1,TooltipGrim:1,TooltipWrink:1,TooltipLump:1,TooltipPlots:1,TooltipPantheon:1,TooltipAscendButton:1,Stats:1,MissingUpgrades:1,UpStats:1,HeavenlyChipsTarget:1,ShowMissedGC:1,Title:1,GeneralSound:1,GCNotification:0,GCFlash:1,ColourGCFlash:"#ffffff",GCSound:1,GCVolume:100,GCSoundURL:"https://freesound.org/data/previews/66/66717_931655-lq.mp3",FortuneNotification:0,FortuneFlash:1,ColourFortuneFlash:"#ffffff",FortuneSound:1,FortuneVolume:100,FortuneSoundURL:"https://freesound.org/data/previews/174/174027_3242494-lq.mp3",SeaNotification:0,SeaFlash:1,ColourSeaFlash:"#ffffff",SeaSound:1,SeaVolume:100,SeaSoundURL:"https://www.freesound.org/data/previews/121/121099_2193266-lq.mp3",GardFlash:1,ColourGardFlash:"#ffffff",GardSound:1,GardVolume:100,GardSoundURL:"https://freesound.org/data/previews/103/103046_861714-lq.mp3",MagicNotification:0,MagicFlash:1,ColourMagicFlash:"#ffffff",MagicSound:1,MagicVolume:100,MagicSoundURL:"https://freesound.org/data/previews/221/221683_1015240-lq.mp3",WrinklerNotification:0,WrinklerFlash:1,ColourWrinklerFlash:"#ffffff",WrinklerSound:1,WrinklerVolume:100,WrinklerSoundURL:"https://freesound.org/data/previews/124/124186_8043-lq.mp3",WrinklerMaxNotification:0,WrinklerMaxFlash:1,ColourWrinklerMaxFlash:"#ffffff",WrinklerMaxSound:1,WrinklerMaxVolume:100,WrinklerMaxSoundURL:"https://freesound.org/data/previews/152/152743_15663-lq.mp3",BulkBuyBlock:0,Header:{Calculation:1,Notation:1,Colours:1,BarsDisplay:1,Tooltip:1,Statistics:1,Notification:1,NotificationGeneral:1,NotificationGC:1,NotificationFC:1,NotificationSea:1,NotificationGard:1,NotificationMagi:1,NotificationWrink:1,NotificationWrinkMax:1,Miscellaneous:1,Lucky:1,Chain:1,Spells:1,Garden:1,Prestige:1,Wrink:1,Sea:1,Misc:1,InfoTab:1}},an=function(e){if(1===e){const e=function(){try{Notification.requestPermission().then()}catch(e){return!1}return!0};"Notification"in window?e()?Notification.requestPermission().then():Notification.requestPermission():console.log("This browser does not support notifications.")}};function rn(){1===e.BotBar?(l("CMBotBar").style.display="",uo()):l("CMBotBar").style.display="none",ho()}function ln(){BeautifyAll(),Game.RefreshStore(),Game.RebuildUpgrades(),uo(),ao(),ro()}function sn(){1===e.Favicon&&y>0?Rt.wrath?l("CMFavicon").href="https://aktanusa.github.io/CookieMonster/favicon/wrathCookie.ico":l("CMFavicon").href="https://aktanusa.github.io/CookieMonster/favicon/goldenCookie.ico":l("CMFavicon").href="https://orteil.dashnet.org/cookieclicker/favicon.ico"}class cn{constructor(e,t){this.type=e,this.group=t}}class dn extends cn{constructor(e,t,o,n,i,a=null){super(e,t),this.label=o,this.desc=n,this.toggle=i,a&&(this.func=a)}}class pn extends cn{constructor(e,t,o){super(e,t),this.desc=o}}class un extends cn{constructor(e,t,o,n){super(e,t),this.label=o,this.desc=n;for(let e=0;e<101;e++)this.label[e]=`${e}%`}}class mn extends cn{constructor(e,t,o,n,i,a){super(e,t),this.label=o,this.desc=n,this.min=i,this.max=a}}const hn={CPSMode:new dn("bool","Calculation",["Current cookies per second","Average cookies per second"],"Calculate times using current cookies per second or average cookies per second",!1),AvgCPSHist:new dn("bool","Calculation",["Average CPS in past 10s","Average CPS in past 15s","Average CPS in past 30s","Average CPS in past 1m","Average CPS in past 5m","Average CPS in past 10m","Average CPS in past 15m","Average CPS in past 30m"],"How much time average Cookies Per Second should consider",!1),AvgClicksHist:new dn("bool","Calculation",["Average clicks in past 1s","Average clicks in past 5s","Average clicks in past 10s","Average clicks in past 15s","Average clicks in past 30s"],"How much time average Cookie Clicks should consider",!1),CalcWrink:new dn("bool","Calculation",["Calculate with wrinklers OFF","Calculate with wrinklers ON","Calculate with single fattest wrinkler ON"],"Calculate times and average Cookies Per Second with (only the single non-shiny fattest) wrinklers",!0,(function(){$e=!0})),Scale:new dn("bool","Notation",["Game's setting scale","Metric","Short scale","Short scale (Abbreviated)","Scientific notation","Engineering notation"],"Change how long numbers are formatted",!1,(function(){ln()})),ScaleDecimals:new dn("bool","Notation",["1 decimals","2 decimals","3 decimals"],'Set the number of decimals used when applicable. This only works with Cookie Monster scales and not with "Game\'s Setting Scale"',!1,(function(){ln()})),ScaleSeparator:new dn("bool","Notation",[". for decimals (standard)",". for thousands"],"Set the separator used for decimals and thousands",!1,(function(){ln()})),ScaleCutoff:new mn("numscale","Notation","Notation cut-off point: ","The number from which Cookie Monster will start formatting numbers based on chosen scale. Standard is 999,999. Setting this above 999,999,999 might break certain notations",1,999999999),TimeFormat:new dn("bool","Notation",["Time XXd, XXh, XXm, XXs","Time XX:XX:XX:XX:XX"],"Change the time format",!1),DetailedTime:new dn("bool","Notation",["Detailed time OFF","Detailed time ON"],"Change how time is displayed in certain statistics and tooltips",!0,(function(){1===e.DetailedTime?Game.sayTime=I:Game.sayTime=S.sayTime})),PPDisplayTime:new dn("bool","Notation",["PP as value (standard)","PP as time unit"],"Display PP as calculated value or as approximate time unit. Note that PP does not translate directly into a time unit and this is therefore only an approximation.",!1),BuildColour:new dn("bool","Colours",["Building colours OFF","Building colours ON"],"Colour code buildings",!0,(function(){ao()})),PPOnlyConsiderBuyable:new dn("bool","Colours",["Don't ignore non-buyable","Ignore non-buyable"],"Makes Cookie Monster label buildings and upgrades you can't buy right now red, useful in those situations where you just want to spend your full bank 'most optimally'",!0),PPExcludeTop:new dn("bool","Colours",["Don't ignore any","Ignore 1st best","Ignore 1st and 2nd best","Ignore 1st, 2nd and 3rd best"],"Makes Cookie Monster ignore the 1st, 2nd or 3rd best buildings in labeling and colouring PP values",!0),PPRigidelMode:new dn("bool","Colours",["Rigidel mode OFF","Rigidel mode ON"],'Makes Cookie Monster ignore all "buy 1" options when colouring PP in order to stay at a total building count ending in 10 for pantheon god Rigidel',!0),PPSecondsLowerLimit:new mn("numscale","Colours","Lower limit for PP (in seconds): ",'If a building or upgrade costs less than the specified seconds of CPS it will also be considered optimal and label it as such ("PP is less than xx seconds of CPS"); setting to 0 ignores this option',0,1/0),ColourBlue:new pn("colour","Colours","Standard colour is blue. Used to show upgrades better than best PP building, for Click Frenzy bar, and for various labels"),ColourGreen:new pn("colour","Colours","Standard colour is green. Used to show best PP building, for Blood Frenzy bar, and for various labels"),ColourYellow:new pn("colour","Colours","Standard colour is yellow. Used to show buildings within the top 10 of PP, for Frenzy bar, and for various labels"),ColourOrange:new pn("colour","Colours","Standard colour is orange. Used to show buildings within the top 20 of PP, for Next Reindeer bar, and for various labels"),ColourRed:new pn("colour","Colours","Standard colour is Red. Used to show buildings within the top 30 of PP, for Clot bar, and for various labels"),ColourPurple:new pn("colour","Colours","Standard colour is purple. Used to show buildings outside of the top 30 of PP, for Next Cookie bar, and for various labels"),ColourGray:new pn("colour","Colours","Standard colour is gray. Used to show negative or infinity PP, and for Next Cookie/Next Reindeer bar"),ColourPink:new pn("colour","Colours","Standard colour is pink. Used for Dragonflight bar"),ColourBrown:new pn("colour","Colours","Standard colour is brown. Used for Dragon Harvest bar"),BotBar:new dn("bool","BarsDisplay",["Bottom bar OFF","Bottom bar ON"],"Building information",!0,(function(){rn()})),TimerBar:new dn("bool","BarsDisplay",["Timer bar OFF","Timer bar ON"],"Bar with timers for golden cookie, season popup, Frenzy (Normal, Clot, Elder), Click Frenzy",!0,(function(){go()})),TimerBarPos:new dn("bool","BarsDisplay",["Timer bar position (top left)","Timer bar position (bottom)"],"Placement of the timer bar",!1,(function(){0===e.TimerBarPos?(l("CMTimerBar").style.width="30%",l("CMTimerBar").style.bottom="",l("game").insertBefore(l("CMTimerBar"),l("sectionLeft"))):(l("CMTimerBar").style.width="100%",l("CMTimerBar").style.bottom="0px",l("wrapper").appendChild(l("CMTimerBar"))),ho()})),TimerBarOverlay:new dn("bool","BarsDisplay",["Timer bar overlay OFF","Timer bar overlay only seconds","Timer bar overlay full"],"Overlay on timers displaying seconds and/or percentage left",!0),UpBarColour:new dn("bool","BarsDisplay",["Upgrade colours/bar OFF","Upgrade colours with bar ON","Upgrade colours without bar ON"],"Colour code upgrades and optionally add a counter bar",!1,(function(){1===e.UpBarColor?(l("CMUpgradeBar").style.display="",ro()):2===e.UpBarColor?(l("CMUpgradeBar").style.display="none",ro()):(l("CMUpgradeBar").style.display="none",Game.RebuildUpgrades())})),UpgradeBarFixedPos:new dn("bool","BarsDisplay",["Upgrade bar fixed position OFF","Upgrade bar fixed position ON"],"Lock the upgrade bar at top of the screen to prevent it from moving ofscreen when scrolling",!0,(function(){1===e.UpgradeBarFixedPos?(l("CMUpgradeBar").style.position="sticky",l("CMUpgradeBar").style.top="0px"):l("CMUpgradeBar").style.position=""})),SortBuildings:new dn("bool","BarsDisplay",["Sort buildings: default","Sort buildings: PP of x1 purchase","Sort buildings: PP of selected bulk mode"],"Sort the display of buildings in either default order or by PP",!1,(function(){ao()})),SortUpgrades:new dn("bool","BarsDisplay",["Sort upgrades: default","Sort upgrades: PP"],"Sort the display of upgrades in either default order or by PP",!1,(function(){ro()})),DragonAuraInfo:new dn("bool","BarsDisplay",["Extra dragon aura info OFF","Extra dragon aura info ON"],"Shows information about changes in CPS and costs in the dragon aura interface.",!0),GrimoireBar:new dn("bool","BarsDisplay",["Grimoire magic meter timer OFF","Grimoire magic meter timer ON"],"A timer on how long before the Grimoire magic meter is full",!0),GCTimer:new dn("bool","BarsDisplay",["Golden cookie timer OFF","Golden cookie timer ON"],"A timer on the golden cookie when it has been spawned",!0,(function(){1===e.GCTimer?Object.keys(re).forEach((e=>{re[e].style.display="block",re[e].style.left=_t[e].l.style.left,re[e].style.top=_t[e].l.style.top})):Object.keys(re).forEach((e=>re[e].style.display="none"))})),Favicon:new dn("bool","BarsDisplay",["Favicon OFF","Favicon ON"],"Update favicon with golden/wrath cookie",!0,(function(){sn()})),WrinklerButtons:new dn("bool","BarsDisplay",["Extra wrinkler buttons OFF","Extra wrinkler buttons ON"],"Show buttons for popping wrinklers at bottom of cookie section",!0,(function(){t()})),HideSectionsButtons:new dn("bool","BarsDisplay",["Hide buildings/upgrades button OFF","Hide buildings/upgrades button ON"],"Show buttons for hiding and showing the buildings and upgrades sections in the right column",!0,(function(){e.HideSectionsButtons?l("CMSectionHidButtons").style.display="":l("CMSectionHidButtons").style.display="none"})),TooltipBuildUpgrade:new dn("bool","Tooltip",["Building/upgrade tooltip information OFF","Building/upgrade  tooltip information ON"],"Extra information in building/upgrade tooltips",!0),TooltipAmor:new dn("bool","Tooltip",["Buildings tooltip amortization information OFF","Buildings tooltip amortization information ON"],"Add amortization information to buildings tooltip",!0),ToolWarnLucky:new dn("bool","Tooltip",["Tooltip lucky warning OFF","Tooltip lucky warning ON"],'A warning when buying if it will put the bank under the amount needed for max "Lucky!" rewards',!0),ToolWarnLuckyFrenzy:new dn("bool","Tooltip",["Tooltip lucky frenzy warning OFF","Tooltip lucky frenzy warning ON"],'A warning when buying if it will put the bank under the amount needed for max "Lucky!" (Frenzy) rewards',!0),ToolWarnConjure:new dn("bool","Tooltip",["Tooltip conjure warning OFF","Tooltip conjure warning ON"],'A warning when buying if it will put the bank under the amount needed for max "Conjure Baked Goods" rewards',!0),ToolWarnConjureFrenzy:new dn("bool","Tooltip",["Tooltip conjure frenzy warning OFF","Tooltip conjure frenzy warning ON"],'A warning when buying if it will put the bank under the amount needed for max "Conjure Baked Goods" rewards with Frenzy active',!0),ToolWarnEdifice:new dn("bool","Tooltip",["Tooltip edifice warning OFF","Tooltip edifice warning ON"],'A warning when buying if it will put the bank under the amount needed for "Spontaneous Edifice" to possibly give you your most expensive building',!0),ToolWarnUser:new mn("numscale","Tooltip","Tooltip warning at x times CPS: ","Use this to show a customized warning if buying it will put the bank under the amount equal to value times cps; setting to 0 disables the function altogether",0,1/0),ToolWarnBon:new dn("bool","Tooltip",["Calculate tooltip warning with bonus CPS OFF","Calculate tooltip warning with bonus CPS ON"],"Calculate the warning with or without the bonus CPS you get from buying",!0),ToolWarnPos:new dn("bool","Tooltip",["Tooltip warning position (left)","Tooltip warning position (bottom)"],"Placement of the warning boxes",!1,(function(){Mo()})),TooltipGrim:new dn("bool","Tooltip",["Grimoire tooltip information OFF","Grimoire tooltip information ON"],"Extra information in tooltip for grimoire",!0),TooltipWrink:new dn("bool","Tooltip",["Wrinkler tooltip OFF","Wrinkler tooltip ON"],"Shows the amount of cookies a wrinkler will give when popping it",!0),TooltipLump:new dn("bool","Tooltip",["Sugar lump tooltip OFF","Sugar lump tooltip ON"],"Shows the current Sugar Lump type in Sugar lump tooltip.",!0),TooltipPlots:new dn("bool","Tooltip",["Garden plots tooltip OFF","Garden plots tooltip ON"],"Shows a tooltip for plants that have a cookie reward.",!0),TooltipPantheon:new dn("bool","Tooltip",["Pantheon tooltip OFF","Pantheon tooltip ON"],"Shows additional info in the pantheon tooltip",!0),TooltipAscendButton:new dn("bool","Tooltip",["Show Extra Info Ascend Tooltip OFF","Show Extra Info Ascend Tooltip ON"],"Shows additional info in the ascend tooltip",!0),Stats:new dn("bool","Statistics",["Statistics OFF","Statistics ON"],"Extra Cookie Monster statistics!",!0),MissingUpgrades:new dn("bool","Statistics",["Missing upgrades OFF","Missing upgrades ON"],"Shows missing upgrades in stats menu",!0),UpStats:new dn("bool","Statistics",["Statistics update rate (default)","Statistics update rate (1s)"],"Default rate is once every 5 seconds",!1),HeavenlyChipsTarget:new mn("numscale","Statistics","Heavenly chips target: ",'Use this to set a heavenly chips target that will be counted towards in the "prestige" statsistics sections',1,1/0),ShowMissedGC:new dn("bool","Statistics",["Missed GC OFF","Missed GC ON"],"Show a stat in the statistics screen that counts how many golden cookies you have missed",!0),Title:new dn("bool","NotificationGeneral",["Title OFF","Title ON","Title pinned tab highlight"],'Update title with colden cookie/season popup timers; pinned tab highlight only changes the title when a golden cookie/season popup spawns; "!" means that golden cookie/reindeer can spawn',!0),GeneralSound:new dn("bool","NotificationGeneral",["Consider game volume setting OFF","Consider game volume setting ON"],'Turning this toggle to "off" makes Cookie Monster no longer consider the volume setting of the base game, allowing mod notifications to play with base game volume turned down',!0),GCNotification:new dn("bool","NotificationGC",["Notification OFF","Notification ON"],"Create a notification when golden cookie spawns",!0,(function(){an(e.GCNotification)})),GCFlash:new dn("bool","NotificationGC",["Flash OFF","Flash ON"],"Flash screen on golden cookie",!0),ColourGCFlash:new pn("colour","NotificationGC","The colour of the GC flash, standard colour is white"),GCSound:new dn("bool","NotificationGC",["Sound OFF","Sound ON"],"Play a sound on golden cookie",!0),GCVolume:new un("vol","NotificationGC",[],"Volume"),GCSoundURL:new dn("url","NotificationGC","Sound URL:","URL of the sound to be played when a golden cookie spawns"),FortuneNotification:new dn("bool","NotificationFC",["Notification OFF","Notification ON"],"Create a notification when fortune cookie is on the ticker",!0,(function(){an(e.FortuneNotification)})),FortuneFlash:new dn("bool","NotificationFC",["Flash OFF","Flash ON"],"Flash screen on fortune cookie spawn",!0),ColourFortuneFlash:new pn("colour","NotificationFC","The colour of the fortune flash, standard colour is white"),FortuneSound:new dn("bool","NotificationFC",["Sound OFF","Sound ON"],"Play a sound on fortune cookie spawn",!0),FortuneVolume:new un("vol","NotificationFC",[],"Volume"),FortuneSoundURL:new dn("url","NotificationFC","Sound URL:","URL of the sound to be played when the ticker has a fortune cookie"),SeaNotification:new dn("bool","NotificationSea",["Notification OFF","Notification ON"],"Create a notification on season popup",!0,(function(){an(e.SeaNotification)})),SeaFlash:new dn("bool","NotificationSea",["Flash OFF","Flash ON"],"Flash screen on season popup",!0),ColourSeaFlash:new pn("colour","NotificationSea","The colour of the season popup flash, standard colour is white"),SeaSound:new dn("bool","NotificationSea",["Sound OFF","Sound ON"],"Play a sound on season popup",!0),SeaVolume:new un("vol","NotificationSea",[],"Volume"),SeaSoundURL:new dn("url","NotificationSea","Sound URL:","URL of the sound to be played when on season popup spawns"),GardFlash:new dn("bool","NotificationGard",["Garden Tick Flash OFF","Flash ON"],"Flash screen on garden tick",!0),ColourGardFlash:new pn("colour","NotificationGard","The colour of the garden flash, standard colour is white"),GardSound:new dn("bool","NotificationGard",["Sound OFF","Sound ON"],"Play a sound on garden tick",!0),GardVolume:new un("vol","NotificationGard",[],"Volume"),GardSoundURL:new dn("url","NotificationGard","Garden Tick Sound URL:","URL of the sound to be played when the garden ticks"),MagicNotification:new dn("bool","NotificationMagi",["Notification OFF","Notification ON"],"Create a notification when magic reaches maximum",!0,(function(){an(e.MagicNotification)})),MagicFlash:new dn("bool","NotificationMagi",["Flash OFF","Flash ON"],"Flash screen when magic reaches maximum",!0),ColourMagicFlash:new pn("colour","NotificationMagi","The colour of the magic flash, standard colour is white"),MagicSound:new dn("bool","NotificationMagi",["Sound OFF","Sound ON"],"Play a sound when magic reaches maximum",!0),MagicVolume:new un("vol","NotificationMagi",[],"Volume"),MagicSoundURL:new dn("url","NotificationMagi","Sound URL:","URL of the sound to be played when magic reaches maxium"),WrinklerNotification:new dn("bool","NotificationWrink",["Notification OFF","Notification ON"],"Create a notification when a wrinkler appears",!0,(function(){an(e.WrinklerNotification)})),WrinklerFlash:new dn("bool","NotificationWrink",["Flash OFF","Flash ON"],"Flash screen when a wrinkler appears",!0),ColourWrinklerFlash:new pn("colour","NotificationWrink","The colour of the wrinkler flash, standard colour is white"),WrinklerSound:new dn("bool","NotificationWrink",["Sound OFF","Sound ON"],"Play a sound when a wrinkler appears",!0),WrinklerVolume:new un("vol","NotificationWrink",[],"Volume"),WrinklerSoundURL:new dn("url","NotificationWrink","Sound URL:","URL of the sound to be played when a wrinkler appears"),WrinklerMaxNotification:new dn("bool","NotificationWrinkMax",["Notification OFF","Notification ON"],"Create a notification when the maximum amount of wrinklers has appeared",!0,(function(){an(e.WrinklerMaxNotification)})),WrinklerMaxFlash:new dn("bool","NotificationWrinkMax",["Flash OFF","Flash ON"],"Flash screen when the maximum amount of Wrinklers has appeared",!0),ColourWrinklerMaxFlash:new pn("colour","NotificationWrinkMax","The colour of the maximum wrinkler flash, standard colour is white"),WrinklerMaxSound:new dn("bool","NotificationWrinkMax",["Sound OFF","Sound ON"],"Play a sound when the maximum amount of wrinklers has appeared",!0),WrinklerMaxVolume:new un("vol","NotificationWrinkMax",[],"Volume"),WrinklerMaxSoundURL:new dn("url","NotificationWrinkMax","Sound URL:","URL of the sound to be played when the maximum amount of wrinklers has appeared"),BulkBuyBlock:new dn("bool","Miscellaneous",["Block Bulk Buying OFF","Block Bulk Buying ON"],"Block clicking bulk buying when you can't buy all. This prevents buying 7 of a building when you are in buy-10 or buy-100 mode.",!0)};function gn(){return JSON.stringify({settings:e,version:"2.031.6"})}function fn(){if(Qe!==Game.dragonLevel||$e){if(Game.dragonLevel<25&&Game.dragonLevels[Game.dragonLevel].buy.toString().includes("sacrifice")){let e=Game.dragonLevels[Game.dragonLevel].buy.toString().match(/Objects\[(.*)\]/)[1];const t=Game.dragonLevels[Game.dragonLevel].buy.toString().match(/sacrifice\((.*?)\)/)[1];if("i"!==e)if(e=e.replaceAll("'",""),Game.Objects[e].amount<t)qe="Not enough buildings to sell";else{let o=0;Go();for(let n=0;n<t;n++){let t=qt[e].basePrice*Game.priceIncrease**Math.max(0,qt[e].amount-1-qt[e].free);t=Game.modifyBuildingPrice(qt[e],t),t=Math.ceil(t),o+=t,qt[e].amount-=1}qe=`Cost to rebuy: ${se(o)}`}else{let o=0;Go(),Object.keys(Game.Objects).forEach((n=>{if(e=n,Game.Objects[e].amount<t)qe="Not enough buildings to sell";else{for(let n=0;n<t;n++){let t=qt[e].basePrice*Game.priceIncrease**Math.max(0,qt[e].amount-1-qt[e].free);t=Game.modifyBuildingPrice(qt[e],t),t=Math.ceil(t),o+=t,qt[e].amount-=1}qe=`Cost to rebuy: ${se(o)}`}}))}}Qe=Game.dragonLevel}}function Cn(e,t){Go(),0===t?He=e:1===t?Ue=e:2===t&&(Re=e);const o=We;return jo(),zo(),o!==We&&jo(),ze-Game.cookiesPs}function yn(){qt=[],Object.keys(Game.Objects).forEach((e=>{qt[e]=bo(e)})),Yt=[],Object.keys(Game.Upgrades).forEach((e=>{Yt[e]=ko(e)})),Kt=[],Object.keys(Game.Achievements).forEach((e=>{Kt[e]=Co(e)})),Go()}let bn=!1;function kn(t,o,n){(1===e[o]||n)&&3===t&&!1===bn||1===t?(l("CMFlashScreen").style.backgroundColor=e[`Colour${o}`],l("CMFlashScreen").style.opacity="0.5",3===t?(l("CMFlashScreen").style.display="inline",setTimeout((function(){kn(2,o,!0)}),1e3/Game.fps)):setTimeout((function(){kn(0,o,!0)}),1e3/Game.fps)):2===t?(l("CMFlashScreen").style.opacity="1",setTimeout((function(){kn(1,o,!0)}),1e3/Game.fps)):0===t&&(l("CMFlashScreen").style.display="none")}function Gn(t,o,n,i){if((1===e[o]||i)&&!1===bn){const o=new Audio(t);e.GeneralSound?o.volume=e[n]/100*(Game.volume/100):o.volume=e[n]/100,o.play()}}function wn(t,o,n){1===e[t]&&"hidden"===document.visibilityState&&!1===bn&&new Notification(o,{body:n,badge:"https://orteil.dashnet.org/cookieclicker/favicon.ico"})}function vn(){if(ie!==Game.OnAscend&&(ie=Game.OnAscend,Game.OnAscend?(l("game").style.bottom="0px",1===e.BotBar&&(l("CMBotBar").style.display="none"),1===e.TimerBar&&(l("CMTimerBar").style.display="none")):(rn(),go()),mo()),!Game.OnAscend&&0===Game.AscendTimer){r!==Object.keys(Game.mods).length&&(yn(),on(),r=Object.keys(Game.mods).length),$e&&(Xo(),Game.Has("Golden switch [off]")?(Go(),Yt["Golden switch [off]"].bought=0,jo(),rt=ze):rt=Game.cookiesPs,en(),Zo(),Ko(),qo(),fn(),function(){for(let e=0;e<11;e+=1)for(let t=0;t<3;t+=1)Qt[e][t]=Cn(e,t)}(),Jo(),function(){let e=0;if(Game.Objects.Bank.minigameLoaded){const t=Game.Objects.Bank.minigame.goods;let o=0;Object.keys(t).forEach((e=>{const n=t[e];o+=n.stock*n.val})),e+=o*Game.cookiesPsRawHighest}e+=function(){let e=0;Go();let t=2;5!==je&&18!==je||(t-=1),5!==De&&18!==De||(t-=1),je=5,De=18;for(let e=0;e<t;++e){let e="Cursor";Object.keys(qt).forEach((t=>{qt[t].amount>0&&(e=t)})),qt[e].amount-=1,Ne-=1}return Object.keys(qt).forEach((t=>{const o=qt[t];e+=io(Game.Objects[o.name],Game.Objects[t].basePrice,o.amount,Game.Objects[t].free,o.amount)})),e}(),Xt=e}(),$e=0);const t=Game.auraMult("Fierce Hoarder")>0;!Ee&&t?(Ee=!0,Be=1):Ee&&!t&&(Ee=!1,Be=1),Be&&(_o(),Be=0),function(){tn(),Wo(),function(){kt=0;let e=0;Object.keys(Game.wrinklers).forEach((t=>{2===Game.wrinklers[t].phase&&(e+=1)}));let t=1;if(qt.Temple.minigameLoaded){const e=Game.hasGod("scorn");1===e?t*=1.15:2===e?t*=1.1:3===e&&(t*=1.05)}Gt=e,kt=e*(.05*e*1.1)*(.05*Game.Has("Sacrilegious corruption")+1)*(.05*Game.Has("Wrinklerspawn")+1)*t}(),Eo(),Yo();const e=Game.HowManyCookiesReset(Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned))+1)-(Game.cookiesEarned+Game.cookiesReset);It=ce(e/lo())}(),G=0,_t={},Object.keys(Game.shimmers).forEach((e=>{_t[Game.shimmers[e].id]=Game.shimmers[e],Game.shimmers[e].spawnLead&&"golden"===Game.shimmers[e].type&&(Rt=Game.shimmers[e],G+=1)})),Object.keys(re).forEach((e=>{void 0===_t[e]&&(re[e].parentNode.removeChild(re[e]),delete re[e])})),y!==Game.shimmerTypes.golden.n?(y=Game.shimmerTypes.golden.n,y&&(b<G&&(kn(3,"GCFlash",!1),Gn(e.GCSoundURL,"GCSound","GCVolume",!1),wn("GCNotification","Golden Cookie Spawned","A Golden Cookie has spawned. Click it now!")),Object.keys(Game.shimmers).forEach((t=>{void 0===re[Game.shimmers[t].id]&&function(t){const o=document.createElement("div");o.id=`GCTimer${t.id}`,o.style.width="96px",o.style.height="96px",o.style.position="absolute",o.style.zIndex="10000000001",o.style.textAlign="center",o.style.lineHeight="96px",o.style.fontFamily='"Kavoon", Georgia, serif',o.style.fontSize="35px",o.style.cursor="pointer",o.style.display="block",0===e.GCTimer&&(o.style.display="none"),o.style.left=t.l.style.left,o.style.top=t.l.style.top,o.onclick=function(){t.pop()},o.onmouseover=function(){t.l.style.filter="brightness(125%) drop-shadow(0px 0px 3px rgba(255,255,255,1))",t.l.style.webkitFilter="brightness(125%) drop-shadow(0px 0px 3px rgba(255,255,255,1))"},o.onmouseout=function(){t.l.style.filter="",t.l.style.webkitFilter=""},re[t.id]=o,l("shimmers").appendChild(o)}(Game.shimmers[t])}))),sn(),b=G,0===G&&(Rt=0)):1===e.GCTimer&&y&&Object.keys(re).forEach((e=>{re[e].style.opacity=_t[e].l.style.opacity,re[e].style.transform=_t[e].l.style.transform,re[e].textContent=Math.ceil(_t[e].life/Game.fps)})),f!==(Game.TickerEffect&&"fortune"===Game.TickerEffect.type)&&(f=Game.TickerEffect&&"fortune"===Game.TickerEffect.type,f&&(kn(3,"FortuneFlash",!1),Gn(e.FortuneSoundURL,"FortuneSound","FortuneVolume",!1),wn("FortuneNotification","Fortune Cookie found","A Fortune Cookie has appeared on the Ticker."))),g!==Game.shimmerTypes.reindeer.spawned&&(g=Game.shimmerTypes.reindeer.spawned,Object.keys(Game.shimmers).forEach((e=>{Game.shimmers[e].spawnLead&&"reindeer"===Game.shimmers[e].type&&($t=Game.shimmers[e])})),kn(3,"SeaFlash",!1),Gn(e.SeaSoundURL,"SeaSound","SeaVolume",!1),wn("SeaNotification","Reindeer sighted!","A Reindeer has spawned. Click it now!")),Game.Objects.Farm.minigameLoaded&&C!==Game.Objects.Farm.minigame.nextStep&&(0!==C&&C<Date.now()&&(kn(3,"GardFlash",!1),Gn(e.GardSoundURL,"GardSound","GardVolume",!1)),C=Game.Objects.Farm.minigame.nextStep),function(){if(Game.Objects["Wizard tower"].minigameLoaded&&1===e.GrimoireBar){const{minigame:t}=Game.Objects["Wizard tower"];t.magic<t.magicM?k=!1:k||(k=!0,kn(3,"MagicFlash",!1),Gn(e.MagicSoundURL,"MagicSound","MagicVolume",!1),wn("MagicNotification","Magic Meter full","Your Magic Meter is full. Cast a spell!"))}}(),function(){if(Game.elderWrath>0){let t=0;Object.keys(Game.wrinklers).forEach((e=>{2===Game.wrinklers[e].phase&&(t+=1)})),t>w?(w=t,t===Game.getWrinklersMax()&&e.WrinklerMaxFlash?kn(3,"WrinklerMaxFlash",!1):kn(3,"WrinklerFlash",!1),t===Game.getWrinklersMax()&&e.WrinklerMaxSound?Gn(e.WrinklerMaxSoundURL,"WrinklerMaxSound","WrinklerMaxVolume",!1):Gn(e.WrinklerSoundURL,"WrinklerSound","WrinklerVolume",!1),t===Game.getWrinklersMax()&&e.WrinklerMaxNotification?wn("WrinklerMaxNotification","Maximum Wrinklers Reached","You have reached your maximum ammount of wrinklers"):wn("WrinklerNotification","A Wrinkler appeared","A new wrinkler has appeared")):w=t}}()}}function xn(){let t="";for(let o=0;o<H.length;o++)t+=`.CMText${H[o]} { color: ${e[`Colour${H[o]}`]}; }\n`;for(let o=0;o<H.length;o++)t+=`.CMBack${H[o]} { background-color: ${e[`Colour${H[o]}`]}; }\n`;for(let o=0;o<H.length;o++)t+=`.CMBorder${H[o]} { border: 1px solid ${e[`Colour${H[o]}`]}; }\n`;l("CMCSS").textContent=t,ao()}function Tn(){const e=b64_to_utf8(unescape(localStorage.getItem("CookieClickerGame")).split("!END!")[0]),t=e.match(/CookieMonster.*(;|$)/);if(null!==t){const o=e.replace(t[0],`CookieMonster:${gn()}`);localStorage.setItem("CookieClickerGame",escape(`${utf8_to_b64(o)}!END!`))}}function Mn(t){if(void 0!==localStorage.CMConfig&&delete localStorage.CMConfig,void 0!==t){e=t,void 0!==e.Colors&&delete e.Colors,void 0!==e.Colours&&delete e.Colours;let o=!1;Object.keys(nn).forEach((t=>{void 0===e[t]?(o=!0,e[t]=nn[t]):"Header"===t&&Object.keys(nn.Header).forEach((n=>{void 0!==e[t][n]&&e[t][n]>-1&&e[t][n]<2||(o=!0,e[t][n]=nn[t][n])}))})),o&&Tn(),vn(),Object.keys(nn).forEach((e=>{"Header"!==e&&void 0!==hn[e].func&&hn[e].func()}))}else Mn(nn);Game.UpdateMenu(),xn()}function Sn(){Object.keys(Game.wrinklers).forEach((e=>{Game.wrinklers[e].sucked>0&&0===Game.wrinklers[e].type&&(Game.wrinklers[e].hp=0)}))}function Pn(t,o){if("b"===t){if(l("tooltip").innerHTML=Game.Objects[o].tooltip(),1===e.TooltipAmor){const e=Lo(Game.Objects[o],Game.Objects[o].basePrice,0,Game.Objects[o].free,Game.Objects[o].amount),t=e-Game.Objects[o].totalCookies;t>0&&(l("tooltip").innerHTML=l("tooltip").innerHTML.split("so far</div>").join(`so far<br/>&bull; <b>${se(t)}</b> ${1===Math.floor(t)?"cookie":"cookies"} left to amortize (${de((e-Game.Objects[o].totalCookies)/(Game.Objects[o].storedTotalCps*Game.globalCpsMult)).text})</div>`))}-1===Game.buyMode&&(l("tooltip").innerHTML=l("tooltip").innerHTML.split(se(Game.Objects[o].bulkPrice)).join(se((Game.Objects[o],Game.Objects[o].basePrice,Game.Objects[o].amount,Game.Objects[o].free,Game.buyBulk,1))))}else if("u"===t){if(!Game.UpgradesInStore[o])return"";l("tooltip").innerHTML=Game.crateTooltip(Game.UpgradesInStore[o],"store")}else"s"===t?l("tooltip").innerHTML=Game.lumpTooltip():"g"===t?l("tooltip").innerHTML=Game.Objects["Wizard tower"].minigame.spellTooltip(o)():"p"===t?l("tooltip").innerHTML=Game.ObjectsById[2].minigame.tileTooltip(o[0],o[1])():"ha"===t?l("tooltip").innerHTML=Game.ObjectsById[2].minigame.toolTooltip(1)():"wb"===t?l("tooltip").innerHTML="":"pag"===t?l("tooltip").innerHTML=Game.Objects.Temple.minigame.godTooltip(o)():"pas"===t&&(l("tooltip").innerHTML=Game.Objects.Temple.minigame.slotTooltip(o[0])());if("b"===t&&1===Game.buyMode||"u"===t||"s"===t||"g"===t||"p"===t&&!Game.keys[16]||"ha"===t||"wb"===t||"pag"===t||"pas"===t&&-1!==o[1]){const e=document.createElement("div");e.id="CMTooltipArea",l("tooltip").appendChild(e)}return te=t,oe=o,So(),l("tooltip").innerHTML}function Bn(){Game.Objects["Wizard tower"].minigameLoaded&&Object.keys(Game.Objects["Wizard tower"].minigame.spellsById).forEach((e=>{null!==l(`grimoireSpell${e}`).onmouseover&&(T[e]=l(`grimoireSpell${e}`).onmouseover,l(`grimoireSpell${e}`).onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Pn("g",`${e}`)}),"this"),Game.tooltip.wobble()})}))}function En(){if(!m&&Game.Objects["Wizard tower"].minigameLoaded){const{minigame:t}=Game.Objects["Wizard tower"];p=t.draw,Game.Objects["Wizard tower"].minigame.draw=function(){p(),1===e.GrimoireBar&&t.magic<t.magicM&&(t.magicBarTextL.innerHTML+=` (${ce(xo(t.magic,t.magicM,t.magicM))})`)},m=!0}}function Nn(){!function(){if(!u&&Game.Objects["Wizard tower"].minigameLoaded){const{minigame:e}=Game.Objects["Wizard tower"];c=e.launch,d=new Function(`return ${e.launch.toString().split("=this").join("= Game.Objects['Wizard tower'].minigame")}`),Game.Objects["Wizard tower"].minigame.launch=function(){d(),Bn(),m=!1,En(),u=!0}}}(),En()}var On=o(877),Fn=o.n(On);const Wn="CMConfig";function Ln(t){null!==l(`slider${t}`)&&(l(`slider${t}right`).innerHTML=`${l(`slider${t}`).value}%`,e[t]=Math.round(l(`slider${t}`).value)),Tn()}function An(t){e.Header[t]+=1,e.Header[t]>1&&(e.Header[t]=0),Tn()}function jn(t,o){const n=document.createElement("div");n.className="title",n.style.padding="0px 16px",n.style.opacity="0.7",n.style.fontSize="17px",n.style.fontFamily='"Kavoon", Georgia, serif',n.appendChild(document.createTextNode(`${t} `));const i=document.createElement("span");return i.style.cursor="pointer",i.style.display="inline-block",i.style.height="14px",i.style.width="14px",i.style.borderRadius="7px",i.style.textAlign="center",i.style.backgroundColor="#C0C0C0",i.style.color="black",i.style.fontSize="13px",i.style.verticalAlign="middle",i.textContent=e.Header[o]?"-":"+",i.onclick=function(){An(o),Game.UpdateMenu()},n.appendChild(i),n}function Dn(e,t,o,n){const i=document.createElement("div");i.className="listing";const a=document.createElement("b");if(a.textContent=t,i.appendChild(a),"withTooltip"===e){i.className="listing",i.appendChild(document.createTextNode(" "));const e=document.createElement("span");e.onmouseout=function(){Game.tooltip.hide()},e.onmouseover=function(){Game.tooltip.draw(this,escape(R[n].innerHTML))},e.style.cursor="default",e.style.display="inline-block",e.style.height="10px",e.style.width="10px",e.style.borderRadius="5px",e.style.textAlign="center",e.style.backgroundColor="#C0C0C0",e.style.color="black",e.style.fontSize="9px",e.style.verticalAlign="bottom",e.textContent="?",i.appendChild(e)}return i.appendChild(document.createTextNode(": ")),i.appendChild(o),i}function Hn(e){const t=document.createDocumentFragment();t.appendChild(document.createTextNode(`${e.length} `));const o=document.createElement("span");o.onmouseout=function(){Game.tooltip.hide()};const n=document.createElement("div"),i=document.createElement("div");i.style.minWidth="140px",i.style.marginBottom="4px";const a=document.createElement("div");return a.className="name",a.style.marginBottom="4px",a.style.textAlign="center",a.textContent="Missing",i.appendChild(a),Object.keys(e).forEach((t=>{const o=document.createElement("div");o.style.textAlign="center",o.appendChild(document.createTextNode(e[t])),i.appendChild(o)})),n.appendChild(i),o.onmouseover=function(){Game.tooltip.draw(this,escape(n.innerHTML))},o.style.cursor="default",o.style.display="inline-block",o.style.height="10px",o.style.width="10px",o.style.borderRadius="5px",o.style.textAlign="center",o.style.backgroundColor="#C0C0C0",o.style.color="black",o.style.fontSize="9px",o.style.verticalAlign="bottom",o.textContent="?",t.appendChild(o),t}function Un(t){const o=document.createElement("div");if(o.className="subsection",o.appendChild(t),o.appendChild(jn("Lucky Cookies","Lucky")),e.Header.Lucky&&o.appendChild(function(){const e=Game.auraMult("Dragon's Fortune")?"GoldCookDragonsFortuneTooltipPlaceholder":"GoldCookTooltipPlaceholder",t=document.createElement("div");t.className="CMStatsLuckySection";const o=Game.cookies+so()<Ye?W:N,n=Game.cookies+so()<Ye?ce((Ye-(Game.cookies+so()))/lo()):"",i=document.createDocumentFragment(),a=document.createElement("span");if(a.style.fontWeight="bold",a.className=P+o,a.textContent=se(Ye),i.appendChild(a),""!==n){const e=document.createElement("small");e.textContent=` (${n})`,i.appendChild(e)}t.appendChild(Dn("withTooltip",'"Lucky!" Cookies Required',i,e));const r=Game.cookies+so()<Ze?W:N,l=Game.cookies+so()<Ze?ce((Ze-(Game.cookies+so()))/lo()):"",s=document.createDocumentFragment(),c=document.createElement("span");if(c.style.fontWeight="bold",c.className=P+r,c.textContent=se(Ze),s.appendChild(c),""!==l){const e=document.createElement("small");e.textContent=` (${l})`,s.appendChild(e)}t.appendChild(Dn("withTooltip",'"Lucky!" Cookies Required (Frenzy)',s,e));const d=Ke!==Je,p=document.createElement("span");p.style.fontWeight="bold",p.className=P+Ke,p.textContent=se(Ke)+(d?` / ${se(Je)}`:""),t.appendChild(Dn("withTooltip",'"Lucky!" Reward (MAX)'+(d?" (Golden / Wrath)":""),p,e));const u=document.createElement("span");u.style.fontWeight="bold",u.className=P+u,u.textContent=se(et)+(d?` / ${se(tt)}`:""),t.appendChild(Dn("withTooltip",'"Lucky!" Reward (MAX) (Frenzy)'+(d?" (Golden / Wrath)":""),u,e));const m=Math.min(.15*(Game.cookies+so()),rt*lt*60*15)+13,h=document.createElement("span");return h.style.fontWeight="bold",h.className=P+h,h.textContent=se(st*m)+(d?` / ${se(ct*m)}`:""),t.appendChild(Dn("withTooltip",'"Lucky!" Reward (CUR)'+(d?" (Golden / Wrath)":""),h,e)),t}()),o.appendChild(jn("Chain Cookies","Chain")),e.Header.Chain&&o.appendChild(function(){const e=Game.auraMult("Dragon's Fortune")?"GoldCookDragonsFortuneTooltipPlaceholder":"GoldCookTooltipPlaceholder",t=document.createElement("div");t.className="CMStatsChainSection";const o=Game.cookies+so()<St?W:N,n=Game.cookies+so()<St?ce((St-(Game.cookies+so()))/lo()):"",i=document.createDocumentFragment(),a=document.createElement("span");if(a.style.fontWeight="bold",a.className=P+o,a.textContent=se(St),i.appendChild(a),""!==n){const e=document.createElement("small");e.textContent=` (${n})`,i.appendChild(e)}t.appendChild(Dn("withTooltip",'"Chain" Cookies Required',i,e));const r=Game.cookies+so()<Et?W:N,l=Game.cookies+so()<Et?ce((Et-(Game.cookies+so()))/lo()):"",s=document.createDocumentFragment(),c=document.createElement("span");if(c.style.fontWeight="bold",c.className=P+r,c.textContent=se(Et),s.appendChild(c),""!==l){const e=document.createElement("small");e.textContent=` (${l})`,s.appendChild(e)}t.appendChild(Dn("withTooltip",'"Chain" Cookies Required (Wrath)',s,e));const d=Game.cookies+so()<Ft?W:N,p=Game.cookies+so()<Ft?ce((Ft-(Game.cookies+so()))/lo()):"",u=document.createDocumentFragment(),m=document.createElement("span");if(m.style.fontWeight="bold",m.className=P+d,m.textContent=se(Ft),u.appendChild(m),""!==p){const e=document.createElement("small");e.textContent=` (${p})`,u.appendChild(e)}t.appendChild(Dn("withTooltip",'"Chain" Cookies Required (Frenzy)',u,e));const h=Game.cookies+so()<At?W:N,g=Game.cookies+so()<At?ce((At-(Game.cookies+so()))/lo()):"",f=document.createDocumentFragment(),C=document.createElement("span");if(C.style.fontWeight="bold",C.className=P+h,C.textContent=se(At),f.appendChild(C),""!==g){const e=document.createElement("small");e.textContent=` (${g})`,f.appendChild(e)}t.appendChild(Dn("withTooltip",'"Chain" Cookies Required (Frenzy) (Wrath)',f,e)),t.appendChild(Dn("withTooltip",'"Chain" Reward (MAX) (Golden / Wrath)',document.createTextNode(`${se(Bt[0])} / ${se(Ot[0])}`),e)),t.appendChild(Dn("withTooltip",'"Chain" Reward (MAX) (Frenzy) (Golden / Wrath)',document.createTextNode(`${se(Lt[0])} / ${se(Lt[0])}`),e));const y=Math.min(60*Game.cookiesPs*60*6*lt,.5*Game.cookies),b=Qo(7,y,st)[0],k=Qo(6,y,ct)[0];return t.appendChild(Dn("withTooltip",'"Chain" Reward (CUR) (Golden / Wrath)',document.createTextNode(`${se(b)} / ${se(k)}`),e)),t.appendChild(Dn("withTooltip","CPS Needed For Next Level (G / W)",document.createTextNode(`${se(Pt)} / ${se(Nt)}`),"ChainNextLevelPlaceholder")),t.appendChild(Dn("withTooltip","CPS Needed For Next Level (Frenzy) (G / W)",document.createTextNode(`${se(Wt)} / ${se(jt)}`),"ChainNextLevelPlaceholder")),t}()),Game.Objects["Wizard tower"].minigameLoaded&&(o.appendChild(jn("Spells","Spells")),e.Header.Spells&&o.appendChild(function(){const e=document.createElement("div");e.className="CMStatsSpellsSection";const t=Game.cookies+so()<ot?W:N,o=Game.cookies+so()<ot?ce((ot-(Game.cookies+so()))/lo()):"",n=document.createDocumentFragment(),i=document.createElement("span");if(i.style.fontWeight="bold",i.className=P+t,i.textContent=se(ot),n.appendChild(i),""!==o){const e=document.createElement("small");e.textContent=` (${o})`,n.appendChild(e)}e.appendChild(Dn("withTooltip",'"Conjure Baked Goods" Cookies Required',n,"GoldCookTooltipPlaceholder")),e.appendChild(Dn("withTooltip",'"Conjure Baked Goods" Reward (MAX)',document.createTextNode(se(nt)),"GoldCookTooltipPlaceholder"));const a=Game.cookies+so()<7*ot?W:N,r=Math.min(.15*(Game.cookies+so()),60*rt*30),l=Game.cookies+so()<7*ot?ce((7*ot-(Game.cookies+so()))/lo()):"",s=document.createDocumentFragment(),c=document.createElement("span");if(c.style.fontWeight="bold",c.className=P+a,c.textContent=se(7*ot),s.appendChild(c),""!==l){const e=document.createElement("small");e.textContent=` (${l})`,s.appendChild(e)}return e.appendChild(Dn("withTooltip",'"Conjure Baked Goods" Cookies Required (Frenzy)',s,"GoldCookTooltipPlaceholder")),e.appendChild(Dn("withTooltip",'"Conjure Baked Goods" Reward (MAX) (Frenzy)',document.createTextNode(se(7*nt)),"GoldCookTooltipPlaceholder")),e.appendChild(Dn("withTooltip",'"Conjure Baked Goods" Reward (CUR)',document.createTextNode(se(r)),"GoldCookTooltipPlaceholder")),it&&e.appendChild(Dn("withTooltip",'"Spontaneous Edifice" Cookies Required (most expensive building)',document.createTextNode(`${se(it)} (${at})`),"GoldCookTooltipPlaceholder")),e}())),Game.Objects.Farm.minigameLoaded&&(o.appendChild(jn("Garden","Garden")),e.Header.Garden&&o.appendChild(function(){const e=document.createElement("div");e.className="CMStatsGardenSection";const t=Game.cookies<60*Game.cookiesPs*10*100?W:N,o=document.createElement("span");o.style.fontWeight="bold",o.className=P+t,o.textContent=se(60*Game.cookiesPs*10*100),e.appendChild(Dn("basic","Cookies required for max reward of Bakeberry: ",o));const n=Game.cookies<60*Game.cookiesPs*100?W:N,i=document.createElement("span");i.style.fontWeight="bold",i.className=P+n,i.textContent=se(60*Game.cookiesPs*100),e.appendChild(Dn("basic","Cookies required for max reward of Chocoroot: ",i));const a=Game.cookies<60*Game.cookiesPs*60*25?W:N,r=document.createElement("span");r.style.fontWeight="bold",r.className=P+a,r.textContent=se(60*Game.cookiesPs*60*25),e.appendChild(Dn("basic","Cookies required for max reward of Queenbeet: ",r));const l=Game.cookies<60*Game.cookiesPs*15*100?W:N,s=document.createElement("span");s.style.fontWeight="bold",s.className=P+l,s.textContent=se(60*Game.cookiesPs*15*100),e.appendChild(Dn("basic","Cookies required for max reward of Duketater: ",s));const c=[];return Object.keys($o).forEach((e=>{Game.HasUnlocked($o[e])||c.push($o[e])})),0!==c.length&&e.appendChild(Dn("basic","Rare plant drops left to unlock",Hn(c))),e}())),o.appendChild(jn("Prestige","Prestige")),e.Header.Prestige&&o.appendChild(function(){const t=document.createElement("div");t.className="CMStatsPrestigeSection";const o=Math.floor(Game.HowMuchPrestige(Ht+Game.cookiesReset+Ct+(Game.HasUnlocked("Chocolate egg")&&!Game.Has("Chocolate egg")?mt:0)));t.appendChild(Dn("withTooltip","Prestige Level (CUR / MAX)",document.createTextNode(`${se(Game.prestige)} / ${se(o)}`),"PrestMaxTooltipPlaceholder"));const n=Math.max(0,Game.HowManyCookiesReset(o+1)-(Ht+Game.cookiesReset+Ct+(Game.HasUnlocked("Chocolate egg")&&!Game.Has("Chocolate egg")&&mt?mt:0))),i=document.createDocumentFragment();i.appendChild(document.createTextNode(se(n)));const a=document.createElement("small");a.textContent=` (${ce(n/Ut,1)})`,i.appendChild(a),t.appendChild(Dn("withTooltip","Cookies To Next Level",i,"NextPrestTooltipPlaceholder")),t.appendChild(Dn("withTooltip","Heavenly Chips (CUR / MAX)",document.createTextNode(`${se(Game.heavenlyChips)} / ${se(o-Game.prestige+Game.heavenlyChips)}`),"HeavenChipMaxTooltipPlaceholder")),t.appendChild(Dn("basic","Heavenly Chips Per Second (last 5 seconds)",document.createTextNode(se(dt,2))));const r=Number(e.HeavenlyChipsTarget);if(!Number.isNaN(r)){const e=r-Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned));e>0&&(t.appendChild(Dn("basic","Heavenly Chips To Target Set In Settings (CUR)",document.createTextNode(se(e)))),t.appendChild(Dn("basic","Time To Target (CUR, Current 5 Second Average)",document.createTextNode(ce(e/dt)))))}const l=function(e){let t=Game.cookiesPs;Go(),0===Yt["Heavenly key"].bought&&(Yt["Heavenly chip secret"].bought=1,Yt["Heavenly cookie stand"].bought=1,Yt["Heavenly bakery"].bought=1,Yt["Heavenly confectionery"].bought=1,Yt["Heavenly key"].bought=1,jo(),t=ze,Go()),Ht>=1e6&&Ao("Sacrifice"),Ht>=1e9&&Ao("Oblivion"),Ht>=1e12&&Ao("From scratch"),Ht>=1e15&&Ao("Nihilism"),Ht>=1e18&&Ao("Dematerialize"),Ht>=1e21&&Ao("Nil zero zilch"),Ht>=1e24&&Ao("Transcendence"),Ht>=1e27&&Ao("Obliterate"),Ht>=1e30&&Ao("Negative void"),Ht>=1e33&&Ao("To crumbs, you say?"),Ht>=1e36&&Ao("You get nothing"),Ht>=1e39&&Ao("Humble rebeginnings"),Ht>=1e42&&Ao("The end of the world"),Ht>=1e45&&Ao("Oh, you're back"),Ht>=1e48&&Ao("Lazarus"),Ht>=1e51&&Ao("Smurf account"),Ht>=1e54&&Ao("If at first you don't succeed"),Yt["Heavenly chip secret"].bought=1,Yt["Heavenly cookie stand"].bought=1,Yt["Heavenly bakery"].bought=1,Yt["Heavenly confectionery"].bought=1,Yt["Heavenly key"].bought=1,Ae=e;const o=We;jo(),zo(),o!==We&&jo();const n=ze-t;return Ae=Game.prestige,n}(o),s=document.createDocumentFragment();s.appendChild(document.createTextNode(se(l)));const c=Math.round(l/Game.cookiesPs*1e4);if(Number.isFinite(c)&&0!==c){const e=document.createElement("small");e.textContent=` (${c/100}% of income)`,s.appendChild(e)}t.appendChild(Dn("withTooltip","Reset Bonus Income",s,"ResetTooltipPlaceholder"));const d=Math.floor(Game.HowMuchPrestige(Game.cookiesReset)),p=Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned)),u=p-d;if(!Game.Has("Lucky digit")){let e=7-p%10;e<0&&(e+=10);const o=u+e,n=p+e,i=document.createDocumentFragment();i.appendChild(document.createTextNode(`${n.toLocaleString()} / ${o.toLocaleString()} (+${e})`)),t.appendChild(Dn("basic",'Next "Lucky Digit" (total / reset)',i))}if(!Game.Has("Lucky number")){let e=777-p%1e3;e<0&&(e+=1e3);const o=u+e,n=p+e,i=document.createDocumentFragment();i.appendChild(document.createTextNode(`${n.toLocaleString()} / ${o.toLocaleString()} (+${e})`)),t.appendChild(Dn("basic",'Next "Lucky Number" (total / reset)',i))}if(!Game.Has("Lucky payout")){let e=777777-p%1e6;e<0&&(e+=1e6);const o=u+e,n=p+e,i=document.createDocumentFragment();i.appendChild(document.createTextNode(`${n.toLocaleString()} / ${o.toLocaleString()} (+${e})`)),t.appendChild(Dn("basic",'Next "Lucky Payout" (total / reset)',i))}return t}()),Game.cpsSucked>0&&(o.appendChild(jn("Wrinklers","Wrink")),e.Header.Wrink)){const e=document.createDocumentFragment();e.appendChild(document.createTextNode(`${se(Ct)} / ${se(yt)} `));const t=document.createElement("a");t.textContent="Pop All Normal",t.className="option",t.onclick=function(){Sn()},e.appendChild(t),o.appendChild(Dn("basic","Rewards of Popping (All/Normal)",e));const n=document.createDocumentFragment();n.appendChild(document.createTextNode(`${se(bt[0])} `));const i=document.createElement("a");i.textContent="Pop Single Fattest",i.className="option",i.onclick=function(){null!==bt[1]&&(Game.wrinklers[bt[1]].hp=0)},n.appendChild(i),o.appendChild(Dn("basic",`Rewards of Popping Single Fattest Non-Shiny Wrinkler (id: ${null!==bt[1]?bt[1]:"None"})`,n))}if(o.appendChild(function(){const t=document.createElement("div");t.className="CMStatsSeasonSection";let o=!1;const n=[];Object.keys(Ho).forEach((e=>{Game.Has(Ho[e])||(n.push(Ho[e]),o=!0)}));const i=[];Object.keys(Uo).forEach((e=>{Game.Has(Uo[e])||(i.push(Uo[e]),o=!0)}));const a=[];Object.keys(Ro).forEach((e=>{Game.Has(Ro[e])||(a.push(Ro[e]),o=!0)}));const r=[];Object.keys(Game.eggDrops).forEach((e=>{Game.HasUnlocked(Game.eggDrops[e])||(r.push(Game.eggDrops[e]),o=!0)}));const l=[];Object.keys(Game.rareEggDrops).forEach((e=>{Game.HasUnlocked(Game.rareEggDrops[e])||(l.push(Game.rareEggDrops[e]),o=!0)}));const s=Game.HasUnlocked("Chocolate egg")&&!Game.Has("Chocolate egg"),c=Game.Has("Century egg");if(("christmas"===Game.season||o||s||c)&&(t.appendChild(jn("Season Specials","Sea")),e.Header.Sea)){if(0!==n.length){t.appendChild(Dn("basic","Halloween cookies left to buy",Hn(n)));let e=.95;if(Game.HasAchiev("Spooky cookies")&&(e=.8),Game.Has("Starterror")&&(e*=.9),e*=1/Game.dropRateMult(),Game.hasGod){const t=Game.hasGod("seasons");1===t?e*=.9:2===t?e*=.95:3===t&&(e*=.97)}const o=n.length/7;t.appendChild(Dn("basic","Chance of receiving a cookie from wrinkler/shiny wrinkler",document.createTextNode(`${se((1-e)*o*100)}% / ${se((1-.9*e)*o*100)}%`)))}if(0!==i.length){t.appendChild(Dn("basic","Christmas cookies left to buy",Hn(i)));let e=.8;if(Game.HasAchiev("Let it snow")&&(e=.6),e*=1/Game.dropRateMult(),Game.Has("Starsnow")&&(e*=.95),Game.hasGod){const t=Game.hasGod("seasons");1===t?e*=.9:2===t?e*=.95:3===t&&(e*=.97)}const o=i.length/7;t.appendChild(Dn("basic","Chance of receiving a cookie from reindeer",document.createTextNode(`${se((1-e)*o*100)}%`)))}0!==a.length&&t.appendChild(Dn("basic","Valentine cookies left to buy",Hn(a)));const e=function(e){let t=e*(1/Game.dropRateMult());if(Game.HasAchiev("Hide & seek champion")&&(t*=.7),Game.Has("Omelette")&&(t*=.9),Game.Has("Starspawn")&&(t*=.9),Game.hasGod){const e=Game.hasGod("seasons");1===e?t*=.9:2===e?t*=.95:3===e&&(t*=.97)}const o=1-t,n=Game.eggDrops.length-r.length,i=Game.rareEggDrops.length-l.length,a=.9*o*(1-n/Game.eggDrops.length),s=.1*o*(1-i/Game.rareEggDrops.length),c=.9*o*(n/Game.eggDrops.length)+.1*o*(i/Game.rareEggDrops.length);return[a+.9*c*(1-n/Game.eggDrops.length),s+.1*c*(1-i/Game.rareEggDrops.length)]};0!==r.length&&(t.appendChild(Dn("basic","Normal easter eggs left to unlock",Hn(r))),t.appendChild(Dn("basic","Chance of receiving an egg from wrinkler/golden cookie",document.createTextNode(`${se(100*e(.98)[0])}% / ${se(100*e(.9)[0])}%`)))),0!==l.length&&(t.appendChild(Dn("basic","Rare easter eggs left to unlock",Hn(l))),t.appendChild(Dn("basic","Chance of receiving a rare egg from wrinkler/golden cookie",document.createTextNode(`${se(100*e(.98)[1])}% / ${se(100*e(.9)[1])}%`)))),"christmas"===Game.season&&t.appendChild(Dn("basic","Reindeer reward",document.createTextNode(se(ut)))),s&&t.appendChild(Dn("withTooltip","Chocolate egg cookies",document.createTextNode(se(mt)),"ChoEggTooltipPlaceholder")),c&&t.appendChild(Dn("basic","Century egg multiplier",document.createTextNode(Math.round(1e4*(pt-1))/100+"%")))}return t}()),o.appendChild(jn("Miscellaneous","Misc")),e.Header.Misc){if(o.appendChild(Dn("basic",`Average Cookies Per Second (Past ${Y[e.AvgCPSHist]<60?`${Y[e.AvgCPSHist]} seconds`:Y[e.AvgCPSHist]/60+(3===e.AvgCPSHist?" minute":" minutes")})`,document.createTextNode(se(lo(),3)))),o.appendChild(Dn("basic",`Average Cookie Clicks Per Second (Past ${K[e.AvgClicksHist]}${0===e.AvgClicksHist?" second":" seconds"})`,document.createTextNode(se(vt,1)))),Game.Has("Fortune cookies")){const e=[];Object.keys(Do).forEach((t=>{Game.Has(Do[t])||e.push(Do[t])})),0!==e.length&&o.appendChild(Dn("basic","Fortune Upgrades Left to Buy",Hn(e)))}if(e.ShowMissedGC&&o.appendChild(Dn("basic","Missed Golden Cookies",document.createTextNode(se(Game.missedGoldenClicks)))),Game.prefs.autosave){const e=document.createElement("span");e.id="CMStatsAutosaveTimer",e.innerText=Game.sayTime(60*Game.fps-(Game.OnAscend?0:Game.T%(60*Game.fps)),4),o.appendChild(Dn("basic","Time till autosave",e))}}l("menu").insertBefore(o,l("menu").childNodes[2]),e.MissingUpgrades&&l("menu").childNodes.forEach((e=>{if(e.children[0])if("Prestige"===e.children[0].innerHTML&&Tt){const t=Tt.match(new RegExp("div","g")||0).length/2,o=document.createElement("div");o.id="CMMissingUpgradesPrestigeTitle",o.className="listing";const n=document.createElement("div");n.innerHTML=`<b>Missing Prestige upgrades:</b> ${t}/${Game.PrestigeUpgrades.length} (${Math.floor(t/Game.PrestigeUpgrades.length*100)}%)`,o.appendChild(n),e.appendChild(o);const i=document.createElement("div");i.className="listing crateBox",i.innerHTML=Tt,e.appendChild(i)}else if("Upgrades"===e.children[0].innerHTML){if(xt){const t=xt.match(new RegExp("div","g")||0).length/2,o=document.createElement("div");o.id="CMMissingUpgradesTitle",o.className="listing";const n=document.createElement("div");n.innerHTML=`<b>Missing normal upgrades:</b> ${t}/${Game.UpgradesByPool[""].length+Game.UpgradesByPool.tech.length} (${Math.floor(t/(Game.UpgradesByPool[""].length+Game.UpgradesByPool.tech.length)*100)}%)`,o.appendChild(n),e.insertBefore(o,e.childNodes[3]);const i=document.createElement("div");i.className="listing crateBox",i.innerHTML=xt,e.insertBefore(i,document.getElementById("CMMissingUpgradesTitle").nextSibling)}if(Mt){const t=Mt.match(new RegExp("div","g")||0).length/2,o=document.createElement("div");o.id="CMMissingUpgradesCookiesTitle",o.className="listing";const n=document.createElement("div");n.innerHTML=`<b>Missing Cookie upgrades:</b> ${t}/${Game.UpgradesByPool.cookie.length} (${Math.floor(t/Game.UpgradesByPool.cookie.length*100)}%)`,o.appendChild(n),e.appendChild(o);const i=document.createElement("div");i.className="listing crateBox",i.innerHTML=Mt,e.appendChild(i)}}}))}const Rn={Calculation:"Calculation",Notation:"Notation",Colours:"Colours and colour coding",BarsDisplay:"Infobars and visual settings",Tooltip:"Tooltips",Statistics:"Statistics",Notification:"Notifications",Miscellaneous:"Miscellaneous"},$n={NotificationGeneral:"General Notifications",NotificationGC:"Golden Cookie",NotificationFC:"Fortune Cookie",NotificationSea:"Season Special",NotificationGard:"Garden Tick",NotificationMagi:"Full Magic Bar",NotificationWrink:"Wrinkler",NotificationWrinkMax:"Maximum Wrinklers"};function In(t,o){const n=document.createElement("div");n.className="title",n.style.opacity="0.7",n.style.fontSize="17px",n.appendChild(document.createTextNode(`${o} `));const i=document.createElement("span");return i.style.cursor="pointer",i.style.display="inline-block",i.style.height="14px",i.style.width="14px",i.style.borderRadius="7px",i.style.textAlign="center",i.style.backgroundColor="#C0C0C0",i.style.color="black",i.style.fontSize="13px",i.style.verticalAlign="middle",i.textContent=e.Header[t]?"-":"+",i.onclick=function(){An(t),Game.UpdateMenu()},n.appendChild(i),n}function zn(t){const o=document.createElement("div");if(o.className="listing","bool"===hn[t].type){const n=document.createElement("a");hn[t].toggle&&0===e[t]?n.className="option off":n.className="option",n.id=Wn+t,n.onclick=function(){!function(t){e[t]+=1,e[t]===hn[t].label.length?(e[t]=0,hn[t].toggle&&(l(Wn+t).className="option off")):l(Wn+t).className="option",void 0!==hn[t].func&&hn[t].func(),l(Wn+t).innerHTML=hn[t].label[e[t]],Tn()}(t)},n.textContent=hn[t].label[e[t]],o.appendChild(n);const i=document.createElement("label");return i.textContent=hn[t].desc,o.appendChild(i),o}if("vol"===hn[t].type){const n=document.createElement("div");n.className="sliderBox";const i=document.createElement("div");i.style.float="left",i.innerHTML=hn[t].desc,n.appendChild(i);const a=document.createElement("div");a.id=`slider${t}right`,a.style.float="right",a.innerHTML=`${e[t]}%`,n.appendChild(a);const r=document.createElement("input");r.className="slider",r.id=`slider${t}`,r.style.clear="both",r.type="range",r.min="0",r.max="100",r.step="1",r.value=e[t],r.oninput=function(){Ln(t)},r.onchange=function(){Ln(t)},n.appendChild(r),o.appendChild(n);const l=document.createElement("a");return l.className="option",l.onclick=function(){Gn(e[t.replace("Volume","SoundURL")],t.replace("Volume","Sound"),t,!0)},l.textContent="Test sound",o.appendChild(l),o}if("url"===hn[t].type){const n=document.createElement("span");n.className="option",n.textContent=`${hn[t].label} `,o.appendChild(n);const i=document.createElement("input");i.id=Wn+t,i.className="option",i.type="text",i.readOnly=!0,i.setAttribute("value",e[t]),i.style.width="300px",o.appendChild(i),o.appendChild(document.createTextNode(" "));const a=document.createElement("input");a.id=`${Wn+t}Prompt`,a.className="option",a.type="text",a.setAttribute("value",e[t]);const r=document.createElement("a");r.className="option",r.onclick=function(){!function(e,t){Game.promptWrapL.className="framed";const o=e;Game.promptL.innerHTML=`${o}<div class="optionBox"></div>`,Object.keys(t).forEach((e=>{const o=document.createElement("a");o.id=`promptOption${e}`,o.className="option",o.onclick=function(){PlaySound("snd/tick.mp3"),t[e][1]()},o.textContent=t[e][0],Game.promptL.children[1].appendChild(o)})),Game.promptAnchorL.style.display="block",Game.darkenL.style.display="block",Game.promptL.focus(),Game.promptOn=1,Game.UpdatePrompt()}(a.outerHTML,[["Save",function(){e[t]=l(`CMConfig${t}Prompt`).value,Tn(),Game.ClosePrompt(),Game.UpdateMenu()}],["Cancel",function(){Game.ClosePrompt()}]])},r.textContent="Edit",o.appendChild(r);const s=document.createElement("label");return s.textContent=hn[t].desc,o.appendChild(s),o}if("colour"===hn[t].type){o.className="";const n=document.createElement("div");n.className="listing";const i=document.createElement("input");i.id=t,i.style.width="65px",i.setAttribute("value",e[t]),n.appendChild(i),new On(i,{hash:!0,position:"right",onInput:function(){e[this.targetElement.id]=this.toHEXString(),xn(),Tn(),Game.UpdateMenu()}});const a=document.createElement("label");if(a.textContent=hn[t].desc,n.appendChild(a),t.includes("Flash")){const e=document.createElement("a");e.className="option",e.onclick=function(){kn(3,t.replace("Colour",""),!0)},e.textContent="Test flash",n.appendChild(e)}return o.appendChild(n),Fn().init(),o}if("numscale"===hn[t].type){const n=document.createElement("span");n.className="option",n.textContent=`${hn[t].label} `,o.appendChild(n);const i=document.createElement("input");i.id=Wn+t,i.className="option",i.type="number",i.value=e[t],i.min=hn[t].min,i.max=hn[t].max,i.oninput=function(){this.value>this.max&&console.log("TEST"),e[t]=this.value,Tn(),ln()},o.appendChild(i),o.appendChild(document.createTextNode(" "));const a=document.createElement("label");return a.textContent=hn[t].desc,o.appendChild(a),o}return o}function Vn(t){if(1===e.TimerBar&&0===e.TimerBarPos){const e=parseInt(l("CMTimerBar").style.height,10);Game.mouseY-=e,t(),Game.mouseY+=e}else t()}function _n(){S.Beautify=Beautify,Beautify=se,S.CalculateGains=Game.CalculateGains,Game.CalculateGains=function(){S.CalculateGains(),$e=1},S.tooltip={},S.tooltip.draw=Game.tooltip.draw,S.tooltip.drawMod=new Function(`return ${Game.tooltip.draw.toString().split("this").join("Game.tooltip")}`)(),Game.tooltip.draw=function(e,t,o){S.tooltip.drawMod(e,t,o)},S.tooltip.update=Game.tooltip.update,S.tooltip.updateMod=new Function(`return ${Game.tooltip.update.toString().split("this.").join("Game.tooltip.")}`)(),Game.tooltip.update=function(){S.tooltip.updateMod(),function(){if("store"===Game.tooltip.origin){let t=0;1===e.ToolWarnLucky&&1===e.ToolWarnPos&&null!==l("CMDispTooltipWarningParent")&&(t=l("CMDispTooltipWarningParent").clientHeight-4),Game.tooltip.tta.style.top=`${Math.min(parseInt(Game.tooltip.tta.style.top,10),l("game").clientHeight+l("topBar").clientHeight-Game.tooltip.tt.clientHeight-t-46)}px`}}()},S.UpdateWrinklers=Game.UpdateWrinklers,Game.UpdateWrinklers=function(){Vn(S.UpdateWrinklers)},S.UpdateSpecial=Game.UpdateSpecial,Game.UpdateSpecial=function(){Vn(S.UpdateSpecial)},l("bigCookie").removeEventListener("click",Game.ClickCookie,!1),l("bigCookie").addEventListener("click",(function(){Vn(Game.ClickCookie)}),!1),S.RebuildUpgrades=Game.RebuildUpgrades,Game.RebuildUpgrades=function(){S.RebuildUpgrades(),M=[],Object.keys(Game.UpgradesInStore).forEach((e=>{null!==l(`upgrade${e}`).onmouseover&&(M[e]=l(`upgrade${e}`).onmouseover,l(`upgrade${e}`).onmouseover=function(){Game.mouseDown||(Game.setOnCrate(this),Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Pn("u",`${e}`)}),"store"),Game.tooltip.wobble())})})),Game.CalculateGains()},S.ClickProduct=Game.ClickProduct,Game.ClickProduct=function(t){(!e.BulkBuyBlock||Game.ObjectsById[t].bulkPrice<Game.cookies||-1===Game.buyMode)&&S.ClickProduct(t)},S.DescribeDragonAura=Game.DescribeDragonAura,Game.DescribeDragonAura=function(t){S.DescribeDragonAura(t),function(t){if(1===e.DragonAuraInfo){const[e,o]=function(e){Go(),l("promptContent").children[0].innerHTML.includes("secondary")?De=e:je=e;let t=0;if(je!==_e||De!==Xe)for(let e=Game.ObjectsById.length-1;e>-1;--e)if(Game.ObjectsById[e].amount>0){const o=qt[Game.ObjectsById[e].name].name;qt[o].amount-=1,Ne-=1,t=qt[o].basePrice*Game.priceIncrease**Math.max(0,qt[o].amount-1-qt[o].free),t=Game.modifyBuildingPrice(qt[o],t),t=Math.ceil(t);break}const o=We;return jo(),zo(),o!==We&&jo(),[ze-Game.cookiesPs,t]}(t),n=ce(o/(e+Game.cookiesPs)),i=se(e/Game.cookiesPs*100);l("dragonAuraInfo").style.minHeight="60px",l("dragonAuraInfo").style.margin="8px",l("dragonAuraInfo").appendChild(document.createElement("div")).className="line";const a=document.createElement("div");a.style.minWidth="200px",a.style.textAlign="center",a.textContent=`Picking this aura will change CPS by ${se(e)} (${i}% of current CPS).`,l("dragonAuraInfo").appendChild(a);const r=document.createElement("div");r.style.minWidth="200px",r.style.textAlign="center",r.textContent=`It will take ${n} to recover the cost.`,l("dragonAuraInfo").appendChild(r)}}(t)},S.ToggleSpecialMenu=Game.ToggleSpecialMenu,Game.ToggleSpecialMenu=function(e){S.ToggleSpecialMenu(e),function(){if(null!==(l("specialPopup").className.match(/onScreen/)&&l("specialPopup").children[0].style.background.match(/dragon/)))for(let e=0;e<l("specialPopup").childNodes.length;e++)"optionBox"===l("specialPopup").childNodes[e].className&&(l("specialPopup").children[e].onmouseover=function(){fn(),Game.tooltip.dynamic=1,Game.tooltip.draw(l("specialPopup"),`<div style="min-width:200px;text-align:center;">${qe}</div>`,"this"),Game.tooltip.wobble()},l("specialPopup").children[e].onmouseout=function(){Game.tooltip.shouldHide=1})}()},S.UpdateMenu=Game.UpdateMenu,Game.UpdateMenu=function(){void 0!==Fn().picker&&void 0!==Fn().picker.owner||(S.UpdateMenu(),function(){const t=document.createElement("div");t.className="title","prefs"===Game.onMenu?(t.textContent="Cookie Monster Settings",function(t){const o=document.createDocumentFragment();o.appendChild(t),Object.keys(Rn).forEach((t=>{const n=In(t,Rn[t]);o.appendChild(n),e.Header[t]&&("Notification"===t?Object.keys($n).forEach((t=>{const n=In(t,$n[t]);n.style.fontSize="15px",n.style.opacity="0.5",o.appendChild(n),e.Header[t]&&Object.keys(hn).forEach((e=>{hn[e].group===t&&o.appendChild(zn(e))}))})):Object.keys(hn).forEach((e=>{hn[e].group===t&&o.appendChild(zn(e))})))}));const n=document.createElement("div");n.className="listing";const i=document.createElement("a");i.className="option",i.onclick=function(){Mn(nn)},i.textContent="Restore Default",n.appendChild(i),o.appendChild(n),l("menu").childNodes[2].insertBefore(o,l("menu").childNodes[2].childNodes[l("menu").childNodes[2].childNodes.length-1])}(t)):"stats"===Game.onMenu?e.Stats&&(t.textContent="Cookie Monster Statistics",Un(t)):"log"===Game.onMenu&&(t.textContent="Cookie Monster ",function(t){const o=document.createElement("div");o.className="subsection";const n=document.createElement("span");if(n.style.cursor="pointer",n.style.display="inline-block",n.style.height="14px",n.style.width="14px",n.style.borderRadius="7px",n.style.textAlign="center",n.style.backgroundColor="#C0C0C0",n.style.color="black",n.style.fontSize="13px",n.style.verticalAlign="middle",n.textContent=e.Header.InfoTab?"-":"+",n.onclick=function(){An("InfoTab"),Game.UpdateMenu()},t.appendChild(n),o.appendChild(t),e.Header.InfoTab){const e=document.createElement("div");e.innerHTML='<div class="listing">\n <a href="https://github.com/Aktanusa/CookieMonster" target="blank">Cookie Monster</a>\n offers a wide range of tools and statistics to enhance your game experience.\n It is not a cheat interface – although it does offer helpers for golden cookies and such, everything can be toggled off at will to only leave how much information you want.</br>\n Progess on new updates and all previous release notes can be found on the GitHub page linked above!</br>\n Please also report any bugs you may find over there!</br>\n </div>\n ',o.appendChild(e);const t=document.createElement("div");t.innerHTML='<div class="listing">\nThe latest update (v 2.031.6) has revamped the way PP is colourcoded. In the previous versions this was not always correct and we had some settings that made it even more incorrect. With this update Cookie Monster compares the PP of a building to all other possibilities (also buying 10 or 100) of a building and colors the PP accordingly.</br>\nThis means that if it is better to buy 10 of a building immediately rather than to buy in increments of 1 Cookie Monster will now show this! In this case the buy 1 option will be yellow, while the buy 10 option will be green.</br>\nThis also means that you will see some more diverse colors. Rather than coloring according to arbitrary "middle values" Cookie Monster now uses a top 10, 20 and 30. While this might take some getting used to we believe the new system is much better in conveying useful information to the user while also being more correct!</br>\n</br>\nThis update also implements the following functions:</br>\n- Added a tooltip displaying the reward to the extra pop wrinkler buttons</br>\n- Added tooltips to the Gods in the Pantheon</br>\n- The tooltip for Elder Pledge now displays correctly, although it takes some time to load after Cookie Monster has been loaded</br>\n- You can now test the volume and sound of notifications in the settings screen</br>\n- The bottom bar will now flicker less and each column has received a bit of padding</br>\n- Introduced new colour scheme for PP, see the explanation of colors in the Readme or the settings</br>\n- Added option to show buttons that can hide the upgrade and building section</br>\n- Added option to display PP as a time unit, note that these are only approximations. PP does not translate directly into time (the name is deceptive, we know...)</br>\n- Statistics page now shows chance for random drops when they have not been achieved</br>\n- In a later hotfix we have added a Rigidel mode that will colour all buy 1 options grey</br>\n</br>\nThis update fixes the following bugs:</br>\n- Incorrect amount for "required for max plant reward" in statistics page</br>\n- Fixed the tooltips of the \'?\' buttons in the statistics page not showing</br>\n- Fixed some unclear settings descriptions </br>\n</br>\nPlease submit any bug reports or feature requests to the <a href="https://github.com/Aktanusa/CookieMonster">GitHub page!</a>\n</br>\n</div>\n',o.appendChild(t)}const i=l("menu").children[1];i.insertBefore(o,i.children[1])}(t))}())},S.sayTime=Game.sayTime,I=function(e,t){return Number.isNaN(e)||e<=0?S.sayTime(e,t):ce(e/Game.fps,1)},S.Logic=Game.Logic,Game.Logic=function(){S.Logic();let t="Cookie Clicker";"fools"===Game.season&&(t="Cookie Baker"),le=`${Game.OnAscend?"Ascending! ":""}${se(Game.cookies)} ${1===Game.cookies?"cookie":"cookies"} - ${t}`,function(){if(Game.OnAscend||0===e.Title)document.title=le;else if(1===e.Title){let e,t,o,n=!1,i=!1;e=Rt?Rt.wrath?`[W${Math.ceil(Rt.life/Game.fps)}]`:`[G${Math.ceil(Rt.life/Game.fps)}]`:Game.Has("Golden switch [off]")?"[GS]":`[${Number(l("CMTimerBarGCMinBar").textContent)<0?"!":""}${Math.ceil((Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.time)/Game.fps)}]`,f&&(n=!0,t="[F]"),"christmas"===Game.season&&(i=!0,o=g?`[R${Math.ceil($t.life/Game.fps)}]`:`[${Number(l("CMTimerBarRenMinBar").textContent)<0?"!":""}${Math.ceil((Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.time)/Game.fps)}]`);let a=le;"["===a.charAt(0)&&(a=a.substring(a.lastIndexOf("]")+1)),document.title=`${e+(n?t:"")+(i?o:"")} ${a}`}else if(2===e.Title){let e="",t=!1;Rt&&(t=!0,Rt.wrath?e+=`[W${Math.ceil(Rt.life/Game.fps)}]`:e+=`[G${Math.ceil(Rt.life/Game.fps)}]`),f&&(t=!0,e+="[F]"),"christmas"===Game.season&&g&&(e+=`[R${Math.ceil($t.life/Game.fps)}]`,t=!0),t&&(e+=" - ");let o="Cookie Clicker";"fools"===Game.season&&(o="Cookie Baker"),e+=o,document.title=e}}(),function(){const t=Math.max(0,Game.HowManyCookiesReset(Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned))+1)-(Game.cookiesEarned+Game.cookiesReset)),o=Game.sayTime((Date.now()-Game.startDate)/1e3*Game.fps,-1);let n="";n+=`You've been on this run for <b>${""===o?"not very long":o}</b>.<br>`,n+='<div class="line"></div>',Game.prestige>0&&(n+=`Your prestige level is currently <b>${se(Game.prestige)}</b>.<br>(CpS +${se(Game.prestige)}%)`,n+='<div class="line"></div>'),n+=Pe<1?"Ascending now would grant you no prestige.":Pe<2?"Ascending now would grant you<br><b>1 prestige level</b> (+1% CpS)<br>and <b>1 heavenly chip</b> to spend.":`Ascending now would grant you<br><b>${se(Pe)} prestige levels</b> (+${se(Pe)}% CpS)<br>and <b>${se(Pe)} heavenly chips</b> to spend.`,n+='<div class="line"></div>',n+=`You need <b>${se(t)} more cookies</b> for the next level.<br>`,n+=e.TooltipAscendButton?`<div class='line'></div>It takes ${It} to reach the next level and you are making ${se(dt,2)} chips on average in the last 5 seconds.<br>`:"",l("ascendTooltip").innerHTML=n}()}}function Xn(){yn(),Zo(),on(),r=Object.keys(Game.mods).length,v=document.createElement("style"),v.type="text/css",v.id="CMCSS",document.head.appendChild(v),function(){const e=document.createElement("div");e.id="CMBotBar",e.style.height="69px",e.style.width="100%",e.style.position="absolute",e.style.display="none",e.style.backgroundColour="#262224",e.style.backgroundImage="linear-gradient(to bottom, #4d4548, #000000)",e.style.borderTop="1px solid black",e.style.overflow="auto",e.style.textShadow="-1px 0 black, 0 1px black, 1px 0 black, 0 -1px black";const t=e.appendChild(document.createElement("table"));t.style.width="100%",t.style.textAlign="center",t.style.whiteSpace="nowrap";const o=t.appendChild(document.createElement("tbody")),n=function(e,t){const o=document.createElement("td");return o.style.textAlign="right",o.className=P+t,o.textContent=e,o},i=o.appendChild(document.createElement("tr"));i.style.fontWeight="bold",i.appendChild(n("CM 2.031.6",O)),o.appendChild(document.createElement("tr")).appendChild(n("Bonus Income",E)),o.appendChild(document.createElement("tr")).appendChild(n("Payback Period",E)),o.appendChild(document.createElement("tr")).appendChild(n("Time Left",E)),l("wrapper").appendChild(e),Object.keys(Game.Objects).forEach((e=>{po(e)}))}(),function(){const e=document.createElement("div");e.id="CMTimerBar",e.style.position="absolute",e.style.display="none",e.style.height="0px",e.style.fontSize="10px",e.style.fontWeight="bold",e.style.backgroundColour="black";const t=co("CMTimerBarGC","Next Cookie",[{id:"CMTimerBarGCMinBar",color:A},{id:"CMTimerBarGCBar",color:L}]);e.appendChild(t);const o=co("CMTimerBarRen","Next Reindeer",[{id:"CMTimerBarRenMinBar",color:A},{id:"CMTimerBarRenBar",color:F}]);e.appendChild(o);const n=document.createElement("div");n.id="CMTimerBarBuffTimers",e.appendChild(n),l("wrapper").appendChild(e)}(),function(){const e=document.createElement("div");e.id="CMUpgradeBar",e.style.width="100%",e.style.backgroundColour="black",e.style.textAlign="center",e.style.fontWeight="bold",e.style.display="none",e.style.zIndex="21",e.onmouseout=function(){Game.tooltip.hide()};const t=document.createElement("div");t.appendChild(function(){const e=document.createElement("div");e.style.minWidth="330px",e.style.marginBottom="4px";const t=document.createElement("div");t.className="name",t.style.marginBottom="4px",t.textContent="Legend",e.appendChild(t);const o=function(e,t){const o=document.createElement("div");o.style.verticalAlign="middle";const n=document.createElement("span");return n.className=B+e,n.style.display="inline-block",n.style.height="10px",n.style.width="10px",n.style.marginRight="4px",o.appendChild(n),o.appendChild(document.createTextNode(t)),o};return e.appendChild(o(E,"Better than the best PP of a building option")),e.appendChild(o(N,"Same as the best PP building option")),e.appendChild(o(O,"Within the top 10 of PP for buildings")),e.appendChild(o(F,"Within the top 20 of PP for buildings")),e.appendChild(o(W,"Within the top 30 of PP for buildings")),e.appendChild(o(L,"Outside of the top 30 of PP for buildings")),e.appendChild(o(A,"Negative or infinity PP")),e}()),e.onmouseover=function(){Game.tooltip.draw(this,escape(t.innerHTML),"store")};const o=function(e,t){const o=document.createElement("span");return o.id=e,o.className=P+t,o.style.width="14.28571428571429%",o.style.display="inline-block",o.textContent="0",o};e.appendChild(o("CMUpgradeBarBlue",E)),e.appendChild(o("CMUpgradeBarGreen",N)),e.appendChild(o("CMUpgradeBarYellow",O)),e.appendChild(o("CMUpgradeBarOrange",F)),e.appendChild(o("CMUpgradeBarRed",W)),e.appendChild(o("CMUpgradeBarPurple",L)),e.appendChild(o("CMUpgradeBarGray",A)),l("upgrades").parentNode.insertBefore(e,l("upgrades").parentNode.childNodes[3])}(),function(){const e=document.createElement("div");e.id="CMFlashScreen",e.style.width="100%",e.style.height="100%",e.style.backgroundColor="white",e.style.display="none",e.style.zIndex="9999999999",e.style.position="absolute",l("wrapper").appendChild(e)}(),function(){const e=document.createElement("div");e.id="CMSectionHidButtons",e.style.textAlign="center";const t=document.createElement("a");t.className="option",t.onclick=function(){"flex"===l("upgrades").style.display?(l("upgrades").style.display="none",l("toggleUpgrades").style.display="none",l("techUpgrades").style.display="none",l("vaultUpgrades").style.display="none"):(l("upgrades").style.display="flex",0!==l("toggleUpgrades").children.length&&(l("toggleUpgrades").style.display="block"),0!==l("techUpgrades").children.length&&(l("techUpgrades").style.display="block"),0!==l("vaultUpgrades").children.length&&(l("vaultUpgrades").style.display="block"))},t.textContent="Hide/Show Upgrades",e.appendChild(t);const o=document.createElement("a");o.className="option",o.onclick=function(){"grid"===l("products").style.display?l("products").style.display="none":l("products").style.display="grid"},o.textContent="Hide/Show Buildings",e.appendChild(o),l("store").insertBefore(e,l("store").childNodes[2])}(),function(){const e=document.createElement("link");e.id="CMFavicon",e.rel="shortcut icon",e.href="https://orteil.dashnet.org/cookieclicker/favicon.ico",document.getElementsByTagName("head")[0].appendChild(e)}(),Object.keys(U).forEach((e=>{!function(e,t,o){const n=document.createElement("div");n.id=e;const i=document.createElement("div");i.style.minWidth=o,i.style.marginBottom="4px";const a=document.createElement("div");a.style.textAlign="left",a.textContent=t,i.appendChild(a),n.appendChild(i),R[e]=n}(U[e][0],U[e][1],U[e][2])})),function(){const e=document.createElement("a");e.id="PopAllNormalWrinklerButton",e.textContent="Pop All Normal",e.className="option",e.onclick=function(){Sn()},e.onmouseout=function(){Game.tooltip.shouldHide=1},e.onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Pn("wb","PopAll")}),"this"),Game.tooltip.wobble()},l("sectionLeftExtra").children[0].append(e);const t=document.createElement("a");t.id="PopFattestWrinklerButton",t.textContent="Pop Single Fattest",t.className="option",t.onclick=function(){null!==bt[1]&&(Game.wrinklers[bt[1]].hp=0)},t.onmouseout=function(){Game.tooltip.shouldHide=1},t.onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Pn("wb","PopFattest")}),"this"),Game.tooltip.wobble()},l("sectionLeftExtra").children[0].append(t)}(),l("products").style.display="grid",l("storeBulk").style.gridRow="1/1",l("upgrades").style.display="flex",l("upgrades").style["flex-wrap"]="wrap",Object.keys(Game.Objects).forEach((e=>{const t=Game.Objects[e];null!==l(`product${t.id}`).onmouseover&&(x[e]=l(`product${t.id}`).onmouseover,l(`product${t.id}`).onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Pn("b",`${e}`)}),"store"),Game.tooltip.wobble()})})),Game.canLumps()&&(s=l("lumps").onmouseover,l("lumps").onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Pn("s","Lump")}),"this"),Game.tooltip.wobble()}),h=Game.LoadMinigames,Game.LoadMinigames=function(){h(),Game.Objects.Farm.minigameLoaded&&(l("gardenTool-1").onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Pn("ha","HarvestAllButton")}),"this"),Game.tooltip.wobble()},Array.from(l("gardenPlot").children).forEach((e=>{const t=e.id.slice(-3);e.onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Pn("p",[`${t[0]}`,`${t[2]}`])}),"this"),Game.tooltip.wobble()}}))),Bn(),function(){if(Game.Objects.Temple.minigameLoaded){for(let e=0;e<11;e+=1)l(`templeGod${e}`).onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Pn("pag",e)}),"this"),Game.tooltip.wobble()};for(let e=0;e<3;e+=1)l(`templeSlot${e}`).onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Pn("pas",[e,Game.Objects.Temple.minigame.slot[e]])}),"this"),Game.tooltip.wobble()}}}(),Nn()},Game.LoadMinigames(),l("backgroundLeftCanvas").onmouseover=function(){z=1},l("backgroundLeftCanvas").onmouseout=function(){z=0,Game.tooltip.hide(),Object.keys(Game.wrinklers).forEach((e=>{_[e]=0}))},_n(),Nn(),Game.CalculateGains(),Mn(),$=Game.OnAscend,Game.prefs.popups?Game.Popup("Cookie Monster version 2.031.6 loaded!"):Game.Notify("Cookie Monster version 2.031.6 loaded!","","",1,1),Game.Win("Third-party")}const Qn={init:function(){bn=!0;let e=!0;Game.version!==Number("2.031")&&(e=confirm("Cookie Monster version 2.031.6 is meant for Game version 2.031.  Loading a different version may cause errors.  Do you still want to load Cookie Monster?")),e&&(Xn(),Game.registerHook("draw",Po),Game.registerHook("logic",vn),bn=!1)},load:function(e){const t=JSON.parse(e);yn(),Mn(t.settings),"2.031.6"!==t.version&&(Game.prefs.popups?Game.Popup("A new version of Cookie Monster has been loaded, check out the release notes in the info tab!"):Game.Notify("A new version of Cookie Monster has been loaded, check out the release notes in the info tab!","","",0,1))},save:gn};Game.registerMod("CookieMonster",Qn)})()})();
+(()=>{var e={877:function(e){!function(t,o){"use strict";"object"!=typeof e.exports?o(t):e.exports=t.document?o(t):function(e){if(!e.document)throw new Error("jscolor needs a window with document");return o(e)}}("undefined"!=typeof window?window:this,(function(e){"use strict";var t,o,n,i,a=((i={initialized:!1,instances:[],readyQueue:[],register:function(){void 0!==e&&e.document&&e.document.addEventListener("DOMContentLoaded",i.pub.init,!1)},installBySelector:function(t,o){if(!(o=o?i.node(o):e.document))throw new Error("Missing root node");for(var n=o.querySelectorAll(t),a=new RegExp("(^|\\s)("+i.pub.lookupClass+")(\\s*(\\{[^}]*\\})|\\s|$)","i"),r=0;r<n.length;r+=1){var l,s;if(!(n[r].jscolor&&n[r].jscolor instanceof i.pub)&&(void 0===n[r].type||"color"!=n[r].type.toLowerCase()||!i.isColorAttrSupported)&&(null!==(l=i.getDataAttr(n[r],"jscolor"))||n[r].className&&(s=n[r].className.match(a)))){var c=n[r],d="";null!==l?d=l:s&&(console.warn('Installation using class name is DEPRECATED. Use data-jscolor="" attribute instead.'+i.docsRef),s[4]&&(d=s[4]));var p=null;if(d.trim())try{p=i.parseOptionsStr(d)}catch(e){console.warn(e+"\n"+d)}try{new i.pub(c,p)}catch(e){console.warn(e)}}}},parseOptionsStr:function(e){var t=null;try{t=JSON.parse(e)}catch(o){if(!i.pub.looseJSON)throw new Error("Could not parse jscolor options as JSON: "+o);try{t=new Function("var opts = ("+e+'); return typeof opts === "object" ? opts : {};')()}catch(e){throw new Error("Could not evaluate jscolor options: "+e)}}return t},getInstances:function(){for(var e=[],t=0;t<i.instances.length;t+=1)i.instances[t]&&i.instances[t].targetElement&&e.push(i.instances[t]);return e},createEl:function(t){var o=e.document.createElement(t);return i.setData(o,"gui",!0),o},node:function(t){if(!t)return null;if("string"==typeof t){var o=t,n=null;try{n=e.document.querySelector(o)}catch(e){return console.warn(e),null}return n||console.warn("No element matches the selector: %s",o),n}return i.isNode(t)?t:(console.warn("Invalid node of type %s: %s",typeof t,t),null)},isNode:function(e){return"object"==typeof Node?e instanceof Node:e&&"object"==typeof e&&"number"==typeof e.nodeType&&"string"==typeof e.nodeName},nodeName:function(e){return!(!e||!e.nodeName)&&e.nodeName.toLowerCase()},removeChildren:function(e){for(;e.firstChild;)e.removeChild(e.firstChild)},isTextInput:function(e){return e&&"input"===i.nodeName(e)&&"text"===e.type.toLowerCase()},isButton:function(e){if(!e)return!1;var t=i.nodeName(e);return"button"===t||"input"===t&&["button","submit","reset"].indexOf(e.type.toLowerCase())>-1},isButtonEmpty:function(e){switch(i.nodeName(e)){case"input":return!e.value||""===e.value.trim();case"button":return""===e.textContent.trim()}return null},isPassiveEventSupported:function(){var t=!1;try{var o=Object.defineProperty({},"passive",{get:function(){t=!0}});e.addEventListener("testPassive",null,o),e.removeEventListener("testPassive",null,o)}catch(e){}return t}(),isColorAttrSupported:(n=e.document.createElement("input"),!(!n.setAttribute||(n.setAttribute("type","color"),"color"!=n.type.toLowerCase()))),dataProp:"_data_jscolor",setData:function(){var e=arguments[0];if(3===arguments.length){var t=e.hasOwnProperty(i.dataProp)?e[i.dataProp]:e[i.dataProp]={},o=arguments[2];return t[a=arguments[1]]=o,!0}if(2===arguments.length&&"object"==typeof arguments[1]){t=e.hasOwnProperty(i.dataProp)?e[i.dataProp]:e[i.dataProp]={};var n=arguments[1];for(var a in n)n.hasOwnProperty(a)&&(t[a]=n[a]);return!0}throw new Error("Invalid arguments")},removeData:function(){var e=arguments[0];if(!e.hasOwnProperty(i.dataProp))return!0;for(var t=1;t<arguments.length;t+=1){var o=arguments[t];delete e[i.dataProp][o]}return!0},getData:function(e,t,o){if(!e.hasOwnProperty(i.dataProp)){if(void 0===o)return;e[i.dataProp]={}}var n=e[i.dataProp];return n.hasOwnProperty(t)||void 0===o||(n[t]=o),n[t]},getDataAttr:function(e,t){var o="data-"+t;return e.getAttribute(o)},setDataAttr:function(e,t,o){var n="data-"+t;e.setAttribute(n,o)},_attachedGroupEvents:{},attachGroupEvent:function(e,t,o,n){i._attachedGroupEvents.hasOwnProperty(e)||(i._attachedGroupEvents[e]=[]),i._attachedGroupEvents[e].push([t,o,n]),t.addEventListener(o,n,!1)},detachGroupEvents:function(e){if(i._attachedGroupEvents.hasOwnProperty(e)){for(var t=0;t<i._attachedGroupEvents[e].length;t+=1){var o=i._attachedGroupEvents[e][t];o[0].removeEventListener(o[1],o[2],!1)}delete i._attachedGroupEvents[e]}},preventDefault:function(e){e.preventDefault&&e.preventDefault(),e.returnValue=!1},captureTarget:function(e){e.setCapture&&(i._capturedTarget=e,i._capturedTarget.setCapture())},releaseTarget:function(){i._capturedTarget&&(i._capturedTarget.releaseCapture(),i._capturedTarget=null)},triggerEvent:function(t,o,n,a){if(t){var r=null;return"function"==typeof Event?r=new Event(o,{bubbles:n,cancelable:a}):(r=e.document.createEvent("Event")).initEvent(o,n,a),!!r&&(i.setData(r,"internal",!0),t.dispatchEvent(r),!0)}},triggerInputEvent:function(e,t,o,n){e&&i.isTextInput(e)&&i.triggerEvent(e,t,o,n)},eventKey:function(e){var t={9:"Tab",13:"Enter",27:"Escape"};return"string"==typeof e.code?e.code:void 0!==e.keyCode&&t.hasOwnProperty(e.keyCode)?t[e.keyCode]:null},strList:function(e){return e?e.replace(/^\s+|\s+$/g,"").split(/\s+/):[]},hasClass:function(e,t){return!!t&&(void 0!==e.classList?e.classList.contains(t):-1!=(" "+e.className.replace(/\s+/g," ")+" ").indexOf(" "+t+" "))},addClass:function(e,t){var o=i.strList(t);if(void 0===e.classList)for(n=0;n<o.length;n+=1)i.hasClass(e,o[n])||(e.className+=(e.className?" ":"")+o[n]);else for(var n=0;n<o.length;n+=1)e.classList.add(o[n])},removeClass:function(e,t){var o=i.strList(t);if(void 0===e.classList)for(a=0;a<o.length;a+=1){var n=new RegExp("^\\s*"+o[a]+"\\s*|\\s*"+o[a]+"\\s*$|\\s+"+o[a]+"(\\s+)","g");e.className=e.className.replace(n,"$1")}else for(var a=0;a<o.length;a+=1)e.classList.remove(o[a])},getCompStyle:function(t){return(e.getComputedStyle?e.getComputedStyle(t):t.currentStyle)||{}},setStyle:function(e,t,o,n){var a=o?"important":"",r=null;for(var l in t)if(t.hasOwnProperty(l)){var s=null;null===t[l]?(r||(r=i.getData(e,"origStyle")),r&&r.hasOwnProperty(l)&&(s=r[l])):(n&&(r||(r=i.getData(e,"origStyle",{})),r.hasOwnProperty(l)||(r[l]=e.style[l])),s=t[l]),null!==s&&e.style.setProperty(l,s,a)}},hexColor:function(e,t,o){return"#"+(("0"+Math.round(e).toString(16)).substr(-2)+("0"+Math.round(t).toString(16)).substr(-2)+("0"+Math.round(o).toString(16)).substr(-2)).toUpperCase()},hexaColor:function(e,t,o,n){return"#"+(("0"+Math.round(e).toString(16)).substr(-2)+("0"+Math.round(t).toString(16)).substr(-2)+("0"+Math.round(o).toString(16)).substr(-2)+("0"+Math.round(255*n).toString(16)).substr(-2)).toUpperCase()},rgbColor:function(e,t,o){return"rgb("+Math.round(e)+","+Math.round(t)+","+Math.round(o)+")"},rgbaColor:function(e,t,o,n){return"rgba("+Math.round(e)+","+Math.round(t)+","+Math.round(o)+","+Math.round(100*(null==n?1:n))/100+")"},linearGradient:(o=function(){for(var t="linear-gradient",o=["","-webkit-","-moz-","-o-","-ms-"],n=e.document.createElement("div"),i=0;i<o.length;i+=1){var a=o[i]+t,r=a+"(to right, rgba(0,0,0,0), rgba(0,0,0,0))";if(n.style.background=r,n.style.background)return a}return t}(),function(){return o+"("+Array.prototype.join.call(arguments,", ")+")"}),setBorderRadius:function(e,t){i.setStyle(e,{"border-radius":t||"0"})},setBoxShadow:function(e,t){i.setStyle(e,{"box-shadow":t||"none"})},getElementPos:function(e,t){var o=0,n=0,a=e.getBoundingClientRect();if(o=a.left,n=a.top,!t){var r=i.getViewPos();o+=r[0],n+=r[1]}return[o,n]},getElementSize:function(e){return[e.offsetWidth,e.offsetHeight]},getAbsPointerPos:function(e){var t=0,o=0;return void 0!==e.changedTouches&&e.changedTouches.length?(t=e.changedTouches[0].clientX,o=e.changedTouches[0].clientY):"number"==typeof e.clientX&&(t=e.clientX,o=e.clientY),{x:t,y:o}},getRelPointerPos:function(e){var t=(e.target||e.srcElement).getBoundingClientRect(),o=0,n=0;return void 0!==e.changedTouches&&e.changedTouches.length?(o=e.changedTouches[0].clientX,n=e.changedTouches[0].clientY):"number"==typeof e.clientX&&(o=e.clientX,n=e.clientY),{x:o-t.left,y:n-t.top}},getViewPos:function(){var t=e.document.documentElement;return[(e.pageXOffset||t.scrollLeft)-(t.clientLeft||0),(e.pageYOffset||t.scrollTop)-(t.clientTop||0)]},getViewSize:function(){var t=e.document.documentElement;return[e.innerWidth||t.clientWidth,e.innerHeight||t.clientHeight]},RGB_HSV:function(e,t,o){e/=255,t/=255,o/=255;var n=Math.min(Math.min(e,t),o),i=Math.max(Math.max(e,t),o),a=i-n;if(0===a)return[null,0,100*i];var r=e===n?3+(o-t)/a:t===n?5+(e-o)/a:1+(t-e)/a;return[60*(6===r?0:r),a/i*100,100*i]},HSV_RGB:function(e,t,o){var n=o/100*255;if(null===e)return[n,n,n];e/=60,t/=100;var i=Math.floor(e),a=n*(1-t),r=n*(1-t*(i%2?e-i:1-(e-i)));switch(i){case 6:case 0:return[n,r,a];case 1:return[r,n,a];case 2:return[a,n,r];case 3:return[a,r,n];case 4:return[r,a,n];case 5:return[n,a,r]}},parseColorString:function(e){var t,o={rgba:null,format:null};if(t=e.match(/^\W*([0-9A-F]{3,8})\W*$/i)){if(8===t[1].length)o.format="hexa",o.rgba=[parseInt(t[1].substr(0,2),16),parseInt(t[1].substr(2,2),16),parseInt(t[1].substr(4,2),16),parseInt(t[1].substr(6,2),16)/255];else if(6===t[1].length)o.format="hex",o.rgba=[parseInt(t[1].substr(0,2),16),parseInt(t[1].substr(2,2),16),parseInt(t[1].substr(4,2),16),null];else{if(3!==t[1].length)return!1;o.format="hex",o.rgba=[parseInt(t[1].charAt(0)+t[1].charAt(0),16),parseInt(t[1].charAt(1)+t[1].charAt(1),16),parseInt(t[1].charAt(2)+t[1].charAt(2),16),null]}return o}if(t=e.match(/^\W*rgba?\(([^)]*)\)\W*$/i)){var n,i,a,r,l=t[1].split(","),s=/^\s*(\d+|\d*\.\d+|\d+\.\d*)\s*$/;if(l.length>=3&&(n=l[0].match(s))&&(i=l[1].match(s))&&(a=l[2].match(s)))return o.format="rgb",o.rgba=[parseFloat(n[1])||0,parseFloat(i[1])||0,parseFloat(a[1])||0,null],l.length>=4&&(r=l[3].match(s))&&(o.format="rgba",o.rgba[3]=parseFloat(r[1])||0),o}return!1},parsePaletteValue:function(e){var t=[];"string"==typeof e?e.replace(/#[0-9A-F]{3}([0-9A-F]{3})?|rgba?\(([^)]*)\)/gi,(function(e){t.push(e)})):Array.isArray(e)&&(t=e);for(var o=[],n=0;n<t.length;n++){var a=i.parseColorString(t[n]);a&&o.push(a)}return o},containsTranparentColor:function(e){for(var t=0;t<e.length;t++){var o=e[t].rgba[3];if(null!==o&&o<1)return!0}return!1},isAlphaFormat:function(e){switch(e.toLowerCase()){case"hexa":case"rgba":return!0}return!1},scaleCanvasForHighDPR:function(t){var o=e.devicePixelRatio||1;t.width*=o,t.height*=o,t.getContext("2d").scale(o,o)},genColorPreviewCanvas:function(e,t,o,n){var a=Math.round(i.pub.previewSeparator.length),r=i.pub.chessboardSize,l=i.pub.chessboardColor1,s=i.pub.chessboardColor2,c=o||2*r,d=2*r,p=i.createEl("canvas"),u=p.getContext("2d");p.width=c,p.height=d,n&&i.scaleCanvasForHighDPR(p),u.fillStyle=l,u.fillRect(0,0,c,d),u.fillStyle=s;for(var m=0;m<c;m+=2*r)u.fillRect(m,0,r,r),u.fillRect(m+r,r,r,r);e&&(u.fillStyle=e,u.fillRect(0,0,c,d));var h=null;switch(t){case"left":h=0,u.clearRect(0,0,a/2,d);break;case"right":h=c-a,u.clearRect(c-a/2,0,a/2,d)}if(null!==h){u.lineWidth=1;for(var g=0;g<i.pub.previewSeparator.length;g+=1)u.beginPath(),u.strokeStyle=i.pub.previewSeparator[g],u.moveTo(.5+h+g,0),u.lineTo(.5+h+g,d),u.stroke()}return{canvas:p,width:c,height:d}},genColorPreviewGradient:function(e,t,o){var n;return n=t&&o?["to "+{left:"right",right:"left"}[t],e+" 0%",e+" "+o+"px","rgba(0,0,0,0) "+(o+1)+"px","rgba(0,0,0,0) 100%"]:["to right",e+" 0%",e+" 100%"],i.linearGradient.apply(this,n)},redrawPosition:function(){if(i.picker&&i.picker.owner){var e,t,o=i.picker.owner;o.fixed?(e=i.getElementPos(o.targetElement,!0),t=[0,0]):(e=i.getElementPos(o.targetElement),t=i.getViewPos());var n,a,r,l=i.getElementSize(o.targetElement),s=i.getViewSize(),c=i.getPickerDims(o),d=[c.borderW,c.borderH];switch(o.position.toLowerCase()){case"left":n=1,a=0,r=-1;break;case"right":n=1,a=0,r=1;break;case"top":n=0,a=1,r=-1;break;default:n=0,a=1,r=1}var p=(l[a]+d[a])/2;if(o.smartPosition)u=[-t[n]+e[n]+d[n]>s[n]&&-t[n]+e[n]+l[n]/2>s[n]/2&&e[n]+l[n]-d[n]>=0?e[n]+l[n]-d[n]:e[n],-t[a]+e[a]+l[a]+d[a]-p+p*r>s[a]?-t[a]+e[a]+l[a]/2>s[a]/2&&e[a]+l[a]-p-p*r>=0?e[a]+l[a]-p-p*r:e[a]+l[a]-p+p*r:e[a]+l[a]-p+p*r>=0?e[a]+l[a]-p+p*r:e[a]+l[a]-p-p*r];else var u=[e[n],e[a]+l[a]-p+p*r];var m=u[n],h=u[a],g=o.fixed?"fixed":"absolute",f=(u[0]+d[0]>e[0]||u[0]<e[0]+l[0])&&u[1]+d[1]<e[1]+l[1];i._drawPosition(o,m,h,g,f)}},_drawPosition:function(e,t,o,n,a){var r=a?0:e.shadowBlur;i.picker.wrap.style.position=n,i.picker.wrap.style.left=t+"px",i.picker.wrap.style.top=o+"px",i.setBoxShadow(i.picker.boxS,e.shadow?new i.BoxShadow(0,r,e.shadowBlur,0,e.shadowColor):null)},getPickerDims:function(e){var t=2*e.controlBorderWidth+e.width,o=2*e.controlBorderWidth+e.height,n=2*e.controlBorderWidth+2*i.getControlPadding(e)+e.sliderSize;i.getSliderChannel(e)&&(t+=n),e.hasAlphaChannel()&&(t+=n);var a=i.getPaletteDims(e,t);a.height&&(o+=a.height+e.padding),e.closeButton&&(o+=2*e.controlBorderWidth+e.padding+e.buttonHeight);var r=t+2*e.padding,l=o+2*e.padding;return{contentW:t,contentH:o,paddedW:r,paddedH:l,borderW:r+2*e.borderWidth,borderH:l+2*e.borderWidth,palette:a}},getPaletteDims:function(e,t){var o=0,n=0,i=0,a=0,r=0,l=e._palette?e._palette.length:0;return l&&(n=(o=e.paletteCols)>0?Math.ceil(l/o):0,i=Math.max(1,Math.floor((t-(o-1)*e.paletteSpacing)/o)),a=e.paletteHeight?Math.min(e.paletteHeight,i):i),n&&(r=n*a+(n-1)*e.paletteSpacing),{cols:o,rows:n,cellW:i,cellH:a,width:t,height:r}},getControlPadding:function(e){return Math.max(e.padding/2,2*e.pointerBorderWidth+e.pointerThickness-e.controlBorderWidth)},getPadYChannel:function(e){switch(e.mode.charAt(1).toLowerCase()){case"v":return"v"}return"s"},getSliderChannel:function(e){if(e.mode.length>2)switch(e.mode.charAt(2).toLowerCase()){case"s":return"s";case"v":return"v"}return null},triggerCallback:function(e,t){if(e[t]){var o=null;if("string"==typeof e[t])try{o=new Function(e[t])}catch(e){console.error(e)}else o=e[t];o&&o.call(e)}},triggerGlobal:function(e){for(var t=i.getInstances(),o=0;o<t.length;o+=1)t[o].trigger(e)},_pointerMoveEvent:{mouse:"mousemove",touch:"touchmove"},_pointerEndEvent:{mouse:"mouseup",touch:"touchend"},_pointerOrigin:null,_capturedTarget:null,onDocumentKeyUp:function(e){-1!==["Tab","Escape"].indexOf(i.eventKey(e))&&i.picker&&i.picker.owner&&i.picker.owner.tryHide()},onWindowResize:function(e){i.redrawPosition()},onWindowScroll:function(e){i.redrawPosition()},onParentScroll:function(e){i.picker&&i.picker.owner&&i.picker.owner.tryHide()},onDocumentMouseDown:function(e){var t=e.target||e.srcElement;t.jscolor&&t.jscolor instanceof i.pub?t.jscolor.showOnClick&&!t.disabled&&t.jscolor.show():i.getData(t,"gui")?i.getData(t,"control")&&i.onControlPointerStart(e,t,i.getData(t,"control"),"mouse"):i.picker&&i.picker.owner&&i.picker.owner.tryHide()},onPickerTouchStart:function(e){var t=e.target||e.srcElement;i.getData(t,"control")&&i.onControlPointerStart(e,t,i.getData(t,"control"),"touch")},onControlPointerStart:function(t,o,n,a){var r=i.getData(o,"instance");i.preventDefault(t),i.captureTarget(o);var l=function(e,r){i.attachGroupEvent("drag",e,i._pointerMoveEvent[a],i.onDocumentPointerMove(t,o,n,a,r)),i.attachGroupEvent("drag",e,i._pointerEndEvent[a],i.onDocumentPointerEnd(t,o,n,a))};if(l(e.document,[0,0]),e.parent&&e.frameElement){var s=e.frameElement.getBoundingClientRect(),c=[-s.left,-s.top];l(e.parent.window.document,c)}var d=i.getAbsPointerPos(t),p=i.getRelPointerPos(t);switch(i._pointerOrigin={x:d.x-p.x,y:d.y-p.y},n){case"pad":"v"===i.getSliderChannel(r)&&0===r.channels.v&&r.fromHSVA(null,null,100,null),i.setPad(r,t,0,0);break;case"sld":i.setSld(r,t,0);break;case"asld":i.setASld(r,t,0)}r.trigger("input")},onDocumentPointerMove:function(e,t,o,n,a){return function(e){var n=i.getData(t,"instance");switch(o){case"pad":i.setPad(n,e,a[0],a[1]);break;case"sld":i.setSld(n,e,a[1]);break;case"asld":i.setASld(n,e,a[1])}n.trigger("input")}},onDocumentPointerEnd:function(e,t,o,n){return function(e){var o=i.getData(t,"instance");i.detachGroupEvents("drag"),i.releaseTarget(),o.trigger("input"),o.trigger("change")}},onPaletteSampleClick:function(e){var t=e.currentTarget,o=i.getData(t,"instance"),n=i.getData(t,"color");"any"===o.format.toLowerCase()&&(o._setFormat(n.format),i.isAlphaFormat(o.getFormat())||(n.rgba[3]=1)),null===n.rgba[3]&&(!0===o.paletteSetsAlpha||"auto"===o.paletteSetsAlpha&&o._paletteHasTransparency)&&(n.rgba[3]=1),o.fromRGBA.apply(o,n.rgba),o.trigger("input"),o.trigger("change"),o.hideOnPaletteClick&&o.hide()},setPad:function(e,t,o,n){var a=i.getAbsPointerPos(t),r=o+a.x-i._pointerOrigin.x-e.padding-e.controlBorderWidth,l=n+a.y-i._pointerOrigin.y-e.padding-e.controlBorderWidth,s=r*(360/(e.width-1)),c=100-l*(100/(e.height-1));switch(i.getPadYChannel(e)){case"s":e.fromHSVA(s,c,null,null);break;case"v":e.fromHSVA(s,null,c,null)}},setSld:function(e,t,o){var n=100-(o+i.getAbsPointerPos(t).y-i._pointerOrigin.y-e.padding-e.controlBorderWidth)*(100/(e.height-1));switch(i.getSliderChannel(e)){case"s":e.fromHSVA(null,n,null,null);break;case"v":e.fromHSVA(null,null,n,null)}},setASld:function(e,t,o){var n=1-(o+i.getAbsPointerPos(t).y-i._pointerOrigin.y-e.padding-e.controlBorderWidth)*(1/(e.height-1));if(n<1){var a=e.getFormat();"any"!==e.format.toLowerCase()||i.isAlphaFormat(a)||e._setFormat("hex"===a?"hexa":"rgba")}e.fromHSVA(null,null,null,n)},createPadCanvas:function(){var e={elm:null,draw:null},t=i.createEl("canvas"),o=t.getContext("2d");return e.elm=t,e.draw=function(e,n,i){t.width=e,t.height=n,o.clearRect(0,0,t.width,t.height);var a=o.createLinearGradient(0,0,t.width,0);a.addColorStop(0,"#F00"),a.addColorStop(1/6,"#FF0"),a.addColorStop(2/6,"#0F0"),a.addColorStop(.5,"#0FF"),a.addColorStop(4/6,"#00F"),a.addColorStop(5/6,"#F0F"),a.addColorStop(1,"#F00"),o.fillStyle=a,o.fillRect(0,0,t.width,t.height);var r=o.createLinearGradient(0,0,0,t.height);switch(i.toLowerCase()){case"s":r.addColorStop(0,"rgba(255,255,255,0)"),r.addColorStop(1,"rgba(255,255,255,1)");break;case"v":r.addColorStop(0,"rgba(0,0,0,0)"),r.addColorStop(1,"rgba(0,0,0,1)")}o.fillStyle=r,o.fillRect(0,0,t.width,t.height)},e},createSliderGradient:function(){var e={elm:null,draw:null},t=i.createEl("canvas"),o=t.getContext("2d");return e.elm=t,e.draw=function(e,n,i,a){t.width=e,t.height=n,o.clearRect(0,0,t.width,t.height);var r=o.createLinearGradient(0,0,0,t.height);r.addColorStop(0,i),r.addColorStop(1,a),o.fillStyle=r,o.fillRect(0,0,t.width,t.height)},e},createASliderGradient:function(){var e={elm:null,draw:null},t=i.createEl("canvas"),o=t.getContext("2d");return e.elm=t,e.draw=function(e,n,a){t.width=e,t.height=n,o.clearRect(0,0,t.width,t.height);var r=t.width/2,l=i.pub.chessboardColor1,s=i.pub.chessboardColor2;if(o.fillStyle=l,o.fillRect(0,0,t.width,t.height),r>0)for(var c=0;c<t.height;c+=2*r)o.fillStyle=s,o.fillRect(0,c,r,r),o.fillRect(r,c+r,r,r);var d=o.createLinearGradient(0,0,0,t.height);d.addColorStop(0,a),d.addColorStop(1,"rgba(0,0,0,0)"),o.fillStyle=d,o.fillRect(0,0,t.width,t.height)},e},BoxShadow:(t=function(e,t,o,n,i,a){this.hShadow=e,this.vShadow=t,this.blur=o,this.spread=n,this.color=i,this.inset=!!a},t.prototype.toString=function(){var e=[Math.round(this.hShadow)+"px",Math.round(this.vShadow)+"px",Math.round(this.blur)+"px",Math.round(this.spread)+"px",this.color];return this.inset&&e.push("inset"),e.join(" ")},t),flags:{leaveValue:1,leaveAlpha:2,leavePreview:4},enumOpts:{format:["auto","any","hex","hexa","rgb","rgba"],previewPosition:["left","right"],mode:["hsv","hvs","hs","hv"],position:["left","right","top","bottom"],alphaChannel:["auto",!0,!1],paletteSetsAlpha:["auto",!0,!1]},deprecatedOpts:{styleElement:"previewElement",onFineChange:"onInput",overwriteImportant:"forceStyle",closable:"closeButton",insetWidth:"controlBorderWidth",insetColor:"controlBorderColor",refine:null},docsRef:" See https://jscolor.com/docs/",pub:function(t,o){var n=this;function a(e,t){if("string"!=typeof e)throw new Error("Invalid value for option name: "+e);if(i.enumOpts.hasOwnProperty(e)&&("string"==typeof t&&(t=t.toLowerCase()),-1===i.enumOpts[e].indexOf(t)))throw new Error("Option '"+e+"' has invalid value: "+t);if(i.deprecatedOpts.hasOwnProperty(e)){var o=e,a=i.deprecatedOpts[e];if(!a)throw new Error("Option '"+e+"' is DEPRECATED");console.warn("Option '%s' is DEPRECATED, using '%s' instead."+i.docsRef,o,a),e=a}var r="set__"+e;if("function"==typeof n[r])return n[r](t),!0;if(e in n)return n[e]=t,!0;throw new Error("Unrecognized configuration option: "+e)}function r(e){if("string"!=typeof e)throw new Error("Invalid value for option name: "+e);if(i.deprecatedOpts.hasOwnProperty(e)){var t=e,o=i.deprecatedOpts[e];if(!o)throw new Error("Option '"+e+"' is DEPRECATED");console.warn("Option '%s' is DEPRECATED, using '%s' instead."+i.docsRef,t,o),e=o}var a="get__"+e;if("function"==typeof n[a])return n[a](value);if(e in n)return n[e];throw new Error("Unrecognized configuration option: "+e)}function l(){n._processParentElementsInDOM(),i.picker||(i.picker={owner:null,wrap:i.createEl("div"),box:i.createEl("div"),boxS:i.createEl("div"),boxB:i.createEl("div"),pad:i.createEl("div"),padB:i.createEl("div"),padM:i.createEl("div"),padCanvas:i.createPadCanvas(),cross:i.createEl("div"),crossBY:i.createEl("div"),crossBX:i.createEl("div"),crossLY:i.createEl("div"),crossLX:i.createEl("div"),sld:i.createEl("div"),sldB:i.createEl("div"),sldM:i.createEl("div"),sldGrad:i.createSliderGradient(),sldPtrS:i.createEl("div"),sldPtrIB:i.createEl("div"),sldPtrMB:i.createEl("div"),sldPtrOB:i.createEl("div"),asld:i.createEl("div"),asldB:i.createEl("div"),asldM:i.createEl("div"),asldGrad:i.createASliderGradient(),asldPtrS:i.createEl("div"),asldPtrIB:i.createEl("div"),asldPtrMB:i.createEl("div"),asldPtrOB:i.createEl("div"),pal:i.createEl("div"),btn:i.createEl("div"),btnT:i.createEl("span")},i.picker.pad.appendChild(i.picker.padCanvas.elm),i.picker.padB.appendChild(i.picker.pad),i.picker.cross.appendChild(i.picker.crossBY),i.picker.cross.appendChild(i.picker.crossBX),i.picker.cross.appendChild(i.picker.crossLY),i.picker.cross.appendChild(i.picker.crossLX),i.picker.padB.appendChild(i.picker.cross),i.picker.box.appendChild(i.picker.padB),i.picker.box.appendChild(i.picker.padM),i.picker.sld.appendChild(i.picker.sldGrad.elm),i.picker.sldB.appendChild(i.picker.sld),i.picker.sldB.appendChild(i.picker.sldPtrOB),i.picker.sldPtrOB.appendChild(i.picker.sldPtrMB),i.picker.sldPtrMB.appendChild(i.picker.sldPtrIB),i.picker.sldPtrIB.appendChild(i.picker.sldPtrS),i.picker.box.appendChild(i.picker.sldB),i.picker.box.appendChild(i.picker.sldM),i.picker.asld.appendChild(i.picker.asldGrad.elm),i.picker.asldB.appendChild(i.picker.asld),i.picker.asldB.appendChild(i.picker.asldPtrOB),i.picker.asldPtrOB.appendChild(i.picker.asldPtrMB),i.picker.asldPtrMB.appendChild(i.picker.asldPtrIB),i.picker.asldPtrIB.appendChild(i.picker.asldPtrS),i.picker.box.appendChild(i.picker.asldB),i.picker.box.appendChild(i.picker.asldM),i.picker.box.appendChild(i.picker.pal),i.picker.btn.appendChild(i.picker.btnT),i.picker.box.appendChild(i.picker.btn),i.picker.boxB.appendChild(i.picker.box),i.picker.wrap.appendChild(i.picker.boxS),i.picker.wrap.appendChild(i.picker.boxB),i.picker.wrap.addEventListener("touchstart",i.onPickerTouchStart,!!i.isPassiveEventSupported&&{passive:!1}));var t,o,a=i.picker,r=!!i.getSliderChannel(n),l=n.hasAlphaChannel(),p=i.getPickerDims(n),u=2*n.pointerBorderWidth+n.pointerThickness+2*n.crossSize,m=i.getControlPadding(n),h=Math.min(n.borderRadius,Math.round(n.padding*Math.PI));a.wrap.className="jscolor-picker-wrap",a.wrap.style.clear="both",a.wrap.style.width=p.borderW+"px",a.wrap.style.height=p.borderH+"px",a.wrap.style.zIndex=n.zIndex,a.box.className="jscolor-picker",a.box.style.width=p.paddedW+"px",a.box.style.height=p.paddedH+"px",a.box.style.position="relative",a.boxS.className="jscolor-picker-shadow",a.boxS.style.position="absolute",a.boxS.style.left="0",a.boxS.style.top="0",a.boxS.style.width="100%",a.boxS.style.height="100%",i.setBorderRadius(a.boxS,h+"px"),a.boxB.className="jscolor-picker-border",a.boxB.style.position="relative",a.boxB.style.border=n.borderWidth+"px solid",a.boxB.style.borderColor=n.borderColor,a.boxB.style.background=n.backgroundColor,i.setBorderRadius(a.boxB,h+"px"),a.padM.style.background="rgba(255,0,0,.2)",a.sldM.style.background="rgba(0,255,0,.2)",a.asldM.style.background="rgba(0,0,255,.2)",a.padM.style.opacity=a.sldM.style.opacity=a.asldM.style.opacity="0",a.pad.style.position="relative",a.pad.style.width=n.width+"px",a.pad.style.height=n.height+"px",a.padCanvas.draw(n.width,n.height,i.getPadYChannel(n)),a.padB.style.position="absolute",a.padB.style.left=n.padding+"px",a.padB.style.top=n.padding+"px",a.padB.style.border=n.controlBorderWidth+"px solid",a.padB.style.borderColor=n.controlBorderColor,a.padM.style.position="absolute",a.padM.style.left="0px",a.padM.style.top="0px",a.padM.style.width=n.padding+2*n.controlBorderWidth+n.width+m+"px",a.padM.style.height=2*n.controlBorderWidth+2*n.padding+n.height+"px",a.padM.style.cursor="crosshair",i.setData(a.padM,{instance:n,control:"pad"}),a.cross.style.position="absolute",a.cross.style.left=a.cross.style.top="0",a.cross.style.width=a.cross.style.height=u+"px",a.crossBY.style.position=a.crossBX.style.position="absolute",a.crossBY.style.background=a.crossBX.style.background=n.pointerBorderColor,a.crossBY.style.width=a.crossBX.style.height=2*n.pointerBorderWidth+n.pointerThickness+"px",a.crossBY.style.height=a.crossBX.style.width=u+"px",a.crossBY.style.left=a.crossBX.style.top=Math.floor(u/2)-Math.floor(n.pointerThickness/2)-n.pointerBorderWidth+"px",a.crossBY.style.top=a.crossBX.style.left="0",a.crossLY.style.position=a.crossLX.style.position="absolute",a.crossLY.style.background=a.crossLX.style.background=n.pointerColor,a.crossLY.style.height=a.crossLX.style.width=u-2*n.pointerBorderWidth+"px",a.crossLY.style.width=a.crossLX.style.height=n.pointerThickness+"px",a.crossLY.style.left=a.crossLX.style.top=Math.floor(u/2)-Math.floor(n.pointerThickness/2)+"px",a.crossLY.style.top=a.crossLX.style.left=n.pointerBorderWidth+"px",a.sld.style.overflow="hidden",a.sld.style.width=n.sliderSize+"px",a.sld.style.height=n.height+"px",a.sldGrad.draw(n.sliderSize,n.height,"#000","#000"),a.sldB.style.display=r?"block":"none",a.sldB.style.position="absolute",a.sldB.style.left=n.padding+n.width+2*n.controlBorderWidth+2*m+"px",a.sldB.style.top=n.padding+"px",a.sldB.style.border=n.controlBorderWidth+"px solid",a.sldB.style.borderColor=n.controlBorderColor,a.sldM.style.display=r?"block":"none",a.sldM.style.position="absolute",a.sldM.style.left=n.padding+n.width+2*n.controlBorderWidth+m+"px",a.sldM.style.top="0px",a.sldM.style.width=n.sliderSize+2*m+2*n.controlBorderWidth+(l?0:Math.max(0,n.padding-m))+"px",a.sldM.style.height=2*n.controlBorderWidth+2*n.padding+n.height+"px",a.sldM.style.cursor="default",i.setData(a.sldM,{instance:n,control:"sld"}),a.sldPtrIB.style.border=a.sldPtrOB.style.border=n.pointerBorderWidth+"px solid "+n.pointerBorderColor,a.sldPtrOB.style.position="absolute",a.sldPtrOB.style.left=-(2*n.pointerBorderWidth+n.pointerThickness)+"px",a.sldPtrOB.style.top="0",a.sldPtrMB.style.border=n.pointerThickness+"px solid "+n.pointerColor,a.sldPtrS.style.width=n.sliderSize+"px",a.sldPtrS.style.height=i.pub.sliderInnerSpace+"px",a.asld.style.overflow="hidden",a.asld.style.width=n.sliderSize+"px",a.asld.style.height=n.height+"px",a.asldGrad.draw(n.sliderSize,n.height,"#000"),a.asldB.style.display=l?"block":"none",a.asldB.style.position="absolute",a.asldB.style.left=n.padding+n.width+2*n.controlBorderWidth+m+(r?n.sliderSize+3*m+2*n.controlBorderWidth:0)+"px",a.asldB.style.top=n.padding+"px",a.asldB.style.border=n.controlBorderWidth+"px solid",a.asldB.style.borderColor=n.controlBorderColor,a.asldM.style.display=l?"block":"none",a.asldM.style.position="absolute",a.asldM.style.left=n.padding+n.width+2*n.controlBorderWidth+m+(r?n.sliderSize+2*m+2*n.controlBorderWidth:0)+"px",a.asldM.style.top="0px",a.asldM.style.width=n.sliderSize+2*m+2*n.controlBorderWidth+Math.max(0,n.padding-m)+"px",a.asldM.style.height=2*n.controlBorderWidth+2*n.padding+n.height+"px",a.asldM.style.cursor="default",i.setData(a.asldM,{instance:n,control:"asld"}),a.asldPtrIB.style.border=a.asldPtrOB.style.border=n.pointerBorderWidth+"px solid "+n.pointerBorderColor,a.asldPtrOB.style.position="absolute",a.asldPtrOB.style.left=-(2*n.pointerBorderWidth+n.pointerThickness)+"px",a.asldPtrOB.style.top="0",a.asldPtrMB.style.border=n.pointerThickness+"px solid "+n.pointerColor,a.asldPtrS.style.width=n.sliderSize+"px",a.asldPtrS.style.height=i.pub.sliderInnerSpace+"px",a.pal.className="jscolor-palette",a.pal.style.display=p.palette.rows?"block":"none",a.pal.style.position="absolute",a.pal.style.left=n.padding+"px",a.pal.style.top=2*n.controlBorderWidth+2*n.padding+n.height+"px",a.pal.innerHTML="";for(var g=i.genColorPreviewCanvas("rgba(0,0,0,0)"),f=0,C=0;C<p.palette.rows;C++)for(var y=0;y<p.palette.cols&&f<n._palette.length;y++,f++){var b=n._palette[f],k=i.rgbaColor.apply(null,b.rgba),G=i.createEl("div");G.style.width=p.palette.cellW-2*n.controlBorderWidth+"px",G.style.height=p.palette.cellH-2*n.controlBorderWidth+"px",G.style.backgroundColor=k;var w=i.createEl("div");w.className="jscolor-palette-sample",w.style.display="block",w.style.position="absolute",w.style.left=(p.palette.cols<=1?0:Math.round(y*((p.contentW-p.palette.cellW)/(p.palette.cols-1))*10)/10)+"px",w.style.top=C*(p.palette.cellH+n.paletteSpacing)+"px",w.style.border=n.controlBorderWidth+"px solid",w.style.borderColor=n.controlBorderColor,w.style.cursor="pointer",null!==b.rgba[3]&&b.rgba[3]<1&&(w.style.backgroundImage="url('"+g.canvas.toDataURL()+"')",w.style.backgroundRepeat="repeat",w.style.backgroundPosition="center center"),i.setData(w,{instance:n,control:"palette-sample",color:b}),w.addEventListener("click",i.onPaletteSampleClick,!1),w.appendChild(G),a.pal.appendChild(w)}a.btn.className="jscolor-btn-close",a.btn.style.display=n.closeButton?"block":"none",a.btn.style.position="absolute",a.btn.style.left=n.padding+"px",a.btn.style.bottom=n.padding+"px",a.btn.style.padding="0 15px",a.btn.style.maxWidth=p.contentW-2*n.controlBorderWidth-30+"px",a.btn.style.overflow="hidden",a.btn.style.height=n.buttonHeight+"px",a.btn.style.whiteSpace="nowrap",a.btn.style.border=n.controlBorderWidth+"px solid",o=(t=n.controlBorderColor.split(/\s+/)).length<2?t[0]:t[1]+" "+t[0]+" "+t[0]+" "+t[1],a.btn.style.borderColor=o,a.btn.style.color=n.buttonColor,a.btn.style.font="12px sans-serif",a.btn.style.textAlign="center",a.btn.style.cursor="pointer",a.btn.onmousedown=function(){n.hide()},a.btnT.style.lineHeight=n.buttonHeight+"px",a.btnT.innerHTML="",a.btnT.appendChild(e.document.createTextNode(n.closeText)),s(),c(),d(),i.picker.owner&&i.picker.owner!==n&&i.removeClass(i.picker.owner.targetElement,i.pub.activeClassName),i.picker.owner=n,n.container===e.document.body?i.redrawPosition():i._drawPosition(n,0,0,"relative",!1),a.wrap.parentNode!==n.container&&n.container.appendChild(a.wrap),i.addClass(n.targetElement,i.pub.activeClassName)}function s(){var e=i.getPadYChannel(n),t=Math.round(n.channels.h/360*(n.width-1)),o=Math.round((1-n.channels[e]/100)*(n.height-1)),a=2*n.pointerBorderWidth+n.pointerThickness+2*n.crossSize,r=-Math.floor(a/2);switch(i.picker.cross.style.left=t+r+"px",i.picker.cross.style.top=o+r+"px",i.getSliderChannel(n)){case"s":var l=i.HSV_RGB(n.channels.h,100,n.channels.v),s=i.HSV_RGB(n.channels.h,0,n.channels.v),c="rgb("+Math.round(l[0])+","+Math.round(l[1])+","+Math.round(l[2])+")",d="rgb("+Math.round(s[0])+","+Math.round(s[1])+","+Math.round(s[2])+")";i.picker.sldGrad.draw(n.sliderSize,n.height,c,d);break;case"v":var p=i.HSV_RGB(n.channels.h,n.channels.s,100);c="rgb("+Math.round(p[0])+","+Math.round(p[1])+","+Math.round(p[2])+")",d="#000",i.picker.sldGrad.draw(n.sliderSize,n.height,c,d)}i.picker.asldGrad.draw(n.sliderSize,n.height,n.toHEXString())}function c(){var e=i.getSliderChannel(n);if(e){var t=Math.round((1-n.channels[e]/100)*(n.height-1));i.picker.sldPtrOB.style.top=t-(2*n.pointerBorderWidth+n.pointerThickness)-Math.floor(i.pub.sliderInnerSpace/2)+"px"}i.picker.asldGrad.draw(n.sliderSize,n.height,n.toHEXString())}function d(){var e=Math.round((1-n.channels.a)*(n.height-1));i.picker.asldPtrOB.style.top=e-(2*n.pointerBorderWidth+n.pointerThickness)-Math.floor(i.pub.sliderInnerSpace/2)+"px"}function p(){return i.picker&&i.picker.owner===n}if(o||(o={}),this.channels={r:255,g:255,b:255,h:0,s:0,v:100,a:1},this.format="auto",this.value=void 0,this.alpha=void 0,this.onChange=void 0,this.onInput=void 0,this.valueElement=void 0,this.alphaElement=void 0,this.previewElement=void 0,this.previewPosition="left",this.previewSize=32,this.previewPadding=8,this.required=!0,this.hash=!0,this.uppercase=!0,this.forceStyle=!0,this.width=181,this.height=101,this.mode="HSV",this.alphaChannel="auto",this.position="bottom",this.smartPosition=!0,this.showOnClick=!0,this.hideOnLeave=!0,this.palette=[],this.paletteCols=10,this.paletteSetsAlpha="auto",this.paletteHeight=16,this.paletteSpacing=4,this.hideOnPaletteClick=!1,this.sliderSize=16,this.crossSize=8,this.closeButton=!1,this.closeText="Close",this.buttonColor="rgba(0,0,0,1)",this.buttonHeight=18,this.padding=12,this.backgroundColor="rgba(255,255,255,1)",this.borderWidth=1,this.borderColor="rgba(187,187,187,1)",this.borderRadius=8,this.controlBorderWidth=1,this.controlBorderColor="rgba(187,187,187,1)",this.shadow=!0,this.shadowBlur=15,this.shadowColor="rgba(0,0,0,0.2)",this.pointerColor="rgba(76,76,76,1)",this.pointerBorderWidth=1,this.pointerBorderColor="rgba(255,255,255,1)",this.pointerThickness=2,this.zIndex=5e3,this.container=void 0,this.minS=0,this.maxS=100,this.minV=0,this.maxV=100,this.minA=0,this.maxA=1,this.option=function(){if(!arguments.length)throw new Error("No option specified");if(1===arguments.length&&"string"==typeof arguments[0]){try{return r(arguments[0])}catch(e){console.warn(e)}return!1}if(arguments.length>=2&&"string"==typeof arguments[0]){try{if(!a(arguments[0],arguments[1]))return!1}catch(e){return console.warn(e),!1}return this.redraw(),this.exposeColor(),!0}if(1===arguments.length&&"object"==typeof arguments[0]){var e=arguments[0],t=!0;for(var o in e)if(e.hasOwnProperty(o))try{a(o,e[o])||(t=!1)}catch(e){console.warn(e),t=!1}return this.redraw(),this.exposeColor(),t}throw new Error("Invalid arguments")},this.channel=function(e,t){if("string"!=typeof e)throw new Error("Invalid value for channel name: "+e);if(void 0===t)return this.channels.hasOwnProperty(e.toLowerCase())?this.channels[e.toLowerCase()]:(console.warn("Getting unknown channel: "+e),!1);var o=!1;switch(e.toLowerCase()){case"r":o=this.fromRGBA(t,null,null,null);break;case"g":o=this.fromRGBA(null,t,null,null);break;case"b":o=this.fromRGBA(null,null,t,null);break;case"h":o=this.fromHSVA(t,null,null,null);break;case"s":o=this.fromHSVA(null,t,null,null);break;case"v":o=this.fromHSVA(null,null,t,null);break;case"a":o=this.fromHSVA(null,null,null,t);break;default:return console.warn("Setting unknown channel: "+e),!1}return!!o&&(this.redraw(),!0)},this.trigger=function(e){for(var t=i.strList(e),o=0;o<t.length;o+=1){var n=t[o].toLowerCase(),a=null;switch(n){case"input":a="onInput";break;case"change":a="onChange"}a&&i.triggerCallback(this,a),i.triggerInputEvent(this.valueElement,n,!0,!0)}},this.fromHSVA=function(e,t,o,n,a){if(void 0===e&&(e=null),void 0===t&&(t=null),void 0===o&&(o=null),void 0===n&&(n=null),null!==e){if(isNaN(e))return!1;this.channels.h=Math.max(0,Math.min(360,e))}if(null!==t){if(isNaN(t))return!1;this.channels.s=Math.max(0,Math.min(100,this.maxS,t),this.minS)}if(null!==o){if(isNaN(o))return!1;this.channels.v=Math.max(0,Math.min(100,this.maxV,o),this.minV)}if(null!==n){if(isNaN(n))return!1;this.channels.a=this.hasAlphaChannel()?Math.max(0,Math.min(1,this.maxA,n),this.minA):1}var r=i.HSV_RGB(this.channels.h,this.channels.s,this.channels.v);return this.channels.r=r[0],this.channels.g=r[1],this.channels.b=r[2],this.exposeColor(a),!0},this.fromRGBA=function(e,t,o,n,a){if(void 0===e&&(e=null),void 0===t&&(t=null),void 0===o&&(o=null),void 0===n&&(n=null),null!==e){if(isNaN(e))return!1;e=Math.max(0,Math.min(255,e))}if(null!==t){if(isNaN(t))return!1;t=Math.max(0,Math.min(255,t))}if(null!==o){if(isNaN(o))return!1;o=Math.max(0,Math.min(255,o))}if(null!==n){if(isNaN(n))return!1;this.channels.a=this.hasAlphaChannel()?Math.max(0,Math.min(1,this.maxA,n),this.minA):1}var r=i.RGB_HSV(null===e?this.channels.r:e,null===t?this.channels.g:t,null===o?this.channels.b:o);null!==r[0]&&(this.channels.h=Math.max(0,Math.min(360,r[0]))),0!==r[2]&&(this.channels.s=Math.max(0,this.minS,Math.min(100,this.maxS,r[1]))),this.channels.v=Math.max(0,this.minV,Math.min(100,this.maxV,r[2]));var l=i.HSV_RGB(this.channels.h,this.channels.s,this.channels.v);return this.channels.r=l[0],this.channels.g=l[1],this.channels.b=l[2],this.exposeColor(a),!0},this.fromHSV=function(e,t,o,n){return console.warn("fromHSV() method is DEPRECATED. Using fromHSVA() instead."+i.docsRef),this.fromHSVA(e,t,o,null,n)},this.fromRGB=function(e,t,o,n){return console.warn("fromRGB() method is DEPRECATED. Using fromRGBA() instead."+i.docsRef),this.fromRGBA(e,t,o,null,n)},this.fromString=function(e,t){if(!this.required&&""===e.trim())return this.setPreviewElementBg(null),this.setValueElementValue(""),!0;var o=i.parseColorString(e);return!!o&&("any"===this.format.toLowerCase()&&(this._setFormat(o.format),i.isAlphaFormat(this.getFormat())||(o.rgba[3]=1)),this.fromRGBA(o.rgba[0],o.rgba[1],o.rgba[2],o.rgba[3],t),!0)},this.toString=function(e){switch(void 0===e&&(e=this.getFormat()),e.toLowerCase()){case"hex":return this.toHEXString();case"hexa":return this.toHEXAString();case"rgb":return this.toRGBString();case"rgba":return this.toRGBAString()}return!1},this.toHEXString=function(){return i.hexColor(this.channels.r,this.channels.g,this.channels.b)},this.toHEXAString=function(){return i.hexaColor(this.channels.r,this.channels.g,this.channels.b,this.channels.a)},this.toRGBString=function(){return i.rgbColor(this.channels.r,this.channels.g,this.channels.b)},this.toRGBAString=function(){return i.rgbaColor(this.channels.r,this.channels.g,this.channels.b,this.channels.a)},this.toGrayscale=function(){return.213*this.channels.r+.715*this.channels.g+.072*this.channels.b},this.toCanvas=function(){return i.genColorPreviewCanvas(this.toRGBAString()).canvas},this.toDataURL=function(){return this.toCanvas().toDataURL()},this.toBackground=function(){return i.pub.background(this.toRGBAString())},this.isLight=function(){return this.toGrayscale()>127.5},this.hide=function(){p()&&(i.removeClass(n.targetElement,i.pub.activeClassName),i.picker.wrap.parentNode.removeChild(i.picker.wrap),delete i.picker.owner)},this.show=function(){l()},this.redraw=function(){p()&&l()},this.getFormat=function(){return this._currentFormat},this._setFormat=function(e){this._currentFormat=e.toLowerCase()},this.hasAlphaChannel=function(){return"auto"===this.alphaChannel?"any"===this.format.toLowerCase()||i.isAlphaFormat(this.getFormat())||void 0!==this.alpha||void 0!==this.alphaElement:this.alphaChannel},this.processValueInput=function(e){this.fromString(e)||this.exposeColor()},this.processAlphaInput=function(e){this.fromHSVA(null,null,null,parseFloat(e))||this.exposeColor()},this.exposeColor=function(e){var t=this.toString(),o=this.getFormat();if(i.setDataAttr(this.targetElement,"current-color",t),e&i.flags.leaveValue||!this.valueElement||("hex"!==o&&"hexa"!==o||(this.uppercase||(t=t.toLowerCase()),this.hash||(t=t.replace(/^#/,""))),this.setValueElementValue(t)),!(e&i.flags.leaveAlpha)&&this.alphaElement){var n=Math.round(100*this.channels.a)/100;this.setAlphaElementValue(n)}e&i.flags.leavePreview||!this.previewElement||((i.isTextInput(this.previewElement)||i.isButton(this.previewElement)&&!i.isButtonEmpty(this.previewElement))&&this.previewPosition,this.setPreviewElementBg(this.toRGBAString())),p()&&(s(),c(),d())},this.setPreviewElementBg=function(e){if(this.previewElement){var t=null,o=null;(i.isTextInput(this.previewElement)||i.isButton(this.previewElement)&&!i.isButtonEmpty(this.previewElement))&&(t=this.previewPosition,o=this.previewSize);var n=[];if(e){n.push({image:i.genColorPreviewGradient(e,t,o?o-i.pub.previewSeparator.length:null),position:"left top",size:"auto",repeat:t?"repeat-y":"repeat",origin:"padding-box"});var a=i.genColorPreviewCanvas("rgba(0,0,0,0)",t?{left:"right",right:"left"}[t]:null,o,!0);n.push({image:"url('"+a.canvas.toDataURL()+"')",position:(t||"left")+" top",size:a.width+"px "+a.height+"px",repeat:t?"repeat-y":"repeat",origin:"padding-box"})}else n.push({image:"none",position:"left top",size:"auto",repeat:"no-repeat",origin:"padding-box"});for(var r={image:[],position:[],size:[],repeat:[],origin:[]},l=0;l<n.length;l+=1)r.image.push(n[l].image),r.position.push(n[l].position),r.size.push(n[l].size),r.repeat.push(n[l].repeat),r.origin.push(n[l].origin);var s={"background-image":r.image.join(", "),"background-position":r.position.join(", "),"background-size":r.size.join(", "),"background-repeat":r.repeat.join(", "),"background-origin":r.origin.join(", ")};i.setStyle(this.previewElement,s,this.forceStyle);var c={left:null,right:null};t&&(c[t]=this.previewSize+this.previewPadding+"px"),s={"padding-left":c.left,"padding-right":c.right},i.setStyle(this.previewElement,s,this.forceStyle,!0)}},this.setValueElementValue=function(e){this.valueElement&&("input"===i.nodeName(this.valueElement)?this.valueElement.value=e:this.valueElement.innerHTML=e)},this.setAlphaElementValue=function(e){this.alphaElement&&("input"===i.nodeName(this.alphaElement)?this.alphaElement.value=e:this.alphaElement.innerHTML=e)},this._processParentElementsInDOM=function(){if(!this._parentElementsProcessed){this._parentElementsProcessed=!0;var e=this.targetElement;do{var t=i.getCompStyle(e);t.position&&"fixed"===t.position.toLowerCase()&&(this.fixed=!0),e!==this.targetElement&&(i.getData(e,"hasScrollListener")||(e.addEventListener("scroll",i.onParentScroll,!1),i.setData(e,"hasScrollListener",!0)))}while((e=e.parentNode)&&"body"!==i.nodeName(e))}},this.tryHide=function(){this.hideOnLeave&&this.hide()},this.set__palette=function(e){this.palette=e,this._palette=i.parsePaletteValue(e),this._paletteHasTransparency=i.containsTranparentColor(this._palette)},i.pub.options)for(var u in i.pub.options)if(i.pub.options.hasOwnProperty(u))try{a(u,i.pub.options[u])}catch(e){console.warn(e)}var m=[];o.preset&&("string"==typeof o.preset?m=o.preset.split(/\s+/):Array.isArray(o.preset)?m=o.preset.slice():console.warn("Unrecognized preset value")),-1===m.indexOf("default")&&m.push("default");for(var h=m.length-1;h>=0;h-=1){var g=m[h];if(g)if(i.pub.presets.hasOwnProperty(g)){for(var u in i.pub.presets[g])if(i.pub.presets[g].hasOwnProperty(u))try{a(u,i.pub.presets[g][u])}catch(e){console.warn(e)}}else console.warn("Unknown preset: %s",g)}var f=["preset"];for(var u in o)if(o.hasOwnProperty(u)&&-1===f.indexOf(u))try{a(u,o[u])}catch(e){console.warn(e)}if(void 0===this.container?this.container=e.document.body:this.container=i.node(this.container),!this.container)throw new Error("Cannot instantiate color picker without a container element");if(this.targetElement=i.node(t),!this.targetElement){if("string"==typeof t&&/^[a-zA-Z][\w:.-]*$/.test(t))throw new Error("If '"+t+"' is supposed to be an ID, please use '#"+t+"' or any valid CSS selector.");throw new Error("Cannot instantiate color picker without a target element")}if(this.targetElement.jscolor&&this.targetElement.jscolor instanceof i.pub)throw new Error("Color picker already installed on this element");if(this.targetElement.jscolor=this,i.addClass(this.targetElement,i.pub.className),i.instances.push(this),i.isButton(this.targetElement)&&("button"!==this.targetElement.type.toLowerCase()&&(this.targetElement.type="button"),i.isButtonEmpty(this.targetElement))){i.removeChildren(this.targetElement),this.targetElement.appendChild(e.document.createTextNode(" "));var C=i.getCompStyle(this.targetElement);(parseFloat(C["min-width"])||0)<this.previewSize&&i.setStyle(this.targetElement,{"min-width":this.previewSize+"px"},this.forceStyle)}if(void 0===this.valueElement?i.isTextInput(this.targetElement)&&(this.valueElement=this.targetElement):null===this.valueElement||(this.valueElement=i.node(this.valueElement)),this.alphaElement&&(this.alphaElement=i.node(this.alphaElement)),void 0===this.previewElement?this.previewElement=this.targetElement:null===this.previewElement||(this.previewElement=i.node(this.previewElement)),this.valueElement&&i.isTextInput(this.valueElement)){var y={onInput:this.valueElement.oninput};this.valueElement.oninput=null,this.valueElement.addEventListener("keydown",(function(e){"Enter"===i.eventKey(e)&&(n.valueElement&&n.processValueInput(n.valueElement.value),n.tryHide())}),!1),this.valueElement.addEventListener("change",(function(e){if(!i.getData(e,"internal")){var t=n.valueElement.value;n.processValueInput(n.valueElement.value),i.triggerCallback(n,"onChange"),n.valueElement.value!==t&&i.triggerInputEvent(n.valueElement,"change",!0,!0)}}),!1),this.valueElement.addEventListener("input",(function(e){i.getData(e,"internal")||(n.valueElement&&n.fromString(n.valueElement.value,i.flags.leaveValue),i.triggerCallback(n,"onInput"))}),!1),y.onInput&&this.valueElement.addEventListener("input",y.onInput,!1),this.valueElement.setAttribute("autocomplete","off"),this.valueElement.setAttribute("autocorrect","off"),this.valueElement.setAttribute("autocapitalize","off"),this.valueElement.setAttribute("spellcheck",!1)}this.alphaElement&&i.isTextInput(this.alphaElement)&&(this.alphaElement.addEventListener("keydown",(function(e){"Enter"===i.eventKey(e)&&(n.alphaElement&&n.processAlphaInput(n.alphaElement.value),n.tryHide())}),!1),this.alphaElement.addEventListener("change",(function(e){if(!i.getData(e,"internal")){var t=n.alphaElement.value;n.processAlphaInput(n.alphaElement.value),i.triggerCallback(n,"onChange"),i.triggerInputEvent(n.valueElement,"change",!0,!0),n.alphaElement.value!==t&&i.triggerInputEvent(n.alphaElement,"change",!0,!0)}}),!1),this.alphaElement.addEventListener("input",(function(e){i.getData(e,"internal")||(n.alphaElement&&n.fromHSVA(null,null,null,parseFloat(n.alphaElement.value),i.flags.leaveAlpha),i.triggerCallback(n,"onInput"),i.triggerInputEvent(n.valueElement,"input",!0,!0))}),!1),this.alphaElement.setAttribute("autocomplete","off"),this.alphaElement.setAttribute("autocorrect","off"),this.alphaElement.setAttribute("autocapitalize","off"),this.alphaElement.setAttribute("spellcheck",!1));var b="FFFFFF";void 0!==this.value?b=this.value:this.valueElement&&void 0!==this.valueElement.value&&(b=this.valueElement.value);var k=void 0;if(void 0!==this.alpha?k=""+this.alpha:this.alphaElement&&void 0!==this.alphaElement.value&&(k=this.alphaElement.value),this._currentFormat=null,["auto","any"].indexOf(this.format.toLowerCase())>-1){var G=i.parseColorString(b);this._currentFormat=G?G.format:"hex"}else this._currentFormat=this.format.toLowerCase();this.processValueInput(b),void 0!==k&&this.processAlphaInput(k)}}).pub.className="jscolor",i.pub.activeClassName="jscolor-active",i.pub.looseJSON=!0,i.pub.presets={},i.pub.presets.default={},i.pub.presets.light={backgroundColor:"rgba(255,255,255,1)",controlBorderColor:"rgba(187,187,187,1)",buttonColor:"rgba(0,0,0,1)"},i.pub.presets.dark={backgroundColor:"rgba(51,51,51,1)",controlBorderColor:"rgba(153,153,153,1)",buttonColor:"rgba(240,240,240,1)"},i.pub.presets.small={width:101,height:101,padding:10,sliderSize:14,paletteCols:8},i.pub.presets.medium={width:181,height:101,padding:12,sliderSize:16,paletteCols:10},i.pub.presets.large={width:271,height:151,padding:12,sliderSize:24,paletteCols:15},i.pub.presets.thin={borderWidth:1,controlBorderWidth:1,pointerBorderWidth:1},i.pub.presets.thick={borderWidth:2,controlBorderWidth:2,pointerBorderWidth:2},i.pub.sliderInnerSpace=3,i.pub.chessboardSize=8,i.pub.chessboardColor1="#666666",i.pub.chessboardColor2="#999999",i.pub.previewSeparator=["rgba(255,255,255,.65)","rgba(128,128,128,.65)"],i.pub.init=function(){if(!i.initialized)for(e.document.addEventListener("mousedown",i.onDocumentMouseDown,!1),e.document.addEventListener("keyup",i.onDocumentKeyUp,!1),e.addEventListener("resize",i.onWindowResize,!1),e.addEventListener("scroll",i.onWindowScroll,!1),i.pub.install(),i.initialized=!0;i.readyQueue.length;)i.readyQueue.shift()()},i.pub.install=function(e){var t=!0;try{i.installBySelector("[data-jscolor]",e)}catch(e){t=!1,console.warn(e)}if(i.pub.lookupClass)try{i.installBySelector("input."+i.pub.lookupClass+", button."+i.pub.lookupClass,e)}catch(e){}return t},i.pub.ready=function(e){return"function"!=typeof e?(console.warn("Passed value is not a function"),!1):(i.initialized?e():i.readyQueue.push(e),!0)},i.pub.trigger=function(e){var t=function(){i.triggerGlobal(e)};i.initialized?t():i.pub.ready(t)},i.pub.hide=function(){i.picker&&i.picker.owner&&i.picker.owner.hide()},i.pub.chessboard=function(e){return e||(e="rgba(0,0,0,0)"),i.genColorPreviewCanvas(e).canvas.toDataURL()},i.pub.background=function(e){var t=[];t.push(i.genColorPreviewGradient(e));var o=i.genColorPreviewCanvas();return t.push(["url('"+o.canvas.toDataURL()+"')","left top","repeat"].join(" ")),t.join(", ")},i.pub.options={},i.pub.lookupClass="jscolor",i.pub.installByClassName=function(){return console.error('jscolor.installByClassName() is DEPRECATED. Use data-jscolor="" attribute instead of a class name.'+i.docsRef),!1},i.register(),i.pub);return void 0===e.jscolor&&(e.jscolor=e.JSColor=a),a}))}},t={};function o(n){var i=t[n];if(void 0!==i)return i.exports;var a=t[n]={exports:{}};return e[n].call(a.exports,a,a.exports,o),a.exports}o.n=e=>{var t=e&&e.__esModule?()=>e.default:()=>e;return o.d(t,{a:t}),t},o.d=(e,t)=>{for(var n in t)o.o(t,n)&&!o.o(e,n)&&Object.defineProperty(e,n,{enumerable:!0,get:t[n]})},o.o=(e,t)=>Object.prototype.hasOwnProperty.call(e,t),(()=>{"use strict";let e={};function t(){e.WrinklerButtons&&Game.elderWrath?(l("PopAllNormalWrinklerButton").style.display="",l("PopFattestWrinklerButton").style.display=""):(l("PopAllNormalWrinklerButton").style.display="none",l("PopFattestWrinklerButton").style.display="none")}const n=["","","M","G","T","P","E","Z","Y"],i=["","","M","B","Tr","Quadr","Quint","Sext","Sept","Oct","Non","Dec","Undec","Duodec","Tredec","Quattuordec","Quindec","Sexdec","Septendec","Octodec","Novemdec","Vigint","Unvigint","Duovigint","Trevigint","Quattuorvigint"],a=["","K","M","B","T","Qa","Qi","Sx","Sp","Oc","No","De","UDe","DDe","TDe","QaDe","QiDe","SxDe","SpDe","ODe","NDe","Vi","UVi","DVi","TVi","QaVi","QiVi","SxVi","SpVi","OVi","NVi","Tr","UTr","DTr","TTr","QaTr","QiTr","SxTr","SpTr","OTr","NTr","Qaa","UQa","DQa","TQa","QaQa","QiQa","SxQa","SpQa","OQa","NQa","Qia","UQi","DQi","TQi","QaQi","QiQi","SxQi","SpQi","OQi","NQi","Sxa","USx","DSx","TSx","QaSx","QiSx","SxSx","SpSx","OSx","NSx","Spa","USp","DSp","TSp","QaSp","QiSp","SxSp","SpSp","OSp","NSp","Oco","UOc","DOc","TOc","QaOc","QiOc","SxOc","SpOc","OOc","NOc","Noa","UNo","DNo","TNo","QaNo","QiNo","SxNo","SpNo","ONo","NNo","Ct","UCt"];let r,s,c,d,p,u,m,h,g,f,C,y,b,k,G,w,v,x=[],T=[],M=[],S={};const P="CMText",B="CMBack",E="Blue",N="Green",O="Yellow",F="Orange",W="Red",L="Purple",A="Gray",j="Pink",D="Brown",H=[A,E,N,O,F,W,L,j,D],U=[["GoldCookTooltipPlaceholder","Calculated with Golden Switch off","200px"],["GoldCookDragonsFortuneTooltipPlaceholder","Calculated with Golden Switch off and at least one golden cookie on-screen","240px"],["PrestMaxTooltipPlaceholder","The MAX prestige is calculated with the cookies gained from popping all wrinklers with Skruuia god in Diamond slot, selling all stock market goods, selling all buildings with Earth Shatterer and Reality Bending auras, and buying Chocolate egg","320px"],["NextPrestTooltipPlaceholder","Calculated with cookies gained from wrinklers and Chocolate egg","200px"],["HeavenChipMaxTooltipPlaceholder","The MAX heavenly chips is calculated with the cookies gained from popping all wrinklers with Skruuia god in Diamond slot, selling all stock market goods, selling all buildings with Earth Shatterer and Reality Bending auras, and buying Chocolate egg","330px"],["ResetTooltipPlaceholder","The bonus income you would get from new prestige levels unlocked at 100% of its potential and from ascension achievements if you have the same buildings/upgrades after reset","370px"],["ChoEggTooltipPlaceholder","The amount of cookies you would get from popping all wrinklers with Skruuia god in Diamond slot, selling all stock market goods, selling all buildings with Earth Shatterer and Reality Bending auras, and then buying Chocolate egg","300px"],["ChainNextLevelPlaceholder","Cheated cookies might break this formula","250px"]],R={};let $,I,z=0,V=-1,_=[],X=1,Q=1,q=1;const Y=[10,15,30,60,300,600,900,1800],K=[1,5,10,15,30],J={Frenzy:O,"Dragon Harvest":D,"Elder frenzy":N,Clot:W,"Click frenzy":E,Dragonflight:j};let Z,ee,te,oe,ne,ie,ae,re={},le="",se=[];function ce(t,o,r){const l=e.ScaleDecimals+1;if(t===1/0)return"Infinity";if(void 0===t)return"0";if(0===e.Scale)return S.Beautify(t,o);if(Number.isFinite(t)){if(t<0)return`-${ce(Math.abs(t))}`;let s="";if(0===t)return t.toString();if(t>.001&&t<e.ScaleCutoff)return s=e.ScaleSeparator?t.toLocaleString("nl"):t.toLocaleString("en"),s;if(4===e.Scale&&!r||4===r)s=t.toExponential(l).toString().replace("e","E");else{const o=t.toExponential().toString(),c=Math.floor(o.slice(o.indexOf("e")+1)/3);s=(t/Number("1e"+3*c)).toFixed(l),1===e.Scale&&!r||1===r?t>=.01&&t<Number("1e"+3*n.length)?s+=` ${n[c]}`:s=ce(t,0,4):2===e.Scale&&!r||2===r?t>=.01&&t<Number("1e"+3*i.length)?s+=` ${i[c]}`:s=ce(t,0,4):3===e.Scale&&!r||3===r?t>=.01&&t<Number("1e"+3*a.length)?s+=` ${a[c]}`:s=ce(t,0,4):(5===e.Scale&&!r||5===r)&&(s+="E"+3*c)}return""===s&&(console.log(`Could not beautify number with Cookie Monster Beautify: ${t}`),s=S.Beautify(t,o)),e.ScaleSeparator&&(s=s.replace(".",",")),s}return console.log(`Could not beautify number with Cookie Monster Beautify: ${t}`),S.Beautify(t,o)}function de(t,o){let n=t;if(t===1/0)return t;if(t<0)return"Negative time period";n=Math.ceil(t);const i=Math.floor(n/31557600),a=Math.floor(n%31557600/86400),r=Math.floor(n%86400/3600),l=Math.floor(n%3600/60),s=Math.floor(n%60);let c="";if(e.TimeFormat){if(n>315576e4)return"XX:XX:XX:XX:XX";c+=(i<10?"0":"")+i+":",c+=(a<10?"0":"")+a+":",c+=(r<10?"0":"")+r+":",c+=(l<10?"0":"")+l+":",c+=(s<10?"0":"")+s}else{if(n>7776e5)return o?"Over 9000 days!":">9000d";c+=i>0?i+(o?1===i?" year":" years":"y")+", ":"",c+=a>0?a+(o?1===a?" day":" days":"d")+", ":"",(c.length>0||r>0)&&(c+=r+(o?1===r?" hour":" hours":"h")+", "),(c.length>0||l>0)&&(c+=l+(o?1===l?" minute":" minutes":"m")+", "),c+=s+(o?1===s?" second":" seconds":"s")}return c}function pe(t){let o,n;return t<=0?(n=e.TimeFormat?"00:00:00:00:00":"Done!",o=N):(n=de(t),o=t>300?W:t>60?F:O),{text:n,color:o}}let ue,me,he,ge,fe,Ce,ye,be,ke,Ge,we,ve,xe,Te,Me,Se,Pe,Be,Ee,Ne,Oe,Fe,We,Le,Ae,je,De,He,Ue,Re,$e,Ie,ze,Ve,_e,Xe=0,Qe=0,qe=0,Ye=0,Ke=0,Je=0,Ze=0,et=0,tt=0,ot=0,nt=0,it=0,at=0,rt=0,lt=0,st=1,ct=1,dt=1,pt=0,ut=0,mt=0,ht=0,gt={},ft={},Ct={},yt=0,bt=0,kt=[0,null],Gt=0,wt=0,vt={},xt={},Tt={},Mt={},St={},Pt=0,Bt=0,Et=[],Nt=0,Ot=0,Ft=[],Wt=0,Lt=0,At=[],jt=0,Dt=0,Ht=[],Ut=0,Rt=0,$t={},It={},zt=0,Vt=0,_t=[],Xt={},Qt=0,qt={0:[0,0,0],1:[0,0,0],2:[0,0,0],3:[0,0,0],4:[0,0,0],5:[0,0,0],6:[0,0,0],7:[0,0,0],8:[0,0,0],9:[0,0,0],10:[0,0,0]},Yt=[],Kt=[],Jt=[];function Zt(e){let t=0;return Game.dragonAuras[De].name!==e&&Game.dragonAuras[He].name!==e||(t=1),"Reality Bending"!==Game.dragonAuras[De].name&&"Reality Bending"!==Game.dragonAuras[He].name||(t+=.1),t}function eo(){let e=.25;return e*=1+Zt("Earth Shatterer"),e}function to(e,t){return void 0===ze[e]?void 0===t?1:t:ze[e]}function oo(e){const t=Kt[e];return(1!==Game.ascensionMode||"prestige"!==t.pool&&"fortune"!==t.tier)&&t?t.bought:0}function no(e){if(Game.hasGod){void 0===Yt.Temple.minigame&&(Yt.Temple.minigame=Game.Objects.Temple.minigame);const t=Yt.Temple.minigame.gods[e];if(Ue===t.id)return 1;if(Re===t.id)return 2;if($e===t.id)return 3}return!1}function io(e,t){let o=t;if(oo("Season savings")&&(o*=.99),oo("Santa's dominion")&&(o*=.99),oo("Faberge egg")&&(o*=.99),oo("Divine discount")&&(o*=.99),oo("Fortune #100")&&(o*=.99),o*=1-.02*Zt("Fierce Hoarder"),Game.hasBuff("Everything must go")&&(o*=.95),Game.hasBuff("Crafty pixies")&&(o*=.98),Game.hasBuff("Nasty goblins")&&(o*=1.02),e.fortune&&oo(e.fortune.name)&&(o*=.93),o*=to("buildingCost"),Yt.Temple.minigameLoaded){const e=no("creation");1===e?o*=.93:2===e?o*=.95:3===e&&(o*=.98)}return o}function ao(e,t,o,n,i,a){const r=void 0===a?0:a;let l=i,s=o,c=0;-1===i&&(l=s),i||(l=Game.buyBulk);for(let o=0;o<l;o++){let o=t*Game.priceIncrease**Math.max(0,s-n);o=r?Game.modifyBuildingPrice(e,o):io(e,o),o=Math.ceil(o);const i=r?e.getSellMultiplier():eo();o=Math.floor(o*i),s>0&&(c+=o,s-=1)}return c}function ro(){let t=Game.buyBulk;if(1===Game.buyMode?Q=t:t=Q,1===t?t=gt:10===t?t=ft:100===t&&(t=Ct),1===Game.buyMode?1===e.BuildColour?Object.keys(t).forEach((o=>{l(`productPrice${Game.Objects[o].id}`).style.color=e[`Colour${t[o].color}`]})):Object.keys(Game.Objects).forEach((e=>{l(`productPrice${Game.Objects[e].id}`).style.removeProperty("color")})):-1===Game.buyMode&&Object.keys(gt).forEach((e=>{const t=Game.Objects[e];l(`productPrice${t.id}`).style.color="",l(`productPrice${t.id}`).innerHTML=ce(ao(t,t.basePrice,t.amount,t.free,Game.buyBulk,1))})),1===Game.buyMode&&e.SortBuildings){let o;1===e.SortBuildings?(o=Object.keys(gt).map((e=>{const t=gt[e];return t.name=e,t.id=Game.Objects[e].id,t})),o.sort((function(e,t){return H.indexOf(e.color)>H.indexOf(t.color)?1:H.indexOf(e.color)<H.indexOf(t.color)||e.pp<t.pp?-1:0}))):2===e.SortBuildings&&(o=Object.keys(t).map((e=>{const o=t[e];return o.name=e,o.id=Game.Objects[e].id,o})),o.sort((function(e,t){return H.indexOf(e.color)>H.indexOf(t.color)?1:H.indexOf(e.color)<H.indexOf(t.color)||e.pp<t.pp?-1:0})));for(let e=0;e<o.length;e++)Game.Objects[o[e].name].l.style.gridRow=`${e+2}/${e+2}`}else{const e=Object.keys(gt).map((e=>{const t=gt[e];return t.name=e,t.id=Game.Objects[e].id,t}));e.sort(((e,t)=>e.id-t.id));for(let t=0;t<e.length;t++)Game.Objects[e[t].name].l.style.gridRow=`${t+2}/${t+2}`}}function lo(){if(e.UpBarColour>0){let e=0,t=0,o=0,n=0,i=0,a=0,r=0;Object.keys(Game.UpgradesInStore).forEach((s=>{const c=Game.UpgradesInStore[s];let d=!1;for(let e=0;e<l(`upgrade${s}`).childNodes.length;e+=1)if(-1!==l(`upgrade${s}`).childNodes[e].className.indexOf(B)){l(`upgrade${s}`).childNodes[e].className=B+vt[c.name].color,d=!0;break}if(!d){const e=document.createElement("div");e.style.width="10px",e.style.height="10px",e.className=B+vt[c.name].color,l(`upgrade${s}`).appendChild(e)}vt[c.name].color===E?e+=1:vt[c.name].color===N?t+=1:vt[c.name].color===O?o+=1:vt[c.name].color===F?n+=1:vt[c.name].color===W?i+=1:vt[c.name].color===L?a+=1:vt[c.name].color===A&&(r+=1)})),l("CMUpgradeBarBlue").textContent=e,l("CMUpgradeBarGreen").textContent=t,l("CMUpgradeBarYellow").textContent=o,l("CMUpgradeBarOrange").textContent=n,l("CMUpgradeBarRed").textContent=i,l("CMUpgradeBarPurple").textContent=a,l("CMUpgradeBarGray").textContent=r}const t=[];for(let e=0;e<Game.UpgradesInStore.length;e+=1){const o={};o.name=Game.UpgradesInStore[e].name,o.price=Game.UpgradesInStore[e].basePrice,o.pp=vt[o.name].pp,t.push(o)}e.SortUpgrades?t.sort((function(e,t){return H.indexOf(e.color)>H.indexOf(t.color)?1:H.indexOf(e.color)<H.indexOf(t.color)||e.pp<t.pp?-1:0})):t.sort(((e,t)=>e.price-t.price));const o=function(e,t){return e.findIndex((e=>e.name===t.name))};for(let e=0;e<Game.UpgradesInStore.length;e+=1)l(`upgrade${e}`).style.order=o(t,Game.UpgradesInStore[e])+1}function so(){return e.CPSMode?Se:0===e.CalcWrink?Game.cookiesPs*(1-Game.cpsSucked):1===e.CalcWrink?Game.cookiesPs*(Gt+(1-.05*wt)):null!==kt[1]&&2===e.CalcWrink&&1===Game.wrinklers[kt[1]].type?Game.cookiesPs*(3*Gt/wt+(1-.05*wt)):Game.cookiesPs*(Gt/wt+(1-.05*wt))}function co(){return 1===e.CalcWrink?yt:2===e.CalcWrink?kt[0]:0}function po(e,t,o){const n=document.createElement("div");n.id=e,n.style.height="12px",n.style.margin="0px 10px",n.style.position="relative";const i=document.createElement("div");i.style.width="100%",i.style.height="10px",i.style.margin="auto",i.style.position="absolute",i.style.left="0px",i.style.top="0px",i.style.right="0px",i.style.bottom="0px";const a=document.createElement("span");a.style.display="inline-block",a.style.textAlign="right",a.style.fontSize="10px",a.style.width="108px",a.style.marginRight="5px",a.style.verticalAlign="text-top",a.textContent=t,i.appendChild(a);for(let e=0;e<o.length;e++){const t=document.createElement("span");t.id=o[e].id,t.style.display="inline-block",t.style.height="10px",t.style.verticalAlign="text-top",t.style.textAlign="center",o.length-1===e&&(t.style.borderTopRightRadius="10px",t.style.borderBottomRightRadius="10px"),void 0!==o[e].color&&(t.className=B+o[e].color),i.appendChild(t)}const r=document.createElement("span");return r.id=`${e}Time`,r.style.marginLeft="5px",r.style.verticalAlign="text-top",i.appendChild(r),n.appendChild(i),n}function uo(e){if(null!==l("CMBotBar")){const t=l("CMBotBar").firstChild.firstChild.childNodes[0],o=l("CMBotBar").firstChild.firstChild.childNodes[1],n=l("CMBotBar").firstChild.firstChild.childNodes[2],i=l("CMBotBar").firstChild.firstChild.childNodes[3],a=e,r=t.appendChild(document.createElement("td"));r.appendChild(document.createTextNode(`${-1!==a.indexOf(" ")?a.substring(0,a.indexOf(" ")):a} (`)),r.appendChild(document.createElement("span")).className="CMTextBlue",r.appendChild(document.createTextNode(")")),t.lastChild.style.paddingLeft="8px",o.appendChild(document.createElement("td")),o.lastChild.style.paddingLeft="8px",n.appendChild(document.createElement("td")),n.lastChild.style.paddingLeft="8px",i.appendChild(document.createElement("td")),i.lastChild.style.paddingLeft="2px"}}function mo(){if(1===e.BotBar&&gt&&1===Game.buyMode){let t=0;Object.keys(gt).forEach((o=>{let n,i=Game.buyBulk;1===Game.buyMode?X=i:i=X,1===i&&(i=gt),10===i&&(i=ft),100===i&&(i=Ct),t+=1,l("CMBotBar").firstChild.firstChild.childNodes[0].childNodes[t].childNodes[1].textContent=Game.Objects[o].amount,l("CMBotBar").firstChild.firstChild.childNodes[1].childNodes[t].textContent=ce(i[o].bonus,2),l("CMBotBar").firstChild.firstChild.childNodes[2].childNodes[t].className=P+i[o].color,n=e.PPDisplayTime?de(Math.round(i[o].pp)):ce(Math.round(i[o].pp),2),l("CMBotBar").firstChild.firstChild.childNodes[2].childNodes[t].textContent=n;const a=pe((Game.Objects[o].bulkPrice-(Game.cookies+co()))/so());l("CMBotBar").firstChild.firstChild.childNodes[3].childNodes[t].className=P+a.color,"Done!"===a.text&&Game.cookies<Game.Objects[o].bulkPrice?l("CMBotBar").firstChild.firstChild.childNodes[3].childNodes[t].textContent=`${a.text} (with Wrink)`:l("CMBotBar").firstChild.firstChild.childNodes[3].childNodes[t].textContent=a.text}))}}function ho(){Game.Background.canvas.width=Game.Background.canvas.parentNode.offsetWidth,Game.Background.canvas.height=Game.Background.canvas.parentNode.offsetHeight,Game.LeftBackground.canvas.width=Game.LeftBackground.canvas.parentNode.offsetWidth,Game.LeftBackground.canvas.height=Game.LeftBackground.canvas.parentNode.offsetHeight}function go(){1===e.BotBar&&1===e.TimerBar&&1===e.TimerBarPos?(l("CMBotBar").style.bottom=l("CMTimerBar").style.height,l("game").style.bottom=`${Number(l("CMTimerBar").style.height.replace("px",""))+70}px`):1===e.BotBar?(l("CMBotBar").style.bottom="0px",l("game").style.bottom="70px"):1===e.TimerBar&&1===e.TimerBarPos?l("game").style.bottom=l("CMTimerBar").style.height:l("game").style.bottom="0px",1===e.TimerBar&&0===e.TimerBarPos?l("sectionLeft").style.top=l("CMTimerBar").style.height:l("sectionLeft").style.top="",ho()}function fo(){1===e.TimerBar?l("CMTimerBar").style.display="":l("CMTimerBar").style.display="none",go()}function Co(){Xe=Game.dragonAura,Qe=Game.dragonAura2}function yo(e){const t=Game.Achievements[e],o={};return o.name=t.name,o}function bo(e){let t=1;if(Object.keys(e.tieredUpgrades).forEach((o=>{!Game.Tiers[e.tieredUpgrades[o].tier].special&&oo(e.tieredUpgrades[o].name)&&(t*=2)})),Object.keys(e.synergies).forEach((o=>{if(oo(e.synergies[o].name)){const n=e.synergies[o];n.buildingTie1.name===e.name?t*=1+.05*n.buildingTie2.amount:n.buildingTie2.name===e.name&&(t*=1+.001*n.buildingTie1.amount)}})),e.fortune&&oo(e.fortune.name)&&(t*=1.07),e.grandma&&oo(e.grandma.name)&&(t*=1+.01*Yt.Grandma.amount*(1/(e.id-1))),"object"==typeof e.tieredUpgrades.misfortune&&1===e.vanilla&&oo(e.tieredUpgrades.misfortune.name))switch(Game.elderWrath){default:t*=1;break;case 1:t*=1.02;break;case 2:t*=1.04;break;case 3:t*=1.06}return t}function ko(e){const t=Game.Objects[e],o={};return"Cursor"===t.name?o.cps=function(e){let t=0;oo("Thousand fingers")&&(t+=.1),oo("Million fingers")&&(t*=5),oo("Billion fingers")&&(t*=10),oo("Trillion fingers")&&(t*=20),oo("Quadrillion fingers")&&(t*=20),oo("Quintillion fingers")&&(t*=20),oo("Sextillion fingers")&&(t*=20),oo("Septillion fingers")&&(t*=20),oo("Octillion fingers")&&(t*=20),oo("Nonillion fingers")&&(t*=20);let o=1,n=0;return Object.keys(Yt).forEach((e=>{"Cursor"!==Yt[e].name&&(n+=Yt[e].amount)})),t*=n,o*=bo(e),o*=Game.magicCpS("Cursor"),o*=to("cursorCps"),Game.ComputeCps(.1,oo("Reinforced index finger")+oo("Carpal tunnel prevention cream")+oo("Ambidextrous"),t)*o}:"Grandma"===t.name?o.cps=function(e){let o=1;if(Object.keys(Game.GrandmaSynergies).forEach((e=>{oo(Game.GrandmaSynergies[e])&&(o*=2)})),oo("Bingo center/Research facility")&&(o*=4),oo("Ritual rolling pins")&&(o*=2),oo("Naughty list")&&(o*=2),oo("Elderwort biscuits")&&(o*=1.02),o*=to("grandmaCps"),oo("Cat ladies"))for(let e=0;e<Game.UpgradesByPool.kitten.length;e++)oo(Game.UpgradesByPool.kitten[e].name)&&(o*=1.29);o*=bo(e);let n=0;oo("One mind")&&(n+=.02*Yt.Grandma.amount),oo("Communal brainsweep")&&(n+=.02*Yt.Grandma.amount),oo("Elder Pact")&&(n+=.05*Yt.Portal.amount);let i=0;return Object.keys(Yt).forEach((e=>{"Grandma"!==Yt[e].name&&(i+=Yt[e].amount)})),o*=1+.01*Zt("Elder Battalion")*i,o*=Game.magicCpS(t.name),(t.baseCps+n)*o}:o.cps=function(e){let t=1;return t*=bo(e),t*=Game.magicCpS(e.name),e.baseCPS*t},o.baseCps=t.baseCps,o.name=t.name,o.tieredUpgrades=t.tieredUpgrades,o.synergies=t.synergies,o.fortune=t.fortune,o.grandma=t.grandma,o.baseCPS=t.baseCps,o.id=t.id,o.vanilla=t.vanilla,o}function Go(e){const t=Game.Upgrades[e],o={};return o.power=t.power,"function"==typeof o.power&&("Sugar crystal cookies"===t.name?o.power=function(){let e=5;return Object.keys(Yt).forEach((t=>{Yt[t].level>=10&&(e+=1)})),e}:o.power=function(){let e=2;if(oo("Starlove")&&(e=3),Game.hasGod){const t=no("seasons");1===t?e*=1.3:2===t?e*=1.2:3===t&&(e*=1.1)}return e}),o.pool=t.pool,o.name=t.name,o}function wo(){Fe=Game.UpgradesOwned,We=Game.pledges,Le=Game.AchievementsOwned,Ae=Game.heavenlyPower,je=Game.prestige,Object.keys(Game.Objects).forEach((e=>{const t=Game.Objects[e];let o=Yt[e];void 0===o&&(Yt[e]=ko(e),o=Yt[e],uo(e)),o.amount=t.amount,o.level=t.level,o.totalCookies=t.totalCookies,o.basePrice=t.basePrice,o.free=t.free,t.minigameLoaded&&("Temple"===t.name&&(Ue=t.minigame.slot[0],Re=t.minigame.slot[1],$e=t.minigame.slot[2]),o.minigameLoaded=t.minigameLoaded,o.minigame=t.minigame),Yt[e]=o})),Object.keys(Game.Upgrades).forEach((e=>{const t=Game.Upgrades[e];let o=Kt[e];void 0===o&&(Kt[e]=Go(e),o=Kt[e]),o.bought=t.bought,Kt[e]=o})),Object.keys(Game.Achievements).forEach((e=>{const t=Game.Achievements[e];let o=Jt[e];void 0===o&&(Jt[e]=yo(e),o=Jt[e]),o.won=t.won,Jt[e]=o})),Co(),De=Xe,He=Qe}function vo(e){const t=document.createElement("div");return t.style.fontWeight="bold",t.id=`${e}Title`,t.className="CMTextBlue",t.textContent=e,t}function xo(e){e.appendChild(vo("Bonus Income"));const t=document.createElement("div");t.style.marginBottom="4px",t.style.color="white",t.id="CMTooltipIncome",e.appendChild(t),e.appendChild(vo("Bonus Cookies per Click")),e.lastChild.style.display="none";const o=document.createElement("div");o.style.marginBottom="4px",o.style.color="white",o.style.display="none",o.id="CMTooltipCookiePerClick",e.appendChild(o),e.appendChild(vo("Payback Period"));const n=document.createElement("div");n.style.marginBottom="4px",n.id="CMTooltipPP",e.appendChild(n),e.appendChild(vo("Time Left"));const i=document.createElement("div");if(i.id="CMTooltipTime",e.appendChild(i),"b"===te){e.appendChild(vo("Production left till next achievement")),e.lastChild.id="CMTooltipProductionHeader";const t=document.createElement("div");t.id="CMTooltipProduction",e.appendChild(t)}}function To(e,t,o){let n=e,i=0;for(;n<o;)n+=.002*Math.max(.002,(n/Math.max(t,100))**.5),i+=1;return i/Game.fps}function Mo(){let e=1;return Object.keys(Game.buffs).forEach((t=>{void 0!==Game.buffs[t].multCpS&&(e*=Game.buffs[t].multCpS)})),e}function So(){null!==l("CMDispTooltipWarningParent")&&(0===e.ToolWarnPos?(l("CMDispTooltipWarningParent").style.top="auto",l("CMDispTooltipWarningParent").style.margin="4px -4px",l("CMDispTooltipWarningParent").style.padding="3px 4px"):(l("CMDispTooltipWarningParent").style.right="auto",l("CMDispTooltipWarningParent").style.margin="4px",l("CMDispTooltipWarningParent").style.padding="4px 3px"))}function Po(){if(wo(),"none"!==l("tooltipAnchor").style.display&&l("CMTooltipArea")){l("CMTooltipArea").innerHTML="";const t=function(){l("tooltip").firstChild.style.paddingBottom="4px";const e=document.createElement("div");return e.style.border="1px solid",e.style.padding="4px",e.style.margin="0px -4px",e.id="CMTooltipBorder",e.className="CMTextGray",e}();l("CMTooltipArea").appendChild(t),"b"===te?function(){if(1===e.TooltipBuildUpgrade&&1===Game.buyMode){let t;if(xo(l("CMTooltipBorder")),1===Game.buyMode?q=t:t=q,1===Game.buyBulk?t=gt:10===Game.buyBulk?t=ft:100===Game.buyBulk&&(t=Ct),Z=Game.Objects[oe].bulkPrice,ee=t[oe].bonus,1===e.TooltipBuildUpgrade&&1===Game.buyMode){l("CMTooltipIncome").textContent=ce(ee,2);const o=Math.round(ee/Game.cookiesPs*1e4);Number.isFinite(o)&&0!==o?l("CMTooltipIncome").textContent+=` (${o/100}% of income)`:l("CMTooltipIncome").textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`,l("CMTooltipBorder").className=P+t[oe].color,e.PPDisplayTime?l("CMTooltipPP").textContent=de(t[oe].pp):l("CMTooltipPP").textContent=ce(t[oe].pp,2),l("CMTooltipPP").className=P+t[oe].color;const n=pe((Z-(Game.cookies+co()))/so());l("CMTooltipTime").textContent=n.text,"Done!"===n.text&&Game.cookies<t[oe].price?l("CMTooltipTime").textContent=`${n.text} (with Wrink)`:l("CMTooltipTime").textContent=n.text,l("CMTooltipTime").className=P+n.color}l("CMTooltipProductionHeader").style.display="none",l("CMTooltipTime").style.marginBottom="0px";for(const e of Object.keys(Game.Objects[oe].productionAchievs))if(!Game.HasAchiev(Game.Objects[oe].productionAchievs[e].achiev.name)){const t=Game.Objects[oe].productionAchievs[e];l("CMTooltipTime").style.marginBottom="4px",l("CMTooltipProductionHeader").style.display="",l("CMTooltipProduction").className=`ProdAchievement${oe}`,l("CMTooltipProduction").textContent=ce(t.pow-Yt[oe].totalCookies,15),l("CMTooltipProduction").style.color="white";break}}else l("CMTooltipArea").style.display="none"}():"u"===te?function(){if(xo(l("CMTooltipBorder")),ee=vt[Game.UpgradesInStore[oe].name].bonus,Z=Game.Upgrades[Game.UpgradesInStore[oe].name].getPrice(),ne=vt[Game.UpgradesInStore[oe].name].bonusMouse,1===e.TooltipBuildUpgrade){l("CMTooltipIncome").textContent=ce(ee,2);const t=Math.round(ee/Game.cookiesPs*1e4);"0"!==l("CMTooltipIncome").textContent||"b"!==te&&"u"!==te?(Number.isFinite(t)&&0!==t?l("CMTooltipIncome").textContent+=` (${t/100}% of income)`:l("CMTooltipIncome").textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`,l("CMTooltipBorder").className=P+vt[Game.UpgradesInStore[oe].name].color,ne&&(l("CMTooltipCookiePerClick").textContent=ce(ne),l("CMTooltipCookiePerClick").style.display="block",l("CMTooltipCookiePerClick").previousSibling.style.display="block"),0===ee&&ne?(l("CMTooltipPP").textContent=`${ce(Z/ne)} Clicks`,l("CMTooltipPP").style.color="white"):(e.PPDisplayTime?l("CMTooltipPP").textContent=de(vt[Game.UpgradesInStore[oe].name].pp):l("CMTooltipPP").textContent=ce(vt[Game.UpgradesInStore[oe].name].pp,2),l("CMTooltipPP").className=P+vt[Game.UpgradesInStore[oe].name].color)):(l("Bonus IncomeTitle").style.display="none",l("CMTooltipIncome").style.display="none",l("Payback PeriodTitle").style.display="none",l("CMTooltipPP").style.display="none");const o=pe((Z-(Game.cookies+co()))/so());if(l("CMTooltipTime").textContent=o.text,"Done!"===o.text&&Game.cookies<Game.UpgradesInStore[oe].getPrice()?l("CMTooltipTime").textContent=`${o.text} (with Wrink)`:l("CMTooltipTime").textContent=o.text,l("CMTooltipTime").className=P+o.color,"Chocolate egg"===Game.UpgradesInStore[oe].name){l("CMTooltipBorder").lastChild.style.marginBottom="4px",l("CMTooltipBorder").appendChild(vo("Cookies to be gained (Currently/Max)"));const e=document.createElement("div");e.style.color="white",e.textContent=`${ce(.05*Game.cookies)} / ${ce(ht)}`,l("CMTooltipBorder").appendChild(e)}}else l("CMTooltipArea").style.display="none"}():"s"===te?function(){if(1===e.TooltipLump){const e=l("CMTooltipBorder");e.appendChild(vo("Current Sugar Lump"));const o=document.createElement("div");o.id="CMTooltipTime",e.appendChild(o);const n=0===(t=Game.lumpCurrentType)?{text:"Normal",color:A}:1===t?{text:"Bifurcated",color:N}:2===t?{text:"Golden",color:O}:3===t?{text:"Meaty",color:F}:4===t?{text:"Caramelized",color:L}:{text:"Unknown Sugar Lump",color:W};o.textContent=n.text,o.className=P+n.color}else l("CMTooltipArea").style.display="none";var t}():"g"===te?function(){const{minigame:t}=Game.Objects["Wizard tower"],o=t.getSpellCost(t.spellsById[oe]);if(1===e.TooltipGrim&&o<=t.magicM){const e=l("CMTooltipBorder");e.appendChild(vo("Time Left"));const n=document.createElement("div");n.id="CMTooltipTime",e.appendChild(n);const i=pe(To(t.magic,t.magicM,o));if(n.textContent=i.text,n.className=P+i.color,o<=t.magic){e.appendChild(vo("Recover Time"));const n=document.createElement("div");n.id="CMTooltipRecover",e.appendChild(n);const i=pe(To(Math.max(0,t.magic-o),t.magicM,t.magic));n.textContent=i.text,n.className=P+i.color}if("0"===oe){e.appendChild(vo("Cookies to be gained/lost"));const t=document.createElement("div");t.id="x",e.appendChild(t);const o=document.createElement("span");o.style.color="#33FF00",o.textContent=ce(Math.min(.15*(Game.cookies+co()),60*lt*30),2),t.appendChild(o);const n=document.createElement("span");n.textContent=" / ",t.appendChild(n);const i=document.createElement("span");i.style.color="red",i.textContent=ce(60*lt*15,2),t.appendChild(i)}l("CMTooltipArea").appendChild(e)}else l("CMTooltipArea").style.display="none"}():"p"===te?function(){const{minigame:t}=Game.Objects.Farm;if(e.TooltipPlots&&0!==t.plot[oe[1]][oe[0]][0]){const e=t.plot[oe[1]][oe[0]][1]>t.plantsById[t.plot[oe[1]][oe[0]][0]-1].mature,o=t.plantsById[t.plot[oe[1]][oe[0]][0]-1].name;l("CMTooltipBorder").appendChild(vo("Reward (Current / Maximum)"));const n=document.createElement("div");n.id="CMTooltipPlantReward",l("CMTooltipBorder").appendChild(n),"Bakeberry"===o?l("CMTooltipPlantReward").textContent=`${e?ce(Math.min(.03*Game.cookies,60*Game.cookiesPs*30)):"0"} / ${ce(60*Game.cookiesPs*30)}`:"Chocoroot"===o||"White chocoroot"===o?l("CMTooltipPlantReward").textContent=`${e?ce(Math.min(.03*Game.cookies,60*Game.cookiesPs*3)):"0"} / ${ce(60*Game.cookiesPs*3)}`:"Queenbeet"===o?l("CMTooltipPlantReward").textContent=`${e?ce(Math.min(.04*Game.cookies,60*Game.cookiesPs*60)):"0"} / ${ce(60*Game.cookiesPs*60)}`:"Duketater"===o?l("CMTooltipPlantReward").textContent=`${e?ce(Math.min(.08*Game.cookies,60*Game.cookiesPs*120)):"0"} / ${ce(60*Game.cookiesPs*120)}`:l("CMTooltipArea").style.display="none"}else l("CMTooltipArea").style.display="none"}():"ha"===te?function(){const{minigame:t}=Game.Objects.Farm;if(e.TooltipLump){l("CMTooltipBorder").appendChild(vo("Cookies gained from harvesting:"));let e=0,o=0;Game.keys[16]&&Game.keys[17]&&(o=1);for(let n=0;n<6;n++)for(let i=0;i<6;i++)if(t.plot[n][i][0]>=1){const a=t.plot[n][i],r=t.plantsById[a[0]-1],l=r.name;let s=!0;o&&r.immortal&&(s=!1),a[1]<r.matureBase&&(s=!1),s&&"Bakeberry"===l?e+=Math.min(.03*Game.cookies,60*Game.cookiesPs*30):s&&"Chocoroot"===l||"White chocoroot"===l?e+=Math.min(.03*Game.cookies,60*Game.cookiesPs*3):s&&"Queenbeet"===l?e+=Math.min(.04*Game.cookies,60*Game.cookiesPs*60):s&&"Duketater"===l&&(e+=Math.min(.08*Game.cookies,60*Game.cookiesPs*120))}l("CMTooltipBorder").appendChild(document.createTextNode(ce(e)))}else l("CMTooltipArea").style.display="none"}():"wb"===te?function(){l("tooltip").innerHTML="",l("tooltip").appendChild(vo("Reward:"));const e=document.createElement("div");e.id="CMWrinklerReward","PopAll"===oe?e.textContent=ce(yt):"PopFattest"===oe&&(e.textContent=ce(kt[0])),l("tooltip").appendChild(e)}():("pag"===te||"pas"===te&&-1!==oe[1])&&function(){if(1===e.TooltipPantheon){const t=l("CMTooltipBorder");let o;o="pas"===te?oe[1]:oe,t.appendChild(vo("Effect in position 1:"));const n=document.createElement("div");if(n.id="CMPantheonTooltipPosition1",0!==qt[o][0]){n.textContent=ce(qt[o][0]);const t=Math.round(qt[o][0]/Game.cookiesPs*1e4);Number.isFinite(t)&&0!==t?n.textContent+=` (${t/100}% of income)`:n.textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`}else n.textContent="No effect to CPS";t.appendChild(n),t.appendChild(vo("Effect in position 2:"));const i=document.createElement("div");if(i.id="CMPantheonTooltipPosition2",0!==qt[o][1]){i.textContent=ce(qt[o][1]);const t=Math.round(qt[o][1]/Game.cookiesPs*1e4);Number.isFinite(t)&&0!==t?i.textContent+=` (${t/100}% of income)`:i.textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`}else i.textContent="No effect to CPS";t.appendChild(i),t.appendChild(vo("Effect in position 3:"));const a=document.createElement("div");if(a.id="CMPantheonTooltipPosition2",0!==qt[o][2]){a.textContent=ce(qt[o][2]);const t=Math.round(qt[o][2]/Game.cookiesPs*1e4);Number.isFinite(t)&&0!==t?a.textContent+=` (${t/100}% of income)`:a.textContent+=` (<0${e.ScaleSeparator?",":"."}01% of income)`}else a.textContent="No effect to CPS";t.appendChild(a),l("CMTooltipArea").appendChild(t)}else l("CMTooltipArea").style.display="none"}(),function(){if("b"===te||"u"===te){null===document.getElementById("CMDispTooltipWarningParent")&&(l("tooltipAnchor").appendChild(function(){const t=document.createElement("div");t.style.position="absolute",t.style.display="block",t.style.left="auto",t.style.bottom="auto",t.id="CMDispTooltipWarningParent";const o=function(e,t,o,n,i){const a=document.createElement("div");a.id=e,a.style.display="none",a.style.transition="opacity 0.1s ease-out",a.className="CMBorder"+t,a.style.padding="2px",a.style.background="#000 url(img/darkNoise.png)";const r=document.createElement("div");a.appendChild(r);const l=document.createElement("span");l.className=P+t,l.style.fontWeight="bold",l.textContent=o,r.appendChild(l),r.appendChild(document.createTextNode(n));const s=document.createElement("div");a.appendChild(s);const c=document.createElement("span");return c.id=i,s.appendChild(document.createTextNode("Deficit: ")),s.appendChild(c),a};return t.appendChild(o("CMDispTooltipWarnLucky",W,"Warning: ",'Purchase of this item will put you under the number of Cookies required for "Lucky!"',"CMDispTooltipWarnLuckyText")),t.firstChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnLuckyFrenzy",O,"Warning: ",'Purchase of this item will put you under the number of Cookies required for "Lucky!" (Frenzy)',"CMDispTooltipWarnLuckyFrenzyText")),t.lastChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnConjure",L,"Warning: ",'Purchase of this item will put you under the number of Cookies required for "Conjure Baked Goods"',"CMDispTooltipWarnConjureText")),t.lastChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnConjureFrenzy",L,"Warning: ",'Purchase of this item will put you under the number of Cookies required for "Conjure Baked Goods" (Frenzy)',"CMDispTooltipWarnConjureFrenzyText")),t.lastChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnEdifice",L,"Warning: ",'Purchase of this item will put you under the number of Cookies needed for "Spontaneous Edifice" to possibly give you your most expensive building"',"CMDispTooltipWarnEdificeText")),t.lastChild.style.marginBottom="4px",t.appendChild(o("CMDispTooltipWarnUser",W,"Warning: ",`Purchase of this item will put you under the number of Cookies equal to ${e.ToolWarnUser} seconds of CPS`,"CMDispTooltipWarnUserText")),t}()),So()),0===e.ToolWarnPos?l("CMDispTooltipWarningParent").style.right="0px":l("CMDispTooltipWarningParent").style.top=`${l("tooltip").offsetHeight}px`,l("CMDispTooltipWarningParent").style.width=l("tooltip").offsetWidth-6+"px";const t=Game.cookies+co()-Z,o=e.ToolWarnBon?ee:0;let n=Ke;if(1===e.ToolWarnBon){let e=ee;e/=Mo(),n+=60*e*15/.15}if(1===e.ToolWarnLucky&&t<n&&("b"!==te||1===Game.buyMode)?(l("CMDispTooltipWarnLucky").style.display="",l("CMDispTooltipWarnLuckyText").textContent=`${ce(n-t)} (${de((n-t)/(so()+o))})`):l("CMDispTooltipWarnLucky").style.display="none",1===e.ToolWarnLuckyFrenzy){const e=7*n;t<e&&("b"!==te||1===Game.buyMode)?(l("CMDispTooltipWarnLuckyFrenzy").style.display="",l("CMDispTooltipWarnLuckyFrenzyText").textContent=`${ce(e-t)} (${de((e-t)/(so()+o))})`):l("CMDispTooltipWarnLuckyFrenzy").style.display="none"}else l("CMDispTooltipWarnLuckyFrenzy").style.display="none";if(1===e.ToolWarnConjure){const e=2*n;t<e&&("b"!==te||1===Game.buyMode)?(l("CMDispTooltipWarnConjure").style.display="",l("CMDispTooltipWarnConjureText").textContent=`${ce(e-t)} (${de((e-t)/(so()+o))})`):l("CMDispTooltipWarnConjure").style.display="none"}else l("CMDispTooltipWarnConjure").style.display="none";if(1===e.ToolWarnConjureFrenzy){const e=2*n*7;t<e&&("b"!==te||1===Game.buyMode)?(l("CMDispTooltipWarnConjureFrenzy").style.display="",l("CMDispTooltipWarnConjureFrenzyText").textContent=`${ce(e-t)} (${de((e-t)/(so()+o))})`):l("CMDispTooltipWarnConjureFrenzy").style.display="none"}else l("CMDispTooltipWarnConjureFrenzy").style.display="none";1===e.ToolWarnEdifice&&Game.Objects["Wizard tower"].minigameLoaded&&at&&t<at&&("b"!==te||1===Game.buyMode)?(l("CMDispTooltipWarnEdifice").style.display="",l("CMDispTooltipWarnEdificeText").textContent=`${ce(at-t)} (${de((at-t)/(so()+o))})`):l("CMDispTooltipWarnEdifice").style.display="none",e.ToolWarnUser>0&&t<e.ToolWarnUser*so()&&("b"!==te||1===Game.buyMode)?(l("CMDispTooltipWarnUser").style.display="",l("CMDispTooltipWarnUser").children[0].textContent=`Purchase of this item will put you under the number of Cookies equal to ${e.ToolWarnUser} seconds of CPS`,l("CMDispTooltipWarnUserText").textContent=`${ce(e.ToolWarnUser*so()-t)} (${de((e.ToolWarnUser*so()-t)/(so()+o))})`):l("CMDispTooltipWarnUser").style.display="none"}else null!==l("CMDispTooltipWarningParent")&&l("CMDispTooltipWarningParent").remove()}()}else null===l("CMTooltipArea")&&null!==l("CMDispTooltipWarningParent")&&l("CMDispTooltipWarningParent").remove()}function Bo(){if(Game.prefs.autosave&&Game.drawT%10==0&&"stats"===Game.onMenu&&e.Stats){const e=document.getElementById("CMStatsAutosaveTimer");e&&(e.innerText=Game.sayTime(60*Game.fps-Game.T%(60*Game.fps),4))}if(ro(),lo(),function(){if(1===e.TimerBar){const t=l("CMTimerBar").offsetWidth-163,o=l("CMTimerBar").offsetWidth-133;let n=0;0!==Game.shimmerTypes.golden.spawned||Game.Has("Golden switch [off]")?l("CMTimerBarGC").style.display="none":(l("CMTimerBarGC").style.display="",l("CMTimerBarGCMinBar").style.width=`${Math.round(Math.max(0,Game.shimmerTypes.golden.minTime-Game.shimmerTypes.golden.time)*t/Game.shimmerTypes.golden.maxTime)}px`,e.TimerBarOverlay>=1?l("CMTimerBarGCMinBar").textContent=Math.ceil((Game.shimmerTypes.golden.minTime-Game.shimmerTypes.golden.time)/Game.fps):l("CMTimerBarGCMinBar").textContent="",Game.shimmerTypes.golden.minTime===Game.shimmerTypes.golden.maxTime?(l("CMTimerBarGCMinBar").style.borderTopRightRadius="10px",l("CMTimerBarGCMinBar").style.borderBottomRightRadius="10px"):(l("CMTimerBarGCMinBar").style.borderTopRightRadius="",l("CMTimerBarGCMinBar").style.borderBottomRightRadius=""),l("CMTimerBarGCBar").style.width=`${Math.round(Math.min(Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.minTime,Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.time)*t/Game.shimmerTypes.golden.maxTime)}px`,e.TimerBarOverlay>=1?l("CMTimerBarGCBar").textContent=Math.ceil(Math.min(Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.minTime,Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.time)/Game.fps):l("CMTimerBarGCBar").textContent="",l("CMTimerBarGCTime").textContent=Math.ceil((Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.time)/Game.fps),n+=1),"christmas"===Game.season&&0===Game.shimmerTypes.reindeer.spawned?(l("CMTimerBarRen").style.display="",l("CMTimerBarRenMinBar").style.width=`${Math.round(Math.max(0,Game.shimmerTypes.reindeer.minTime-Game.shimmerTypes.reindeer.time)*t/Game.shimmerTypes.reindeer.maxTime)}px`,e.TimerBarOverlay>=1?l("CMTimerBarRenMinBar").textContent=Math.ceil((Game.shimmerTypes.reindeer.minTime-Game.shimmerTypes.reindeer.time)/Game.fps):l("CMTimerBarRenMinBar").textContent="",l("CMTimerBarRenBar").style.width=`${Math.round(Math.min(Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.minTime,Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.time)*t/Game.shimmerTypes.reindeer.maxTime)}px`,e.TimerBarOverlay>=1?l("CMTimerBarRenBar").textContent=Math.ceil(Math.min(Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.minTime,Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.time)/Game.fps):l("CMTimerBarRenBar").textContent="",l("CMTimerBarRenTime").textContent=Math.ceil((Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.time)/Game.fps),n+=1):l("CMTimerBarRen").style.display="none";const i={};l("CMTimerBarBuffTimers").innerHTML="",Object.keys(Game.buffs).forEach((t=>{if(Game.buffs[t]){const a=po(Game.buffs[t].name,Game.buffs[t].name,[{id:`${Game.buffs[t].name}Bar`}]);a.style.display="";let r="";r=void 0!==J[Game.buffs[t].name]?J[Game.buffs[t].name]:L,a.lastChild.children[1].className=B+r,a.lastChild.children[1].style.color="black",2===e.TimerBarOverlay?a.lastChild.children[1].textContent=`${Math.round(Game.buffs[t].time/Game.buffs[t].maxTime*100)}%`:a.lastChild.children[1].textContent="",a.lastChild.children[1].style.width=`${Math.round(Game.buffs[t].time*(o-8*Math.ceil(Game.buffs[t].time/Game.fps).toString().length)/Game.buffs[t].maxTime)}px`,a.lastChild.children[2].textContent=Math.ceil(Game.buffs[t].time/Game.fps),n+=1,i[Game.buffs[t].name]=a}})),Object.keys(i).forEach((e=>{l("CMTimerBarBuffTimers").appendChild(i[e])})),0!==n&&(l("CMTimerBar").style.height=12*n+2+"px"),ae!==n&&(ae=n,go())}}(),mo(),Po(),function(){if(1===e.TooltipWrink&&1===z){let e=!1;Object.keys(Game.wrinklers).forEach((t=>{const o=Game.wrinklers[t];if(o.phase>0&&o.selected){if(e=!0,0===_[t]||void 0===_[t]){const e=document.createElement("div"),o=document.createElement("div");o.style.minWidth="120px",o.style.marginBottom="4px";const n=document.createElement("div");n.style.textAlign="center",n.id="CMTooltipWrinkler",o.appendChild(n),e.appendChild(o),Game.tooltip.draw(this,escape(e.innerHTML)),V=t,_[t]=1}}else _[t]=0})),e||Game.tooltip.hide()}}(),function(){if(1===e.TooltipWrink&&null!==l("CMTooltipWrinkler")){let{sucked:e}=Game.wrinklers[V],t=1.1;if(Game.Has("Sacrilegious corruption")&&(t*=1.05),1===Game.wrinklers[V].type&&(t*=3),e*=t,Game.Has("Wrinklerspawn")&&(e*=1.05),Yt.Temple.minigameLoaded){const t=Game.hasGod("scorn");1===t?e*=1.15:2===t?e*=1.1:3===t&&(e*=1.05)}l("CMTooltipWrinkler").textContent=ce(e)}}(),e.UpStats&&"stats"===Game.onMenu&&(Game.drawT-1)%(5*Game.fps)!=0&&(Game.drawT-1)%Game.fps==0&&Game.UpdateMenu(),t(),e.Scale){let e=l("cookies").innerHTML.replace(/.*(?=<br>)/i,ce(Game.cookies));Game.prefs.monospace&&(e=`<span class="monospace">${e}</span>`),l("cookies").innerHTML=e}}class Eo{constructor(e){this.maxLength=e,this.queue=[]}addLatest(e){this.queue.push(e)>this.maxLength&&this.queue.shift()}calcAverage(e){let t=e;t>this.maxLength&&(t=this.maxLength),t>this.queue.length&&(t=this.queue.length);let o=0;for(let e=this.queue.length-1;e>=0&&e>this.queue.length-1-t;e--)o+=this.queue[e];return 0===o?0:o/t}}function No(){const t=Math.floor(Date.now()/1e3);if(Game.T/Game.fps%1==0){let o=Game.cookies+Qt;Game.cpsSucked>0&&(o+=yt),Ut=Math.max(Game.cookiesEarned,o),o*=.05;const n=t-ye,i=Math.max(0,Game.cookies-be)/n,a=Math.max(0,yt-ke)/n,r=Math.max(0,kt[0]-Ge)/n,l=Math.max(0,o-ht)/n,s=(Game.cookieClicks-we)/n;for(let e=0;e<n;e++)ue.addLatest(i),me.addLatest(a),he.addLatest(r),ge.addLatest(l),fe.addLatest(s);ye=t,be=Game.cookies,ke=yt,Ge=kt[0],ht=o,we=Game.cookieClicks;const c=Y[e.AvgCPSHist];ve=ue.calcAverage(c),xe=me.calcAverage(c),Te=he.calcAverage(c),Me=ge.calcAverage(c),Se=ve,1===e.CalcWrink&&(Se+=xe),2===e.CalcWrink&&(Se+=Te);const d=Game.HasUnlocked("Chocolate egg")&&!Game.Has("Chocolate egg");Rt=d||0===e.CalcWrink?ve+xe+(d?Me:0):Se,xt=fe.calcAverage(K[e.AvgClicksHist])}}function Oo(t,o){let n="";return n=t.pp<=0||t.pp===1/0?A:t.pp<Vt?E:t.pp===Vt?N:t.pp<_t[10][0]?O:t.pp<_t[20][0]?F:t.pp<_t[30][0]?W:L,0!==Number(e.PPSecondsLowerLimit)&&o/so()<Number(e.PPSecondsLowerLimit)&&(n=E),e.PPOnlyConsiderBuyable&&o-Game.cookies>0&&(n=W),n}function Fo(t,o){Object.keys(t).forEach((n=>{if(e.PPRigidelMode&&1===o)t[n].color=A;else{t[n].color=Oo(t[n],Game.Objects[n].getSumPrice(o));for(let o=0;o<e.PPExcludeTop;o++)t[n].pp===_t[o][0]&&(t[n].color=A)}}))}function Wo(t,o){Object.keys(t).forEach((n=>{const i=Game.Objects[n].getSumPrice(o);Game.cookiesPs?t[n].pp=Math.max(i-(Game.cookies+co()),0)/Game.cookiesPs+i/t[n].bonus:t[n].pp=i/t[n].bonus,e.PPRigidelMode&&1===o||_t.push([t[n].pp,o])}))}function Lo(){!function(){Vt=1/0,_t=[],void 0===e.PPExcludeTop&&(e.PPExcludeTop=0),Wo(gt,1),Wo(ft,10),Wo(Ct,100),_t.sort(((e,t)=>e[0]-t[0]));let t=0;if(e.PPOnlyConsiderBuyable)for(;_t[t][1]>Game.cookies&&(t+=1,_t.length!==t+1););Vt=_t[e.PPExcludeTop][t],Fo(gt,1),Fo(ft,10),Fo(Ct,100)}(),Object.keys(vt).forEach((e=>{Game.cookiesPs?vt[e].pp=Math.max(Game.Upgrades[e].getPrice()-(Game.cookies+co()),0)/Game.cookiesPs+Game.Upgrades[e].getPrice()/vt[e].bonus:vt[e].pp=Game.Upgrades[e].getPrice()/vt[e].bonus,Number.isNaN(vt[e].pp)&&(vt[e].pp=1/0),vt[e].color=Oo(vt[e],Game.Upgrades[e].getPrice())}))}function Ao(e,t,o,n,i){let a=o,r=0;for(let o=0;o<i;o+=1){let o=t*Game.priceIncrease**Math.max(0,a-n);o=Game.modifyBuildingPrice(e,o),o=Math.ceil(o),r+=o,a+=1}return r}function jo(e){Jt[e]&&0===Jt[e].won&&(Jt[e].won=1,"shadow"!==Game.Achievements[e].pool&&(Le+=1))}function Do(){Ve=0;let e=1;const t={};Object.keys(Game.Objects).forEach((e=>{if(Game.Objects[e].minigameLoaded&&Game.Objects[e].minigame.effs){const o=Game.Objects[e].minigame.effs;Object.keys(o).forEach((e=>{t[e]?t[e]*=o[e]:t[e]=o[e]}))}})),ze=t,1!==Game.ascensionMode&&(e+=.01*parseFloat(je)*Ae*function(){let e=0;if(oo("Heavenly chip secret")&&(e+=.05),oo("Heavenly cookie stand")&&(e+=.2),oo("Heavenly bakery")&&(e+=.25),oo("Heavenly confectionery")&&(e+=.25),oo("Heavenly key")&&(e+=.25),e*=1+.05*Zt("Dragon God"),oo("Lucky digit")&&(e*=1.01),oo("Lucky number")&&(e*=1.01),oo("Lucky payout")&&(e*=1.01),Game.hasGod){const t=no("creation");1===t?e*=.7:2===t?e*=.8:3===t&&(e*=.9)}return e}()),e*=to("cps"),oo("Heralds")&&1!==Game.ascensionMode&&(e*=1+.01*Game.heralds),Object.keys(Game.cookieUpgrades).forEach((t=>{const o=Game.cookieUpgrades[t];oo(o.name)&&("function"==typeof o.power?e*=1+.01*Kt[o.name].power(Kt[o.name]):e*=1+.01*o.power)})),oo("Specialized chocolate chips")&&(e*=1.01),oo("Designer cocoa beans")&&(e*=1.02),oo("Underworld ovens")&&(e*=1.03),oo("Exotic nuts")&&(e*=1.04),oo("Arcane sugar")&&(e*=1.05),oo("Increased merriness")&&(e*=1.15),oo("Improved jolliness")&&(e*=1.15),oo("A lump of coal")&&(e*=1.01),oo("An itchy sweater")&&(e*=1.01),oo("Santa's dominion")&&(e*=1.2),oo("Fortune #100")&&(e*=1.01),oo("Fortune #101")&&(e*=1.07),oo("Dragon scale")&&(e*=1.03);let o=1;if(no){let t=no("asceticism");1===t?e*=1.15:2===t?e*=1.1:3===t&&(e*=1.05),t=no("ages"),1===t?e*=1+.15*Math.sin(Date.now()/1e3/10800*Math.PI*2):2===t?e*=1+.15*Math.sin(Date.now()/1e3/43200*Math.PI*2):3===t&&(e*=1+.15*Math.sin(Date.now()/1e3/86400*Math.PI*2)),t=no("decadence"),1===t?o*=.93:2===t?o*=.95:3===t&&(o*=.98),t=no("industry"),1===t?o*=1.1:2===t?o*=1.06:3===t&&(o*=1.03),t=no("labor"),1===t?o*=.97:2===t?o*=.98:3===t&&(o*=.99)}oo("Santa's legacy")&&(e*=1+.03*(Game.santaLevel+1));const n=Le/25;let i=1;if(oo("Santa's milk and cookies")&&(i*=1.05),i*=1+.05*Zt("Breath of Milk"),no){const e=no("mother");1===e?i*=1.1:2===e?i*=1.05:3===e&&(i*=1.03)}i*=to("milk");let a=1;oo("Kitten helpers")&&(a*=1+.1*n*i),oo("Kitten workers")&&(a*=1+.125*n*i),oo("Kitten engineers")&&(a*=1+.15*n*i),oo("Kitten overseers")&&(a*=1+.175*n*i),oo("Kitten managers")&&(a*=1+.2*n*i),oo("Kitten accountants")&&(a*=1+.2*n*i),oo("Kitten specialists")&&(a*=1+.2*n*i),oo("Kitten experts")&&(a*=1+.2*n*i),oo("Kitten consultants")&&(a*=1+.2*n*i),oo("Kitten assistants to the regional manager")&&(a*=1+.175*n*i),oo("Kitten marketeers")&&(a*=1+.15*n*i),oo("Kitten analysts")&&(a*=1+.125*n*i),oo("Kitten executives")&&(a*=1+.115*n*i),oo("Kitten angels")&&(a*=1+.1*n*i),oo("Fortune #103")&&(a*=1+.05*n*i),Object.keys(Yt).forEach((e=>{const t=Yt[e];let a=t.cps(t);1!==Game.ascensionMode&&(a*=(1+.01*t.level)*o),"Grandma"===t.name&&oo("Milkhelp&reg; lactose intolerance relief tablets")&&(a*=1+.05*n*i),Ve+=t.amount*a})),oo('"egg"')&&(Ve+=9),e*=a;let r=1;if(oo("Chicken egg")&&(r*=1.01),oo("Duck egg")&&(r*=1.01),oo("Turkey egg")&&(r*=1.01),oo("Quail egg")&&(r*=1.01),oo("Robin egg")&&(r*=1.01),oo("Ostrich egg")&&(r*=1.01),oo("Cassowary egg")&&(r*=1.01),oo("Salmon roe")&&(r*=1.01),oo("Frogspawn")&&(r*=1.01),oo("Shark egg")&&(r*=1.01),oo("Turtle egg")&&(r*=1.01),oo("Ant larva")&&(r*=1.01),oo("Century egg")){let e=10*Math.floor((Date.now()-Game.startDate)/1e3/10)/60/60/24;e=Math.min(e,100),ut=1+.1*(1-(1-e/100)**3),r*=ut}e*=r,oo("Sugar baking")&&(e*=1+.01*Math.min(100,Game.lumps)),e*=1+Zt("Radiant Appetite");const l=Ve*e;Object.keys(Game.CpsAchievements).forEach((e=>{l>=Game.CpsAchievements[e].threshold&&jo(Game.CpsAchievements[e].name)})),_e=l;const{n:s}=Game.shimmerTypes.golden,c=Zt("Dragon's Fortune");for(let t=0;t<s;t++)e*=1+1.23*c;const d=Game.bakeryName.toLowerCase();if("orteil"===d?e*=.99:"ortiel"===d&&(e*=.98),oo("Elder Covenant")&&(e*=.95),oo("Golden switch [off]")){let t=1.5;if(oo("Residual luck")){const e=Game.goldenCookieUpgrades;Object.keys(e).forEach((o=>{oo(e[o])&&(t+=.1)}))}e*=t}if(oo("Shimmering veil [off]")){let t=.5;oo("Reinforced membrane")&&(t+=.1),e*=1+t}oo("Magic shenanigans")&&(e*=1e3),oo("Occult obstruction")&&(e*=0),Ve=Game.runModHookOnValue("cps",Ve),e*=Mo(),Ve*=e}const Ho=["Fortune #001","Fortune #002","Fortune #003","Fortune #004","Fortune #005","Fortune #006","Fortune #007","Fortune #008","Fortune #009","Fortune #010","Fortune #011","Fortune #012","Fortune #013","Fortune #014","Fortune #015","Fortune #016","Fortune #017","Fortune #018","Fortune #100","Fortune #101","Fortune #102","Fortune #103","Fortune #104"],Uo=["Skull cookies","Ghost cookies","Bat cookies","Slime cookies","Pumpkin cookies","Eyeball cookies","Spider cookies"],Ro=["Christmas tree biscuits","Snowflake biscuits","Snowman biscuits","Holly biscuits","Candy cane biscuits","Bell biscuits","Present biscuits"],$o=["Pure heart biscuits","Ardent heart biscuits","Sour heart biscuits","Weeping heart biscuits","Golden heart biscuits","Eternal heart biscuits","Prism heart biscuits"],Io=["Elderwort biscuits","Bakeberry cookies","Duketater cookies","Green yeast digestives","Wheat slims","Fern tea","Ichor syrup"];function zo(e){return Jt[e]?Jt[e].won:0}function Vo(){let e=0;Object.keys(Game.GrandmaSynergies).forEach((t=>{oo(Game.GrandmaSynergies[t])&&(e+=1)})),!zo("Elder")&&e>=7&&jo("Elder"),!zo("Veteran")&&e>=14&&jo("Veteran");let t=0,o=1,n=1,i=1e5;Object.keys(Yt).forEach((e=>{t+=Yt[e].amount,i=Math.min(Yt[e].amount,i),zo("Mathematician")||Yt[e].amount<Math.min(128,2**(Game.ObjectsById.length-Game.Objects[e].id-1))&&(o=0),zo("Base 10")||Yt[e].amount<10*(Game.ObjectsById.length-Game.Objects[e].id)&&(n=0)})),i>=1&&jo("One with everything"),1===o&&jo("Mathematician"),1===n&&jo("Base 10"),i>=100&&jo("Centennial"),i>=150&&jo("Centennial and a half"),i>=200&&jo("Bicentennial"),i>=250&&jo("Bicentennial and a half"),i>=300&&jo("Tricentennial"),i>=350&&jo("Tricentennial and a half"),i>=400&&jo("Quadricentennial"),i>=450&&jo("Quadricentennial and a half"),i>=500&&jo("Quincentennial"),i>=550&&jo("Quincentennial and a half"),i>=600&&jo("Sexcentennial"),t>=100&&jo("Builder"),t>=500&&jo("Architect"),t>=1e3&&jo("Engineer"),t>=2e3&&jo("Lord of Constructs"),t>=4e3&&jo("Grand design"),t>=8e3&&jo("Ecumenopolis"),Fe>=20&&jo("Enhancer"),Fe>=50&&jo("Augmenter"),Fe>=100&&jo("Upgrader"),Fe>=200&&jo("Lord of Progress"),Fe>=300&&jo("The full picture"),Fe>=400&&jo("When there's nothing left to add"),t>=4e3&&Fe>=300&&jo("Polymath"),t>=8e3&&Fe>=400&&jo("Renaissance baker"),Yt.Cursor.amount+Yt.Grandma.amount>=777&&jo("The elder scrolls");let a=!0;Object.keys(Uo).forEach((e=>{oo(Uo[e])||(a=!1)})),a&&jo("Spooky cookies");let r=!0;if(Object.keys(Ro).forEach((e=>{oo(Ro[e])||(r=!1)})),r&&jo("Let it snow"),oo("Fortune cookies")){const e=Game.Tiers.fortune.upgrades;let t=0;Object.keys(e).forEach((o=>{oo(e[o].name)&&(t+=1)})),t>=e.length&&jo("O Fortuna")}}function _o(e){const t={};return Object.keys(Game.Objects).forEach((o=>{t[o]={},t[o].bonus=function(e,t){wo(),Yt[e].amount+=t;const o=Yt[e];"Cursor"===e?(o.amount>=1&&jo("Click"),o.amount>=2&&jo("Double-click"),o.amount>=50&&jo("Mouse wheel"),o.amount>=100&&jo("Of Mice and Men"),o.amount>=200&&jo("The Digital"),o.amount>=300&&jo("Extreme polydactyly"),o.amount>=400&&jo("Dr. T"),o.amount>=500&&jo("Thumbs, phalanges, metacarpals"),o.amount>=600&&jo("With her finger and her thumb"),o.amount>=700&&jo("Gotta hand it to you"),o.amount>=800&&jo("The devil's workshop")):Object.keys(Game.Objects[o.name].tieredAchievs).forEach((e=>{o.amount>=Game.Tiers[Game.Objects[o.name].tieredAchievs[e].tier].achievUnlock&&jo(Game.Objects[o.name].tieredAchievs[e].name)}));const n=Le;return Do(),Vo(),n!==Le&&Do(),Ve-Game.cookiesPs}(o,e),1!==e&&(Ee=1)})),t}function Xo(){Object.keys(Game.Objects).forEach((e=>{gt[e].price=Ao(Game.Objects[e],Game.Objects[e].basePrice,Game.Objects[e].amount,Game.Objects[e].free,1),ft[e].price=Ao(Game.Objects[e],Game.Objects[e].basePrice,Game.Objects[e].amount,Game.Objects[e].free,10),Ct[e].price=Ao(Game.Objects[e],Game.Objects[e].basePrice,Game.Objects[e].amount,Game.Objects[e].free,100)}))}function Qo(){gt=_o(1),ft=_o(10),Ct=_o(100),vt=[],Object.keys(Game.Upgrades).forEach((t=>{const o=function(e){if("toggle"===Game.Upgrades[e].pool||0===Game.Upgrades[e].bought&&Game.Upgrades[e].unlocked&&"prestige"!==Game.Upgrades[e].pool){wo(),"Shimmering veil [on]"===Kt[e].name?Kt["Shimmering veil [off]"].bought=0:"Golden switch [on]"===Kt[e].name?Kt["Golden switch [off]"].bought=0:Kt[e].bought=(Kt[e].bought+1)%2,Kt[e],Game.CountsAsUpgradeOwned(Game.Upgrades[e].pool)&&(Fe+=1),"Elder Pledge"===e?(We+=1,We>0&&jo("Elder nap"),We>=5&&jo("Elder slumber")):"Elder Covenant"===e?jo("Elder calm"):"Prism heart biscuits"===e?jo("Lovely cookies"):"Heavenly key"===e&&jo("Wholesome");const t=Le;Do(),Vo(),t!==Le&&Do();const o=function(){let e=0;oo("Thousand fingers")&&(e+=.1),oo("Million fingers")&&(e*=5),oo("Billion fingers")&&(e*=10),oo("Trillion fingers")&&(e*=20),oo("Quadrillion fingers")&&(e*=20),oo("Quintillion fingers")&&(e*=20),oo("Sextillion fingers")&&(e*=20),oo("Septillion fingers")&&(e*=20),oo("Octillion fingers")&&(e*=20),oo("Nonillion fingers")&&(e*=20);let t=0;Object.keys(Yt).forEach((e=>{t+=Yt[e].amount})),t-=Yt.Cursor.amount,e*=t,oo("Plastic mouse")&&(e+=.01*Ve),oo("Iron mouse")&&(e+=.01*Ve),oo("Titanium mouse")&&(e+=.01*Ve),oo("Adamantium mouse")&&(e+=.01*Ve),oo("Unobtainium mouse")&&(e+=.01*Ve),oo("Eludium mouse")&&(e+=.01*Ve),oo("Wishalloy mouse")&&(e+=.01*Ve),oo("Fantasteel mouse")&&(e+=.01*Ve),oo("Nevercrack mouse")&&(e+=.01*Ve),oo("Armythril mouse")&&(e+=.01*Ve),oo("Technobsidian mouse")&&(e+=.01*Ve),oo("Plasmarble mouse")&&(e+=.01*Ve),oo("Miraculite mouse")&&(e+=.01*Ve),oo("Fortune #104")&&(e+=.01*Ve);let o=1;if(oo("Santa's helpers")&&(o*=1.1),oo("Cookie egg")&&(o*=1.1),oo("Halo gloves")&&(o*=1.1),oo("Dragon claw")&&(o*=1.03),oo("Aura gloves")&&(o*=1+.05*Math.min(Game.Objects.Cursor.level,oo("Luminous gloves")?20:10)),o*=to("click"),Yt.Temple.minigameLoaded&&no){const e=no("labor");1===e?o*=1.15:2===e?o*=1.1:3===e&&(o*=1.05)}Object.keys(Game.buffs).forEach((e=>{void 0!==Game.buffs[e].multClick&&(o*=Game.buffs[e].multClick)})),o*=1+.05*Zt("Dragon Cursor");let n=o*Game.ComputeCps(1,oo("Reinforced index finger")+oo("Carpal tunnel prevention cream")+oo("Ambidextrous"),e);return n=Game.runModHookOnValue("cookiesPerClick",n),Game.hasBuff("Cursed finger")&&(n=Game.buffs["Cursed finger"].power),n}()-Game.computedMouseCps;return o?[Ve-Game.cookiesPs,o]:[Ve-Game.cookiesPs]}return[]}(t);"Elder Pledge"===t?(vt[t]={bonus:Game.cookiesPs-ve},1===e.CalcWrink?vt[t].bonus-=xe:2===e.CalcWrink&&(vt[t].bonus-=Te),Number.isFinite(vt[t].bonus)||(vt[t].bonus=0)):(vt[t]={},o[0]&&(vt[t].bonus=o[0]),o[1]&&(vt[t].bonusMouse=o[1]))}))}function qo(e,t,o){let n=0,i=0,a=0,r=0,l=1+Math.max(0,Math.ceil(Math.log(Game.cookies)/Math.LN10)-10);for(;a<t*o;)i=Math.max(e,Math.min(Math.floor(1/9*10**l*e*o),t*o)),a=Math.max(e,Math.min(Math.floor(1/9*10**(l+1)*e*o),t*o)),r=Math.floor(1/9*10**(l+1)*e*o),n+=i,l+=1;return[n,i,r]}function Yo(){let e=60*lt*60*6*st;const t=Mo();t>0?e/=t:e=0,Et=qo(7,e,ct),Pt=2*Et[1]/ct,Bt=Et[2]/60/60/6/st,Ft=qo(6,e,dt),Nt=2*Ft[1]/dt,Ot=Ft[2]/60/60/6/st,At=qo(7,7*e,ct),Wt=2*At[1]/ct,Lt=At[2]/60/60/6/st,Ht=qo(6,7*e,dt),jt=2*Ht[1]/dt,Dt=Ht[2]/60/60/6/st}function Ko(){const e=Math.floor(Date.now()/1e3);if(Game.T/Game.fps%1==0){const t=Game.HowMuchPrestige(Game.cookiesReset),o=Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned))-Math.floor(t),n=e-Pe,i=Math.max(0,o-Be)/n;for(let e=0;e<n;e++)Ce.addLatest(i);Pe=e,Be=o,pt=Ce.calcAverage(5)}}function Jo(){Tt="",St="",Mt="";const e=[];Object.keys(Game.Upgrades).forEach((t=>{e.push(Game.Upgrades[t])})),e.sort((function(e,t){return e.order>t.order?1:e.order<t.order?-1:0})),Object.keys(e).forEach((t=>{const o=e[t];if(0===o.bought){let e="";e+=function(e){let t="crate upgrade missing";"prestige"===e.pool&&(t+=" heavenly");let o=0;Game.prefs.crates||(o=1),o&&(t+=" noFrame");let{icon:n}=e;e.iconFunction&&(n=e.iconFunction());const i=`function() {return Game.crateTooltip(Game.UpgradesById[${e.id}], 'stats');}`;return`<div class="${t}"\n\t${Game.getDynamicTooltip(i,"top",!0)}\n\tstyle = "${n[2]?`background-image: url(${n[2]});`:""}background-position:${48*-n[0]}px ${48*-n[1]}px;">\n\t</div>`}(o),"prestige"===o.pool?Mt+=e:"cookie"===o.pool?St+=e:"toggle"!==o.pool&&"unused"!==o.pool&&"debug"!==o.pool&&(Tt+=e)}}))}function Zo(){if("christmas"===Game.season){let e=60*Game.cookiesPs;Game.hasBuff("Elder frenzy")&&(e*=.5),Game.hasBuff("Frenzy")&&(e*=.75),mt=Math.max(25,e),Game.Has("Ho ho ho-flavored frosting")&&(mt*=2)}}function en(){Ke=900*lt/.15,Ke*=st;const e=Mo();e>0?Ke/=e:Ke=0,Je=ct*(.15*Ke)+13,Ze=dt*(.15*Ke)+13,et=7*Ke,tt=ct*(.15*et)+13,ot=dt*(.15*et)+13,nt=2*Ke,it=.15*nt,at=0;let t=0,o=0;Object.keys(Game.Objects).forEach((e=>{Game.Objects[e].amount>t&&(t=Game.Objects[e].amount),Game.Objects[e].amount>0&&(o+=1)})),Object.keys(Game.Objects).forEach((e=>{(Game.Objects[e].amount<t||1===o)&&Game.Objects[e].amount<400&&2*Game.Objects[e].price>at&&(at=2*Game.Objects[e].price,rt=e)}))}function tn(){let e=1,t=1,o=1;oo("Green yeast digestives")&&(o*=1.01),oo("Dragon fang")&&(o*=1.03),e*=1+.1*Game.auraMult("Ancestral Metamorphosis"),e*=Game.eff("goldenCookieGain"),t*=1+.1*Game.auraMult("Unholy Dominion"),t*=Game.eff("wrathCookieGain"),ct=o*e,dt=o*t,st=1,0===Game.shimmerTypes.golden.n&&(st*=1+1.23*Game.auraMult("Dragon's Fortune"))}function on(){yt=0,bt=0,kt=[0,null];for(let e=0;e<Game.wrinklers.length;e++){let{sucked:t}=Game.wrinklers[e],o=1.1;if(Game.Has("Sacrilegious corruption")&&(o*=1.05),1===Game.wrinklers[e].type&&(o*=3),t*=o,Game.Has("Wrinklerspawn")&&(t*=1.05),Yt.Temple.minigameLoaded){const e=Game.hasGod("scorn");1===e?t*=1.15:2===e?t*=1.1:3===e&&(t*=1.05)}yt+=t,0===Game.wrinklers[e].type&&(bt+=t,t>kt[0]&&(kt=[t,e]))}}function nn(){Co(),on(),en(),tn(),Yo(),Jo(),Zo(),ue=new Eo(Y[Y.length-1]),me=new Eo(Y[Y.length-1]),he=new Eo(Y[Y.length-1]),ge=new Eo(Y[Y.length-1]),fe=new Eo(K[K.length-1]),Ce=new Eo(5),Ko(),No(),Qo(),Xo(),Lo()}const an={CPSMode:1,AvgCPSHist:3,AvgClicksHist:0,CalcWrink:0,Scale:2,ScaleDecimals:2,ScaleSeparator:0,ScaleCutoff:999999,TimeFormat:0,DetailedTime:1,PPDisplayTime:0,BuildColour:1,PPOnlyConsiderBuyable:0,PPExcludeTop:0,PPRigidelMode:0,PPSecondsLowerLimit:0,ColourBlue:"#4bb8f0",ColourGreen:"#00ff00",ColourYellow:"#ffff00",ColourOrange:"#ff7f00",ColourRed:"#ff0000",ColourPurple:"#ff00ff",ColourGray:"#b3b3b3",ColourPink:"#ff1493",ColourBrown:"#8b4513",BotBar:1,TimerBar:1,TimerBarPos:0,TimerBarOverlay:2,UpBarColour:1,UpgradeBarFixedPos:1,SortBuildings:0,SortUpgrades:0,DragonAuraInfo:1,GrimoireBar:1,GCTimer:1,Favicon:1,WrinklerButtons:1,HideSectionsButtons:0,TooltipBuildUpgrade:1,TooltipAmor:0,ToolWarnLucky:1,ToolWarnLuckyFrenzy:1,ToolWarnConjure:1,ToolWarnConjureFrenzy:1,ToolWarnEdifice:1,ToolWarnUser:0,ToolWarnBon:1,ToolWarnPos:1,TooltipGrim:1,TooltipWrink:1,TooltipLump:1,TooltipPlots:1,TooltipPantheon:1,TooltipAscendButton:1,Stats:1,MissingUpgrades:1,UpStats:1,HeavenlyChipsTarget:1,ShowMissedGC:1,Title:1,GeneralSound:1,GCNotification:0,GCFlash:1,ColourGCFlash:"#ffffff",GCSound:1,GCVolume:100,GCSoundURL:"https://freesound.org/data/previews/66/66717_931655-lq.mp3",FortuneNotification:0,FortuneFlash:1,ColourFortuneFlash:"#ffffff",FortuneSound:1,FortuneVolume:100,FortuneSoundURL:"https://freesound.org/data/previews/174/174027_3242494-lq.mp3",SeaNotification:0,SeaFlash:1,ColourSeaFlash:"#ffffff",SeaSound:1,SeaVolume:100,SeaSoundURL:"https://www.freesound.org/data/previews/121/121099_2193266-lq.mp3",GardFlash:1,ColourGardFlash:"#ffffff",GardSound:1,GardVolume:100,GardSoundURL:"https://freesound.org/data/previews/103/103046_861714-lq.mp3",MagicNotification:0,MagicFlash:1,ColourMagicFlash:"#ffffff",MagicSound:1,MagicVolume:100,MagicSoundURL:"https://freesound.org/data/previews/221/221683_1015240-lq.mp3",WrinklerNotification:0,WrinklerFlash:1,ColourWrinklerFlash:"#ffffff",WrinklerSound:1,WrinklerVolume:100,WrinklerSoundURL:"https://freesound.org/data/previews/124/124186_8043-lq.mp3",WrinklerMaxNotification:0,WrinklerMaxFlash:1,ColourWrinklerMaxFlash:"#ffffff",WrinklerMaxSound:1,WrinklerMaxVolume:100,WrinklerMaxSoundURL:"https://freesound.org/data/previews/152/152743_15663-lq.mp3",BulkBuyBlock:0,FavouriteSettings:1,Header:{Favourite:1,Calculation:1,Notation:1,Colours:1,BarsDisplay:1,Tooltip:1,Statistics:1,Notification:1,NotificationGeneral:1,NotificationGC:1,NotificationFC:1,NotificationSea:1,NotificationGard:1,NotificationMagi:1,NotificationWrink:1,NotificationWrinkMax:1,Miscellaneous:1,Lucky:1,Chain:1,Spells:1,Garden:1,Prestige:1,Wrink:1,Sea:1,Misc:1,InfoTab:1}},rn=function(e){if(1===e){const e=function(){try{Notification.requestPermission().then()}catch(e){return!1}return!0};"Notification"in window?e()?Notification.requestPermission().then():Notification.requestPermission():console.log("This browser does not support notifications.")}};function ln(){1===e.BotBar?(l("CMBotBar").style.display="",mo()):l("CMBotBar").style.display="none",go()}function sn(){BeautifyAll(),Game.RefreshStore(),Game.RebuildUpgrades(),mo(),ro(),lo()}function cn(){1===e.Favicon&&y>0?$t.wrath?l("CMFavicon").href="https://aktanusa.github.io/CookieMonster/favicon/wrathCookie.ico":l("CMFavicon").href="https://aktanusa.github.io/CookieMonster/favicon/goldenCookie.ico":l("CMFavicon").href="https://orteil.dashnet.org/cookieclicker/favicon.ico"}class dn{constructor(e,t){this.type=e,this.group=t}}class pn extends dn{constructor(e,t,o,n,i,a=null){super(e,t),this.label=o,this.desc=n,this.toggle=i,a&&(this.func=a)}}class un extends dn{constructor(e,t,o){super(e,t),this.desc=o}}class mn extends dn{constructor(e,t,o,n){super(e,t),this.label=o,this.desc=n;for(let e=0;e<101;e++)this.label[e]=`${e}%`}}class hn extends dn{constructor(e,t,o,n,i,a){super(e,t),this.label=o,this.desc=n,this.min=i,this.max=a}}const gn={CPSMode:new pn("bool","Calculation",["Current cookies per second","Average cookies per second"],"Calculate times using current cookies per second or average cookies per second",!1),AvgCPSHist:new pn("bool","Calculation",["Average CPS in past 10s","Average CPS in past 15s","Average CPS in past 30s","Average CPS in past 1m","Average CPS in past 5m","Average CPS in past 10m","Average CPS in past 15m","Average CPS in past 30m"],"How much time average Cookies Per Second should consider",!1),AvgClicksHist:new pn("bool","Calculation",["Average clicks in past 1s","Average clicks in past 5s","Average clicks in past 10s","Average clicks in past 15s","Average clicks in past 30s"],"How much time average Cookie Clicks should consider",!1),CalcWrink:new pn("bool","Calculation",["Calculate with wrinklers OFF","Calculate with wrinklers ON","Calculate with single fattest wrinkler ON"],"Calculate times and average Cookies Per Second with (only the single non-shiny fattest) wrinklers",!0,(function(){Ie=!0})),Scale:new pn("bool","Notation",["Game's setting scale","Metric","Short scale","Short scale (Abbreviated)","Scientific notation","Engineering notation"],"Change how long numbers are formatted",!1,(function(){sn()})),ScaleDecimals:new pn("bool","Notation",["1 decimals","2 decimals","3 decimals"],'Set the number of decimals used when applicable. This only works with Cookie Monster scales and not with "Game\'s Setting Scale"',!1,(function(){sn()})),ScaleSeparator:new pn("bool","Notation",[". for decimals (standard)",". for thousands"],"Set the separator used for decimals and thousands",!1,(function(){sn()})),ScaleCutoff:new hn("numscale","Notation","Notation cut-off point: ","The number from which Cookie Monster will start formatting numbers based on chosen scale. Standard is 999,999. Setting this above 999,999,999 might break certain notations",1,999999999),TimeFormat:new pn("bool","Notation",["Time XXd, XXh, XXm, XXs","Time XX:XX:XX:XX:XX"],"Change the time format",!1),DetailedTime:new pn("bool","Notation",["Detailed time OFF","Detailed time ON"],"Change how time is displayed in certain statistics and tooltips",!0,(function(){1===e.DetailedTime?Game.sayTime=I:Game.sayTime=S.sayTime})),PPDisplayTime:new pn("bool","Notation",["PP as value (standard)","PP as time unit"],"Display PP as calculated value or as approximate time unit. Note that PP does not translate directly into a time unit and this is therefore only an approximation.",!1),BuildColour:new pn("bool","Colours",["Building colours OFF","Building colours ON"],"Colour code buildings",!0,(function(){ro()})),PPOnlyConsiderBuyable:new pn("bool","Colours",["Don't ignore non-buyable","Ignore non-buyable"],"Makes Cookie Monster label buildings and upgrades you can't buy right now red, useful in those situations where you just want to spend your full bank 'most optimally'",!0),PPExcludeTop:new pn("bool","Colours",["Don't ignore any","Ignore 1st best","Ignore 1st and 2nd best","Ignore 1st, 2nd and 3rd best"],"Makes Cookie Monster ignore the 1st, 2nd or 3rd best buildings in labeling and colouring PP values",!0),PPRigidelMode:new pn("bool","Colours",["Rigidel mode OFF","Rigidel mode ON"],'Makes Cookie Monster ignore all "buy 1" options when colouring PP in order to stay at a total building count ending in 10 for pantheon god Rigidel',!0),PPSecondsLowerLimit:new hn("numscale","Colours","Lower limit for PP (in seconds): ",'If a building or upgrade costs less than the specified seconds of CPS it will also be considered optimal and label it as such ("PP is less than xx seconds of CPS"); setting to 0 ignores this option',0,1/0),ColourBlue:new un("colour","Colours","Standard colour is blue. Used to show upgrades better than best PP building, for Click Frenzy bar, and for various labels"),ColourGreen:new un("colour","Colours","Standard colour is green. Used to show best PP building, for Blood Frenzy bar, and for various labels"),ColourYellow:new un("colour","Colours","Standard colour is yellow. Used to show buildings within the top 10 of PP, for Frenzy bar, and for various labels"),ColourOrange:new un("colour","Colours","Standard colour is orange. Used to show buildings within the top 20 of PP, for Next Reindeer bar, and for various labels"),ColourRed:new un("colour","Colours","Standard colour is Red. Used to show buildings within the top 30 of PP, for Clot bar, and for various labels"),ColourPurple:new un("colour","Colours","Standard colour is purple. Used to show buildings outside of the top 30 of PP, for Next Cookie bar, and for various labels"),ColourGray:new un("colour","Colours","Standard colour is gray. Used to show negative or infinity PP, and for Next Cookie/Next Reindeer bar"),ColourPink:new un("colour","Colours","Standard colour is pink. Used for Dragonflight bar"),ColourBrown:new un("colour","Colours","Standard colour is brown. Used for Dragon Harvest bar"),BotBar:new pn("bool","BarsDisplay",["Bottom bar OFF","Bottom bar ON"],"Building information",!0,(function(){ln()})),TimerBar:new pn("bool","BarsDisplay",["Timer bar OFF","Timer bar ON"],"Bar with timers for golden cookie, season popup, Frenzy (Normal, Clot, Elder), Click Frenzy",!0,(function(){fo()})),TimerBarPos:new pn("bool","BarsDisplay",["Timer bar position (top left)","Timer bar position (bottom)"],"Placement of the timer bar",!1,(function(){0===e.TimerBarPos?(l("CMTimerBar").style.width="30%",l("CMTimerBar").style.bottom="",l("game").insertBefore(l("CMTimerBar"),l("sectionLeft"))):(l("CMTimerBar").style.width="100%",l("CMTimerBar").style.bottom="0px",l("wrapper").appendChild(l("CMTimerBar"))),go()})),TimerBarOverlay:new pn("bool","BarsDisplay",["Timer bar overlay OFF","Timer bar overlay only seconds","Timer bar overlay full"],"Overlay on timers displaying seconds and/or percentage left",!0),UpBarColour:new pn("bool","BarsDisplay",["Upgrade colours/bar OFF","Upgrade colours with bar ON","Upgrade colours without bar ON"],"Colour code upgrades and optionally add a counter bar",!1,(function(){1===e.UpBarColor?(l("CMUpgradeBar").style.display="",lo()):2===e.UpBarColor?(l("CMUpgradeBar").style.display="none",lo()):(l("CMUpgradeBar").style.display="none",Game.RebuildUpgrades())})),UpgradeBarFixedPos:new pn("bool","BarsDisplay",["Upgrade bar fixed position OFF","Upgrade bar fixed position ON"],"Lock the upgrade bar at top of the screen to prevent it from moving ofscreen when scrolling",!0,(function(){1===e.UpgradeBarFixedPos?(l("CMUpgradeBar").style.position="sticky",l("CMUpgradeBar").style.top="0px"):l("CMUpgradeBar").style.position=""})),SortBuildings:new pn("bool","BarsDisplay",["Sort buildings: default","Sort buildings: PP of x1 purchase","Sort buildings: PP of selected bulk mode"],"Sort the display of buildings in either default order or by PP",!1,(function(){ro()})),SortUpgrades:new pn("bool","BarsDisplay",["Sort upgrades: default","Sort upgrades: PP"],"Sort the display of upgrades in either default order or by PP",!1,(function(){lo()})),DragonAuraInfo:new pn("bool","BarsDisplay",["Extra dragon aura info OFF","Extra dragon aura info ON"],"Shows information about changes in CPS and costs in the dragon aura interface.",!0),GrimoireBar:new pn("bool","BarsDisplay",["Grimoire magic meter timer OFF","Grimoire magic meter timer ON"],"A timer on how long before the Grimoire magic meter is full",!0),GCTimer:new pn("bool","BarsDisplay",["Golden cookie timer OFF","Golden cookie timer ON"],"A timer on the golden cookie when it has been spawned",!0,(function(){1===e.GCTimer?Object.keys(re).forEach((e=>{re[e].style.display="block",re[e].style.left=Xt[e].l.style.left,re[e].style.top=Xt[e].l.style.top})):Object.keys(re).forEach((e=>re[e].style.display="none"))})),Favicon:new pn("bool","BarsDisplay",["Favicon OFF","Favicon ON"],"Update favicon with golden/wrath cookie",!0,(function(){cn()})),WrinklerButtons:new pn("bool","BarsDisplay",["Extra wrinkler buttons OFF","Extra wrinkler buttons ON"],"Show buttons for popping wrinklers at bottom of cookie section",!0,(function(){t()})),HideSectionsButtons:new pn("bool","BarsDisplay",["Hide buildings/upgrades button OFF","Hide buildings/upgrades button ON"],"Show buttons for hiding and showing the buildings and upgrades sections in the right column",!0,(function(){e.HideSectionsButtons?l("CMSectionHidButtons").style.display="":l("CMSectionHidButtons").style.display="none"})),TooltipBuildUpgrade:new pn("bool","Tooltip",["Building/upgrade tooltip information OFF","Building/upgrade  tooltip information ON"],"Extra information in building/upgrade tooltips",!0),TooltipAmor:new pn("bool","Tooltip",["Buildings tooltip amortization information OFF","Buildings tooltip amortization information ON"],"Add amortization information to buildings tooltip",!0),ToolWarnLucky:new pn("bool","Tooltip",["Tooltip lucky warning OFF","Tooltip lucky warning ON"],'A warning when buying if it will put the bank under the amount needed for max "Lucky!" rewards',!0),ToolWarnLuckyFrenzy:new pn("bool","Tooltip",["Tooltip lucky frenzy warning OFF","Tooltip lucky frenzy warning ON"],'A warning when buying if it will put the bank under the amount needed for max "Lucky!" (Frenzy) rewards',!0),ToolWarnConjure:new pn("bool","Tooltip",["Tooltip conjure warning OFF","Tooltip conjure warning ON"],'A warning when buying if it will put the bank under the amount needed for max "Conjure Baked Goods" rewards',!0),ToolWarnConjureFrenzy:new pn("bool","Tooltip",["Tooltip conjure frenzy warning OFF","Tooltip conjure frenzy warning ON"],'A warning when buying if it will put the bank under the amount needed for max "Conjure Baked Goods" rewards with Frenzy active',!0),ToolWarnEdifice:new pn("bool","Tooltip",["Tooltip edifice warning OFF","Tooltip edifice warning ON"],'A warning when buying if it will put the bank under the amount needed for "Spontaneous Edifice" to possibly give you your most expensive building',!0),ToolWarnUser:new hn("numscale","Tooltip","Tooltip warning at x times CPS: ","Use this to show a customized warning if buying it will put the bank under the amount equal to value times cps; setting to 0 disables the function altogether",0,1/0),ToolWarnBon:new pn("bool","Tooltip",["Calculate tooltip warning with bonus CPS OFF","Calculate tooltip warning with bonus CPS ON"],"Calculate the warning with or without the bonus CPS you get from buying",!0),ToolWarnPos:new pn("bool","Tooltip",["Tooltip warning position (left)","Tooltip warning position (bottom)"],"Placement of the warning boxes",!1,(function(){So()})),TooltipGrim:new pn("bool","Tooltip",["Grimoire tooltip information OFF","Grimoire tooltip information ON"],"Extra information in tooltip for grimoire",!0),TooltipWrink:new pn("bool","Tooltip",["Wrinkler tooltip OFF","Wrinkler tooltip ON"],"Shows the amount of cookies a wrinkler will give when popping it",!0),TooltipLump:new pn("bool","Tooltip",["Sugar lump tooltip OFF","Sugar lump tooltip ON"],"Shows the current Sugar Lump type in Sugar lump tooltip.",!0),TooltipPlots:new pn("bool","Tooltip",["Garden plots tooltip OFF","Garden plots tooltip ON"],"Shows a tooltip for plants that have a cookie reward.",!0),TooltipPantheon:new pn("bool","Tooltip",["Pantheon tooltip OFF","Pantheon tooltip ON"],"Shows additional info in the pantheon tooltip",!0),TooltipAscendButton:new pn("bool","Tooltip",["Show Extra Info Ascend Tooltip OFF","Show Extra Info Ascend Tooltip ON"],"Shows additional info in the ascend tooltip",!0),Stats:new pn("bool","Statistics",["Statistics OFF","Statistics ON"],"Extra Cookie Monster statistics!",!0),MissingUpgrades:new pn("bool","Statistics",["Missing upgrades OFF","Missing upgrades ON"],"Shows missing upgrades in stats menu",!0),UpStats:new pn("bool","Statistics",["Statistics update rate (default)","Statistics update rate (1s)"],"Default rate is once every 5 seconds",!1),HeavenlyChipsTarget:new hn("numscale","Statistics","Heavenly chips target: ",'Use this to set a heavenly chips target that will be counted towards in the "prestige" statsistics sections',1,1/0),ShowMissedGC:new pn("bool","Statistics",["Missed GC OFF","Missed GC ON"],"Show a stat in the statistics screen that counts how many golden cookies you have missed",!0),Title:new pn("bool","NotificationGeneral",["Title OFF","Title ON","Title pinned tab highlight"],'Update title with colden cookie/season popup timers; pinned tab highlight only changes the title when a golden cookie/season popup spawns; "!" means that golden cookie/reindeer can spawn',!0),GeneralSound:new pn("bool","NotificationGeneral",["Consider game volume setting OFF","Consider game volume setting ON"],'Turning this toggle to "off" makes Cookie Monster no longer consider the volume setting of the base game, allowing mod notifications to play with base game volume turned down',!0),GCNotification:new pn("bool","NotificationGC",["Notification OFF","Notification ON"],"Create a notification when golden cookie spawns",!0,(function(){rn(e.GCNotification)})),GCFlash:new pn("bool","NotificationGC",["Flash OFF","Flash ON"],"Flash screen on golden cookie",!0),ColourGCFlash:new un("colour","NotificationGC","The colour of the GC flash, standard colour is white"),GCSound:new pn("bool","NotificationGC",["Sound OFF","Sound ON"],"Play a sound on golden cookie",!0),GCVolume:new mn("vol","NotificationGC",[],"Volume"),GCSoundURL:new pn("url","NotificationGC","Sound URL:","URL of the sound to be played when a golden cookie spawns"),FortuneNotification:new pn("bool","NotificationFC",["Notification OFF","Notification ON"],"Create a notification when fortune cookie is on the ticker",!0,(function(){rn(e.FortuneNotification)})),FortuneFlash:new pn("bool","NotificationFC",["Flash OFF","Flash ON"],"Flash screen on fortune cookie spawn",!0),ColourFortuneFlash:new un("colour","NotificationFC","The colour of the fortune flash, standard colour is white"),FortuneSound:new pn("bool","NotificationFC",["Sound OFF","Sound ON"],"Play a sound on fortune cookie spawn",!0),FortuneVolume:new mn("vol","NotificationFC",[],"Volume"),FortuneSoundURL:new pn("url","NotificationFC","Sound URL:","URL of the sound to be played when the ticker has a fortune cookie"),SeaNotification:new pn("bool","NotificationSea",["Notification OFF","Notification ON"],"Create a notification on season popup",!0,(function(){rn(e.SeaNotification)})),SeaFlash:new pn("bool","NotificationSea",["Flash OFF","Flash ON"],"Flash screen on season popup",!0),ColourSeaFlash:new un("colour","NotificationSea","The colour of the season popup flash, standard colour is white"),SeaSound:new pn("bool","NotificationSea",["Sound OFF","Sound ON"],"Play a sound on season popup",!0),SeaVolume:new mn("vol","NotificationSea",[],"Volume"),SeaSoundURL:new pn("url","NotificationSea","Sound URL:","URL of the sound to be played when on season popup spawns"),GardFlash:new pn("bool","NotificationGard",["Garden Tick Flash OFF","Flash ON"],"Flash screen on garden tick",!0),ColourGardFlash:new un("colour","NotificationGard","The colour of the garden flash, standard colour is white"),GardSound:new pn("bool","NotificationGard",["Sound OFF","Sound ON"],"Play a sound on garden tick",!0),GardVolume:new mn("vol","NotificationGard",[],"Volume"),GardSoundURL:new pn("url","NotificationGard","Garden Tick Sound URL:","URL of the sound to be played when the garden ticks"),MagicNotification:new pn("bool","NotificationMagi",["Notification OFF","Notification ON"],"Create a notification when magic reaches maximum",!0,(function(){rn(e.MagicNotification)})),MagicFlash:new pn("bool","NotificationMagi",["Flash OFF","Flash ON"],"Flash screen when magic reaches maximum",!0),ColourMagicFlash:new un("colour","NotificationMagi","The colour of the magic flash, standard colour is white"),MagicSound:new pn("bool","NotificationMagi",["Sound OFF","Sound ON"],"Play a sound when magic reaches maximum",!0),MagicVolume:new mn("vol","NotificationMagi",[],"Volume"),MagicSoundURL:new pn("url","NotificationMagi","Sound URL:","URL of the sound to be played when magic reaches maxium"),WrinklerNotification:new pn("bool","NotificationWrink",["Notification OFF","Notification ON"],"Create a notification when a wrinkler appears",!0,(function(){rn(e.WrinklerNotification)})),WrinklerFlash:new pn("bool","NotificationWrink",["Flash OFF","Flash ON"],"Flash screen when a wrinkler appears",!0),ColourWrinklerFlash:new un("colour","NotificationWrink","The colour of the wrinkler flash, standard colour is white"),WrinklerSound:new pn("bool","NotificationWrink",["Sound OFF","Sound ON"],"Play a sound when a wrinkler appears",!0),WrinklerVolume:new mn("vol","NotificationWrink",[],"Volume"),WrinklerSoundURL:new pn("url","NotificationWrink","Sound URL:","URL of the sound to be played when a wrinkler appears"),WrinklerMaxNotification:new pn("bool","NotificationWrinkMax",["Notification OFF","Notification ON"],"Create a notification when the maximum amount of wrinklers has appeared",!0,(function(){rn(e.WrinklerMaxNotification)})),WrinklerMaxFlash:new pn("bool","NotificationWrinkMax",["Flash OFF","Flash ON"],"Flash screen when the maximum amount of Wrinklers has appeared",!0),ColourWrinklerMaxFlash:new un("colour","NotificationWrinkMax","The colour of the maximum wrinkler flash, standard colour is white"),WrinklerMaxSound:new pn("bool","NotificationWrinkMax",["Sound OFF","Sound ON"],"Play a sound when the maximum amount of wrinklers has appeared",!0),WrinklerMaxVolume:new mn("vol","NotificationWrinkMax",[],"Volume"),WrinklerMaxSoundURL:new pn("url","NotificationWrinkMax","Sound URL:","URL of the sound to be played when the maximum amount of wrinklers has appeared"),BulkBuyBlock:new pn("bool","Miscellaneous",["Block bulk buying OFF","Block bulk buying ON"],"Block clicking bulk buying when you can't buy all. This prevents buying 7 of a building when you are in buy-10 or buy-100 mode.",!0),FavouriteSettings:new pn("bool","Miscellaneous",["Favourite settings section OFF","Favourite settings section ON","Favourite settings section ON (Locked)"],"Show stars before each setting which allows selecting it for a 'favourites' section at the top of the Cookie Monster settings. Setting this to Locked removes the stars but shows the 'favourites' section",!0,(function(){Game.UpdateMenu()}))};function fn(){return JSON.stringify({favouriteSettings:se,settings:e,version:"2.031.6"})}function Cn(){if(qe!==Game.dragonLevel||Ie){if(Game.dragonLevel<25&&Game.dragonLevels[Game.dragonLevel].buy.toString().includes("sacrifice")){let e=Game.dragonLevels[Game.dragonLevel].buy.toString().match(/Objects\[(.*)\]/)[1];const t=Game.dragonLevels[Game.dragonLevel].buy.toString().match(/sacrifice\((.*?)\)/)[1];if("i"!==e)if(e=e.replaceAll("'",""),Game.Objects[e].amount<t)Ye="Not enough buildings to sell";else{let o=0;wo();for(let n=0;n<t;n++){let t=Yt[e].basePrice*Game.priceIncrease**Math.max(0,Yt[e].amount-1-Yt[e].free);t=Game.modifyBuildingPrice(Yt[e],t),t=Math.ceil(t),o+=t,Yt[e].amount-=1}Ye=`Cost to rebuy: ${ce(o)}`}else{let o=0;wo(),Object.keys(Game.Objects).forEach((n=>{if(e=n,Game.Objects[e].amount<t)Ye="Not enough buildings to sell";else{for(let n=0;n<t;n++){let t=Yt[e].basePrice*Game.priceIncrease**Math.max(0,Yt[e].amount-1-Yt[e].free);t=Game.modifyBuildingPrice(Yt[e],t),t=Math.ceil(t),o+=t,Yt[e].amount-=1}Ye=`Cost to rebuy: ${ce(o)}`}}))}}qe=Game.dragonLevel}}function yn(e,t){wo(),0===t?Ue=e:1===t?Re=e:2===t&&($e=e);const o=Le;return Do(),Vo(),o!==Le&&Do(),Ve-Game.cookiesPs}function bn(){Yt=[],Object.keys(Game.Objects).forEach((e=>{Yt[e]=ko(e)})),Kt=[],Object.keys(Game.Upgrades).forEach((e=>{Kt[e]=Go(e)})),Jt=[],Object.keys(Game.Achievements).forEach((e=>{Jt[e]=yo(e)})),wo()}let kn=!1;function Gn(t,o,n){(1===e[o]||n)&&3===t&&!1===kn||1===t?(l("CMFlashScreen").style.backgroundColor=e[`Colour${o}`],l("CMFlashScreen").style.opacity="0.5",3===t?(l("CMFlashScreen").style.display="inline",setTimeout((function(){Gn(2,o,!0)}),1e3/Game.fps)):setTimeout((function(){Gn(0,o,!0)}),1e3/Game.fps)):2===t?(l("CMFlashScreen").style.opacity="1",setTimeout((function(){Gn(1,o,!0)}),1e3/Game.fps)):0===t&&(l("CMFlashScreen").style.display="none")}function wn(t,o,n,i){if((1===e[o]||i)&&!1===kn){const o=new Audio(t);e.GeneralSound?o.volume=e[n]/100*(Game.volume/100):o.volume=e[n]/100,o.play()}}function vn(t,o,n){1===e[t]&&"hidden"===document.visibilityState&&!1===kn&&new Notification(o,{body:n,badge:"https://orteil.dashnet.org/cookieclicker/favicon.ico"})}function xn(){if(ie!==Game.OnAscend&&(ie=Game.OnAscend,Game.OnAscend?(l("game").style.bottom="0px",1===e.BotBar&&(l("CMBotBar").style.display="none"),1===e.TimerBar&&(l("CMTimerBar").style.display="none")):(ln(),fo()),ho()),!Game.OnAscend&&0===Game.AscendTimer){r!==Object.keys(Game.mods).length&&(bn(),nn(),r=Object.keys(Game.mods).length),Ie&&(Qo(),Game.Has("Golden switch [off]")?(wo(),Kt["Golden switch [off]"].bought=0,Do(),lt=Ve):lt=Game.cookiesPs,tn(),en(),Jo(),Yo(),Cn(),function(){for(let e=0;e<11;e+=1)for(let t=0;t<3;t+=1)qt[e][t]=yn(e,t)}(),Zo(),function(){let e=0;if(Game.Objects.Bank.minigameLoaded){const t=Game.Objects.Bank.minigame.goods;let o=0;Object.keys(t).forEach((e=>{const n=t[e];o+=n.stock*n.val})),e+=o*Game.cookiesPsRawHighest}e+=function(){let e=0;wo();let t=2;5!==De&&18!==De||(t-=1),5!==He&&18!==He||(t-=1),De=5,He=18;for(let e=0;e<t;++e){let e="Cursor";Object.keys(Yt).forEach((t=>{Yt[t].amount>0&&(e=t)})),Yt[e].amount-=1,Oe-=1}return Object.keys(Yt).forEach((t=>{const o=Yt[t];e+=ao(Game.Objects[o.name],Game.Objects[t].basePrice,o.amount,Game.Objects[t].free,o.amount)})),e}(),Qt=e}(),Ie=0);const t=Game.auraMult("Fierce Hoarder")>0;!Ne&&t?(Ne=!0,Ee=1):Ne&&!t&&(Ne=!1,Ee=1),Ee&&(Xo(),Ee=0),function(){on(),Lo(),function(){Gt=0;let e=0;Object.keys(Game.wrinklers).forEach((t=>{2===Game.wrinklers[t].phase&&(e+=1)}));let t=1;if(Yt.Temple.minigameLoaded){const e=Game.hasGod("scorn");1===e?t*=1.15:2===e?t*=1.1:3===e&&(t*=1.05)}wt=e,Gt=e*(.05*e*1.1)*(.05*Game.Has("Sacrilegious corruption")+1)*(.05*Game.Has("Wrinklerspawn")+1)*t}(),No(),Ko();const e=Game.HowManyCookiesReset(Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned))+1)-(Game.cookiesEarned+Game.cookiesReset);zt=de(e/so())}(),G=0,Xt={},Object.keys(Game.shimmers).forEach((e=>{Xt[Game.shimmers[e].id]=Game.shimmers[e],Game.shimmers[e].spawnLead&&"golden"===Game.shimmers[e].type&&($t=Game.shimmers[e],G+=1)})),Object.keys(re).forEach((e=>{void 0===Xt[e]&&(re[e].parentNode.removeChild(re[e]),delete re[e])})),y!==Game.shimmerTypes.golden.n?(y=Game.shimmerTypes.golden.n,y&&(b<G&&(Gn(3,"GCFlash",!1),wn(e.GCSoundURL,"GCSound","GCVolume",!1),vn("GCNotification","Golden Cookie Spawned","A Golden Cookie has spawned. Click it now!")),Object.keys(Game.shimmers).forEach((t=>{void 0===re[Game.shimmers[t].id]&&function(t){const o=document.createElement("div");o.id=`GCTimer${t.id}`,o.style.width="96px",o.style.height="96px",o.style.position="absolute",o.style.zIndex="10000000001",o.style.textAlign="center",o.style.lineHeight="96px",o.style.fontFamily='"Kavoon", Georgia, serif',o.style.fontSize="35px",o.style.cursor="pointer",o.style.display="block",0===e.GCTimer&&(o.style.display="none"),o.style.left=t.l.style.left,o.style.top=t.l.style.top,o.onclick=function(){t.pop()},o.onmouseover=function(){t.l.style.filter="brightness(125%) drop-shadow(0px 0px 3px rgba(255,255,255,1))",t.l.style.webkitFilter="brightness(125%) drop-shadow(0px 0px 3px rgba(255,255,255,1))"},o.onmouseout=function(){t.l.style.filter="",t.l.style.webkitFilter=""},re[t.id]=o,l("shimmers").appendChild(o)}(Game.shimmers[t])}))),cn(),b=G,0===G&&($t=0)):1===e.GCTimer&&y&&Object.keys(re).forEach((e=>{re[e].style.opacity=Xt[e].l.style.opacity,re[e].style.transform=Xt[e].l.style.transform,re[e].textContent=Math.ceil(Xt[e].life/Game.fps)})),f!==(Game.TickerEffect&&"fortune"===Game.TickerEffect.type)&&(f=Game.TickerEffect&&"fortune"===Game.TickerEffect.type,f&&(Gn(3,"FortuneFlash",!1),wn(e.FortuneSoundURL,"FortuneSound","FortuneVolume",!1),vn("FortuneNotification","Fortune Cookie found","A Fortune Cookie has appeared on the Ticker."))),g!==Game.shimmerTypes.reindeer.spawned&&(g=Game.shimmerTypes.reindeer.spawned,Object.keys(Game.shimmers).forEach((e=>{Game.shimmers[e].spawnLead&&"reindeer"===Game.shimmers[e].type&&(It=Game.shimmers[e])})),Gn(3,"SeaFlash",!1),wn(e.SeaSoundURL,"SeaSound","SeaVolume",!1),vn("SeaNotification","Reindeer sighted!","A Reindeer has spawned. Click it now!")),Game.Objects.Farm.minigameLoaded&&C!==Game.Objects.Farm.minigame.nextStep&&(0!==C&&C<Date.now()&&(Gn(3,"GardFlash",!1),wn(e.GardSoundURL,"GardSound","GardVolume",!1)),C=Game.Objects.Farm.minigame.nextStep),function(){if(Game.Objects["Wizard tower"].minigameLoaded&&1===e.GrimoireBar){const{minigame:t}=Game.Objects["Wizard tower"];t.magic<t.magicM?k=!1:k||(k=!0,Gn(3,"MagicFlash",!1),wn(e.MagicSoundURL,"MagicSound","MagicVolume",!1),vn("MagicNotification","Magic Meter full","Your Magic Meter is full. Cast a spell!"))}}(),function(){if(Game.elderWrath>0){let t=0;Object.keys(Game.wrinklers).forEach((e=>{2===Game.wrinklers[e].phase&&(t+=1)})),t>w?(w=t,t===Game.getWrinklersMax()&&e.WrinklerMaxFlash?Gn(3,"WrinklerMaxFlash",!1):Gn(3,"WrinklerFlash",!1),t===Game.getWrinklersMax()&&e.WrinklerMaxSound?wn(e.WrinklerMaxSoundURL,"WrinklerMaxSound","WrinklerMaxVolume",!1):wn(e.WrinklerSoundURL,"WrinklerSound","WrinklerVolume",!1),t===Game.getWrinklersMax()&&e.WrinklerMaxNotification?vn("WrinklerMaxNotification","Maximum Wrinklers Reached","You have reached your maximum ammount of wrinklers"):vn("WrinklerNotification","A Wrinkler appeared","A new wrinkler has appeared")):w=t}}()}}function Tn(){let t="";for(let o=0;o<H.length;o++)t+=`.CMText${H[o]} { color: ${e[`Colour${H[o]}`]}; }\n`;for(let o=0;o<H.length;o++)t+=`.CMBack${H[o]} { background-color: ${e[`Colour${H[o]}`]}; }\n`;for(let o=0;o<H.length;o++)t+=`.CMBorder${H[o]} { border: 1px solid ${e[`Colour${H[o]}`]}; }\n`;l("CMCSS").textContent=t,ro()}function Mn(){const e=b64_to_utf8(unescape(localStorage.getItem("CookieClickerGame")).split("!END!")[0]),t=e.match(/CookieMonster.*(;|$)/);if(null!==t){const o=e.replace(t[0],`CookieMonster:${fn()}`);localStorage.setItem("CookieClickerGame",escape(`${utf8_to_b64(o)}!END!`))}}function Sn(t){if(void 0!==localStorage.CMConfig&&delete localStorage.CMConfig,void 0!==t){e=t,void 0!==e.Colors&&delete e.Colors,void 0!==e.Colours&&delete e.Colours;let o=!1;Object.keys(an).forEach((t=>{void 0===e[t]?(o=!0,e[t]=an[t]):"Header"===t&&Object.keys(an.Header).forEach((n=>{void 0!==e[t][n]&&e[t][n]>-1&&e[t][n]<2||(o=!0,e[t][n]=an[t][n])}))})),o&&Mn(),xn(),Object.keys(an).forEach((e=>{"Header"!==e&&void 0!==gn[e].func&&gn[e].func()}))}else Sn(an);Game.UpdateMenu(),Tn()}function Pn(){Object.keys(Game.wrinklers).forEach((e=>{Game.wrinklers[e].sucked>0&&0===Game.wrinklers[e].type&&(Game.wrinklers[e].hp=0)}))}function Bn(t,o){if("b"===t){if(l("tooltip").innerHTML=Game.Objects[o].tooltip(),1===e.TooltipAmor){const e=Ao(Game.Objects[o],Game.Objects[o].basePrice,0,Game.Objects[o].free,Game.Objects[o].amount),t=e-Game.Objects[o].totalCookies;t>0&&(l("tooltip").innerHTML=l("tooltip").innerHTML.split("so far</div>").join(`so far<br/>&bull; <b>${ce(t)}</b> ${1===Math.floor(t)?"cookie":"cookies"} left to amortize (${pe((e-Game.Objects[o].totalCookies)/(Game.Objects[o].storedTotalCps*Game.globalCpsMult)).text})</div>`))}-1===Game.buyMode&&(l("tooltip").innerHTML=l("tooltip").innerHTML.split(ce(Game.Objects[o].bulkPrice)).join(ce((Game.Objects[o],Game.Objects[o].basePrice,Game.Objects[o].amount,Game.Objects[o].free,Game.buyBulk,1))))}else if("u"===t){if(!Game.UpgradesInStore[o])return"";l("tooltip").innerHTML=Game.crateTooltip(Game.UpgradesInStore[o],"store")}else"s"===t?l("tooltip").innerHTML=Game.lumpTooltip():"g"===t?l("tooltip").innerHTML=Game.Objects["Wizard tower"].minigame.spellTooltip(o)():"p"===t?l("tooltip").innerHTML=Game.ObjectsById[2].minigame.tileTooltip(o[0],o[1])():"ha"===t?l("tooltip").innerHTML=Game.ObjectsById[2].minigame.toolTooltip(1)():"wb"===t?l("tooltip").innerHTML="":"pag"===t?l("tooltip").innerHTML=Game.Objects.Temple.minigame.godTooltip(o)():"pas"===t&&(l("tooltip").innerHTML=Game.Objects.Temple.minigame.slotTooltip(o[0])());if("b"===t&&1===Game.buyMode||"u"===t||"s"===t||"g"===t||"p"===t&&!Game.keys[16]||"ha"===t||"wb"===t||"pag"===t||"pas"===t&&-1!==o[1]){const e=document.createElement("div");e.id="CMTooltipArea",l("tooltip").appendChild(e)}return te=t,oe=o,Po(),l("tooltip").innerHTML}function En(){Game.Objects["Wizard tower"].minigameLoaded&&Object.keys(Game.Objects["Wizard tower"].minigame.spellsById).forEach((e=>{null!==l(`grimoireSpell${e}`).onmouseover&&(T[e]=l(`grimoireSpell${e}`).onmouseover,l(`grimoireSpell${e}`).onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Bn("g",`${e}`)}),"this"),Game.tooltip.wobble()})}))}function Nn(){if(!m&&Game.Objects["Wizard tower"].minigameLoaded){const{minigame:t}=Game.Objects["Wizard tower"];p=t.draw,Game.Objects["Wizard tower"].minigame.draw=function(){p(),1===e.GrimoireBar&&t.magic<t.magicM&&(t.magicBarTextL.innerHTML+=` (${de(To(t.magic,t.magicM,t.magicM))})`)},m=!0}}function On(){!function(){if(!u&&Game.Objects["Wizard tower"].minigameLoaded){const{minigame:e}=Game.Objects["Wizard tower"];c=e.launch,d=new Function(`return ${e.launch.toString().split("=this").join("= Game.Objects['Wizard tower'].minigame")}`),Game.Objects["Wizard tower"].minigame.launch=function(){d(),En(),m=!1,Nn(),u=!0}}}(),Nn()}var Fn=o(877),Wn=o.n(Fn);const Ln="CMConfig";function An(t){null!==l(`slider${t}`)&&(l(`slider${t}right`).innerHTML=`${l(`slider${t}`).value}%`,e[t]=Math.round(l(`slider${t}`).value)),Mn()}function jn(t){e.Header[t]+=1,e.Header[t]>1&&(e.Header[t]=0),Mn()}function Dn(t,o){const n=document.createElement("div");n.className="title",n.style.padding="0px 16px",n.style.opacity="0.7",n.style.fontSize="17px",n.style.fontFamily='"Kavoon", Georgia, serif',n.appendChild(document.createTextNode(`${t} `));const i=document.createElement("span");return i.style.cursor="pointer",i.style.display="inline-block",i.style.height="14px",i.style.width="14px",i.style.borderRadius="7px",i.style.textAlign="center",i.style.backgroundColor="#C0C0C0",i.style.color="black",i.style.fontSize="13px",i.style.verticalAlign="middle",i.textContent=e.Header[o]?"-":"+",i.onclick=function(){jn(o),Game.UpdateMenu()},n.appendChild(i),n}function Hn(e,t,o,n){const i=document.createElement("div");i.className="listing";const a=document.createElement("b");if(a.textContent=t,i.appendChild(a),"withTooltip"===e){i.className="listing",i.appendChild(document.createTextNode(" "));const e=document.createElement("span");e.onmouseout=function(){Game.tooltip.hide()},e.onmouseover=function(){Game.tooltip.draw(this,escape(R[n].innerHTML))},e.style.cursor="default",e.style.display="inline-block",e.style.height="10px",e.style.width="10px",e.style.borderRadius="5px",e.style.textAlign="center",e.style.backgroundColor="#C0C0C0",e.style.color="black",e.style.fontSize="9px",e.style.verticalAlign="bottom",e.textContent="?",i.appendChild(e)}return i.appendChild(document.createTextNode(": ")),i.appendChild(o),i}function Un(e){const t=document.createDocumentFragment();t.appendChild(document.createTextNode(`${e.length} `));const o=document.createElement("span");o.onmouseout=function(){Game.tooltip.hide()};const n=document.createElement("div"),i=document.createElement("div");i.style.minWidth="140px",i.style.marginBottom="4px";const a=document.createElement("div");return a.className="name",a.style.marginBottom="4px",a.style.textAlign="center",a.textContent="Missing",i.appendChild(a),Object.keys(e).forEach((t=>{const o=document.createElement("div");o.style.textAlign="center",o.appendChild(document.createTextNode(e[t])),i.appendChild(o)})),n.appendChild(i),o.onmouseover=function(){Game.tooltip.draw(this,escape(n.innerHTML))},o.style.cursor="default",o.style.display="inline-block",o.style.height="10px",o.style.width="10px",o.style.borderRadius="5px",o.style.textAlign="center",o.style.backgroundColor="#C0C0C0",o.style.color="black",o.style.fontSize="9px",o.style.verticalAlign="bottom",o.textContent="?",t.appendChild(o),t}function Rn(t){const o=document.createElement("div");if(o.className="subsection",o.appendChild(t),o.appendChild(Dn("Lucky Cookies","Lucky")),e.Header.Lucky&&o.appendChild(function(){const e=Game.auraMult("Dragon's Fortune")?"GoldCookDragonsFortuneTooltipPlaceholder":"GoldCookTooltipPlaceholder",t=document.createElement("div");t.className="CMStatsLuckySection";const o=Game.cookies+co()<Ke?W:N,n=Game.cookies+co()<Ke?de((Ke-(Game.cookies+co()))/so()):"",i=document.createDocumentFragment(),a=document.createElement("span");if(a.style.fontWeight="bold",a.className=P+o,a.textContent=ce(Ke),i.appendChild(a),""!==n){const e=document.createElement("small");e.textContent=` (${n})`,i.appendChild(e)}t.appendChild(Hn("withTooltip",'"Lucky!" Cookies Required',i,e));const r=Game.cookies+co()<et?W:N,l=Game.cookies+co()<et?de((et-(Game.cookies+co()))/so()):"",s=document.createDocumentFragment(),c=document.createElement("span");if(c.style.fontWeight="bold",c.className=P+r,c.textContent=ce(et),s.appendChild(c),""!==l){const e=document.createElement("small");e.textContent=` (${l})`,s.appendChild(e)}t.appendChild(Hn("withTooltip",'"Lucky!" Cookies Required (Frenzy)',s,e));const d=Je!==Ze,p=document.createElement("span");p.style.fontWeight="bold",p.className=P+Je,p.textContent=ce(Je)+(d?` / ${ce(Ze)}`:""),t.appendChild(Hn("withTooltip",'"Lucky!" Reward (MAX)'+(d?" (Golden / Wrath)":""),p,e));const u=document.createElement("span");u.style.fontWeight="bold",u.className=P+u,u.textContent=ce(tt)+(d?` / ${ce(ot)}`:""),t.appendChild(Hn("withTooltip",'"Lucky!" Reward (MAX) (Frenzy)'+(d?" (Golden / Wrath)":""),u,e));const m=Math.min(.15*(Game.cookies+co()),lt*st*60*15)+13,h=document.createElement("span");return h.style.fontWeight="bold",h.className=P+h,h.textContent=ce(ct*m)+(d?` / ${ce(dt*m)}`:""),t.appendChild(Hn("withTooltip",'"Lucky!" Reward (CUR)'+(d?" (Golden / Wrath)":""),h,e)),t}()),o.appendChild(Dn("Chain Cookies","Chain")),e.Header.Chain&&o.appendChild(function(){const e=Game.auraMult("Dragon's Fortune")?"GoldCookDragonsFortuneTooltipPlaceholder":"GoldCookTooltipPlaceholder",t=document.createElement("div");t.className="CMStatsChainSection";const o=Game.cookies+co()<Pt?W:N,n=Game.cookies+co()<Pt?de((Pt-(Game.cookies+co()))/so()):"",i=document.createDocumentFragment(),a=document.createElement("span");if(a.style.fontWeight="bold",a.className=P+o,a.textContent=ce(Pt),i.appendChild(a),""!==n){const e=document.createElement("small");e.textContent=` (${n})`,i.appendChild(e)}t.appendChild(Hn("withTooltip",'"Chain" Cookies Required',i,e));const r=Game.cookies+co()<Nt?W:N,l=Game.cookies+co()<Nt?de((Nt-(Game.cookies+co()))/so()):"",s=document.createDocumentFragment(),c=document.createElement("span");if(c.style.fontWeight="bold",c.className=P+r,c.textContent=ce(Nt),s.appendChild(c),""!==l){const e=document.createElement("small");e.textContent=` (${l})`,s.appendChild(e)}t.appendChild(Hn("withTooltip",'"Chain" Cookies Required (Wrath)',s,e));const d=Game.cookies+co()<Wt?W:N,p=Game.cookies+co()<Wt?de((Wt-(Game.cookies+co()))/so()):"",u=document.createDocumentFragment(),m=document.createElement("span");if(m.style.fontWeight="bold",m.className=P+d,m.textContent=ce(Wt),u.appendChild(m),""!==p){const e=document.createElement("small");e.textContent=` (${p})`,u.appendChild(e)}t.appendChild(Hn("withTooltip",'"Chain" Cookies Required (Frenzy)',u,e));const h=Game.cookies+co()<jt?W:N,g=Game.cookies+co()<jt?de((jt-(Game.cookies+co()))/so()):"",f=document.createDocumentFragment(),C=document.createElement("span");if(C.style.fontWeight="bold",C.className=P+h,C.textContent=ce(jt),f.appendChild(C),""!==g){const e=document.createElement("small");e.textContent=` (${g})`,f.appendChild(e)}t.appendChild(Hn("withTooltip",'"Chain" Cookies Required (Frenzy) (Wrath)',f,e)),t.appendChild(Hn("withTooltip",'"Chain" Reward (MAX) (Golden / Wrath)',document.createTextNode(`${ce(Et[0])} / ${ce(Ft[0])}`),e)),t.appendChild(Hn("withTooltip",'"Chain" Reward (MAX) (Frenzy) (Golden / Wrath)',document.createTextNode(`${ce(At[0])} / ${ce(At[0])}`),e));const y=Math.min(60*Game.cookiesPs*60*6*st,.5*Game.cookies),b=qo(7,y,ct)[0],k=qo(6,y,dt)[0];return t.appendChild(Hn("withTooltip",'"Chain" Reward (CUR) (Golden / Wrath)',document.createTextNode(`${ce(b)} / ${ce(k)}`),e)),t.appendChild(Hn("withTooltip","CPS Needed For Next Level (G / W)",document.createTextNode(`${ce(Bt)} / ${ce(Ot)}`),"ChainNextLevelPlaceholder")),t.appendChild(Hn("withTooltip","CPS Needed For Next Level (Frenzy) (G / W)",document.createTextNode(`${ce(Lt)} / ${ce(Dt)}`),"ChainNextLevelPlaceholder")),t}()),Game.Objects["Wizard tower"].minigameLoaded&&(o.appendChild(Dn("Spells","Spells")),e.Header.Spells&&o.appendChild(function(){const e=document.createElement("div");e.className="CMStatsSpellsSection";const t=Game.cookies+co()<nt?W:N,o=Game.cookies+co()<nt?de((nt-(Game.cookies+co()))/so()):"",n=document.createDocumentFragment(),i=document.createElement("span");if(i.style.fontWeight="bold",i.className=P+t,i.textContent=ce(nt),n.appendChild(i),""!==o){const e=document.createElement("small");e.textContent=` (${o})`,n.appendChild(e)}e.appendChild(Hn("withTooltip",'"Conjure Baked Goods" Cookies Required',n,"GoldCookTooltipPlaceholder")),e.appendChild(Hn("withTooltip",'"Conjure Baked Goods" Reward (MAX)',document.createTextNode(ce(it)),"GoldCookTooltipPlaceholder"));const a=Game.cookies+co()<7*nt?W:N,r=Math.min(.15*(Game.cookies+co()),60*lt*30),l=Game.cookies+co()<7*nt?de((7*nt-(Game.cookies+co()))/so()):"",s=document.createDocumentFragment(),c=document.createElement("span");if(c.style.fontWeight="bold",c.className=P+a,c.textContent=ce(7*nt),s.appendChild(c),""!==l){const e=document.createElement("small");e.textContent=` (${l})`,s.appendChild(e)}return e.appendChild(Hn("withTooltip",'"Conjure Baked Goods" Cookies Required (Frenzy)',s,"GoldCookTooltipPlaceholder")),e.appendChild(Hn("withTooltip",'"Conjure Baked Goods" Reward (MAX) (Frenzy)',document.createTextNode(ce(7*it)),"GoldCookTooltipPlaceholder")),e.appendChild(Hn("withTooltip",'"Conjure Baked Goods" Reward (CUR)',document.createTextNode(ce(r)),"GoldCookTooltipPlaceholder")),at&&e.appendChild(Hn("withTooltip",'"Spontaneous Edifice" Cookies Required (most expensive building)',document.createTextNode(`${ce(at)} (${rt})`),"GoldCookTooltipPlaceholder")),e}())),Game.Objects.Farm.minigameLoaded&&(o.appendChild(Dn("Garden","Garden")),e.Header.Garden&&o.appendChild(function(){const e=document.createElement("div");e.className="CMStatsGardenSection";const t=Game.cookies<60*Game.cookiesPs*10*100?W:N,o=document.createElement("span");o.style.fontWeight="bold",o.className=P+t,o.textContent=ce(60*Game.cookiesPs*10*100),e.appendChild(Hn("basic","Cookies required for max reward of Bakeberry: ",o));const n=Game.cookies<60*Game.cookiesPs*100?W:N,i=document.createElement("span");i.style.fontWeight="bold",i.className=P+n,i.textContent=ce(60*Game.cookiesPs*100),e.appendChild(Hn("basic","Cookies required for max reward of Chocoroot: ",i));const a=Game.cookies<60*Game.cookiesPs*60*25?W:N,r=document.createElement("span");r.style.fontWeight="bold",r.className=P+a,r.textContent=ce(60*Game.cookiesPs*60*25),e.appendChild(Hn("basic","Cookies required for max reward of Queenbeet: ",r));const l=Game.cookies<60*Game.cookiesPs*15*100?W:N,s=document.createElement("span");s.style.fontWeight="bold",s.className=P+l,s.textContent=ce(60*Game.cookiesPs*15*100),e.appendChild(Hn("basic","Cookies required for max reward of Duketater: ",s));const c=[];return Object.keys(Io).forEach((e=>{Game.HasUnlocked(Io[e])||c.push(Io[e])})),0!==c.length&&e.appendChild(Hn("basic","Rare plant drops left to unlock",Un(c))),e}())),o.appendChild(Dn("Prestige","Prestige")),e.Header.Prestige&&o.appendChild(function(){const t=document.createElement("div");t.className="CMStatsPrestigeSection";const o=Math.floor(Game.HowMuchPrestige(Ut+Game.cookiesReset+yt+(Game.HasUnlocked("Chocolate egg")&&!Game.Has("Chocolate egg")?ht:0)));t.appendChild(Hn("withTooltip","Prestige Level (CUR / MAX)",document.createTextNode(`${ce(Game.prestige)} / ${ce(o)}`),"PrestMaxTooltipPlaceholder"));const n=Math.max(0,Game.HowManyCookiesReset(o+1)-(Ut+Game.cookiesReset+yt+(Game.HasUnlocked("Chocolate egg")&&!Game.Has("Chocolate egg")&&ht?ht:0))),i=document.createDocumentFragment();i.appendChild(document.createTextNode(ce(n)));const a=document.createElement("small");a.textContent=` (${de(n/Rt,1)})`,i.appendChild(a),t.appendChild(Hn("withTooltip","Cookies To Next Level",i,"NextPrestTooltipPlaceholder")),t.appendChild(Hn("withTooltip","Heavenly Chips (CUR / MAX)",document.createTextNode(`${ce(Game.heavenlyChips)} / ${ce(o-Game.prestige+Game.heavenlyChips)}`),"HeavenChipMaxTooltipPlaceholder")),t.appendChild(Hn("basic","Heavenly Chips Per Second (last 5 seconds)",document.createTextNode(ce(pt,2))));const r=Number(e.HeavenlyChipsTarget);if(!Number.isNaN(r)){const e=r-Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned));e>0&&(t.appendChild(Hn("basic","Heavenly Chips To Target Set In Settings (CUR)",document.createTextNode(ce(e)))),t.appendChild(Hn("basic","Time To Target (CUR, Current 5 Second Average)",document.createTextNode(de(e/pt)))))}const l=function(e){let t=Game.cookiesPs;wo(),0===Kt["Heavenly key"].bought&&(Kt["Heavenly chip secret"].bought=1,Kt["Heavenly cookie stand"].bought=1,Kt["Heavenly bakery"].bought=1,Kt["Heavenly confectionery"].bought=1,Kt["Heavenly key"].bought=1,Do(),t=Ve,wo()),Ut>=1e6&&jo("Sacrifice"),Ut>=1e9&&jo("Oblivion"),Ut>=1e12&&jo("From scratch"),Ut>=1e15&&jo("Nihilism"),Ut>=1e18&&jo("Dematerialize"),Ut>=1e21&&jo("Nil zero zilch"),Ut>=1e24&&jo("Transcendence"),Ut>=1e27&&jo("Obliterate"),Ut>=1e30&&jo("Negative void"),Ut>=1e33&&jo("To crumbs, you say?"),Ut>=1e36&&jo("You get nothing"),Ut>=1e39&&jo("Humble rebeginnings"),Ut>=1e42&&jo("The end of the world"),Ut>=1e45&&jo("Oh, you're back"),Ut>=1e48&&jo("Lazarus"),Ut>=1e51&&jo("Smurf account"),Ut>=1e54&&jo("If at first you don't succeed"),Kt["Heavenly chip secret"].bought=1,Kt["Heavenly cookie stand"].bought=1,Kt["Heavenly bakery"].bought=1,Kt["Heavenly confectionery"].bought=1,Kt["Heavenly key"].bought=1,je=e;const o=Le;Do(),Vo(),o!==Le&&Do();const n=Ve-t;return je=Game.prestige,n}(o),s=document.createDocumentFragment();s.appendChild(document.createTextNode(ce(l)));const c=Math.round(l/Game.cookiesPs*1e4);if(Number.isFinite(c)&&0!==c){const e=document.createElement("small");e.textContent=` (${c/100}% of income)`,s.appendChild(e)}t.appendChild(Hn("withTooltip","Reset Bonus Income",s,"ResetTooltipPlaceholder"));const d=Math.floor(Game.HowMuchPrestige(Game.cookiesReset)),p=Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned)),u=p-d;if(!Game.Has("Lucky digit")){let e=7-p%10;e<0&&(e+=10);const o=u+e,n=p+e,i=document.createDocumentFragment();i.appendChild(document.createTextNode(`${n.toLocaleString()} / ${o.toLocaleString()} (+${e})`)),t.appendChild(Hn("basic",'Next "Lucky Digit" (total / reset)',i))}if(!Game.Has("Lucky number")){let e=777-p%1e3;e<0&&(e+=1e3);const o=u+e,n=p+e,i=document.createDocumentFragment();i.appendChild(document.createTextNode(`${n.toLocaleString()} / ${o.toLocaleString()} (+${e})`)),t.appendChild(Hn("basic",'Next "Lucky Number" (total / reset)',i))}if(!Game.Has("Lucky payout")){let e=777777-p%1e6;e<0&&(e+=1e6);const o=u+e,n=p+e,i=document.createDocumentFragment();i.appendChild(document.createTextNode(`${n.toLocaleString()} / ${o.toLocaleString()} (+${e})`)),t.appendChild(Hn("basic",'Next "Lucky Payout" (total / reset)',i))}return t}()),Game.cpsSucked>0&&(o.appendChild(Dn("Wrinklers","Wrink")),e.Header.Wrink)){const e=document.createDocumentFragment();e.appendChild(document.createTextNode(`${ce(yt)} / ${ce(bt)} `));const t=document.createElement("a");t.textContent="Pop All Normal",t.className="option",t.onclick=function(){Pn()},e.appendChild(t),o.appendChild(Hn("basic","Rewards of Popping (All/Normal)",e));const n=document.createDocumentFragment();n.appendChild(document.createTextNode(`${ce(kt[0])} `));const i=document.createElement("a");i.textContent="Pop Single Fattest",i.className="option",i.onclick=function(){null!==kt[1]&&(Game.wrinklers[kt[1]].hp=0)},n.appendChild(i),o.appendChild(Hn("basic",`Rewards of Popping Single Fattest Non-Shiny Wrinkler (id: ${null!==kt[1]?kt[1]:"None"})`,n))}if(o.appendChild(function(){const t=document.createElement("div");t.className="CMStatsSeasonSection";let o=!1;const n=[];Object.keys(Uo).forEach((e=>{Game.Has(Uo[e])||(n.push(Uo[e]),o=!0)}));const i=[];Object.keys(Ro).forEach((e=>{Game.Has(Ro[e])||(i.push(Ro[e]),o=!0)}));const a=[];Object.keys($o).forEach((e=>{Game.Has($o[e])||(a.push($o[e]),o=!0)}));const r=[];Object.keys(Game.eggDrops).forEach((e=>{Game.HasUnlocked(Game.eggDrops[e])||(r.push(Game.eggDrops[e]),o=!0)}));const l=[];Object.keys(Game.rareEggDrops).forEach((e=>{Game.HasUnlocked(Game.rareEggDrops[e])||(l.push(Game.rareEggDrops[e]),o=!0)}));const s=Game.HasUnlocked("Chocolate egg")&&!Game.Has("Chocolate egg"),c=Game.Has("Century egg");if(("christmas"===Game.season||o||s||c)&&(t.appendChild(Dn("Season Specials","Sea")),e.Header.Sea)){if(0!==n.length){t.appendChild(Hn("basic","Halloween cookies left to buy",Un(n)));let e=.95;if(Game.HasAchiev("Spooky cookies")&&(e=.8),Game.Has("Starterror")&&(e*=.9),e*=1/Game.dropRateMult(),Game.hasGod){const t=Game.hasGod("seasons");1===t?e*=.9:2===t?e*=.95:3===t&&(e*=.97)}const o=n.length/7;t.appendChild(Hn("basic","Chance of receiving a cookie from wrinkler/shiny wrinkler",document.createTextNode(`${ce((1-e)*o*100)}% / ${ce((1-.9*e)*o*100)}%`)))}if(0!==i.length){t.appendChild(Hn("basic","Christmas cookies left to buy",Un(i)));let e=.8;if(Game.HasAchiev("Let it snow")&&(e=.6),e*=1/Game.dropRateMult(),Game.Has("Starsnow")&&(e*=.95),Game.hasGod){const t=Game.hasGod("seasons");1===t?e*=.9:2===t?e*=.95:3===t&&(e*=.97)}const o=i.length/7;t.appendChild(Hn("basic","Chance of receiving a cookie from reindeer",document.createTextNode(`${ce((1-e)*o*100)}%`)))}0!==a.length&&t.appendChild(Hn("basic","Valentine cookies left to buy",Un(a)));const e=function(e){let t=e*(1/Game.dropRateMult());if(Game.HasAchiev("Hide & seek champion")&&(t*=.7),Game.Has("Omelette")&&(t*=.9),Game.Has("Starspawn")&&(t*=.9),Game.hasGod){const e=Game.hasGod("seasons");1===e?t*=.9:2===e?t*=.95:3===e&&(t*=.97)}const o=1-t,n=Game.eggDrops.length-r.length,i=Game.rareEggDrops.length-l.length,a=.9*o*(1-n/Game.eggDrops.length),s=.1*o*(1-i/Game.rareEggDrops.length),c=.9*o*(n/Game.eggDrops.length)+.1*o*(i/Game.rareEggDrops.length);return[a+.9*c*(1-n/Game.eggDrops.length),s+.1*c*(1-i/Game.rareEggDrops.length)]};0!==r.length&&(t.appendChild(Hn("basic","Normal easter eggs left to unlock",Un(r))),t.appendChild(Hn("basic","Chance of receiving an egg from wrinkler/golden cookie",document.createTextNode(`${ce(100*e(.98)[0])}% / ${ce(100*e(.9)[0])}%`)))),0!==l.length&&(t.appendChild(Hn("basic","Rare easter eggs left to unlock",Un(l))),t.appendChild(Hn("basic","Chance of receiving a rare egg from wrinkler/golden cookie",document.createTextNode(`${ce(100*e(.98)[1])}% / ${ce(100*e(.9)[1])}%`)))),"christmas"===Game.season&&t.appendChild(Hn("basic","Reindeer reward",document.createTextNode(ce(mt)))),s&&t.appendChild(Hn("withTooltip","Chocolate egg cookies",document.createTextNode(ce(ht)),"ChoEggTooltipPlaceholder")),c&&t.appendChild(Hn("basic","Century egg multiplier",document.createTextNode(Math.round(1e4*(ut-1))/100+"%")))}return t}()),o.appendChild(Dn("Miscellaneous","Misc")),e.Header.Misc){if(o.appendChild(Hn("basic",`Average Cookies Per Second (Past ${Y[e.AvgCPSHist]<60?`${Y[e.AvgCPSHist]} seconds`:Y[e.AvgCPSHist]/60+(3===e.AvgCPSHist?" minute":" minutes")})`,document.createTextNode(ce(so(),3)))),o.appendChild(Hn("basic",`Average Cookie Clicks Per Second (Past ${K[e.AvgClicksHist]}${0===e.AvgClicksHist?" second":" seconds"})`,document.createTextNode(ce(xt,1)))),Game.Has("Fortune cookies")){const e=[];Object.keys(Ho).forEach((t=>{Game.Has(Ho[t])||e.push(Ho[t])})),0!==e.length&&o.appendChild(Hn("basic","Fortune Upgrades Left to Buy",Un(e)))}if(e.ShowMissedGC&&o.appendChild(Hn("basic","Missed Golden Cookies",document.createTextNode(ce(Game.missedGoldenClicks)))),Game.prefs.autosave){const e=document.createElement("span");e.id="CMStatsAutosaveTimer",e.innerText=Game.sayTime(60*Game.fps-(Game.OnAscend?0:Game.T%(60*Game.fps)),4),o.appendChild(Hn("basic","Time till autosave",e))}}l("menu").insertBefore(o,l("menu").childNodes[2]),e.MissingUpgrades&&l("menu").childNodes.forEach((e=>{if(e.children[0])if("Prestige"===e.children[0].innerHTML&&Mt){const t=Mt.match(new RegExp("div","g")||0).length/2,o=document.createElement("div");o.id="CMMissingUpgradesPrestigeTitle",o.className="listing";const n=document.createElement("div");n.innerHTML=`<b>Missing Prestige upgrades:</b> ${t}/${Game.PrestigeUpgrades.length} (${Math.floor(t/Game.PrestigeUpgrades.length*100)}%)`,o.appendChild(n),e.appendChild(o);const i=document.createElement("div");i.className="listing crateBox",i.innerHTML=Mt,e.appendChild(i)}else if("Upgrades"===e.children[0].innerHTML){if(Tt){const t=Tt.match(new RegExp("div","g")||0).length/2,o=document.createElement("div");o.id="CMMissingUpgradesTitle",o.className="listing";const n=document.createElement("div");n.innerHTML=`<b>Missing normal upgrades:</b> ${t}/${Game.UpgradesByPool[""].length+Game.UpgradesByPool.tech.length} (${Math.floor(t/(Game.UpgradesByPool[""].length+Game.UpgradesByPool.tech.length)*100)}%)`,o.appendChild(n),e.insertBefore(o,e.childNodes[3]);const i=document.createElement("div");i.className="listing crateBox",i.innerHTML=Tt,e.insertBefore(i,document.getElementById("CMMissingUpgradesTitle").nextSibling)}if(St){const t=St.match(new RegExp("div","g")||0).length/2,o=document.createElement("div");o.id="CMMissingUpgradesCookiesTitle",o.className="listing";const n=document.createElement("div");n.innerHTML=`<b>Missing Cookie upgrades:</b> ${t}/${Game.UpgradesByPool.cookie.length} (${Math.floor(t/Game.UpgradesByPool.cookie.length*100)}%)`,o.appendChild(n),e.appendChild(o);const i=document.createElement("div");i.className="listing crateBox",i.innerHTML=St,e.appendChild(i)}}}))}const $n={Favourite:"Favourite Settings",Calculation:"Calculation",Notation:"Notation",Colours:"Colours and colour coding",BarsDisplay:"Infobars and visual settings",Tooltip:"Tooltips",Statistics:"Statistics",Notification:"Notifications",Miscellaneous:"Miscellaneous"},In={NotificationGeneral:"General Notifications",NotificationGC:"Golden Cookie",NotificationFC:"Fortune Cookie",NotificationSea:"Season Special",NotificationGard:"Garden Tick",NotificationMagi:"Full Magic Bar",NotificationWrink:"Wrinkler",NotificationWrinkMax:"Maximum Wrinklers"};function zn(t,o){const n=document.createElement("div");n.className="title",n.style.opacity="0.7",n.style.fontSize="17px",n.appendChild(document.createTextNode(`${o} `));const i=document.createElement("span");return i.style.cursor="pointer",i.style.display="inline-block",i.style.height="14px",i.style.width="14px",i.style.borderRadius="7px",i.style.textAlign="center",i.style.backgroundColor="#C0C0C0",i.style.color="black",i.style.fontSize="13px",i.style.verticalAlign="middle",i.textContent=e.Header[t]?"-":"+",i.onclick=function(){jn(t),Game.UpdateMenu()},n.appendChild(i),n}function Vn(t){const o=document.createElement("div");if(o.className="listing",1===e.FavouriteSettings){const e=document.createElement("a");se.includes(t)?e.innerText="★":e.innerText="☆",e.className="option",e.onclick=function(){!function(e){se.includes(e)?se=se.filter((function(t){return t!==e})):se.push(e)}(t),Mn(),Game.UpdateMenu()},o.appendChild(e),o.appendChild(document.createTextNode(" "))}if("bool"===gn[t].type){const n=document.createElement("a");gn[t].toggle&&0===e[t]?n.className="option off":n.className="option",n.id=Ln+t,n.onclick=function(){!function(t){e[t]+=1,e[t]===gn[t].label.length?(e[t]=0,gn[t].toggle&&(l(Ln+t).className="option off")):l(Ln+t).className="option",void 0!==gn[t].func&&gn[t].func(),l(Ln+t).innerHTML=gn[t].label[e[t]],Mn()}(t)},n.textContent=gn[t].label[e[t]],o.appendChild(n);const i=document.createElement("label");return i.textContent=gn[t].desc,o.appendChild(i),o}if("vol"===gn[t].type){const n=document.createElement("div");n.className="sliderBox";const i=document.createElement("div");i.style.float="left",i.innerHTML=gn[t].desc,n.appendChild(i);const a=document.createElement("div");a.id=`slider${t}right`,a.style.float="right",a.innerHTML=`${e[t]}%`,n.appendChild(a);const r=document.createElement("input");r.className="slider",r.id=`slider${t}`,r.style.clear="both",r.type="range",r.min="0",r.max="100",r.step="1",r.value=e[t],r.oninput=function(){An(t)},r.onchange=function(){An(t)},n.appendChild(r),o.appendChild(n);const l=document.createElement("a");return l.className="option",l.onclick=function(){wn(e[t.replace("Volume","SoundURL")],t.replace("Volume","Sound"),t,!0)},l.textContent="Test sound",o.appendChild(l),o}if("url"===gn[t].type){const n=document.createElement("span");n.className="option",n.textContent=`${gn[t].label} `,o.appendChild(n);const i=document.createElement("input");i.id=Ln+t,i.className="option",i.type="text",i.readOnly=!0,i.setAttribute("value",e[t]),i.style.width="300px",o.appendChild(i),o.appendChild(document.createTextNode(" "));const a=document.createElement("input");a.id=`${Ln+t}Prompt`,a.className="option",a.type="text",a.setAttribute("value",e[t]);const r=document.createElement("a");r.className="option",r.onclick=function(){!function(e,t){Game.promptWrapL.className="framed";const o=e;Game.promptL.innerHTML=`${o}<div class="optionBox"></div>`,Object.keys(t).forEach((e=>{const o=document.createElement("a");o.id=`promptOption${e}`,o.className="option",o.onclick=function(){PlaySound("snd/tick.mp3"),t[e][1]()},o.textContent=t[e][0],Game.promptL.children[1].appendChild(o)})),Game.promptAnchorL.style.display="block",Game.darkenL.style.display="block",Game.promptL.focus(),Game.promptOn=1,Game.UpdatePrompt()}(a.outerHTML,[["Save",function(){e[t]=l(`CMConfig${t}Prompt`).value,Mn(),Game.ClosePrompt(),Game.UpdateMenu()}],["Cancel",function(){Game.ClosePrompt()}]])},r.textContent="Edit",o.appendChild(r);const s=document.createElement("label");return s.textContent=gn[t].desc,o.appendChild(s),o}if("colour"===gn[t].type){const n=document.createElement("span");n.className="option";const i=document.createElement("input");i.id=t,i.style.width="65px",i.setAttribute("value",e[t]),n.appendChild(i),new Fn(i,{hash:!0,position:"right",onInput:function(){e[this.targetElement.id]=this.toHEXString(),Tn(),Mn(),Game.UpdateMenu()}});const a=document.createElement("label");if(a.textContent=gn[t].desc,n.appendChild(a),t.includes("Flash")){const e=document.createElement("a");e.className="option",e.onclick=function(){Gn(3,t.replace("Colour",""),!0)},e.textContent="Test flash",n.appendChild(e)}return o.appendChild(n),Wn().init(),o}if("numscale"===gn[t].type){const n=document.createElement("span");n.className="option",n.textContent=`${gn[t].label} `,o.appendChild(n);const i=document.createElement("input");i.id=Ln+t,i.className="option",i.type="number",i.value=e[t],i.min=gn[t].min,i.max=gn[t].max,i.oninput=function(){this.value>this.max&&console.log("TEST"),e[t]=this.value,Mn(),sn()},o.appendChild(i),o.appendChild(document.createTextNode(" "));const a=document.createElement("label");return a.textContent=gn[t].desc,o.appendChild(a),o}return o}function _n(t){if(1===e.TimerBar&&0===e.TimerBarPos){const e=parseInt(l("CMTimerBar").style.height,10);Game.mouseY-=e,t(),Game.mouseY+=e}else t()}function Xn(){S.Beautify=Beautify,Beautify=ce,S.CalculateGains=Game.CalculateGains,Game.CalculateGains=function(){S.CalculateGains(),Ie=1},S.tooltip={},S.tooltip.draw=Game.tooltip.draw,S.tooltip.drawMod=new Function(`return ${Game.tooltip.draw.toString().split("this").join("Game.tooltip")}`)(),Game.tooltip.draw=function(e,t,o){S.tooltip.drawMod(e,t,o)},S.tooltip.update=Game.tooltip.update,S.tooltip.updateMod=new Function(`return ${Game.tooltip.update.toString().split("this.").join("Game.tooltip.")}`)(),Game.tooltip.update=function(){S.tooltip.updateMod(),function(){if("store"===Game.tooltip.origin){let t=0;1===e.ToolWarnLucky&&1===e.ToolWarnPos&&null!==l("CMDispTooltipWarningParent")&&(t=l("CMDispTooltipWarningParent").clientHeight-4),Game.tooltip.tta.style.top=`${Math.min(parseInt(Game.tooltip.tta.style.top,10),l("game").clientHeight+l("topBar").clientHeight-Game.tooltip.tt.clientHeight-t-46)}px`}}()},S.UpdateWrinklers=Game.UpdateWrinklers,Game.UpdateWrinklers=function(){_n(S.UpdateWrinklers)},S.UpdateSpecial=Game.UpdateSpecial,Game.UpdateSpecial=function(){_n(S.UpdateSpecial)},l("bigCookie").removeEventListener("click",Game.ClickCookie,!1),l("bigCookie").addEventListener("click",(function(){_n(Game.ClickCookie)}),!1),S.RebuildUpgrades=Game.RebuildUpgrades,Game.RebuildUpgrades=function(){S.RebuildUpgrades(),M=[],Object.keys(Game.UpgradesInStore).forEach((e=>{null!==l(`upgrade${e}`).onmouseover&&(M[e]=l(`upgrade${e}`).onmouseover,l(`upgrade${e}`).onmouseover=function(){Game.mouseDown||(Game.setOnCrate(this),Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Bn("u",`${e}`)}),"store"),Game.tooltip.wobble())})})),Game.CalculateGains()},S.ClickProduct=Game.ClickProduct,Game.ClickProduct=function(t){(!e.BulkBuyBlock||Game.ObjectsById[t].bulkPrice<Game.cookies||-1===Game.buyMode)&&S.ClickProduct(t)},S.DescribeDragonAura=Game.DescribeDragonAura,Game.DescribeDragonAura=function(t){S.DescribeDragonAura(t),function(t){if(1===e.DragonAuraInfo){const[e,o]=function(e){wo(),l("promptContent").children[0].innerHTML.includes("secondary")?He=e:De=e;let t=0;if(De!==Xe||He!==Qe)for(let e=Game.ObjectsById.length-1;e>-1;--e)if(Game.ObjectsById[e].amount>0){const o=Yt[Game.ObjectsById[e].name].name;Yt[o].amount-=1,Oe-=1,t=Yt[o].basePrice*Game.priceIncrease**Math.max(0,Yt[o].amount-1-Yt[o].free),t=Game.modifyBuildingPrice(Yt[o],t),t=Math.ceil(t);break}const o=Le;return Do(),Vo(),o!==Le&&Do(),[Ve-Game.cookiesPs,t]}(t),n=de(o/(e+Game.cookiesPs)),i=ce(e/Game.cookiesPs*100);l("dragonAuraInfo").style.minHeight="60px",l("dragonAuraInfo").style.margin="8px",l("dragonAuraInfo").appendChild(document.createElement("div")).className="line";const a=document.createElement("div");a.style.minWidth="200px",a.style.textAlign="center",a.textContent=`Picking this aura will change CPS by ${ce(e)} (${i}% of current CPS).`,l("dragonAuraInfo").appendChild(a);const r=document.createElement("div");r.style.minWidth="200px",r.style.textAlign="center",r.textContent=`It will take ${n} to recover the cost.`,l("dragonAuraInfo").appendChild(r)}}(t)},S.ToggleSpecialMenu=Game.ToggleSpecialMenu,Game.ToggleSpecialMenu=function(e){S.ToggleSpecialMenu(e),function(){if(null!==(l("specialPopup").className.match(/onScreen/)&&l("specialPopup").children[0].style.background.match(/dragon/)))for(let e=0;e<l("specialPopup").childNodes.length;e++)"optionBox"===l("specialPopup").childNodes[e].className&&(l("specialPopup").children[e].onmouseover=function(){Cn(),Game.tooltip.dynamic=1,Game.tooltip.draw(l("specialPopup"),`<div style="min-width:200px;text-align:center;">${Ye}</div>`,"this"),Game.tooltip.wobble()},l("specialPopup").children[e].onmouseout=function(){Game.tooltip.shouldHide=1})}()},S.UpdateMenu=Game.UpdateMenu,Game.UpdateMenu=function(){void 0!==Wn().picker&&void 0!==Wn().picker.owner||(S.UpdateMenu(),function(){const t=document.createElement("div");t.className="title","prefs"===Game.onMenu?(t.textContent="Cookie Monster Settings",function(t){const o=document.createDocumentFragment();o.appendChild(t),Object.keys($n).forEach((t=>{if("Favourite"===t){if(0!==se.length&&e.FavouriteSettings>0&&(o.appendChild(zn(t,$n[t])),e.Header[t]))for(let e=0;e<se.length;e++)o.appendChild(Vn(se[e]))}else o.appendChild(zn(t,$n[t])),e.Header[t]&&("Notification"===t?Object.keys(In).forEach((t=>{const n=zn(t,In[t]);n.style.fontSize="15px",n.style.opacity="0.5",o.appendChild(n),e.Header[t]&&Object.keys(gn).forEach((e=>{gn[e].group===t&&o.appendChild(Vn(e))}))})):Object.keys(gn).forEach((e=>{gn[e].group===t&&o.appendChild(Vn(e))})))}));const n=document.createElement("div");n.className="listing";const i=document.createElement("a");i.className="option",i.onclick=function(){Sn(an)},i.textContent="Restore Default",n.appendChild(i),o.appendChild(n),l("menu").childNodes[2].insertBefore(o,l("menu").childNodes[2].childNodes[l("menu").childNodes[2].childNodes.length-1])}(t)):"stats"===Game.onMenu?e.Stats&&(t.textContent="Cookie Monster Statistics",Rn(t)):"log"===Game.onMenu&&(t.textContent="Cookie Monster ",function(t){const o=document.createElement("div");o.className="subsection";const n=document.createElement("span");if(n.style.cursor="pointer",n.style.display="inline-block",n.style.height="14px",n.style.width="14px",n.style.borderRadius="7px",n.style.textAlign="center",n.style.backgroundColor="#C0C0C0",n.style.color="black",n.style.fontSize="13px",n.style.verticalAlign="middle",n.textContent=e.Header.InfoTab?"-":"+",n.onclick=function(){jn("InfoTab"),Game.UpdateMenu()},t.appendChild(n),o.appendChild(t),e.Header.InfoTab){const e=document.createElement("div");e.innerHTML='<div class="listing">\n <a href="https://github.com/Aktanusa/CookieMonster" target="blank">Cookie Monster</a>\n offers a wide range of tools and statistics to enhance your game experience.\n It is not a cheat interface – although it does offer helpers for golden cookies and such, everything can be toggled off at will to only leave how much information you want.</br>\n Progess on new updates and all previous release notes can be found on the GitHub page linked above!</br>\n Please also report any bugs you may find over there!</br>\n </div>\n ',o.appendChild(e);const t=document.createElement("div");t.innerHTML='<div class="listing">\nThe latest update (v 2.031.6) has revamped the way PP is colourcoded. In the previous versions this was not always correct and we had some settings that made it even more incorrect. With this update Cookie Monster compares the PP of a building to all other possibilities (also buying 10 or 100) of a building and colors the PP accordingly.</br>\nThis means that if it is better to buy 10 of a building immediately rather than to buy in increments of 1 Cookie Monster will now show this! In this case the buy 1 option will be yellow, while the buy 10 option will be green.</br>\nThis also means that you will see some more diverse colors. Rather than coloring according to arbitrary "middle values" Cookie Monster now uses a top 10, 20 and 30. While this might take some getting used to we believe the new system is much better in conveying useful information to the user while also being more correct!</br>\n</br>\nThis update also implements the following functions:</br>\n- Added a tooltip displaying the reward to the extra pop wrinkler buttons</br>\n- Added tooltips to the Gods in the Pantheon</br>\n- The tooltip for Elder Pledge now displays correctly, although it takes some time to load after Cookie Monster has been loaded</br>\n- You can now test the volume and sound of notifications in the settings screen</br>\n- The bottom bar will now flicker less and each column has received a bit of padding</br>\n- Introduced new colour scheme for PP, see the explanation of colors in the Readme or the settings</br>\n- Added option to show buttons that can hide the upgrade and building section</br>\n- Added option to display PP as a time unit, note that these are only approximations. PP does not translate directly into time (the name is deceptive, we know...)</br>\n- Statistics page now shows chance for random drops when they have not been achieved</br>\n- In a later hotfix we have added a Rigidel mode that will colour all buy 1 options grey</br>\n</br>\nThis update fixes the following bugs:</br>\n- Incorrect amount for "required for max plant reward" in statistics page</br>\n- Fixed the tooltips of the \'?\' buttons in the statistics page not showing</br>\n- Fixed some unclear settings descriptions </br>\n</br>\nPlease submit any bug reports or feature requests to the <a href="https://github.com/Aktanusa/CookieMonster">GitHub page!</a>\n</br>\n</div>\n',o.appendChild(t)}const i=l("menu").children[1];i.insertBefore(o,i.children[1])}(t))}())},S.sayTime=Game.sayTime,I=function(e,t){return Number.isNaN(e)||e<=0?S.sayTime(e,t):de(e/Game.fps,1)},S.Logic=Game.Logic,Game.Logic=function(){S.Logic();let t="Cookie Clicker";"fools"===Game.season&&(t="Cookie Baker"),le=`${Game.OnAscend?"Ascending! ":""}${ce(Game.cookies)} ${1===Game.cookies?"cookie":"cookies"} - ${t}`,function(){if(Game.OnAscend||0===e.Title)document.title=le;else if(1===e.Title){let e,t,o,n=!1,i=!1;e=$t?$t.wrath?`[W${Math.ceil($t.life/Game.fps)}]`:`[G${Math.ceil($t.life/Game.fps)}]`:Game.Has("Golden switch [off]")?"[GS]":`[${Number(l("CMTimerBarGCMinBar").textContent)<0?"!":""}${Math.ceil((Game.shimmerTypes.golden.maxTime-Game.shimmerTypes.golden.time)/Game.fps)}]`,f&&(n=!0,t="[F]"),"christmas"===Game.season&&(i=!0,o=g?`[R${Math.ceil(It.life/Game.fps)}]`:`[${Number(l("CMTimerBarRenMinBar").textContent)<0?"!":""}${Math.ceil((Game.shimmerTypes.reindeer.maxTime-Game.shimmerTypes.reindeer.time)/Game.fps)}]`);let a=le;"["===a.charAt(0)&&(a=a.substring(a.lastIndexOf("]")+1)),document.title=`${e+(n?t:"")+(i?o:"")} ${a}`}else if(2===e.Title){let e="",t=!1;$t&&(t=!0,$t.wrath?e+=`[W${Math.ceil($t.life/Game.fps)}]`:e+=`[G${Math.ceil($t.life/Game.fps)}]`),f&&(t=!0,e+="[F]"),"christmas"===Game.season&&g&&(e+=`[R${Math.ceil(It.life/Game.fps)}]`,t=!0),t&&(e+=" - ");let o="Cookie Clicker";"fools"===Game.season&&(o="Cookie Baker"),e+=o,document.title=e}}(),function(){const t=Math.max(0,Game.HowManyCookiesReset(Math.floor(Game.HowMuchPrestige(Game.cookiesReset+Game.cookiesEarned))+1)-(Game.cookiesEarned+Game.cookiesReset)),o=Game.sayTime((Date.now()-Game.startDate)/1e3*Game.fps,-1);let n="";n+=`You've been on this run for <b>${""===o?"not very long":o}</b>.<br>`,n+='<div class="line"></div>',Game.prestige>0&&(n+=`Your prestige level is currently <b>${ce(Game.prestige)}</b>.<br>(CpS +${ce(Game.prestige)}%)`,n+='<div class="line"></div>'),n+=Be<1?"Ascending now would grant you no prestige.":Be<2?"Ascending now would grant you<br><b>1 prestige level</b> (+1% CpS)<br>and <b>1 heavenly chip</b> to spend.":`Ascending now would grant you<br><b>${ce(Be)} prestige levels</b> (+${ce(Be)}% CpS)<br>and <b>${ce(Be)} heavenly chips</b> to spend.`,n+='<div class="line"></div>',n+=`You need <b>${ce(t)} more cookies</b> for the next level.<br>`,n+=e.TooltipAscendButton?`<div class='line'></div>It takes ${zt} to reach the next level and you are making ${ce(pt,2)} chips on average in the last 5 seconds.<br>`:"",l("ascendTooltip").innerHTML=n}()}}function Qn(){bn(),en(),nn(),r=Object.keys(Game.mods).length,v=document.createElement("style"),v.type="text/css",v.id="CMCSS",document.head.appendChild(v),function(){const e=document.createElement("div");e.id="CMBotBar",e.style.height="69px",e.style.width="100%",e.style.position="absolute",e.style.display="none",e.style.backgroundColour="#262224",e.style.backgroundImage="linear-gradient(to bottom, #4d4548, #000000)",e.style.borderTop="1px solid black",e.style.overflow="auto",e.style.textShadow="-1px 0 black, 0 1px black, 1px 0 black, 0 -1px black";const t=e.appendChild(document.createElement("table"));t.style.width="100%",t.style.textAlign="center",t.style.whiteSpace="nowrap";const o=t.appendChild(document.createElement("tbody")),n=function(e,t){const o=document.createElement("td");return o.style.textAlign="right",o.className=P+t,o.textContent=e,o},i=o.appendChild(document.createElement("tr"));i.style.fontWeight="bold",i.appendChild(n("CM 2.031.6",O)),o.appendChild(document.createElement("tr")).appendChild(n("Bonus Income",E)),o.appendChild(document.createElement("tr")).appendChild(n("Payback Period",E)),o.appendChild(document.createElement("tr")).appendChild(n("Time Left",E)),l("wrapper").appendChild(e),Object.keys(Game.Objects).forEach((e=>{uo(e)}))}(),function(){const e=document.createElement("div");e.id="CMTimerBar",e.style.position="absolute",e.style.display="none",e.style.height="0px",e.style.fontSize="10px",e.style.fontWeight="bold",e.style.backgroundColour="black";const t=po("CMTimerBarGC","Next Cookie",[{id:"CMTimerBarGCMinBar",color:A},{id:"CMTimerBarGCBar",color:L}]);e.appendChild(t);const o=po("CMTimerBarRen","Next Reindeer",[{id:"CMTimerBarRenMinBar",color:A},{id:"CMTimerBarRenBar",color:F}]);e.appendChild(o);const n=document.createElement("div");n.id="CMTimerBarBuffTimers",e.appendChild(n),l("wrapper").appendChild(e)}(),function(){const e=document.createElement("div");e.id="CMUpgradeBar",e.style.width="100%",e.style.backgroundColour="black",e.style.textAlign="center",e.style.fontWeight="bold",e.style.display="none",e.style.zIndex="21",e.onmouseout=function(){Game.tooltip.hide()};const t=document.createElement("div");t.appendChild(function(){const e=document.createElement("div");e.style.minWidth="330px",e.style.marginBottom="4px";const t=document.createElement("div");t.className="name",t.style.marginBottom="4px",t.textContent="Legend",e.appendChild(t);const o=function(e,t){const o=document.createElement("div");o.style.verticalAlign="middle";const n=document.createElement("span");return n.className=B+e,n.style.display="inline-block",n.style.height="10px",n.style.width="10px",n.style.marginRight="4px",o.appendChild(n),o.appendChild(document.createTextNode(t)),o};return e.appendChild(o(E,"Better than the best PP of a building option")),e.appendChild(o(N,"Same as the best PP building option")),e.appendChild(o(O,"Within the top 10 of PP for buildings")),e.appendChild(o(F,"Within the top 20 of PP for buildings")),e.appendChild(o(W,"Within the top 30 of PP for buildings")),e.appendChild(o(L,"Outside of the top 30 of PP for buildings")),e.appendChild(o(A,"Negative or infinity PP")),e}()),e.onmouseover=function(){Game.tooltip.draw(this,escape(t.innerHTML),"store")};const o=function(e,t){const o=document.createElement("span");return o.id=e,o.className=P+t,o.style.width="14.28571428571429%",o.style.display="inline-block",o.textContent="0",o};e.appendChild(o("CMUpgradeBarBlue",E)),e.appendChild(o("CMUpgradeBarGreen",N)),e.appendChild(o("CMUpgradeBarYellow",O)),e.appendChild(o("CMUpgradeBarOrange",F)),e.appendChild(o("CMUpgradeBarRed",W)),e.appendChild(o("CMUpgradeBarPurple",L)),e.appendChild(o("CMUpgradeBarGray",A)),l("upgrades").parentNode.insertBefore(e,l("upgrades").parentNode.childNodes[3])}(),function(){const e=document.createElement("div");e.id="CMFlashScreen",e.style.width="100%",e.style.height="100%",e.style.backgroundColor="white",e.style.display="none",e.style.zIndex="9999999999",e.style.position="absolute",l("wrapper").appendChild(e)}(),function(){const e=document.createElement("div");e.id="CMSectionHidButtons",e.style.textAlign="center";const t=document.createElement("a");t.className="option",t.onclick=function(){"flex"===l("upgrades").style.display?(l("upgrades").style.display="none",l("toggleUpgrades").style.display="none",l("techUpgrades").style.display="none",l("vaultUpgrades").style.display="none"):(l("upgrades").style.display="flex",0!==l("toggleUpgrades").children.length&&(l("toggleUpgrades").style.display="block"),0!==l("techUpgrades").children.length&&(l("techUpgrades").style.display="block"),0!==l("vaultUpgrades").children.length&&(l("vaultUpgrades").style.display="block"))},t.textContent="Hide/Show Upgrades",e.appendChild(t);const o=document.createElement("a");o.className="option",o.onclick=function(){"grid"===l("products").style.display?l("products").style.display="none":l("products").style.display="grid"},o.textContent="Hide/Show Buildings",e.appendChild(o),l("store").insertBefore(e,l("store").childNodes[2])}(),function(){const e=document.createElement("link");e.id="CMFavicon",e.rel="shortcut icon",e.href="https://orteil.dashnet.org/cookieclicker/favicon.ico",document.getElementsByTagName("head")[0].appendChild(e)}(),Object.keys(U).forEach((e=>{!function(e,t,o){const n=document.createElement("div");n.id=e;const i=document.createElement("div");i.style.minWidth=o,i.style.marginBottom="4px";const a=document.createElement("div");a.style.textAlign="left",a.textContent=t,i.appendChild(a),n.appendChild(i),R[e]=n}(U[e][0],U[e][1],U[e][2])})),function(){const e=document.createElement("a");e.id="PopAllNormalWrinklerButton",e.textContent="Pop All Normal",e.className="option",e.onclick=function(){Pn()},e.onmouseout=function(){Game.tooltip.shouldHide=1},e.onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Bn("wb","PopAll")}),"this"),Game.tooltip.wobble()},l("sectionLeftExtra").children[0].append(e);const t=document.createElement("a");t.id="PopFattestWrinklerButton",t.textContent="Pop Single Fattest",t.className="option",t.onclick=function(){null!==kt[1]&&(Game.wrinklers[kt[1]].hp=0)},t.onmouseout=function(){Game.tooltip.shouldHide=1},t.onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Bn("wb","PopFattest")}),"this"),Game.tooltip.wobble()},l("sectionLeftExtra").children[0].append(t)}(),l("products").style.display="grid",l("storeBulk").style.gridRow="1/1",l("upgrades").style.display="flex",l("upgrades").style["flex-wrap"]="wrap",Object.keys(Game.Objects).forEach((e=>{const t=Game.Objects[e];null!==l(`product${t.id}`).onmouseover&&(x[e]=l(`product${t.id}`).onmouseover,l(`product${t.id}`).onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Bn("b",`${e}`)}),"store"),Game.tooltip.wobble()})})),Game.canLumps()&&(s=l("lumps").onmouseover,l("lumps").onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Bn("s","Lump")}),"this"),Game.tooltip.wobble()}),h=Game.LoadMinigames,Game.LoadMinigames=function(){h(),Game.Objects.Farm.minigameLoaded&&(l("gardenTool-1").onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Bn("ha","HarvestAllButton")}),"this"),Game.tooltip.wobble()},Array.from(l("gardenPlot").children).forEach((e=>{const t=e.id.slice(-3);e.onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Bn("p",[`${t[0]}`,`${t[2]}`])}),"this"),Game.tooltip.wobble()}}))),En(),function(){if(Game.Objects.Temple.minigameLoaded){for(let e=0;e<11;e+=1)l(`templeGod${e}`).onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Bn("pag",e)}),"this"),Game.tooltip.wobble()};for(let e=0;e<3;e+=1)l(`templeSlot${e}`).onmouseover=function(){Game.tooltip.dynamic=1,Game.tooltip.draw(this,(function(){return Bn("pas",[e,Game.Objects.Temple.minigame.slot[e]])}),"this"),Game.tooltip.wobble()}}}(),On()},Game.LoadMinigames(),l("backgroundLeftCanvas").onmouseover=function(){z=1},l("backgroundLeftCanvas").onmouseout=function(){z=0,Game.tooltip.hide(),Object.keys(Game.wrinklers).forEach((e=>{_[e]=0}))},Xn(),On(),Game.CalculateGains(),Sn(),$=Game.OnAscend,Game.prefs.popups?Game.Popup("Cookie Monster version 2.031.6 loaded!"):Game.Notify("Cookie Monster version 2.031.6 loaded!","","",1,1),Game.Win("Third-party")}const qn={init:function(){kn=!0;let e=!0;Game.version!==Number("2.031")&&(e=confirm("Cookie Monster version 2.031.6 is meant for Game version 2.031.  Loading a different version may cause errors.  Do you still want to load Cookie Monster?")),e&&(Qn(),Game.registerHook("draw",Bo),Game.registerHook("logic",xn),kn=!1)},load:function(e){const t=JSON.parse(e);bn(),void 0!==t.favouriteSettings&&(se=t.favouriteSettings),Sn(t.settings),"2.031.6"!==t.version&&(Game.prefs.popups?Game.Popup("A new version of Cookie Monster has been loaded, check out the release notes in the info tab!"):Game.Notify("A new version of Cookie Monster has been loaded, check out the release notes in the info tab!","","",0,1))},save:fn};Game.registerMod("CookieMonster",qn)})()})();
